Derive copy number feature distributions

This function summarise each copy-number profile using a number of different feature distributions: sigment size, breakpoint number (per ten megabase), change-point copy-number, segment copy-number, breakpoint number (per chromosome arm), length of segments with oscilating copy-number.

get_features(CN_data, cores = 1, genome_build = c("hg19", "hg38"))

Arguments

CN_data
a list contains multiple data.frames (recommended), each data.frame stores copy-number profile for one sample with 'chromosome', 'start', 'end' and 'segVal' these four necessary columns. Of note, 'segVal' column shoule be absolute copy number values.

cores
number of compute cores to run this task. You can use parallel::detectCores() function to check how many cores you can use.

genome_build
genome build version, should be 'hg19' or 'hg38'.

Value

a list contains six copy number feature distributions.

See also

Examples

# NOT RUN {
# Load copy number list
load(system.file("extdata", "toy_cnlist.RData",
  package = "sigminer", mustWork = TRUE
))
# }# NOT RUN {
cn_features <- get_features(cn_list, cores = 1)
# }