Obtain Reference Signatures

The signatures and their aetiologies mainly obtained from COSMIC database and cleaned before saving into sigminer package.

get_sig_db(sig_db = "legacy")

Arguments

sig_db
can be 'legacy' (for COSMIC v2 'SBS'), 'SBS', 'DBS', 'ID' and 'TSB' (for SBS transcriptional strand bias signatures). Default 'legacy'.

Value

a list.

See also

Examples

s1 <- get_sig_db()
s2 <- get_sig_db("DBS")
s1

#> $db
#>             COSMIC_1     COSMIC_2    COSMIC_3 COSMIC_4    COSMIC_5 COSMIC_6
#> A[C>A]A 1.109833e-02 6.827080e-04 0.022172307   0.0365 0.014941548   0.0017
#> A[C>A]C 9.149341e-03 6.191070e-04 0.017871675   0.0309 0.008960918   0.0028
#> A[C>A]G 1.490070e-03 9.927900e-05 0.002138340   0.0183 0.002207846   0.0005
#> A[C>A]T 6.233885e-03 3.238910e-04 0.016265146   0.0243 0.009206905   0.0019
#> A[C>G]A 1.801068e-03 2.634810e-04 0.024002615   0.0097 0.011671022   0.0013
#> A[C>G]C 2.580909e-03 2.698660e-04 0.012160304   0.0054 0.007292089   0.0012
#> A[C>G]G 5.925480e-04 2.192340e-04 0.005275419   0.0031 0.002303839   0.0000
#> A[C>G]T 2.963986e-03 6.109730e-04 0.023277656   0.0054 0.011696246   0.0018
#> A[C>T]A 2.951453e-02 7.441557e-03 0.017872171   0.0120 0.021839188   0.0312
#> A[C>T]C 1.432275e-02 2.726312e-03 0.008896034   0.0075 0.012756106   0.0163
#> A[C>T]G 1.716469e-01 3.322083e-03 0.003572708   0.0028 0.016760177   0.0908
#> A[C>T]T 1.262376e-02 3.326528e-03 0.014797612   0.0059 0.016477955   0.0149
#> A[T>A]A 4.021520e-03 1.323770e-07 0.008428564   0.0048 0.008902904   0.0006
#> A[T>A]C 2.371144e-03 1.130700e-04 0.007373097   0.0039 0.007399202   0.0033
#> A[T>A]G 2.810910e-03 5.329290e-04 0.007357300   0.0100 0.011508051   0.0000
#> A[T>A]T 8.360909e-03 1.491110e-04 0.008753939   0.0030 0.011193623   0.0053
#> A[T>C]A 1.391577e-02 1.303510e-03 0.013035709   0.0084 0.035366735   0.0075
#> A[T>C]C 6.274961e-03 4.255770e-04 0.009186290   0.0020 0.013771102   0.0056
#> A[T>C]G 1.013764e-02 5.754730e-04 0.011716920   0.0081 0.028449052   0.0217
#> A[T>C]T 9.256316e-03 1.488164e-03 0.016978871   0.0036 0.027302974   0.0023
#> A[T>G]A 1.587636e-03 3.371840e-05 0.002351368   0.0000 0.003462152   0.0000
#> A[T>G]C 1.784091e-03 2.484480e-05 0.001464231   0.0002 0.002246593   0.0017
#> A[T>G]G 1.385831e-03 2.734540e-04 0.009053778   0.0015 0.005490217   0.0007
#> A[T>G]T 3.158539e-03 2.176950e-04 0.007030976   0.0002 0.003821227   0.0029
#> C[C>A]A 6.595870e-03 6.774450e-04 0.018781726   0.0461 0.009674904   0.0101
#> C[C>A]C 7.342368e-03 2.136810e-04 0.015760458   0.0614 0.004952301   0.0241
#> C[C>A]G 8.928400e-04 6.770460e-06 0.001963390   0.0088 0.002800627   0.0091
#> C[C>A]T 7.186582e-03 4.163330e-04 0.014722861   0.0432 0.011013466   0.0571
#> C[C>G]A 1.284983e-03 2.777210e-05 0.016832572   0.0105 0.007537523   0.0000
#> C[C>G]C 7.021350e-04 2.796440e-04 0.013531442   0.0097 0.007633227   0.0000
#> C[C>G]G 5.062900e-04 1.916160e-05 0.004176458   0.0063 0.002613760   0.0000
#> C[C>G]T 1.381543e-03 3.127820e-04 0.024046390   0.0094 0.009417098   0.0002
#> C[C>T]A 2.089645e-02 1.501951e-02 0.014395059   0.0210 0.022769209   0.0085
#> C[C>T]C 1.850171e-02 3.516918e-03 0.008544781   0.0144 0.017509476   0.0099
#> C[C>T]G 9.557722e-02 4.979276e-03 0.003518466   0.0076 0.012862238   0.0901
#> C[C>T]T 1.711331e-02 8.956553e-03 0.016075546   0.0201 0.020429513   0.0087
#> C[T>A]A 1.182587e-03 1.545990e-04 0.007571320   0.0075 0.005043164   0.0001
#> C[T>A]C 1.903167e-03 4.640190e-04 0.012725464   0.0111 0.006208844   0.0026
#> C[T>A]G 1.487961e-03 2.304100e-04 0.011508562   0.0342 0.010507569   0.0008
#> C[T>A]T 2.179344e-03 5.748860e-04 0.016456178   0.0115 0.009784641   0.0011
#> C[T>C]A 4.176675e-03 5.466190e-04 0.007895013   0.0052 0.014238630   0.0062
#> C[T>C]C 5.252593e-03 3.923330e-04 0.014431170   0.0026 0.012489500   0.0040
#> C[T>C]G 7.013225e-03 3.620250e-04 0.008422972   0.0100 0.018741631   0.0270
#> C[T>C]T 6.713813e-03 5.609000e-04 0.011932430   0.0054 0.019054083   0.0033
#> C[T>G]A 3.026910e-04 1.137850e-04 0.001974181   0.0000 0.002041649   0.0001
#> C[T>G]C 2.098502e-03 2.209510e-05 0.005824296   0.0013 0.003478502   0.0040
#> C[T>G]G 1.599549e-03 2.282460e-04 0.010464655   0.0046 0.007147456   0.0050
#> C[T>G]T 2.758538e-03 6.711130e-05 0.008724387   0.0012 0.011486812   0.0086
#> G[C>A]A 8.232604e-03 3.520130e-04 0.009696540   0.0376 0.011892169   0.0024
#> G[C>A]C 5.758021e-03 1.338170e-04 0.010843341   0.0399 0.009247857   0.0058
#> G[C>A]G 6.163350e-04 1.784420e-04 0.000929141   0.0227 0.002809188   0.0021
#> G[C>A]T 4.459080e-03 1.228320e-04 0.012215383   0.0258 0.010301267   0.0087
#> G[C>G]A 6.021230e-04 4.483850e-05 0.011916845   0.0070 0.005559423   0.0000
#> G[C>G]C 2.393352e-03 1.452010e-05 0.009823653   0.0091 0.005388801   0.0030
#> G[C>G]G 2.485340e-07 4.066590e-05 0.001671054   0.0062 0.001100590   0.0000
#> G[C>G]T 8.900810e-04 2.684290e-04 0.017914334   0.0060 0.006041248   0.0017
#> G[C>T]A 2.494381e-02 6.390808e-03 0.016127273   0.0087 0.020038259   0.0653
#> G[C>T]C 2.716149e-02 1.995817e-03 0.008208632   0.0080 0.018022312   0.0773
#> G[C>T]G 1.035708e-01 3.030210e-04 0.001212923   0.0023 0.013193818   0.1339
#> G[C>T]T 1.768985e-02 3.265818e-03 0.010611649   0.0082 0.019502915   0.0524
#> G[T>A]A 6.892890e-04 1.149940e-04 0.004434783   0.0069 0.006740388   0.0000
#> G[T>A]C 5.524100e-04 2.938620e-04 0.005615432   0.0052 0.004021596   0.0028
#> G[T>A]G 1.200229e-03 8.872130e-05 0.008070136   0.0133 0.006814352   0.0006
#> G[T>A]T 2.107137e-03 2.157550e-04 0.008679122   0.0045 0.005101033   0.0021
#> G[T>C]A 1.124784e-02 4.673850e-06 0.006850056   0.0061 0.016294983   0.0122
#> G[T>C]C 6.999724e-03 1.860660e-04 0.006260749   0.0016 0.009574561   0.0059
#> G[T>C]G 4.977593e-03 4.950440e-07 0.006098763   0.0042 0.014103929   0.0115
#> G[T>C]T 1.066741e-02 5.785980e-04 0.007509344   0.0024 0.015666875   0.0042
#> G[T>G]A 9.904500e-05 9.555240e-05 0.004144488   0.0000 0.001627665   0.0000
#> G[T>G]C 2.023660e-04 4.700240e-05 0.004501985   0.0000 0.000327735   0.0016
#> G[T>G]G 1.188353e-03 1.099260e-04 0.016391453   0.0018 0.005948880   0.0010
#> G[T>G]T 8.007230e-04 8.647720e-05 0.007067237   0.0002 0.003307467   0.0035
#> T[C>A]A 1.225006e-02 1.512743e-02 0.011653225   0.0330 0.014774016   0.0017
#> T[C>A]C 1.116223e-02 6.532492e-03 0.016606775   0.0538 0.012043465   0.0029
#> T[C>A]G 2.275496e-03 1.656455e-03 0.001357239   0.0104 0.003902362   0.0011
#> T[C>A]T 1.525910e-02 1.239461e-02 0.016328076   0.0370 0.018243396   0.0058
#> T[C>G]A 1.874853e-03 3.724185e-02 0.016041405   0.0032 0.002681057   0.0000
#> T[C>G]C 2.067419e-03 1.941340e-05 0.020149920   0.0105 0.007924048   0.0002
#> T[C>G]G 3.048970e-04 1.625465e-03 0.002527911   0.0031 0.001319076   0.0000
#> T[C>G]T 3.151574e-03 6.687990e-02 0.032673641   0.0050 0.006644596   0.0001
#> T[C>T]A 1.449210e-02 4.199414e-01 0.008880209   0.0035 0.010997560   0.0074
#> T[C>T]C 1.768078e-02 8.197250e-02 0.013529573   0.0070 0.020644965   0.0067
#> T[C>T]G 7.600222e-02 4.772019e-02 0.001705409   0.0011 0.007534490   0.0391
#> T[C>T]T 1.376170e-02 2.286749e-01 0.010304417   0.0077 0.011787462   0.0047
#> T[T>A]A 5.600155e-03 8.075250e-05 0.007132681   0.0045 0.009205637   0.0002
#> T[T>A]C 1.999079e-03 4.802900e-06 0.009102798   0.0046 0.006835344   0.0008
#> T[T>A]G 1.090066e-03 6.660500e-05 0.006565863   0.0082 0.007144209   0.0000
#> T[T>A]T 3.981023e-03 2.763160e-04 0.014712135   0.0045 0.010240850   0.0007
#> T[T>C]A 8.073616e-03 1.019320e-04 0.009115495   0.0028 0.017059641   0.0059
#> T[T>C]C 4.857381e-03 4.703440e-04 0.010953791   0.0016 0.014195709   0.0035
#> T[T>C]G 8.325454e-03 1.923540e-04 0.006113033   0.0036 0.012596502   0.0106
#> T[T>C]T 6.257106e-03 5.853300e-04 0.010774336   0.0022 0.017375139   0.0029
#> T[T>G]A 1.397554e-03 7.173700e-05 0.005427184   0.0000 0.005202874   0.0009
#> T[T>G]C 1.291737e-03 1.428150e-05 0.006160250   0.0003 0.005131608   0.0019
#> T[T>G]G 2.031077e-03 2.066150e-04 0.011076526   0.0030 0.006055254   0.0011
#> T[T>G]T 4.030128e-03 2.359820e-05 0.013000984   0.0011 0.013369936   0.0072
#>         COSMIC_7    COSMIC_8 COSMIC_9 COSMIC_10 COSMIC_11 COSMIC_12
#> A[C>A]A   0.0004 0.036718004   0.0120    0.0007    0.0002    0.0077
#> A[C>A]C   0.0005 0.033245722   0.0067    0.0010    0.0010    0.0047
#> A[C>A]G   0.0000 0.002525311   0.0005    0.0003    0.0000    0.0017
#> A[C>A]T   0.0004 0.033598550   0.0068    0.0092    0.0002    0.0046
#> A[C>G]A   0.0000 0.008356845   0.0048    0.0005    0.0007    0.0031
#> A[C>G]C   0.0000 0.004306380   0.0023    0.0003    0.0003    0.0015
#> A[C>G]G   0.0000 0.000584415   0.0000    0.0000    0.0000    0.0000
#> A[C>G]T   0.0001 0.008634907   0.0038    0.0002    0.0009    0.0025
#> A[C>T]A   0.0000 0.018066687   0.0093    0.0000    0.0225    0.0121
#> A[C>T]C   0.0197 0.005650015   0.0056    0.0032    0.1099    0.0054
#> A[C>T]G   0.0001 0.019265105   0.0125    0.0126    0.0072    0.0019
#> A[C>T]T   0.0043 0.020805967   0.0076    0.0047    0.0639    0.0067
#> A[T>A]A   0.0010 0.013365083   0.0121    0.0000    0.0002    0.0058
#> A[T>A]C   0.0008 0.012431202   0.0042    0.0002    0.0001    0.0065
#> A[T>A]G   0.0009 0.014036657   0.0068    0.0000    0.0007    0.0046
#> A[T>A]T   0.0035 0.024109917   0.0185    0.0012    0.0003    0.0060
#> A[T>C]A   0.0005 0.016206510   0.0252    0.0031    0.0006    0.0466
#> A[T>C]C   0.0001 0.007787803   0.0105    0.0052    0.0015    0.0317
#> A[T>C]G   0.0007 0.009852584   0.0176    0.0026    0.0008    0.0487
#> A[T>C]T   0.0011 0.021829464   0.0247    0.0040    0.0003    0.0250
#> A[T>G]A   0.0000 0.004219719   0.0335    0.0000    0.0000    0.0012
#> A[T>G]C   0.0000 0.000793671   0.0051    0.0040    0.0000    0.0010
#> A[T>G]G   0.0009 0.006865800   0.0052    0.0015    0.0001    0.0018
#> A[T>G]T   0.0000 0.002939922   0.0193    0.0137    0.0000    0.0030
#> C[C>A]A   0.0012 0.031723757   0.0098    0.0031    0.0007    0.0135
#> C[C>A]C   0.0006 0.025505407   0.0057    0.0009    0.0017    0.0112
#> C[C>A]G   0.0000 0.001159624   0.0000    0.0007    0.0010    0.0028
#> C[C>A]T   0.0013 0.028791173   0.0091    0.0160    0.0014    0.0071
#> C[C>G]A   0.0001 0.006619077   0.0018    0.0000    0.0000    0.0019
#> C[C>G]C   0.0004 0.006077703   0.0020    0.0000    0.0002    0.0019
#> C[C>G]G   0.0006 0.000656033   0.0000    0.0000    0.0008    0.0000
#> C[C>G]T   0.0003 0.007805027   0.0039    0.0000    0.0002    0.0019
#> C[C>T]A   0.0754 0.004865810   0.0098    0.0012    0.0310    0.0187
#> C[C>T]C   0.1007 0.003980150   0.0069    0.0024    0.1518    0.0094
#> C[C>T]G   0.0208 0.008338639   0.0076    0.0109    0.0135    0.0046
#> C[C>T]T   0.0788 0.018843627   0.0097    0.0043    0.0816    0.0205
#> C[T>A]A   0.0000 0.012024218   0.0055    0.0000    0.0004    0.0017
#> C[T>A]C   0.0014 0.017881092   0.0049    0.0001    0.0011    0.0086
#> C[T>A]G   0.0007 0.016356775   0.0051    0.0003    0.0006    0.0031
#> C[T>A]T   0.0018 0.026247802   0.0069    0.0009    0.0004    0.0054
#> C[T>C]A   0.0010 0.005173842   0.0154    0.0005    0.0001    0.0312
#> C[T>C]C   0.0008 0.010936894   0.0118    0.0019    0.0004    0.0385
#> C[T>C]G   0.0010 0.005657774   0.0154    0.0015    0.0001    0.0415
#> C[T>C]T   0.0045 0.011120389   0.0275    0.0020    0.0009    0.0324
#> C[T>G]A   0.0000 0.002424051   0.0216    0.0022    0.0000    0.0013
#> C[T>G]C   0.0008 0.002096211   0.0064    0.0018    0.0001    0.0049
#> C[T>G]G   0.0009 0.006604046   0.0126    0.0037    0.0011    0.0045
#> C[T>G]T   0.0013 0.004866714   0.0509    0.0182    0.0009    0.0063
#> G[C>A]A   0.0003 0.023682329   0.0118    0.0014    0.0004    0.0062
#> G[C>A]C   0.0001 0.015821896   0.0092    0.0022    0.0010    0.0056
#> G[C>A]G   0.0000 0.000850914   0.0000    0.0002    0.0000    0.0015
#> G[C>A]T   0.0001 0.021061236   0.0085    0.0088    0.0006    0.0027
#> G[C>G]A   0.0000 0.003724568   0.0011    0.0000    0.0000    0.0012
#> G[C>G]C   0.0004 0.003046597   0.0029    0.0002    0.0002    0.0011
#> G[C>G]G   0.0000 0.000321213   0.0000    0.0000    0.0000    0.0000
#> G[C>G]T   0.0000 0.005775664   0.0044    0.0004    0.0000    0.0009
#> G[C>T]A   0.0000 0.002261263   0.0062    0.0000    0.0261    0.0124
#> G[C>T]C   0.0210 0.001616649   0.0069    0.0134    0.0975    0.0092
#> G[C>T]G   0.0002 0.011606712   0.0088    0.0270    0.0090    0.0023
#> G[C>T]T   0.0161 0.006284929   0.0101    0.0152    0.0522    0.0115
#> G[T>A]A   0.0000 0.008139179   0.0030    0.0000    0.0003    0.0015
#> G[T>A]C   0.0000 0.007910080   0.0029    0.0001    0.0003    0.0039
#> G[T>A]G   0.0006 0.009057263   0.0028    0.0000    0.0004    0.0019
#> G[T>A]T   0.0010 0.017812239   0.0036    0.0004    0.0001    0.0031
#> G[T>C]A   0.0002 0.005720779   0.0091    0.0064    0.0010    0.0573
#> G[T>C]C   0.0001 0.005181052   0.0097    0.0101    0.0016    0.0274
#> G[T>C]G   0.0003 0.005040707   0.0094    0.0054    0.0019    0.0410
#> G[T>C]T   0.0081 0.008488220   0.0126    0.0124    0.0023    0.0405
#> G[T>G]A   0.0000 0.000974579   0.0072    0.0000    0.0000    0.0000
#> G[T>G]C   0.0000 0.000524822   0.0006    0.0020    0.0000    0.0004
#> G[T>G]G   0.0017 0.006087753   0.0050    0.0009    0.0010    0.0011
#> G[T>G]T   0.0009 0.005427434   0.0185    0.0030    0.0000    0.0032
#> T[C>A]A   0.0010 0.027032383   0.0222    0.0374    0.0009    0.0066
#> T[C>A]C   0.0020 0.018089773   0.0043    0.0103    0.0025    0.0099
#> T[C>A]G   0.0002 0.001694875   0.0000    0.0031    0.0001    0.0019
#> T[C>A]T   0.0013 0.038141331   0.0322    0.3083    0.0004    0.0049
#> T[C>G]A   0.0002 0.004117879   0.0033    0.0000    0.0001    0.0024
#> T[C>G]C   0.0010 0.003799650   0.0025    0.0000    0.0011    0.0027
#> T[C>G]G   0.0003 0.000025653   0.0000    0.0000    0.0003    0.0000
#> T[C>G]T   0.0007 0.006247554   0.0049    0.0001    0.0008    0.0015
#> T[C>T]A   0.1202 0.007180669   0.0050    0.0037    0.0302    0.0138
#> T[C>T]C   0.2887 0.004061033   0.0084    0.0211    0.1589    0.0095
#> T[C>T]G   0.0992 0.005535122   0.0047    0.2141    0.0080    0.0010
#> T[C>T]T   0.0844 0.012209070   0.0096    0.0392    0.0954    0.0090
#> T[T>A]A   0.0010 0.015139996   0.0215    0.0000    0.0003    0.0057
#> T[T>A]C   0.0015 0.012093555   0.0027    0.0008    0.0000    0.0132
#> T[T>A]G   0.0010 0.008362547   0.0140    0.0000    0.0000    0.0037
#> T[T>A]T   0.0050 0.027653296   0.0139    0.0028    0.0008    0.0094
#> T[T>C]A   0.0023 0.004901072   0.0156    0.0030    0.0002    0.0500
#> T[T>C]C   0.0018 0.006081011   0.0137    0.0097    0.0009    0.0342
#> T[T>C]G   0.0019 0.001712485   0.0098    0.0065    0.0007    0.0320
#> T[T>C]T   0.0024 0.010003442   0.0309    0.0099    0.0005    0.0371
#> T[T>G]A   0.0000 0.001743221   0.0502    0.0050    0.0000    0.0019
#> T[T>G]C   0.0010 0.002549838   0.0081    0.0092    0.0000    0.0027
#> T[T>G]G   0.0010 0.006030395   0.0088    0.0022    0.0003    0.0011
#> T[T>G]T   0.0014 0.007223999   0.0545    0.0633    0.0003    0.0032
#>            COSMIC_13 COSMIC_14 COSMIC_15 COSMIC_16    COSMIC_17    COSMIC_18
#> A[C>A]A 3.347570e-04    0.0001    0.0013    0.0161 1.832019e-03 5.053642e-02
#> A[C>A]C 6.487360e-04    0.0042    0.0040    0.0097 3.422360e-04 1.093982e-02
#> A[C>A]G 3.814460e-05    0.0005    0.0000    0.0022 1.576230e-06 2.288073e-03
#> A[C>A]T 8.466580e-04    0.0296    0.0057    0.0088 3.179665e-03 1.942409e-02
#> A[C>G]A 3.775165e-03    0.0001    0.0011    0.0048 1.661704e-03 1.516888e-03
#> A[C>G]C 9.208250e-04    0.0000    0.0001    0.0024 1.626898e-03 2.498785e-03
#> A[C>G]G 1.989050e-05    0.0000    0.0006    0.0000 2.580190e-05 2.614509e-03
#> A[C>G]T 3.860632e-03    0.0001    0.0010    0.0073 1.328508e-03 3.983011e-03
#> A[C>T]A 1.480181e-03    0.0293    0.0117    0.0135 9.350488e-03 3.825624e-03
#> A[C>T]C 4.050870e-05    0.0321    0.0169    0.0076 4.224289e-03 2.465214e-03
#> A[C>T]G 3.165950e-04    0.0496    0.0309    0.0035 9.229213e-03 8.463957e-03
#> A[C>T]T 0.000000e+00    0.0277    0.0066    0.0101 5.735961e-03 5.605740e-03
#> A[T>A]A 1.243618e-03    0.0000    0.0021    0.0081 6.130020e-04 3.037211e-03
#> A[T>A]C 1.060002e-03    0.0008    0.0090    0.0073 1.878587e-03 8.111670e-04
#> A[T>A]G 1.025540e-04    0.0000    0.0009    0.0080 1.087350e-03 4.503207e-03
#> A[T>A]T 9.718120e-04    0.0009    0.0079    0.0084 2.045494e-03 2.736523e-03
#> A[T>C]A 5.494540e-04    0.0026    0.0115    0.0717 2.697314e-03 2.933054e-03
#> A[T>C]C 1.455070e-04    0.0046    0.0073    0.0171 8.078150e-03 1.513285e-03
#> A[T>C]G 2.791780e-04    0.0035    0.0073    0.0348 4.432871e-03 3.329669e-03
#> A[T>C]T 2.674610e-04    0.0025    0.0063    0.0497 3.140370e-03 5.491318e-03
#> A[T>G]A 3.862000e-04    0.0000    0.0000    0.0052 5.313920e-05 3.861380e-04
#> A[T>G]C 1.123830e-04    0.0020    0.0006    0.0042 6.468661e-03 1.254982e-03
#> A[T>G]G 3.600070e-05    0.0007    0.0012    0.0061 3.481549e-03 2.369216e-03
#> A[T>G]T 1.924940e-04    0.0056    0.0029    0.0076 3.092638e-02 3.386030e-04
#> C[C>A]A 1.710090e-03    0.0056    0.0106    0.0159 1.032430e-03 8.876811e-02
#> C[C>A]C 1.159257e-03    0.0102    0.0084    0.0100 4.218800e-04 2.064139e-02
#> C[C>A]G 2.441670e-04    0.0009    0.0015    0.0022 2.974630e-04 1.717840e-02
#> C[C>A]T 1.256768e-03    0.1257    0.0228    0.0084 3.147940e-05 3.767696e-02
#> C[C>G]A 5.333651e-03    0.0000    0.0009    0.0040 1.924979e-03 4.584640e-04
#> C[C>G]C 9.123030e-04    0.0000    0.0003    0.0050 7.093680e-04 3.654228e-03
#> C[C>G]G 2.854670e-04    0.0000    0.0004    0.0028 3.517870e-06 5.405774e-03
#> C[C>G]T 6.383098e-03    0.0000    0.0017    0.0080 9.537760e-04 5.519605e-03
#> C[C>T]A 3.647278e-03    0.0060    0.0051    0.0210 2.126521e-02 9.767511e-03
#> C[C>T]C 1.004229e-03    0.0167    0.0088    0.0098 3.847235e-03 5.654828e-03
#> C[C>T]G 1.292020e-04    0.0293    0.0155    0.0016 1.317611e-02 1.618322e-02
#> C[C>T]T 6.117700e-04    0.0186    0.0096    0.0211 9.648025e-03 5.153319e-03
#> C[T>A]A 2.851590e-04    0.0000    0.0016    0.0054 3.007920e-05 4.123460e-04
#> C[T>A]C 2.222870e-04    0.0034    0.0043    0.0128 1.022375e-02 2.491411e-03
#> C[T>A]G 2.869520e-04    0.0002    0.0027    0.0130 4.467036e-03 5.523046e-03
#> C[T>A]T 8.004460e-07    0.0020    0.0007    0.0096 3.124488e-02 3.313713e-03
#> C[T>C]A 4.275150e-05    0.0006    0.0032    0.0146 7.427000e-03 1.921690e-04
#> C[T>C]C 4.330750e-05    0.0053    0.0050    0.0194 3.965700e-02 3.050193e-03
#> C[T>C]G 5.427250e-05    0.0016    0.0128    0.0190 2.860970e-02 2.596464e-03
#> C[T>C]T 7.639440e-07    0.0027    0.0035    0.0263 1.018697e-01 1.310029e-03
#> C[T>G]A 1.509420e-05    0.0000    0.0000    0.0002 9.146980e-05 6.100250e-05
#> C[T>G]C 1.476250e-04    0.0020    0.0019    0.0082 1.981429e-02 2.015697e-03
#> C[T>G]G 4.598730e-05    0.0005    0.0007    0.0067 1.344984e-02 5.311967e-03
#> C[T>G]T 4.637150e-04    0.0063    0.0045    0.0186 2.614566e-01 2.341629e-03
#> G[C>A]A 1.321100e-04    0.0018    0.0024    0.0096 6.535405e-03 1.287242e-01
#> G[C>A]C 7.542440e-04    0.0114    0.0099    0.0094 1.293804e-03 1.609288e-02
#> G[C>A]G 9.682080e-05    0.0011    0.0013    0.0036 1.283663e-03 9.271417e-03
#> G[C>A]T 5.511040e-04    0.0736    0.0309    0.0063 3.103767e-03 7.201707e-02
#> G[C>G]A 1.349420e-03    0.0002    0.0022    0.0031 1.296324e-03 9.238800e-04
#> G[C>G]C 7.888080e-04    0.0001    0.0005    0.0032 2.607610e-04 6.486480e-04
#> G[C>G]G 8.179120e-05    0.0000    0.0006    0.0000 4.377550e-06 5.428400e-04
#> G[C>G]T 2.680123e-03    0.0003    0.0018    0.0029 1.820393e-03 1.376725e-03
#> G[C>T]A 2.934138e-03    0.0557    0.0476    0.0125 2.068755e-03 2.703069e-03
#> G[C>T]C 2.515171e-03    0.0618    0.1345    0.0092 3.254830e-04 4.199308e-03
#> G[C>T]G 3.205265e-03    0.0978    0.1829    0.0009 7.511101e-03 1.189736e-02
#> G[C>T]T 1.687731e-03    0.0555    0.0800    0.0116 9.256970e-04 1.913381e-03
#> G[T>A]A 1.379600e-04    0.0000    0.0000    0.0030 8.545690e-06 1.875434e-03
#> G[T>A]C 2.528670e-04    0.0025    0.0038    0.0040 2.573467e-03 1.510310e-03
#> G[T>A]G 1.535170e-04    0.0000    0.0026    0.0051 3.538823e-03 1.505369e-03
#> G[T>A]T 1.227460e-04    0.0009    0.0043    0.0033 3.879010e-03 3.611025e-03
#> G[T>C]A 1.054081e-03    0.0078    0.0195    0.0161 1.140130e-05 5.291260e-04
#> G[T>C]C 5.299000e-04    0.0127    0.0499    0.0099 9.657713e-03 2.279131e-03
#> G[T>C]G 6.300960e-04    0.0082    0.0154    0.0148 4.654129e-03 5.871057e-03
#> G[T>C]T 6.761880e-04    0.0086    0.0094    0.0142 8.149235e-03 4.771566e-03
#> G[T>G]A 1.848990e-05    0.0000    0.0000    0.0000 3.919380e-05 1.341532e-03
#> G[T>G]C 9.337350e-05    0.0019    0.0022    0.0032 9.078462e-03 9.204310e-07
#> G[T>G]G 1.057920e-05    0.0012    0.0037    0.0008 4.782775e-03 9.695000e-03
#> G[T>G]T 5.374120e-05    0.0038    0.0190    0.0044 6.349776e-02 2.890554e-03
#> T[C>A]A 4.816605e-02    0.0005    0.0057    0.0122 1.824558e-03 6.708064e-02
#> T[C>A]C 1.732956e-02    0.0166    0.0062    0.0145 1.286518e-03 4.186901e-02
#> T[C>A]G 2.293157e-03    0.0000    0.0025    0.0040 2.172942e-03 1.516238e-02
#> T[C>A]T 1.863458e-02    0.0516    0.0145    0.0157 5.012316e-03 1.178778e-01
#> T[C>G]A 2.802349e-01    0.0000    0.0009    0.0050 3.926850e-04 1.802969e-03
#> T[C>G]C 6.388529e-02    0.0000    0.0026    0.0064 6.556020e-04 7.863007e-03
#> T[C>G]G 9.528486e-03    0.0000    0.0000    0.0000 1.364150e-05 1.513223e-03
#> T[C>G]T 3.302255e-01    0.0001    0.0015    0.0169 2.852380e-03 3.567102e-03
#> T[C>T]A 1.138421e-01    0.0092    0.0112    0.0172 1.610079e-02 1.492339e-02
#> T[C>T]C 1.502484e-02    0.0283    0.0156    0.0131 1.677998e-02 9.962744e-03
#> T[C>T]G 6.102098e-03    0.0094    0.0104    0.0036 1.627210e-02 1.154971e-02
#> T[C>T]T 2.814066e-02    0.0229    0.0076    0.0193 1.982742e-02 5.398014e-03
#> T[T>A]A 4.032610e-04    0.0000    0.0008    0.0085 6.988120e-04 4.663278e-03
#> T[T>A]C 7.718130e-04    0.0016    0.0019    0.0087 2.579359e-03 2.578156e-03
#> T[T>A]G 2.451940e-04    0.0000    0.0006    0.0063 2.113360e-05 2.016738e-03
#> T[T>A]T 1.603460e-04    0.0012    0.0003    0.0080 3.883131e-03 2.818679e-03
#> T[T>C]A 7.840560e-04    0.0018    0.0068    0.0246 1.708587e-03 3.023040e-04
#> T[T>C]C 1.329750e-04    0.0056    0.0099    0.0183 1.107415e-02 6.260791e-03
#> T[T>C]G 3.740240e-05    0.0034    0.0039    0.0165 5.932003e-03 1.541265e-03
#> T[T>C]T 1.797030e-04    0.0026    0.0047    0.0174 7.208218e-03 1.357797e-03
#> T[T>G]A 5.465820e-04    0.0000    0.0000    0.0068 1.334110e-04 3.631840e-05
#> T[T>G]C 2.353470e-04    0.0015    0.0004    0.0069 9.613345e-03 3.233838e-03
#> T[T>G]G 4.790220e-07    0.0002    0.0009    0.0049 4.522462e-03 7.546020e-04
#> T[T>G]T 6.705880e-04    0.0025    0.0033    0.0163 5.804041e-02 2.126442e-03
#>         COSMIC_19    COSMIC_20 COSMIC_21   COSMIC_22    COSMIC_23    COSMIC_24
#> A[C>A]A    0.0107 1.179962e-03    0.0001 0.001504070 0.0004533610 2.864599e-02
#> A[C>A]C    0.0074 2.211505e-03    0.0007 0.002451011 0.0003668000 2.021464e-02
#> A[C>A]G    0.0005 1.616910e-07    0.0000 0.000000000 0.0000000000 2.047900e-02
#> A[C>A]T    0.0074 3.008010e-03    0.0006 0.000922453 0.0000000000 2.460014e-02
#> A[C>G]A    0.0058 6.970410e-04    0.0005 0.000527644 0.0000000000 1.199304e-02
#> A[C>G]C    0.0019 2.059310e-03    0.0008 0.000000000 0.0000000000 8.425120e-03
#> A[C>G]G    0.0000 1.273500e-05    0.0000 0.000000000 0.0000000000 0.000000e+00
#> A[C>G]T    0.0072 8.485870e-04    0.0018 0.000299455 0.0000000000 3.881383e-03
#> A[C>T]A    0.0221 3.285805e-02    0.0051 0.003113390 0.0197671540 6.315518e-03
#> A[C>T]C    0.0203 2.219605e-02    0.0062 0.001524860 0.0818805799 3.844327e-03
#> A[C>T]G    0.0368 3.457201e-02    0.0256 0.003625641 0.0092487150 3.110664e-03
#> A[C>T]T    0.0266 1.881956e-02    0.0030 0.000000000 0.0332096519 3.420419e-03
#> A[T>A]A    0.0011 5.744010e-04    0.0083 0.049619230 0.0000000000 1.166747e-03
#> A[T>A]C    0.0053 1.971462e-03    0.0292 0.011666740 0.0000000000 1.517310e-05
#> A[T>A]G    0.0045 1.905870e-04    0.0027 0.067915453 0.0002374190 6.332456e-03
#> A[T>A]T    0.0033 7.421104e-03    0.0043 0.014599918 0.0000000000 4.967220e-06
#> A[T>C]A    0.0145 1.639506e-02    0.0265 0.016839931 0.0010684490 4.726009e-03
#> A[T>C]C    0.0048 1.925179e-02    0.0165 0.001108262 0.0002215730 6.279490e-04
#> A[T>C]G    0.0135 5.478492e-02    0.0224 0.004913616 0.0004988020 2.723759e-03
#> A[T>C]T    0.0160 7.341270e-03    0.0189 0.002630524 0.0003132410 7.122368e-03
#> A[T>G]A    0.0014 1.765290e-06    0.0000 0.001651999 0.0000000000 0.000000e+00
#> A[T>G]C    0.0000 3.813880e-04    0.0031 0.000000000 0.0000000000 1.463951e-03
#> A[T>G]G    0.0010 6.587580e-04    0.0016 0.001086873 0.0000000000 8.476130e-04
#> A[T>G]T    0.0017 5.321250e-04    0.0018 0.000000000 0.0000000000 1.075455e-03
#> C[C>A]A    0.0112 1.737711e-02    0.0020 0.004549693 0.0001647390 6.355928e-02
#> C[C>A]C    0.0159 3.650246e-02    0.0014 0.003764474 0.0007368750 3.375700e-02
#> C[C>A]G    0.0018 1.248259e-02    0.0027 0.000900163 0.0001639540 2.242899e-02
#> C[C>A]T    0.0096 1.034012e-01    0.0056 0.004439846 0.0007227320 2.008651e-02
#> C[C>G]A    0.0002 1.331630e-05    0.0001 0.000582018 0.0001990490 1.347682e-02
#> C[C>G]C    0.0028 5.481450e-04    0.0000 0.000825300 0.0002837760 1.610569e-02
#> C[C>G]G    0.0002 0.000000e+00    0.0000 0.000382115 0.0000000000 7.974720e-04
#> C[C>G]T    0.0008 1.207310e-04    0.0003 0.001760646 0.0001957750 1.038867e-02
#> C[C>T]A    0.0438 1.114518e-02    0.0034 0.007731350 0.0605726279 6.035447e-03
#> C[C>T]C    0.0753 1.815275e-02    0.0091 0.007110851 0.1863756337 0.000000e+00
#> C[C>T]G    0.0238 4.438494e-02    0.0318 0.005502031 0.0151490830 2.924419e-03
#> C[C>T]T    0.1159 1.384714e-02    0.0040 0.004169777 0.1560153968 3.971202e-03
#> C[T>A]A    0.0043 4.504240e-05    0.0000 0.086898545 0.0005945280 2.766635e-03
#> C[T>A]C    0.0053 4.389056e-03    0.0037 0.053706847 0.0000000000 2.694462e-03
#> C[T>A]G    0.0021 1.127836e-03    0.0025 0.213220865 0.0001731200 6.911165e-03
#> C[T>A]T    0.0022 2.385541e-03    0.0010 0.061827507 0.0004578130 2.915200e-03
#> C[T>C]A    0.0066 1.486770e-02    0.0219 0.004027346 0.0000000000 1.848877e-03
#> C[T>C]C    0.0070 1.835616e-02    0.0100 0.001352444 0.0000000000 4.453690e-03
#> C[T>C]G    0.0040 7.415056e-02    0.0211 0.002749666 0.0003207050 0.000000e+00
#> C[T>C]T    0.0053 1.100112e-02    0.0204 0.002363354 0.0000000000 1.597389e-03
#> C[T>G]A    0.0000 2.326410e-05    0.0000 0.001120357 0.0000000000 0.000000e+00
#> C[T>G]C    0.0011 8.476495e-03    0.0005 0.001064584 0.0001901440 2.185437e-03
#> C[T>G]G    0.0013 1.538378e-02    0.0004 0.004188568 0.0000000000 1.139100e-04
#> C[T>G]T    0.0019 2.185395e-03    0.0032 0.001629410 0.0002108350 2.533718e-03
#> G[C>A]A    0.0032 1.116124e-03    0.0001 0.001298370 0.0003499070 5.467645e-02
#> G[C>A]C    0.0031 3.271429e-03    0.0038 0.001809622 0.0005683380 1.099460e-01
#> G[C>A]G    0.0000 3.751613e-03    0.0003 0.000333838 0.0001494180 4.652487e-02
#> G[C>A]T    0.0072 1.293719e-02    0.0023 0.000519148 0.0000000000 5.042871e-02
#> G[C>G]A    0.0020 1.339769e-03    0.0004 0.000000000 0.0000000000 4.945061e-03
#> G[C>G]C    0.0000 1.811151e-03    0.0042 0.000000000 0.0001375010 9.369876e-03
#> G[C>G]G    0.0000 6.375380e-07    0.0000 0.000214696 0.0000000000 5.625727e-03
#> G[C>G]T    0.0000 1.238264e-03    0.0004 0.000538439 0.0000000000 3.055770e-03
#> G[C>T]A    0.0374 5.679395e-02    0.0163 0.001915371 0.0532820589 2.367905e-02
#> G[C>T]C    0.0614 5.442986e-02    0.0392 0.003229134 0.1629183437 1.591920e-02
#> G[C>T]G    0.0278 4.592689e-03    0.0019 0.003300599 0.0130464950 1.264808e-02
#> G[C>T]T    0.0917 4.958086e-02    0.0166 0.001114060 0.1075903998 1.812978e-02
#> G[T>A]A    0.0026 1.631300e-04    0.0000 0.053709146 0.0000000000 2.225127e-03
#> G[T>A]C    0.0000 3.799180e-03    0.0142 0.009489397 0.0001832040 4.853149e-03
#> G[T>A]G    0.0016 8.283340e-04    0.0031 0.050293902 0.0000000000 1.957590e-04
#> G[T>A]T    0.0052 1.900747e-03    0.0041 0.010601457 0.0000000000 0.000000e+00
#> G[T>C]A    0.0039 2.469317e-03    0.1426 0.004217454 0.0002511690 1.269750e-03
#> G[T>C]C    0.0030 1.565633e-03    0.0665 0.000000000 0.0005382180 6.710340e-03
#> G[T>C]G    0.0060 2.734175e-02    0.0884 0.001474885 0.0000000000 2.776850e-04
#> G[T>C]T    0.0099 1.498265e-03    0.1003 0.001561043 0.0009957700 4.238478e-03
#> G[T>G]A    0.0000 3.579150e-06    0.0000 0.000000000 0.0000000000 0.000000e+00
#> G[T>G]C    0.0000 2.830575e-03    0.0006 0.000000000 0.0000000000 1.073632e-03
#> G[T>G]G    0.0043 2.733219e-03    0.0004 0.001877589 0.0001871320 2.924697e-03
#> G[T>G]T    0.0025 3.558559e-03    0.0008 0.000000000 0.0000000000 8.250980e-04
#> T[C>A]A    0.0051 6.883000e-04    0.0001 0.004051635 0.0020712900 3.145126e-02
#> T[C>A]C    0.0063 4.058229e-03    0.0005 0.002739771 0.0006964100 6.253829e-02
#> T[C>A]G    0.0018 1.001102e-03    0.0000 0.000513151 0.0000993936 1.280259e-02
#> T[C>A]T    0.0087 1.222910e-02    0.0004 0.003605651 0.0002074300 3.029768e-02
#> T[C>G]A    0.0005 2.429850e-05    0.0002 0.000225491 0.0002306090 1.022086e-02
#> T[C>G]C    0.0011 1.519990e-04    0.0003 0.000628695 0.0001657870 7.578365e-03
#> T[C>G]G    0.0000 0.000000e+00    0.0000 0.000395808 0.0000000000 1.592714e-03
#> T[C>G]T    0.0105 6.830600e-04    0.0003 0.000257375 0.0004389560 6.100727e-03
#> T[C>T]A    0.0218 1.071032e-02    0.0064 0.002016722 0.0145586140 2.657189e-02
#> T[C>T]C    0.0330 8.175510e-03    0.0086 0.001990834 0.0435104469 1.205569e-02
#> T[C>T]G    0.0090 2.316390e-02    0.0126 0.001498473 0.0027526410 5.363360e-03
#> T[C>T]T    0.0267 9.540087e-03    0.0064 0.001182227 0.0243287850 5.532499e-03
#> T[T>A]A    0.0039 9.245480e-04    0.0027 0.040810503 0.0000000000 0.000000e+00
#> T[T>A]C    0.0048 7.082070e-04    0.0045 0.025570396 0.0000000000 7.994129e-03
#> T[T>A]G    0.0052 9.700310e-05    0.0000 0.069845217 0.0002315260 1.120214e-02
#> T[T>A]T    0.0085 2.313489e-03    0.0008 0.023579062 0.0000000000 2.705790e-04
#> T[T>C]A    0.0052 5.734230e-03    0.0389 0.004803670 0.0000000000 1.682291e-03
#> T[T>C]C    0.0020 1.123425e-02    0.0313 0.000912158 0.0002306090 2.344528e-03
#> T[T>C]G    0.0057 2.376761e-02    0.0249 0.001781736 0.0002395140 4.812740e-04
#> T[T>C]T    0.0038 5.395186e-03    0.0274 0.000000000 0.0000000000 3.581480e-04
#> T[T>G]A    0.0032 2.691470e-07    0.0000 0.001651699 0.0000000000 0.000000e+00
#> T[T>G]C    0.0018 3.772340e-04    0.0018 0.000000000 0.0000000000 0.000000e+00
#> T[T>G]G    0.0011 5.154220e-04    0.0003 0.002572752 0.0002475020 1.360505e-03
#> T[T>G]T    0.0013 6.156570e-04    0.0003 0.000000000 0.0000000000 6.951580e-05
#>           COSMIC_25    COSMIC_26    COSMIC_27    COSMIC_28   COSMIC_29
#> A[C>A]A 0.009896768 2.039773e-03 0.0052056270 0.0013974390 0.069981987
#> A[C>A]C 0.006998929 1.487162e-03 0.0047382270 0.0009171880 0.055152357
#> A[C>A]G 0.001448443 2.839460e-04 0.0007826980 0.0000000000 0.017846984
#> A[C>A]T 0.004966565 5.978660e-04 0.0027182420 0.0005134100 0.026804716
#> A[C>G]A 0.008032568 1.272881e-03 0.0013241630 0.0002551950 0.009301971
#> A[C>G]C 0.001635970 1.528195e-03 0.0017710840 0.0002688990 0.003479049
#> A[C>G]G 0.000000000 3.072460e-04 0.0000000000 0.0000000000 0.000154434
#> A[C>G]T 0.003428450 2.498252e-03 0.0013227200 0.0003076980 0.003976790
#> A[C>T]A 0.020987618 5.907232e-03 0.0137232861 0.0054337920 0.005197221
#> A[C>T]C 0.013141284 1.062646e-02 0.0054243060 0.0018154150 0.008976470
#> A[C>T]G 0.019659894 1.993032e-02 0.0067064290 0.0054320390 0.031638996
#> A[C>T]T 0.005728519 1.133453e-02 0.0044756140 0.0016292130 0.004769667
#> A[T>A]A 0.020430527 4.459340e-03 0.1430763896 0.0009640310 0.000000000
#> A[T>A]C 0.008892561 1.282224e-02 0.0014640850 0.0033656680 0.003049596
#> A[T>A]G 0.006241948 1.172344e-03 0.0012262810 0.0017857540 0.003909982
#> A[T>A]T 0.014850878 3.993138e-03 0.0025351980 0.0056530890 0.005105527
#> A[T>C]A 0.020663283 5.502865e-02 0.0050473010 0.0040927470 0.010119181
#> A[T>C]C 0.009502478 2.759460e-02 0.0018456910 0.0078807290 0.006349683
#> A[T>C]G 0.019911492 5.179104e-02 0.0056888140 0.0055208340 0.014659610
#> A[T>C]T 0.014488390 3.907224e-02 0.0025927550 0.0032324840 0.007892413
#> A[T>G]A 0.005281680 8.065340e-05 0.0011363370 0.0063646760 0.000000000
#> A[T>G]C 0.000000000 1.631710e-04 0.0009589760 0.0166109060 0.000000000
#> A[T>G]G 0.007136422 1.255080e-03 0.0000676606 0.0065928420 0.002925171
#> A[T>G]T 0.000000000 1.850383e-03 0.0002560280 0.0863928968 0.000000000
#> C[C>A]A 0.014832948 3.705850e-03 0.0050650730 0.0011685160 0.051410212
#> C[C>A]C 0.007822175 3.980723e-03 0.0022341530 0.0003342920 0.025825651
#> C[C>A]G 0.001276977 8.117420e-04 0.0002663120 0.0000536520 0.014496183
#> C[C>A]T 0.012563655 1.903843e-02 0.0031005700 0.0001866720 0.040355074
#> C[C>G]A 0.007200790 1.279106e-03 0.0047306490 0.0001033660 0.002618863
#> C[C>G]C 0.004598473 1.214977e-03 0.0012616450 0.0001443110 0.001111512
#> C[C>G]G 0.000823709 2.078330e-04 0.0003608570 0.0000000000 0.001520047
#> C[C>G]T 0.000196360 2.297026e-03 0.0002680260 0.0002910000 0.000139751
#> C[C>T]A 0.014333995 6.594445e-03 0.0174384851 0.0043915330 0.009126472
#> C[C>T]C 0.005936644 6.510632e-03 0.0071448700 0.0026476610 0.007067603
#> C[C>T]G 0.011559392 1.190494e-02 0.0075035620 0.0022845640 0.019875180
#> C[C>T]T 0.013632597 6.238556e-03 0.0071480280 0.0024239840 0.001634059
#> C[T>A]A 0.026296323 3.561423e-03 0.3306010183 0.0025665800 0.001765448
#> C[T>A]C 0.015799883 3.901611e-03 0.0028135990 0.0196108970 0.000383359
#> C[T>A]G 0.062860606 2.390304e-03 0.0043968570 0.0088336000 0.003556845
#> C[T>A]T 0.036622351 1.635923e-03 0.0054308920 0.0109426100 0.002232755
#> C[T>C]A 0.009028015 3.788938e-02 0.0048990790 0.0207380030 0.000000000
#> C[T>C]C 0.010771632 2.174074e-02 0.0057244480 0.0432837969 0.002774954
#> C[T>C]G 0.013350661 4.724010e-02 0.0056614790 0.0370882619 0.006902415
#> C[T>C]T 0.008961030 2.074062e-02 0.0039005890 0.0050527160 0.003041425
#> C[T>G]A 0.000000000 9.539630e-05 0.0000000000 0.0095898420 0.001116855
#> C[T>G]C 0.002886454 3.919455e-03 0.0003812450 0.0339132239 0.000000000
#> C[T>G]G 0.011769492 5.497357e-03 0.0004032570 0.0185164820 0.002169693
#> C[T>G]T 0.008550807 6.788719e-03 0.0014390960 0.1189671027 0.001143663
#> G[C>A]A 0.013465295 1.375312e-03 0.0107558720 0.0021366290 0.078046610
#> G[C>A]C 0.006468298 1.961639e-03 0.0124423350 0.0004292080 0.073460915
#> G[C>A]G 0.000658103 1.322970e-05 0.0016296280 0.0000000000 0.017383297
#> G[C>A]T 0.010473080 1.934996e-03 0.0074123550 0.0013185070 0.059602472
#> G[C>G]A 0.004169656 1.320607e-03 0.0014181660 0.0000815914 0.000265760
#> G[C>G]C 0.002392805 1.845501e-03 0.0013467170 0.0001052190 0.000945013
#> G[C>G]G 0.000418194 2.050610e-04 0.0000000000 0.0000000000 0.000000000
#> G[C>G]T 0.001994012 1.225533e-03 0.0045606860 0.0001596790 0.004796745
#> G[C>T]A 0.010038112 9.606515e-03 0.0142153141 0.0033842410 0.011638793
#> G[C>T]C 0.006953875 1.950743e-02 0.0201436461 0.0016874140 0.003781752
#> G[C>T]G 0.010468363 2.250272e-02 0.0093684680 0.0021893090 0.019684388
#> G[C>T]T 0.009823821 1.730747e-02 0.0075896260 0.0010439940 0.005761337
#> G[T>A]A 0.036217801 2.242887e-03 0.1098518724 0.0005514610 0.002930098
#> G[T>A]C 0.005847110 5.207287e-03 0.0000234557 0.0031639880 0.000000000
#> G[T>A]G 0.038241179 1.358122e-03 0.0015659370 0.0021528470 0.000533632
#> G[T>A]T 0.020167954 2.513113e-03 0.0000000000 0.0020077330 0.000000000
#> G[T>C]A 0.017489824 9.805283e-02 0.0047773800 0.0036550480 0.007090804
#> G[T>C]C 0.004045612 4.022636e-02 0.0019649540 0.0070711750 0.007111328
#> G[T>C]G 0.013402246 4.462070e-02 0.0012291680 0.0082322380 0.005643064
#> G[T>C]T 0.014159694 5.552807e-02 0.0037472250 0.0004078850 0.011237674
#> G[T>G]A 0.005546227 3.739320e-05 0.0000000000 0.0000000000 0.000000000
#> G[T>G]C 0.000000000 2.460705e-03 0.0000507906 0.0000000000 0.000000000
#> G[T>G]G 0.008685244 8.172020e-04 0.0047429190 0.0042178080 0.003640594
#> G[T>G]T 0.002768572 7.833561e-03 0.0022984760 0.0316489969 0.006181024
#> T[C>A]A 0.006938147 2.680062e-03 0.0070792850 0.0007759030 0.026894407
#> T[C>A]C 0.010246607 2.031644e-03 0.0062688010 0.0008292920 0.031608137
#> T[C>A]G 0.000000000 2.650720e-04 0.0014464030 0.0000000000 0.012500468
#> T[C>A]T 0.024625787 3.017187e-03 0.0098062780 0.0019980740 0.041207176
#> T[C>G]A 0.009798725 4.201897e-03 0.0028704340 0.0005213700 0.000820445
#> T[C>G]C 0.003266507 2.807762e-03 0.0031079680 0.0006673260 0.000444119
#> T[C>G]G 0.002057437 6.197000e-09 0.0000128104 0.0000000000 0.000143297
#> T[C>G]T 0.016400917 8.019383e-03 0.0021305880 0.0012457800 0.001369869
#> T[C>T]A 0.011077521 1.130329e-02 0.0187577761 0.0076222580 0.012076583
#> T[C>T]C 0.006448788 1.080753e-02 0.0079891840 0.0051166190 0.000000000
#> T[C>T]G 0.014083416 1.036433e-02 0.0046091310 0.0019180280 0.015830197
#> T[C>T]T 0.012418586 7.249145e-03 0.0118809320 0.0061931590 0.013462137
#> T[T>A]A 0.021830329 6.281450e-04 0.0528138872 0.0002032380 0.001191207
#> T[T>A]C 0.003405687 5.073754e-03 0.0000470918 0.0101809650 0.001682664
#> T[T>A]G 0.010191520 2.013840e-05 0.0000000000 0.0032504270 0.002438756
#> T[T>A]T 0.014527539 1.235538e-03 0.0019297710 0.0164928120 0.000000000
#> T[T>C]A 0.014515050 2.878968e-02 0.0017210150 0.0099295520 0.011498666
#> T[T>C]C 0.007406083 3.663949e-02 0.0034892130 0.0290105629 0.003584215
#> T[T>C]G 0.004892531 1.914369e-02 0.0011528470 0.0156832190 0.001694297
#> T[T>C]T 0.017014781 2.307169e-02 0.0015279570 0.0172915640 0.003799706
#> T[T>G]A 0.002081480 8.534940e-06 0.0011890220 0.0093896590 0.000000000
#> T[T>G]C 0.000578979 2.718510e-03 0.0002802950 0.0301170769 0.000000000
#> T[T>G]G 0.009429196 1.369161e-03 0.0023530560 0.0126987510 0.000353696
#> T[T>G]T 0.007869672 2.568077e-03 0.0001395610 0.2336597824 0.006104834
#>         COSMIC_30
#> A[C>A]A 0.0000000
#> A[C>A]C 0.0000000
#> A[C>A]G 0.0019673
#> A[C>A]T 0.0000000
#> A[C>G]A 0.0000000
#> A[C>G]C 0.0000000
#> A[C>G]G 0.0048200
#> A[C>G]T 0.0000000
#> A[C>T]A 0.0651190
#> A[C>T]C 0.0543970
#> A[C>T]G 0.0204604
#> A[C>T]T 0.0219359
#> A[T>A]A 0.0075743
#> A[T>A]C 0.0037380
#> A[T>A]G 0.0065906
#> A[T>A]T 0.0093447
#> A[T>C]A 0.0000000
#> A[T>C]C 0.0087547
#> A[T>C]G 0.0090498
#> A[T>C]T 0.0000000
#> A[T>G]A 0.0053118
#> A[T>G]C 0.0024592
#> A[T>G]G 0.0062955
#> A[T>G]T 0.0095416
#> C[C>A]A 0.0000000
#> C[C>A]C 0.0000000
#> C[C>A]G 0.0022624
#> C[C>A]T 0.0000000
#> C[C>G]A 0.0000000
#> C[C>G]C 0.0000000
#> C[C>G]G 0.0032461
#> C[C>G]T 0.0000000
#> C[C>T]A 0.0694472
#> C[C>T]C 0.0638403
#> C[C>T]G 0.0173126
#> C[C>T]T 0.0342318
#> C[T>A]A 0.0053118
#> C[T>A]C 0.0067873
#> C[T>A]G 0.0082628
#> C[T>A]T 0.0000000
#> C[T>C]A 0.0000000
#> C[T>C]C 0.0000000
#> C[T>C]G 0.0000000
#> C[T>C]T 0.0000000
#> C[T>G]A 0.0046233
#> C[T>G]C 0.0060004
#> C[T>G]G 0.0073775
#> C[T>G]T 0.0091481
#> G[C>A]A 0.0088530
#> G[C>A]C 0.0093449
#> G[C>A]G 0.0008853
#> G[C>A]T 0.0081645
#> G[C>G]A 0.0073775
#> G[C>G]C 0.0065906
#> G[C>G]G 0.0015739
#> G[C>G]T 0.0000000
#> G[C>T]A 0.0485934
#> G[C>T]C 0.0494787
#> G[C>T]G 0.0157387
#> G[C>T]T 0.0188865
#> G[T>A]A 0.0037380
#> G[T>A]C 0.0033445
#> G[T>A]G 0.0050167
#> G[T>A]T 0.0063939
#> G[T>C]A 0.0000000
#> G[T>C]C 0.0065906
#> G[T>C]G 0.0078694
#> G[T>C]T 0.0084596
#> G[T>G]A 0.0032461
#> G[T>G]C 0.0018690
#> G[T>G]G 0.0033445
#> G[T>G]T 0.0056069
#> T[C>A]A 0.0000000
#> T[C>A]C 0.0000000
#> T[C>A]G 0.0016722
#> T[C>A]T 0.0000000
#> T[C>G]A 0.0000000
#> T[C>G]C 0.0000000
#> T[C>G]G 0.0019673
#> T[C>G]T 0.0000000
#> T[C>T]A 0.0849892
#> T[C>T]C 0.0903010
#> T[C>T]G 0.0151485
#> T[C>T]T 0.0455440
#> T[T>A]A 0.0076726
#> T[T>A]C 0.0064922
#> T[T>A]G 0.0036396
#> T[T>A]T 0.0000000
#> T[T>C]A 0.0000000
#> T[T>C]C 0.0091481
#> T[T>C]G 0.0060004
#> T[T>C]T 0.0000000
#> T[T>G]A 0.0086563
#> T[T>G]C 0.0043282
#> T[T>G]G 0.0082628
#> T[T>G]T 0.0000000
#> 
#> $aetiology
#>                                             aetiology
#> COSMIC_1  spontaneous deamination of 5-methylcytosine
#> COSMIC_2              APOBEC Cytidine Deaminase (C>T)
#> COSMIC_3              defects in DNA-DSB repair by HR
#> COSMIC_4       exposure to tobacco (smoking) mutagens
#> COSMIC_5                                      Unknown
#> COSMIC_6                defective DNA mismatch repair
#> COSMIC_7                                  UV exposure
#> COSMIC_8                                      Unknown
#> COSMIC_9                    defects in polymerase-eta
#> COSMIC_10                  defects in polymerase POLE
#> COSMIC_11               exposure to alkylating agents
#> COSMIC_12                                     Unknown
#> COSMIC_13             APOBEC Cytidine Deaminase (C>G)
#> COSMIC_14                                     Unknown
#> COSMIC_15               defective DNA mismatch repair
#> COSMIC_16                                     Unknown
#> COSMIC_17                                     Unknown
#> COSMIC_18                                     Unknown
#> COSMIC_19                                     Unknown
#> COSMIC_20               defective DNA mismatch repair
#> COSMIC_21                                     unknown
#> COSMIC_22               exposure to aristolochic acid
#> COSMIC_23                                     Unknown
#> COSMIC_24                      exposures to aflatoxin
#> COSMIC_25                                     Unknown
#> COSMIC_26               defective DNA mismatch repair
#> COSMIC_27                                     Unknown
#> COSMIC_28                                     Unknown
#> COSMIC_29      exposure to tobacco (chewing) mutagens
#> COSMIC_30                                     Unknown
#> 

s2

#> $db
#>               DBS1         DBS2         DBS3         DBS4         DBS5
#> AC>CA 4.971263e-05 3.740792e-04 6.347323e-03 4.141857e-03 2.139822e-03
#> AC>CG 2.640671e-05 3.170671e-05 4.648040e-05 4.962225e-05 4.619616e-04
#> AC>CT 4.291091e-03 1.260267e-04 3.018727e-03 8.323733e-18 2.019832e-03
#> AC>GA 1.270323e-04 2.370502e-04 5.847534e-03 4.401974e-03 5.579536e-04
#> AC>GG 5.711452e-05 4.170883e-18 4.498103e-04 7.013145e-18 1.109908e-17
#> AC>GT 2.960753e-03 2.800593e-04 1.009574e-02 1.120503e-08 8.219316e-04
#> AC>TA 1.200305e-03 1.480313e-03 1.009574e-03 1.270570e-03 1.479877e-03
#> AC>TG 1.720437e-04 6.361346e-06 1.109532e-17 5.312382e-05 1.119907e-03
#> AC>TT 2.650674e-02 2.920618e-03 1.779250e-06 1.610722e-04 7.469379e-03
#> AT>CA 4.821225e-04 3.780800e-05 7.037032e-03 9.164110e-04 1.079910e-04
#> AT>CC 2.970755e-04 4.170883e-18 9.076172e-03 7.003141e-18 2.129823e-05
#> AT>CG 3.981012e-06 4.170883e-18 1.459385e-02 9.464244e-18 1.049913e-17
#> AT>GA 2.120539e-03 1.680356e-04 1.809237e-03 3.241454e-04 1.109908e-03
#> AT>GC 1.740442e-04 5.851238e-06 5.247787e-03 1.760790e-05 1.049913e-17
#> AT>TA 2.340595e-03 4.580969e-04 2.349009e-03 4.532032e-03 7.929340e-04
#> CC>AA 1.060269e-05 7.241532e-01 9.086168e-03 7.003141e-18 1.049913e-17
#> CC>AG 1.230313e-05 7.261537e-02 1.109532e-17 5.432436e-03 2.429798e-02
#> CC>AT 1.870475e-02 1.000212e-01 5.887517e-03 1.570704e-02 2.429798e-03
#> CC>GA 1.340341e-04 1.770375e-02 6.347323e-04 3.551593e-03 6.759438e-05
#> CC>GG 9.822496e-05 4.500952e-04 7.246944e-05 7.303275e-05 1.189901e-05
#> CC>GT 6.921759e-03 1.100233e-03 1.139519e-17 7.993585e-05 2.479794e-05
#> CC>TA 1.040264e-02 1.600339e-02 2.079123e-02 1.700763e-04 1.269894e-03
#> CC>TG 2.550648e-03 2.370502e-04 2.898778e-05 8.303724e-06 4.809600e-03
#> CC>TT 8.582181e-01 4.170883e-18 1.109532e-17 7.003141e-18 1.049913e-17
#> CG>AT 6.031533e-07 1.150243e-03 2.838803e-04 2.901301e-04 1.639864e-03
#> CG>GC 5.421378e-06 3.690781e-04 1.109532e-17 5.102288e-06 9.919175e-05
#> CG>GT 1.700432e-06 1.280271e-03 1.109532e-17 7.003141e-18 6.479461e-03
#> CG>TA 3.560905e-04 4.170883e-18 9.785873e-02 7.003141e-18 1.049913e-17
#> CG>TC 1.260320e-05 2.150455e-03 7.216956e-04 7.003141e-18 2.979752e-05
#> CG>TT 1.240315e-04 5.561177e-03 7.166978e-03 3.611620e-04 2.219815e-03
#> CT>AA 2.770704e-05 3.210679e-03 1.029566e-04 7.003141e-18 4.659612e-01
#> CT>AC 4.191065e-04 4.170883e-18 2.908773e-02 7.003141e-18 2.089826e-01
#> CT>AG 2.850725e-18 5.871242e-04 1.139519e-01 3.631629e-04 2.669778e-02
#> CT>GA 8.922268e-05 1.810383e-05 1.719275e-05 6.733020e-04 1.609866e-02
#> CT>GC 2.520641e-04 1.340284e-04 2.688866e-05 7.223239e-05 4.849597e-03
#> CT>GG 1.080275e-05 8.981901e-05 5.987475e-05 9.264155e-04 5.699526e-04
#> CT>TA 7.811985e-03 3.380715e-03 1.839224e-04 7.003141e-18 3.339722e-03
#> CT>TC 1.050267e-02 3.840813e-05 6.337327e-03 7.003141e-18 2.369803e-04
#> CT>TG 2.240569e-03 3.900825e-04 5.217800e-02 3.081382e-05 8.739273e-04
#> GC>AA 1.580402e-03 5.081075e-03 2.159089e-05 3.281472e-01 1.329889e-04
#> GC>AG 1.070272e-04 3.250688e-04 5.217800e-04 1.130507e-02 2.659779e-03
#> GC>AT 4.871238e-05 9.532017e-04 3.988318e-03 9.244146e-03 1.569869e-03
#> GC>CA 5.761464e-05 1.060224e-03 2.369001e-04 1.190534e-02 4.579619e-03
#> GC>CG 5.151309e-06 6.691416e-05 1.109532e-17 9.834410e-07 3.229731e-04
#> GC>TA 8.272102e-07 7.361558e-04 4.228217e-03 1.770794e-02 2.219815e-04
#> TA>AT 5.051284e-04 2.750582e-03 1.449389e-03 3.041364e-02 1.869844e-03
#> TA>CG 7.551919e-06 1.890400e-05 3.338592e-03 3.891745e-04 7.889344e-06
#> TA>CT 7.021785e-04 2.800593e-04 3.438550e-04 8.723912e-03 2.269811e-03
#> TA>GC 2.010511e-05 8.221740e-05 1.989161e-03 7.023150e-05 5.219566e-05
#> TA>GG 7.031787e-05 7.891670e-05 5.457698e-03 2.220996e-03 1.629864e-03
#> TA>GT 1.150292e-03 1.300275e-03 3.618474e-04 8.063616e-03 1.909841e-03
#> TC>AA 1.090277e-05 1.450307e-03 5.487686e-03 3.791700e-01 4.039664e-02
#> TC>AG 3.090786e-06 2.000423e-04 6.137412e-06 1.870839e-03 1.489876e-02
#> TC>AT 2.460625e-03 4.170883e-18 1.119528e-03 8.413773e-03 7.479378e-04
#> TC>CA 4.421124e-06 6.821443e-04 6.547239e-02 8.463796e-02 1.429881e-03
#> TC>CG 3.300839e-05 3.890823e-06 1.159511e-04 4.001795e-04 1.059912e-17
#> TC>CT 1.900483e-03 3.350709e-03 1.939182e-02 8.283715e-06 1.049913e-03
#> TC>GA 6.421632e-05 8.851873e-04 7.776720e-02 3.641633e-02 2.459795e-05
#> TC>GG 4.111045e-04 9.872089e-06 1.769254e-04 1.040467e-05 3.479711e-05
#> TC>GT 5.101297e-03 8.571814e-05 4.338171e-02 7.003141e-18 3.859679e-05
#> TG>AA 1.370348e-03 2.960626e-03 6.677184e-05 7.633423e-03 1.799850e-02
#> TG>AC 2.370603e-04 2.200466e-04 8.036611e-05 1.180529e-04 1.459879e-03
#> TG>AT 2.410613e-03 4.290908e-03 9.575962e-04 7.003141e-18 2.039830e-02
#> TG>CA 2.790709e-04 1.440305e-04 1.589330e-02 8.043607e-04 4.759604e-04
#> TG>CC 1.030262e-04 5.941257e-05 4.837960e-04 1.200538e-04 5.059579e-04
#> TG>CT 1.850470e-03 3.070650e-03 6.067441e-03 8.213684e-05 3.279727e-04
#> TG>GA 4.071035e-04 1.470311e-03 5.227795e-04 7.853522e-03 3.269728e-03
#> TG>GC 8.892260e-05 2.720576e-04 4.977901e-04 6.042710e-04 2.919757e-04
#> TG>GT 1.900483e-03 6.401355e-03 1.029566e-02 3.231449e-05 8.269312e-02
#> TT>AA 5.491396e-03 4.310912e-03 3.218643e-03 1.800808e-05 4.919591e-03
#> TT>AC 7.201830e-04 3.450730e-04 2.428976e-03 6.602961e-07 9.659196e-04
#> TT>AG 4.181063e-04 2.150455e-04 2.289035e-02 7.003141e-18 8.929257e-04
#> TT>CA 7.251843e-04 4.170883e-18 6.937074e-04 7.003141e-18 1.049913e-17
#> TT>CC 5.421378e-03 2.980631e-06 8.156560e-03 7.003141e-18 7.729357e-04
#> TT>CG 8.382130e-05 1.520322e-05 3.178659e-02 7.253253e-18 1.049913e-17
#> TT>GA 1.860473e-03 4.791014e-05 1.229482e-02 7.003141e-18 2.989751e-05
#> TT>GC 1.060269e-03 1.540326e-06 6.647197e-02 7.003141e-18 5.659529e-05
#> TT>GG 1.610409e-03 4.170883e-18 1.709279e-01 7.003141e-18 1.049913e-17
#>               DBS6         DBS7         DBS8         DBS9        DBS10
#> AC>CA 2.660571e-03 4.588013e-02 2.149236e-01 2.459182e-02 9.630417e-03
#> AC>CG 1.460313e-04 6.247295e-03 9.726540e-04 2.069312e-03 1.770077e-05
#> AC>CT 2.400515e-04 1.359411e-02 3.178869e-01 2.849052e-02 1.350059e-02
#> AC>GA 8.581841e-03 9.725788e-03 1.159588e-03 2.529159e-02 5.350232e-03
#> AC>GG 1.040223e-04 2.538901e-03 8.097120e-04 7.127629e-03 5.900256e-04
#> AC>GT 1.320283e-04 4.498052e-03 1.509463e-04 1.209598e-02 2.870124e-02
#> AC>TA 1.010217e-02 5.417654e-03 1.239559e-02 4.118630e-02 1.050046e-04
#> AC>TG 4.490964e-04 9.975680e-04 2.569086e-03 1.789405e-02 7.660332e-06
#> AC>TT 3.180682e-03 1.179489e-03 2.679047e-04 4.438523e-02 7.100308e-04
#> AT>CA 1.140245e-04 1.109520e-02 2.779012e-05 1.269578e-02 3.530153e-03
#> AT>CC 6.391371e-05 6.257290e-03 1.429492e-04 1.779408e-03 1.160050e-03
#> AT>CG 1.820391e-06 1.109520e-03 2.659054e-04 9.216934e-18 2.590112e-03
#> AT>GA 1.620348e-03 1.439377e-02 1.619424e-04 6.367881e-03 8.060349e-05
#> AT>GC 1.720369e-05 9.285979e-03 2.259196e-04 1.809398e-03 4.080177e-03
#> AT>TA 2.200472e-03 3.398528e-02 3.148880e-04 9.196940e-03 2.930127e-03
#> CC>AA 8.301781e-18 9.036087e-18 3.808645e-02 9.017000e-18 1.370059e-02
#> CC>AG 6.161322e-02 8.296407e-05 1.549449e-02 2.639122e-02 7.130309e-05
#> CC>AT 4.380940e-04 8.516312e-04 2.878976e-02 1.949351e-02 1.050046e-02
#> CC>GA 3.550762e-02 3.018693e-04 8.087123e-04 1.099634e-02 3.390147e-03
#> CC>GG 1.810388e-02 2.049113e-03 5.598009e-05 3.058982e-03 7.410321e-18
#> CC>GT 3.440738e-03 1.629294e-04 2.289186e-04 1.789405e-02 1.080047e-03
#> CC>TA 2.220476e-03 9.036087e-18 7.157454e-04 6.867715e-04 2.390104e-02
#> CC>TG 6.871474e-03 2.358978e-04 4.208503e-03 1.549484e-02 7.410321e-18
#> CC>TT 8.301781e-18 9.036087e-18 4.378443e-03 9.017000e-18 2.060089e-02
#> CG>AT 7.541618e-03 6.207312e-05 3.218855e-04 2.189272e-04 7.410321e-18
#> CG>GC 5.901266e-04 1.869191e-04 1.129598e-17 1.199601e-05 1.900082e-05
#> CG>GT 3.220691e-03 2.069104e-05 1.209570e-03 9.936694e-05 1.020044e-05
#> CG>TA 8.301781e-18 9.036087e-18 5.058201e-04 9.017000e-18 3.580155e-01
#> CG>TC 8.911912e-03 4.867892e-05 9.046782e-04 1.329558e-04 2.650115e-03
#> CG>TT 1.870401e-02 3.128645e-05 3.288830e-04 8.287243e-05 7.770337e-03
#> CT>AA 4.480961e-03 9.036087e-18 2.059268e-02 9.017000e-18 3.930170e-03
#> CT>AC 1.950419e-03 1.009563e-17 1.399502e-02 9.017000e-18 1.300056e-02
#> CT>AG 3.950848e-02 1.259455e-03 2.039275e-02 5.818064e-04 2.650115e-02
#> CT>GA 7.341575e-03 1.869191e-03 5.568019e-03 2.839055e-03 7.410321e-18
#> CT>GC 1.790384e-03 1.559325e-02 2.369157e-04 3.108966e-03 7.410321e-18
#> CT>GG 6.811462e-03 3.998269e-03 3.058912e-04 1.039654e-02 7.410321e-18
#> CT>TA 8.841897e-05 9.036087e-18 1.089612e-02 9.017000e-18 4.030175e-03
#> CT>TC 8.301781e-18 6.317264e-02 1.779367e-02 2.999002e-05 2.750119e-03
#> CT>TG 1.120240e-02 9.036087e-18 9.656565e-04 9.456854e-18 4.570198e-02
#> GC>AA 8.301781e-18 9.036087e-18 1.289541e-02 9.017000e-18 4.080177e-03
#> GC>AG 3.660785e-03 7.296840e-05 8.516971e-03 1.239588e-01 7.410321e-18
#> GC>AT 1.880403e-03 3.498485e-04 5.508041e-03 1.259581e-01 7.750336e-03
#> GC>CA 8.691865e-03 6.697100e-04 4.828283e-03 2.739089e-02 1.050046e-04
#> GC>CG 9.041940e-18 2.648853e-03 2.419140e-04 2.499169e-03 7.410321e-18
#> GC>TA 8.291779e-04 6.237299e-06 8.556956e-03 1.399534e-02 2.810122e-03
#> TA>AT 8.571839e-03 5.577585e-02 9.696551e-03 1.549484e-02 1.040045e-02
#> TA>CG 2.590556e-04 3.048680e-03 2.089257e-05 1.349551e-03 4.110178e-03
#> TA>CT 2.610560e-02 2.908740e-02 4.098542e-03 1.419528e-02 1.620070e-03
#> TA>GC 2.640567e-04 1.509346e-03 7.887195e-04 5.788074e-04 7.410321e-18
#> TA>GG 2.180468e-04 4.308134e-03 2.169228e-04 3.248919e-03 8.590372e-03
#> TA>GT 6.251341e-05 2.628862e-02 1.529456e-04 7.777412e-03 3.610156e-03
#> TC>AA 4.580983e-02 1.959152e-02 3.878620e-02 3.688773e-03 7.560328e-03
#> TC>AG 8.281777e-03 3.998269e-06 1.159588e-02 4.668447e-02 2.710117e-04
#> TC>AT 5.941275e-04 2.289009e-03 1.849342e-02 5.178277e-02 3.090134e-03
#> TC>CA 1.180253e-02 2.708827e-03 1.759374e-02 4.878377e-02 2.520109e-02
#> TC>CG 8.621850e-04 1.629294e-03 1.089612e-02 7.787409e-03 8.420365e-04
#> TC>CT 9.392015e-03 9.186022e-03 8.756885e-03 5.138290e-02 2.620114e-02
#> TC>GA 7.341575e-03 9.036087e-18 9.606583e-03 1.729425e-02 4.380190e-02
#> TC>GG 1.510324e-03 7.056944e-04 1.909321e-03 3.728759e-03 7.410321e-18
#> TC>GT 8.301781e-18 9.036087e-18 2.759019e-04 9.146957e-18 4.590199e-02
#> TG>AA 2.550547e-02 2.259022e-02 8.646924e-03 1.579475e-02 1.420062e-03
#> TG>AC 1.010217e-02 2.618866e-03 3.848631e-03 7.267582e-03 2.070090e-04
#> TG>AT 2.210474e-01 1.289442e-03 8.217077e-03 7.457519e-04 2.500108e-03
#> TG>CA 1.140245e-02 2.388965e-02 5.058201e-03 2.279242e-02 2.920127e-02
#> TG>CC 7.891693e-03 1.049546e-03 1.249556e-03 1.179608e-02 1.900082e-04
#> TG>CT 2.750590e-01 9.036087e-18 1.949307e-02 9.676781e-04 4.330188e-03
#> TG>GA 7.441597e-03 8.246429e-03 2.709036e-04 6.177945e-03 2.020088e-05
#> TG>GC 5.601202e-03 3.878321e-03 3.538741e-03 3.908700e-03 7.410321e-18
#> TG>GT 2.740588e-02 8.706230e-03 7.857205e-03 3.338889e-03 2.170094e-03
#> TT>AA 1.930414e-04 2.348983e-01 3.578727e-03 3.408866e-04 1.450063e-02
#> TT>AC 1.280275e-04 3.998269e-02 1.129598e-17 6.657785e-03 7.410321e-18
#> TT>AG 2.120455e-03 5.387667e-02 1.459481e-02 9.176947e-04 6.960302e-03
#> TT>CA 8.301781e-18 9.765771e-02 6.287763e-03 9.026997e-18 1.290056e-03
#> TT>CC 8.301781e-18 3.538468e-02 2.938955e-04 5.158284e-03 2.880125e-03
#> TT>CG 1.120240e-05 2.668844e-03 1.159588e-04 1.979341e-05 8.580372e-03
#> TT>GA 8.301781e-18 4.358113e-02 3.628709e-03 9.166950e-18 8.690377e-03
#> TT>GC 4.090878e-07 3.668411e-03 2.858983e-04 3.558816e-05 2.580112e-02
#> TT>GG 8.301781e-18 9.036087e-18 5.378087e-04 9.017000e-18 8.470367e-02
#>              DBS11
#> AC>CA 1.369907e-03
#> AC>CG 1.939868e-04
#> AC>CT 7.559485e-04
#> AC>GA 4.419699e-04
#> AC>GG 2.249847e-03
#> AC>GT 5.709611e-03
#> AC>TA 2.639820e-04
#> AC>TG 2.389837e-04
#> AC>TT 2.819808e-03
#> AT>CA 9.479354e-03
#> AT>CC 1.899870e-03
#> AT>CG 1.739881e-05
#> AT>GA 1.119924e-02
#> AT>GC 4.179715e-03
#> AT>TA 3.839738e-03
#> CC>AA 7.349499e-18
#> CC>AG 9.719337e-04
#> CC>AT 3.449765e-02
#> CC>GA 9.529350e-04
#> CC>GG 3.449765e-03
#> CC>GT 6.949526e-02
#> CC>TA 5.879599e-02
#> CC>TG 3.229780e-02
#> CC>TT 3.319774e-01
#> CG>AT 1.819876e-03
#> CG>GC 1.659887e-04
#> CG>GT 1.179920e-03
#> CG>TA 7.379497e-18
#> CG>TC 6.849533e-04
#> CG>TT 6.989523e-04
#> CT>AA 4.399700e-03
#> CT>AC 2.349840e-03
#> CT>AG 4.959662e-04
#> CT>GA 1.159921e-02
#> CT>GC 6.109583e-03
#> CT>GG 4.019726e-03
#> CT>TA 4.169716e-02
#> CT>TC 2.989796e-03
#> CT>TG 1.479899e-03
#> GC>AA 9.999318e-04
#> GC>AG 3.289776e-03
#> GC>AT 2.719815e-02
#> GC>CA 2.109856e-03
#> GC>CG 4.039725e-03
#> GC>TA 2.429834e-04
#> TA>AT 2.229848e-03
#> TA>CG 5.839602e-04
#> TA>CT 3.379770e-03
#> TA>GC 1.019930e-03
#> TA>GG 3.749744e-04
#> TA>GT 2.669818e-03
#> TC>AA 2.169852e-04
#> TC>AG 4.739677e-02
#> TC>AT 8.459423e-02
#> TC>CA 2.819808e-04
#> TC>CG 3.239779e-03
#> TC>CT 2.009863e-02
#> TC>GA 2.349840e-04
#> TC>GG 7.329500e-03
#> TC>GT 1.519896e-02
#> TG>AA 7.829466e-03
#> TG>AC 8.659410e-04
#> TG>AT 1.049928e-02
#> TG>CA 5.609618e-02
#> TG>CC 3.129787e-03
#> TG>CT 5.579620e-03
#> TG>GA 7.309502e-03
#> TG>GC 6.119583e-03
#> TG>GT 5.239643e-03
#> TT>AA 4.769675e-03
#> TT>AC 1.039929e-03
#> TT>AG 1.269913e-03
#> TT>CA 2.019862e-03
#> TT>CC 3.999727e-03
#> TT>CG 9.959321e-18
#> TT>GA 7.069518e-04
#> TT>GC 5.779606e-06
#> TT>GG 7.339500e-18
#> 
#> $aetiology
#>                                              aetiology
#> DBS1                                       UV exposure
#> DBS2            exposure to tobacco (smoking) mutagens
#> DBS3   Polymerase epsilon exonuclease domain mutations
#> DBS4                                           Unknown
#> DBS5  Prior chemotherapy treatment with platinum drugs
#> DBS6                                           Unknown
#> DBS7                     Defective DNA mismatch repair
#> DBS8                                           Unknown
#> DBS9                                           Unknown
#> DBS10                    Defective DNA mismatch repair
#> DBS11           Possibly related to APOBEC mutagenesis
#> 
#> $date
#> [1] "2020/05/03"
#>