Get Signature Exposure from 'Signature' Object

Get Signature Exposure from 'Signature' Object

get_sig_exposure(
  Signature,
  type = c("absolute", "relative"),
  rel_threshold = 0.01
)

Arguments

Signature
a Signature object obtained either from sig_extract or sig_auto_extract, or just a raw exposure matrix with column representing samples (patients) and row representing signatures.

type
'absolute' for signature exposure and 'relative' for signature relative exposure.

rel_threshold
used when type is 'relative', relative exposure less than this value will be set to 0 and the remaining signature exposure will be scaled to make sum as 1 accordingly.

Value

a data.table

Examples

# Load mutational signature
load(system.file("extdata", "toy_mutational_signature.RData",
  package = "sigminer", mustWork = TRUE
))
# Get signature exposure
get_sig_exposure(sig2)

#>            sample      Sig1      Sig2      Sig3
#>   1: TCGA-AB-2802 0.0000000  7.602352 1.2566069
#>   2: TCGA-AB-2803 0.0000000 12.007936 0.7412961
#>   3: TCGA-AB-2804 0.0000000  4.838900 0.0000000
#>   4: TCGA-AB-2805 0.2000784 12.007936 0.0000000
#>   5: TCGA-AB-2806 0.2923331 10.273787 0.0000000
#>  ---                                           
#> 184: TCGA-AB-3007 0.0000000  6.691893 0.0000000
#> 185: TCGA-AB-3008 0.2000784  2.940509 0.0000000
#> 186: TCGA-AB-3009 0.0000000 25.388449 2.4198934
#> 187: TCGA-AB-3011 0.0000000  4.838900 0.0000000
#> 188: TCGA-AB-3012 0.0000000  5.770868 0.0000000

get_sig_exposure(sig2, type = "relative")

#> Filtering the samples with no signature exposure:

#> TCGA-AB-2834 TCGA-AB-2883

#>            sample       Sig1      Sig2       Sig3
#>   1: TCGA-AB-2802 0.00000000 0.8581541 0.14184589
#>   2: TCGA-AB-2803 0.00000000 0.9418556 0.05814437
#>   3: TCGA-AB-2804 0.00000000 1.0000000 0.00000000
#>   4: TCGA-AB-2805 0.01638910 0.9836109 0.00000000
#>   5: TCGA-AB-2806 0.02766703 0.9723330 0.00000000
#>  ---                                             
#> 182: TCGA-AB-3007 0.00000000 1.0000000 0.00000000
#> 183: TCGA-AB-3008 0.06370731 0.9362927 0.00000000
#> 184: TCGA-AB-3009 0.00000000 0.9129796 0.08702041
#> 185: TCGA-AB-3011 0.00000000 1.0000000 0.00000000
#> 186: TCGA-AB-3012 0.00000000 1.0000000 0.00000000