Calculate Similarity between Identified Signatures and Reference Signatures

The reference signatures can be either a Signature object specified by Ref argument or known COSMIC signatures specified by sig_db argument. Two COSMIC databases are used for comparisons - "legacy" which includes 30 signaures, and "SBS" - which includes updated/refined 65 signatures. This function is modified from compareSignatures() in maftools package.

get_sig_similarity(
  Signature,
  Ref = NULL,
  sig_db = "legacy",
  method = "cosine",
  normalize = c("row", "feature"),
  feature_setting = sigminer::CN.features,
  verbose = TRUE
)

Arguments

Signature
a Signature object or a component-by-signature matrix (sum of each column is 1). More please see examples.

Ref
default is NULL, can be a same object as Signature.

sig_db
can be 'legacy' or 'SBS'. Default 'legacy'.

method
default is 'cosine' for cosine similarity.

normalize
one of "row" and "feature". "row" is typically used for mutational signatures. "feature" is designed by me to use when input are copy number signatures.

feature_setting
a data.frame used for classification. Only used when method is "Wang" ("W"). Default is CN.features. Users can also set custom input with "feature", "min" and "max" columns avaiable. Valid features can be printed by unique(CN.features$feature).

verbose
if TRUE, print extra info.

Value

a list containing smilarities, aetiologies if available, and best match.

Examples

# Load mutational signature
load(system.file("extdata", "toy_mutational_signature.RData",
  package = "sigminer", mustWork = TRUE
))

get_sig_similarity(sig2, Ref = sig2)

#> -Comparing against Custom signatures

#> ------------------------------------

#> --Found Sig1 most similar to Sig1 [similarity: 1]

#> --Found Sig2 most similar to Sig2 [similarity: 1]

#> --Found Sig3 most similar to Sig3 [similarity: 1]

#> ------------------------------------

#> Return result invisiblely.

# \donttest{
get_sig_similarity(sig2)

#> -Comparing against COSMIC signatures

#> ------------------------------------

#> --Found Sig1 most similar to COSMIC_28

#>    Aetiology: Unknown [similarity: 0.393]

#> --Found Sig2 most similar to COSMIC_1

#>    Aetiology: spontaneous deamination of 5-methylcytosine [similarity: 0.913]

#> --Found Sig3 most similar to COSMIC_10

#>    Aetiology: defects in polymerase POLE [similarity: 0.531]

#> ------------------------------------

#> Return result invisiblely.

get_sig_similarity(sig2, sig_db = "SBS")

#> -Comparing against COSMIC signatures

#> ------------------------------------

#> --Found Sig1 most similar to SBS28

#>    Aetiology: Unknown [similarity: 0.383]

#> --Found Sig2 most similar to SBS1

#>    Aetiology: spontaneous or enzymatic deamination of 5-methylcytosine [similarity: 0.868]

#> --Found Sig3 most similar to SBS10b

#>    Aetiology: Polymerase epsilon exonuclease domain mutations [similarity: 0.865]

#> ------------------------------------

#> Return result invisiblely.

# }