Get Tidy Parameter from Flexmix Model

When users derive copy number features, it is useful to know the parameters of the fit components, including mean, sd and coefficient of variation. This function is used by sig_tally function and exported to users for extra usage.

get_tidy_parameter(x)

Arguments

x
a flexmix object or a list of flexmix objects.

Value

a tibble.

Examples

load(system.file("extdata", "toy_copynumber_tally_M.RData",
  package = "sigminer", mustWork = TRUE
))
# Get all parameters
d1 <- get_tidy_parameter(cn_tally_M$components)
d1

#> # A tibble: 20 x 7
#>    feature     components   dist      mean      sd      cv n_obs
#>    <chr>       <chr>        <chr>    <dbl>   <dbl>   <dbl> <dbl>
#>  1 bp10MB      bp10MB1      pois  2.60e-16 1.61e-8 6.21e+7  2888
#>  2 bp10MB      bp10MB2      pois  1.03e+ 0 1.02e+0 9.83e-1   112
#>  3 bpchrarm    bpchrarm1    pois  2.72e- 2 1.65e-1 6.06e+0   370
#>  4 bpchrarm    bpchrarm2    pois  2.51e+ 0 1.58e+0 6.32e-1    70
#>  5 changepoint changepoint1 norm  6.21e- 1 6.94e-1 1.12e+0   321
#>  6 changepoint changepoint2 norm  5.04e+ 0 3.38e+0 6.72e-1    18
#>  7 copynumber  copynumber1  norm  1.92e+ 0 5.88e-1 3.07e-1   354
#>  8 copynumber  copynumber2  norm  3.77e+ 0 2.94e+0 7.82e-1    46
#>  9 osCN        osCN1        pois  8.87e-11 9.42e-6 1.06e+5   226
#> 10 osCN        osCN2        pois  1.40e+ 0 1.18e+0 8.46e-1     7
#> 11 segsize     segsize1     norm  4.09e+ 5 3.16e+5 7.74e-1    56
#> 12 segsize     segsize2     norm  7.93e+ 6 5.46e+6 6.89e-1    74
#> 13 segsize     segsize3     norm  3.22e+ 7 2.01e+5 6.25e-3    18
#> 14 segsize     segsize4     norm  4.48e+ 7 2.03e+7 4.53e-1    79
#> 15 segsize     segsize5     norm  8.43e+ 7 9.51e+6 1.13e-1    69
#> 16 segsize     segsize6     norm  1.14e+ 8 3.08e+7 2.69e-1    14
#> 17 segsize     segsize7     norm  1.34e+ 8 1.16e+6 8.67e-3    26
#> 18 segsize     segsize8     norm  1.52e+ 8 1.47e+7 9.66e-2    23
#> 19 segsize     segsize9     norm  1.85e+ 8 9.85e+6 5.34e-2    28
#> 20 segsize     segsize10    norm  2.43e+ 8 1.36e+6 5.59e-3    13

# Get parameters for segsize feature
d2 <- get_tidy_parameter(cn_tally_M$components$segsize)
d2

#> # A tibble: 10 x 5
#>    dist        mean        sd      cv n_obs
#>    <chr>      <dbl>     <dbl>   <dbl> <dbl>
#>  1 norm     408603.   316408. 0.774      56
#>  2 norm    7925979.  5459508. 0.689      74
#>  3 norm   32155630.   201036. 0.00625    18
#>  4 norm   44776707. 20267870. 0.453      79
#>  5 norm   84272127.  9513170. 0.113      69
#>  6 norm  114214740. 30768304. 0.269      14
#>  7 norm  133528486.  1158161. 0.00867    26
#>  8 norm  151978846. 14681542. 0.0966     23
#>  9 norm  184546599.  9849617. 0.0534     28
#> 10 norm  243182065.  1360105. 0.00559    13