Show Copy Number Feature Distributions

Show Copy Number Feature Distributions

show_cn_features(
  features,
  method = "Wang",
  rm_outlier = FALSE,
  ylab = NULL,
  log_y = FALSE,
  return_plotlist = FALSE,
  base_size = 12,
  nrow = 2,
  align = "hv",
  ...
)

Arguments

features

a feature list generate from sig_tally function.

method

method for feature classification, can be one of "Wang" ("W"), "S" (for method described in Steele et al. 2019), "X" (for method described in Tao et al. 2023).

rm_outlier

default is FALSE, if TRUE, remove outliers. Only works when method is "Wang" ("W").

ylab

lab of y axis.

log_y

logical, if TRUE, show log10 based y axis, only works for input from "Wang" ("W") method.

return_plotlist

if TRUE, return a list of ggplot objects but a combined plot.

base_size

overall font size.

nrow

(optional) Number of rows in the plot grid.

align

(optional) Specifies whether graphs in the grid should be horizontally ("h") or vertically ("v") aligned. Options are "none" (default), "hv" (align in both directions), "h", and "v".

...

other options pass to plot_grid function of cowplot package.

Value

a ggplot object