Show Copy Number Variation Frequency Profile with Circos

Show Copy Number Variation Frequency Profile with Circos

show_cn_freq_circos(
  data,
  groups = NULL,
  cutoff = 2L,
  resolution_factor = 1L,
  title = c("AMP", "DEL"),
  chrs = paste0("chr", 1:22),
  genome_build = c("hg19", "hg38", "mm10", "mm9"),
  cols = NULL,
  plot_ideogram = TRUE,
  track_height = 0.5,
  ideogram_height = 1,
  ...
)

Arguments

data

a CopyNumber object or a data.frame containing at least 'chromosome', 'start', 'end', 'segVal', 'sample' these columns.

groups

a named list or a column name for specifying groups.

cutoff

copy number value cutoff for splitting data into AMP and DEL. The values equal to cutoff are discarded. Default is 2, you can also set a length-2 vector, e.g. c(2, 2).

resolution_factor

an integer to control the resolution. When it is 1 (default), compute frequency in each cytoband. When it is 2, use compute frequency in each half cytoband.

title

length-2 titles for AMP and DEL.

chrs

chromosomes start with 'chr'.

genome_build

genome build version, used when data is a data.frame, should be 'hg19' or 'hg38'.

cols

length-2 colors for AMP and DEL.

plot_ideogram

default is TRUE, show ideogram.

track_height

track height in mm unit.

ideogram_height

ideogram height in mm unit.

...

other parameters passing to circlize::circos.genomicLines.

Value

Nothing.

Examples

# \donttest{
load(system.file("extdata", "toy_copynumber.RData",
  package = "sigminer", mustWork = TRUE
))

show_cn_freq_circos(cn)
ss <- unique(cn@data$sample)
show_cn_freq_circos(cn, groups = list(a = ss[1:5], b = ss[6:10]), cols = c("red", "green"))
# }