Draw Corrplot for Signature Exposures and Other Features

Draw Corrplot for Signature Exposures and Other Features

show_sig_feature_corrplot(
  tidy_cor,
  feature_list,
  sort_features = FALSE,
  return_plotlist = FALSE,
  p_val = 0.05,
  xlab = "Signatures",
  ylab = "Features",
  gradient_colors = scale_color_gradient2(low = "blue", mid = "white", high = "red",
    midpoint = 0),
  align = c("row", "column"),
  plot_ratio = "auto",
  breaks_count = c(0L, 200L, 400L, 600L, 800L, 1020L)
)

Arguments

tidy_cor
data returned by get_tidy_association.

feature_list
a character vector contains features want to be plotted. If missing, all features will be used.

sort_features
default is FALSE, use feature order obtained from the previous step. If TRUE, sort features as feature_list.

return_plotlist
if TRUE, return as a list of ggplot objects.

p_val
p value threshold.

xlab
label for x axis.

ylab
label for y axis.

gradient_colors
a Scale object representing gradient colors used to plot.

align
one of "row" and "column".

plot_ratio
a length-2 numeric vector to set the height/width ratio.

breaks_count
breaks for sample count.

Value

a ggplot2 object

See also