Extract Signatures through the Automatic Relevance Determination Technique

A bayesian variant of NMF algorithm to enable optimal inferences for the number of signatures through the automatic relevance determination technique. This functions delevers highly interpretable and sparse representations for both signature profiles and attributions at a balance between data fitting and model complexity (this method may introduce more signatures than expected, especially for copy number signatures (thus I don't recommend you to use this feature to extract copy number signatures)). See detail part and references for more.

sig_auto_extract(
  nmf_matrix = NULL,
  result_prefix = "BayesNMF",
  destdir = tempdir(),
  method = c("L1W.L2H", "L1KL", "L2KL"),
  strategy = c("stable", "optimal", "ms"),
  ref_sigs = NULL,
  K0 = 25,
  nrun = 10,
  niter = 2e+05,
  tol = 1e-07,
  cores = 1,
  optimize = FALSE,
  skip = FALSE,
  recover = FALSE
)

Arguments

nmf_matrix

a matrix used for NMF decomposition with rows indicate samples and columns indicate components.

result_prefix

prefix for result data files.

destdir

path to save data runs, default is tempdir().

method

default is "L1W.L2H", which uses an exponential prior for W and a half-normal prior for H (This method is used by PCAWG project, see reference #3). You can also use "L1KL" to set expoential priors for both W and H, and "L2KL" to set half-normal priors for both W and H. The latter two methods are originally implemented by SignatureAnalyzer software.

strategy

the selection strategy for returned data. Set 'stable' for getting optimal result from the most frequent K. Set 'optimal' for getting optimal result from all Ks. Set 'ms' for getting result with maximum mean cosine similarity with provided reference signatures. See ref_sigs option for details. If you want select other solution, please check get_bayesian_result.

ref_sigs

A Signature object or matrix or string for specifying reference signatures, only used when strategy = 'ms'. See Signature and sig_db options in get_sig_similarity for details.

K0

number of initial signatures.

nrun

number of independent simulations.

niter

the maximum number of iterations.

tol

tolerance for convergence.

cores

number of cpu cores to run NMF.

optimize

if TRUE, then refit the denovo signatures with QP method, see sig_fit.

skip

if TRUE, it will skip running a previous stored result. This can be used to extend run times, e.g. you try running 10 times firstly and then you want to extend it to 20 times.

recover

if TRUE, try to recover result from previous runs based on input result_prefix, destdir and nrun. This is pretty useful for reproducing result. Please use skip if you want to recover an unfinished job.

Value

a list with Signature class.

Details

There are three methods available in this function: "L1W.L2H", "L1KL" and "L2KL". They use different priors for the bayesian variant of NMF algorithm (see method parameter) written by reference #1 and implemented in SignatureAnalyzer software (reference #2).

I copied source code for the three methods from Broad Institute and supplementary files of reference #3, and wrote this higher function. It is more friendly for users to extract, visualize and analyze signatures by combining with other powerful functions in sigminer package. Besides, I implemented parallel computation to speed up the calculation process and a similar input and output structure like sig_extract().

References

Tan, Vincent YF, and Cédric Févotte. "Automatic relevance determination in nonnegative matrix factorization with the/spl beta/-divergence." IEEE Transactions on Pattern Analysis and Machine Intelligence 35.7 (2012): 1592-1605.

Kim, Jaegil, et al. "Somatic ERCC2 mutations are associated with a distinct genomic signature in urothelial tumors." Nature genetics 48.6 (2016): 600.

Alexandrov, Ludmil, et al. "The repertoire of mutational signatures in human cancer." BioRxiv (2018): 322859.

See also

Author

Shixiang Wang

Examples

# \donttest{
load(system.file("extdata", "toy_copynumber_tally_W.RData",
  package = "sigminer", mustWork = TRUE
))
res <- sig_auto_extract(cn_tally_W$nmf_matrix, result_prefix = "Test_copynumber", nrun = 1)
#> Select Run 1, which K = 2 as best solution.
# At default, all run files are stored in tempdir()
dir(tempdir(), pattern = "Test_copynumber")
#> [1] "Test_copynumber.1.rds"

laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read_maf(maf = laml.maf)
#> -Reading
#> -Validating
#> -Silent variants: 475 
#> -Summarizing
#> -Processing clinical data
#> --Missing clinical data
#> -Finished in 0.151s elapsed (0.197s cpu) 
mt_tally <- sig_tally(
  laml,
  ref_genome = "BSgenome.Hsapiens.UCSC.hg19",
  use_syn = TRUE
)
#> ℹ [2025-11-30 14:30:09.617392]: Started.
#> ℹ [2025-11-30 14:30:09.621784]: We would assume you marked all variants' position in + strand.
#> ✔ [2025-11-30 14:30:09.623841]: Reference genome loaded.
#> ✔ [2025-11-30 14:30:09.62827]: Variants from MAF object queried.
#> ✔ [2025-11-30 14:30:09.630812]: Chromosome names checked.
#> ✔ [2025-11-30 14:30:09.634339]: Sex chromosomes properly handled.
#> ✔ [2025-11-30 14:30:09.636167]: Only variants located in standard chromosomes (1:22, X, Y, M/MT) are kept.
#> ✔ [2025-11-30 14:30:09.641137]: Variant start and end position checked.
#> ✔ [2025-11-30 14:30:09.645094]: Variant data for matrix generation preprocessed.
#> ℹ [2025-11-30 14:30:09.646633]: SBS matrix generation - start.
#> ℹ [2025-11-30 14:30:09.648605]: Extracting 5' and 3' adjacent bases.
#> ℹ [2025-11-30 14:30:09.935082]: Extracting +/- 20bp around mutated bases for background C>T estimation.
#> ℹ [2025-11-30 14:30:10.263496]: Estimating APOBEC enrichment scores.
#> ℹ [2025-11-30 14:30:10.265592]: Performing one-way Fisher's test for APOBEC enrichment.
#> ✔ [2025-11-30 14:30:10.338105]: APOBEC related mutations are enriched in 3.297% of samples (APOBEC enrichment score > 2; 6 of 182 samples)
#> ℹ [2025-11-30 14:30:10.341436]: Creating SBS sample-by-component matrices.
#> ✔ [2025-11-30 14:30:10.346091]: SBS-6 matrix created.
#> ✔ [2025-11-30 14:30:10.351291]: SBS-96 matrix created.
#> ✔ [2025-11-30 14:30:10.364386]: SBS-1536 matrix created.
#> ℹ [2025-11-30 14:30:10.366189]: Return SBS-96 as major matrix.
#> ✔ [2025-11-30 14:30:10.368997]: Done.
#> ℹ [2025-11-30 14:30:10.370668]: 0.753 secs elapsed.

x <- sig_auto_extract(mt_tally$nmf_matrix,
  strategy = "ms", nrun = 3, ref_sigs = "legacy"
)
#> The follow samples dropped due to null catalogue:
#> 	TCGA-AB-2823, TCGA-AB-2840, TCGA-AB-2848, TCGA-AB-2909, TCGA-AB-2942
#> Solutions ordered by mean cosine similarity to references:
#> Run#3:K#3 Run#2:K#2 Run#1:K#3 
#> 0.6716667 0.6395000 0.6053333 
#> Details:
#> $`Run#1:K#3`
#>  COSMIC_1 COSMIC_10 COSMIC_20 
#>     0.904     0.448     0.464 
#> 
#> $`Run#2:K#2`
#>  COSMIC_1 COSMIC_18 
#>     0.966     0.313 
#> 
#> $`Run#3:K#3`
#>  COSMIC_1 COSMIC_15 COSMIC_10 
#>     0.908     0.556     0.551 
#> 
#> Select Run 3, which K = 3 as best solution.
x
#> $Signature
#>                  Sig1          Sig2          Sig3
#> A[C>A]A  1.754082e-79  1.141872e+01 9.940065e-315
#> A[C>A]C  2.683076e+01 1.011444e-120 1.184611e-241
#> A[C>A]G  3.700795e+00 7.403353e-116  0.000000e+00
#> A[C>A]T  7.401589e+00 1.866393e-184 3.070082e-183
#> A[C>G]A  2.682892e-39  9.403649e+00  0.000000e+00
#> A[C>G]C  5.551192e+00  2.023179e-76 5.432420e-312
#> A[C>G]G  7.401589e+00 1.766526e-293  0.000000e+00
#> A[C>G]T  1.234290e-07  5.373514e+00  0.000000e+00
#> A[C>T]A  4.058175e+01  4.794021e+00 1.355512e-260
#> A[C>T]C  4.191258e+01  3.827845e+00 5.069421e-120
#> A[C>T]G  2.137209e+02  8.592716e-45 3.377965e-252
#> A[C>T]T  2.405516e+01 1.531645e-126  0.000000e+00
#> A[T>A]A  0.000000e+00  1.284361e+00  0.000000e+00
#> A[T>A]C  1.110238e+01  1.737439e-32  1.061064e-51
#> A[T>A]G  5.551192e+00  0.000000e+00 4.997092e-293
#> A[T>A]T  3.700795e+00 5.615185e-265  0.000000e+00
#> A[T>C]A  2.590556e+01 4.173111e-128  0.000000e+00
#> A[T>C]C  1.110238e+01  0.000000e+00 9.653289e-171
#> A[T>C]G  2.498036e+01 2.143056e-226 6.817122e-252
#> A[T>C]T  2.349137e+01  3.767757e+00 5.483424e-309
#> A[T>G]A 9.723643e-119  4.030135e+00  0.000000e+00
#> A[T>G]C  2.091221e-90 1.235164e-322  0.000000e+00
#> A[T>G]G  2.775596e+00  0.000000e+00  0.000000e+00
#> A[T>G]T  2.775596e+00 3.028034e-204  0.000000e+00
#> C[C>A]A  1.665358e+01  4.994172e-81 3.399690e-286
#> C[C>A]C  1.387798e+01 3.824967e-248 2.324849e-271
#> C[C>A]G  1.295278e+01 4.909341e-261  0.000000e+00
#> C[C>A]T  1.665358e+01  0.000000e+00  0.000000e+00
#> C[C>G]A  1.202758e+01 4.822308e-261 6.067007e-133
#> C[C>G]C  3.700795e+00 1.951872e-256  0.000000e+00
#> C[C>G]G  1.387798e+01  2.920090e-15  1.117240e-84
#> C[C>G]T  6.476391e+00  0.000000e+00  0.000000e+00
#> C[C>T]A  3.793314e+01 5.780051e-169  0.000000e+00
#> C[C>T]C  2.775596e+01  4.749816e-10  0.000000e+00
#> C[C>T]G  1.896657e+02  1.719069e-71 6.623729e-188
#> C[C>T]T  5.551192e+01 3.402650e-178 7.905114e-251
#> C[T>A]A  9.933357e-51  0.000000e+00  3.819727e-01
#> C[T>A]C 1.003547e-128  6.716892e+00 1.859619e-201
#> C[T>A]G  1.110238e+01 8.835885e-140 6.062977e-213
#> C[T>A]T  4.625993e+00  0.000000e+00  0.000000e+00
#> C[T>C]A  4.056794e-87  4.701825e+00  0.000000e+00
#> C[T>C]C  1.665358e+01 1.016288e-117  0.000000e+00
#> C[T>C]G  1.349182e+01  2.967110e+00  0.000000e+00
#> C[T>C]T  8.897874e+00  3.615530e+00 1.813842e-121
#> C[T>G]A 7.407716e-109  6.716892e-01  0.000000e+00
#> C[T>G]C  7.401589e+00 2.136071e-256 1.457348e-194
#> C[T>G]G  9.251987e+00 1.490293e-187  0.000000e+00
#> C[T>G]T  4.625993e+00 6.850656e-274 4.305078e-111
#> G[C>A]A  1.893518e-27  1.141872e+01 4.080755e-203
#> G[C>A]C  1.878162e-25  6.045203e+00 6.269584e-195
#> G[C>A]G  4.145868e-53  3.358446e+00  0.000000e+00
#> G[C>A]T  8.326788e+00 4.667561e-142  0.000000e+00
#> G[C>G]A  1.017719e+01  7.772128e-20 4.217862e-250
#> G[C>G]C  6.476391e+00 6.963854e-243 9.903832e-260
#> G[C>G]G  6.476391e+00 1.004081e-265  0.000000e+00
#> G[C>G]T  0.000000e+00  4.030135e+00 4.816197e-313
#> G[C>T]A  1.197639e-71  1.813561e+01 1.277332e-211
#> G[C>T]C  4.533473e+01 5.284995e-177 8.938562e-201
#> G[C>T]G  1.392790e+02  2.583348e+01 1.791415e-258
#> G[C>T]T  4.718513e+01 1.718363e-188 1.376814e-273
#> G[T>A]A 2.589963e-129  1.257508e-84  9.996284e-01
#> G[T>A]C  2.883698e-78  6.045203e+00 5.287714e-165
#> G[T>A]G  3.700795e+00  8.355948e-81 4.817495e-276
#> G[T>A]T  5.551192e+00 1.200606e-155 2.086095e-308
#> G[T>C]A  2.420945e+00  2.272543e+00  0.000000e+00
#> G[T>C]C  5.861007e-91  6.045203e+00  6.872234e-97
#> G[T>C]G  1.017719e+01 8.510163e-306  0.000000e+00
#> G[T>C]T  1.110238e+01  1.115982e-89 2.015221e-211
#> G[T>G]A  1.649547e-98 2.494312e-159  1.999257e+00
#> G[T>G]C  2.775596e+00 7.086483e-158 1.692522e-292
#> G[T>G]G  5.551192e+00  0.000000e+00  0.000000e+00
#> G[T>G]T 2.880486e-111  1.343378e+00  0.000000e+00
#> T[C>A]A  1.141113e-05  5.601604e+00  5.351991e-98
#> T[C>A]C  1.518822e-11  1.544885e+01 4.441403e-268
#> T[C>A]G  7.401589e+00  5.935841e-25  0.000000e+00
#> T[C>A]T  2.590556e+01  1.271110e-25 1.650145e-196
#> T[C>G]A  4.625993e+00 2.250403e-200 7.199593e-126
#> T[C>G]C  7.401589e+00 4.640572e-215 5.702354e-151
#> T[C>G]G  1.017719e+01 2.156932e-129 9.093875e-215
#> T[C>G]T  5.804083e+00  1.831470e+00 1.220342e-321
#> T[C>T]A  3.978354e+01 5.756015e-175 5.258892e-210
#> T[C>T]C  2.842809e+01  8.243999e+00 6.437675e-321
#> T[C>T]G  2.333522e-22 1.044562e-103  5.547938e+01
#> T[C>T]T  2.498036e+01 7.995473e-222 1.141326e-225
#> T[T>A]A  0.000000e+00  0.000000e+00  4.998142e-01
#> T[T>A]C  3.700795e+00 1.107650e-261  4.463792e-64
#> T[T>A]G  0.000000e+00 1.913786e-243  1.999257e+00
#> T[T>A]T  4.625993e+00  2.693225e-22 9.425559e-105
#> T[T>C]A  0.000000e+00  1.343378e+00 1.729230e-322
#> T[T>C]C  8.326788e+00 7.152293e-102 7.695634e-295
#> T[T>C]G  1.263118e-73  6.716892e+00  0.000000e+00
#> T[T>C]T  1.076724e-26 6.901522e-165  9.996284e+00
#> T[T>G]A  2.775596e+00 3.198556e-153  0.000000e+00
#> T[T>G]C  4.625993e+00 1.002991e-163  0.000000e+00
#> T[T>G]G 3.921155e-314  1.343378e+00  0.000000e+00
#> T[T>G]T 5.098291e-121 1.092453e-175  2.499071e+00
#> 
#> $Signature.norm
#>                  Sig1          Sig2          Sig3
#> A[C>A]A  1.210732e-82  6.085745e-02 1.345895e-316
#> A[C>A]C  1.851958e-02 5.390616e-123 1.603976e-243
#> A[C>A]G  2.554425e-03 3.945708e-118  0.000000e+00
#> A[C>A]T  5.108849e-03 9.947168e-187 4.156924e-185
#> A[C>G]A  1.851831e-42  5.011790e-02  0.000000e+00
#> A[C>G]C  3.831637e-03  1.078278e-78 7.355555e-314
#> A[C>G]G  5.108849e-03 9.414917e-296  0.000000e+00
#> A[C>G]T  8.519525e-11  2.863880e-02  0.000000e+00
#> A[C>T]A  2.801101e-02  2.555032e-02 1.835377e-262
#> A[C>T]C  2.892960e-02  2.040097e-02 6.864050e-122
#> A[C>T]G  1.475180e-01  4.579593e-47 4.573801e-254
#> A[C>T]T  1.660376e-02 8.163091e-129  0.000000e+00
#> A[T>A]A  0.000000e+00  6.845161e-03  0.000000e+00
#> A[T>A]C  7.663274e-03  9.259895e-35  1.436692e-53
#> A[T>A]G  3.831637e-03  0.000000e+00 6.766115e-295
#> A[T>A]T  2.554425e-03 2.992682e-267  0.000000e+00
#> A[T>C]A  1.788097e-02 2.224111e-130  0.000000e+00
#> A[T>C]C  7.663274e-03  0.000000e+00 1.307066e-172
#> A[T>C]G  1.724237e-02 1.142168e-228 9.230456e-254
#> A[T>C]T  1.621461e-02  2.008073e-02 7.424614e-311
#> A[T>G]A 6.711616e-122  2.147910e-02  0.000000e+00
#> A[T>G]C  1.443437e-93  0.000000e+00  0.000000e+00
#> A[T>G]G  1.915818e-03  0.000000e+00  0.000000e+00
#> A[T>G]T  1.915818e-03 1.613828e-206  0.000000e+00
#> C[C>A]A  1.149491e-02  2.661705e-83 4.603216e-288
#> C[C>A]C  9.579092e-03 2.038563e-250 3.147870e-273
#> C[C>A]G  8.940486e-03 2.616493e-263  0.000000e+00
#> C[C>A]T  1.149491e-02  0.000000e+00  0.000000e+00
#> C[C>G]A  8.301880e-03 2.570108e-263 8.214793e-135
#> C[C>G]C  2.554425e-03 1.040274e-258  0.000000e+00
#> C[C>G]G  9.579092e-03  1.556298e-17  1.512755e-86
#> C[C>G]T  4.470243e-03  0.000000e+00  0.000000e+00
#> C[C>T]A  2.618285e-02 3.080549e-171  0.000000e+00
#> C[C>T]C  1.915818e-02  2.531472e-12  0.000000e+00
#> C[C>T]G  1.309143e-01  9.161987e-74 8.968600e-190
#> C[C>T]T  3.831637e-02 1.813484e-180 1.070361e-252
#> C[T>A]A  6.856368e-54  0.000000e+00  5.171950e-03
#> C[T>A]C 6.926849e-132  3.579850e-02 2.517944e-203
#> C[T>A]G  7.663274e-03 4.709193e-142 8.209336e-215
#> C[T>A]T  3.193031e-03  0.000000e+00  0.000000e+00
#> C[T>C]A  2.800149e-90  2.505895e-02  0.000000e+00
#> C[T>C]C  1.149491e-02 5.416431e-120  0.000000e+00
#> C[T>C]G  9.312547e-03  1.581358e-02  0.000000e+00
#> C[T>C]T  6.141640e-03  1.926941e-02 2.455961e-123
#> C[T>G]A 5.113078e-112  3.579850e-03  0.000000e+00
#> C[T>G]C  5.108849e-03 1.138445e-258 1.973265e-196
#> C[T>G]G  6.386061e-03 7.942699e-190  0.000000e+00
#> C[T>G]T  3.193031e-03 3.651141e-276 5.829122e-113
#> G[C>A]A  1.306976e-30  6.085745e-02 5.525385e-205
#> G[C>A]C  1.296377e-28  3.221865e-02 8.489083e-197
#> G[C>A]G  2.861630e-56  1.789925e-02  0.000000e+00
#> G[C>A]T  5.747455e-03 2.487634e-144  0.000000e+00
#> G[C>G]A  7.024667e-03  4.142251e-22 5.711029e-252
#> G[C>G]C  4.470243e-03 3.711471e-245 1.340989e-261
#> G[C>G]G  4.470243e-03 5.351374e-268  0.000000e+00
#> G[C>G]T  0.000000e+00  2.147910e-02 6.521183e-315
#> G[C>T]A  8.266544e-75  9.665596e-02 1.729521e-213
#> G[C>T]C  3.129170e-02 2.816703e-179 1.210291e-202
#> G[C>T]G  9.613547e-02  1.376827e-01 2.425595e-260
#> G[C>T]T  3.256891e-02 9.158228e-191 1.864221e-275
#> G[T>A]A 1.787688e-132  6.702042e-87  1.353508e-02
#> G[T>A]C  1.990434e-81  3.221865e-02 7.159621e-167
#> G[T>A]G  2.554425e-03  4.453405e-83 6.522940e-278
#> G[T>A]T  3.831637e-03 6.398774e-158 2.824595e-310
#> G[T>C]A  1.671025e-03  1.211179e-02  0.000000e+00
#> G[T>C]C  4.045482e-94  3.221865e-02  9.305079e-99
#> G[T>C]G  7.024667e-03 4.535596e-308  0.000000e+00
#> G[T>C]T  7.663274e-03  5.947765e-92 2.728631e-213
#> G[T>G]A 1.138578e-101 1.329374e-161  2.707015e-02
#> G[T>G]C  1.915818e-03 3.776828e-160 2.291693e-294
#> G[T>G]G  3.831637e-03  0.000000e+00  0.000000e+00
#> G[T>G]T 1.988218e-114  7.159700e-03  0.000000e+00
#> T[C>A]A  7.876379e-09  2.985443e-02 7.246653e-100
#> T[C>A]C  1.048347e-14  8.233655e-02 6.013707e-270
#> T[C>A]G  5.108849e-03  3.163579e-27  0.000000e+00
#> T[C>A]T  1.788097e-02  6.774538e-28 2.234315e-198
#> T[C>G]A  3.193031e-03 1.199380e-202 9.748326e-128
#> T[C>G]C  5.108849e-03 2.473250e-217 7.721048e-153
#> T[C>G]G  7.024667e-03 1.149564e-131 1.231320e-216
#> T[C>G]T  4.006192e-03  9.761044e-03 1.482197e-323
#> T[C>T]A  2.746006e-02 3.067739e-177 7.120595e-212
#> T[C>T]C  1.962211e-02  4.393740e-02 8.893182e-323
#> T[C>T]G  1.610683e-25 5.567119e-106  7.511967e-01
#> T[C>T]T  1.724237e-02 4.261286e-224 1.545368e-227
#> T[T>A]A  0.000000e+00  0.000000e+00  6.767538e-03
#> T[T>A]C  2.554425e-03 5.903355e-264  6.044021e-66
#> T[T>A]G  0.000000e+00 1.019976e-245  2.707015e-02
#> T[T>A]T  3.193031e-03  1.435387e-24 1.276231e-106
#> T[T>C]A  0.000000e+00  7.159700e-03  0.000000e+00
#> T[T>C]C  5.747455e-03 3.811902e-104 1.041997e-296
#> T[T>C]G  8.718507e-77  3.579850e-02  0.000000e+00
#> T[T>C]T  7.431943e-30 3.678251e-167  1.353508e-01
#> T[T>G]A  1.915818e-03 1.704710e-155  0.000000e+00
#> T[T>G]C  3.193031e-03 5.345561e-166  0.000000e+00
#> T[T>G]G 2.706526e-317  7.159700e-03  0.000000e+00
#> T[T>G]T 3.519028e-124 5.822364e-178  3.383769e-02
#> 
#> $Exposure
#>       TCGA-AB-2802  TCGA-AB-2803  TCGA-AB-2804  TCGA-AB-2805 TCGA-AB-2806
#> Sig1  9.354442e+00  1.194702e+01  6.671630e+00  1.279268e+01    10.449942
#> Sig2 5.617517e-124 2.910809e-149 2.394159e-217 8.184612e-145     2.408221
#> Sig3  0.000000e+00  6.175842e-01  0.000000e+00  0.000000e+00     0.000000
#>      TCGA-AB-2807 TCGA-AB-2808 TCGA-AB-2809 TCGA-AB-2810 TCGA-AB-2812
#> Sig1    20.935465 8.470661e+00     2.711285   11.0923000     5.512876
#> Sig2     1.972799 3.784334e-71     1.009666    0.0000000     1.587820
#> Sig3     0.000000 0.000000e+00     0.000000    0.9991224     0.000000
#>      TCGA-AB-2813 TCGA-AB-2814 TCGA-AB-2816  TCGA-AB-2817 TCGA-AB-2818
#> Sig1   10.3204112    1.5974453 5.755611e+00  8.470661e+00    8.2972131
#> Sig2    2.5073629    3.1690068 1.233842e-93 4.759504e-116    0.9821667
#> Sig3    0.6175842    0.9991224 6.175842e-01  6.175842e-01    0.0000000
#>      TCGA-AB-2819 TCGA-AB-2820 TCGA-AB-2821  TCGA-AB-2822 TCGA-AB-2824
#> Sig1    8.7986823     11.74236     9.354442  1.769108e+01    3.3778414
#> Sig2    2.8263862      1.01378     0.000000 7.628788e-186    0.4859137
#> Sig3    0.9991224      0.00000     1.559996  6.175842e-01    0.0000000
#>      TCGA-AB-2825 TCGA-AB-2826 TCGA-AB-2827 TCGA-AB-2828 TCGA-AB-2829
#> Sig1    0.0000000     3.888454     7.127190   10.2282111     9.354442
#> Sig2    0.8432031     0.000000     2.315584    0.0000000     0.000000
#> Sig3    0.0000000     0.000000     1.301536    0.6175842     0.000000
#>      TCGA-AB-2830 TCGA-AB-2831 TCGA-AB-2832 TCGA-AB-2833  TCGA-AB-2834
#> Sig1   11.0124677     1.971331 3.834034e-27 6.671630e+00  0.000000e+00
#> Sig2    0.9101625     0.000000 4.833646e+00 5.392740e-40 3.194988e-244
#> Sig3    0.6175842     0.000000 0.000000e+00 6.175842e-01  0.000000e+00
#>      TCGA-AB-2835 TCGA-AB-2836  TCGA-AB-2838 TCGA-AB-2839  TCGA-AB-2841
#> Sig1    0.0000000    0.9927296  1.609008e+01   12.3956955  3.888454e+00
#> Sig2    0.0000000    0.0000000 5.219567e-139    2.3020527 1.244919e-105
#> Sig3    0.9991224    0.6175842  9.991224e-01    0.6175842  0.000000e+00
#>      TCGA-AB-2842 TCGA-AB-2843 TCGA-AB-2844 TCGA-AB-2845 TCGA-AB-2846
#> Sig1     1.971331    5.5686973    9.3544425 2.936390e+00    7.3033794
#> Sig2     0.000000    0.9874826    0.0000000 2.243634e-72    2.2181026
#> Sig3     0.000000    1.5020691    0.9991224 0.000000e+00    0.6175842
#>      TCGA-AB-2847 TCGA-AB-2849  TCGA-AB-2850 TCGA-AB-2851  TCGA-AB-2853
#> Sig1     7.752680    14.728527  3.888454e+00    4.6320087  6.671630e+00
#> Sig2     1.382442     4.400544 1.265679e-105    0.9924791 4.532800e-168
#> Sig3     0.000000     0.000000  0.000000e+00    0.0000000  6.175842e-01
#>      TCGA-AB-2854 TCGA-AB-2855 TCGA-AB-2857 TCGA-AB-2858 TCGA-AB-2859
#> Sig1     8.936610      2.93639   11.9470237    8.6798863     8.162009
#> Sig2     1.171535      0.00000    0.0000000    1.9369303     1.705246
#> Sig3     0.000000      0.00000    0.6175842    0.6175842     0.000000
#>      TCGA-AB-2860  TCGA-AB-2861 TCGA-AB-2862 TCGA-AB-2863 TCGA-AB-2864
#> Sig1     2.884302  1.609008e+01     7.792324   11.3491207   10.8244163
#> Sig2     2.582979 1.477205e-261     1.934460    2.4147406    1.0623203
#> Sig3     1.301536  6.175842e-01     0.000000    0.6175842    0.9991224
#>      TCGA-AB-2865 TCGA-AB-2866 TCGA-AB-2867 TCGA-AB-2868 TCGA-AB-2869
#> Sig1   13.2623149    0.9927296     5.755611     10.81177     6.361940
#> Sig2    1.1489645    0.0000000     0.000000      1.07229     2.236439
#> Sig3    0.9991224    0.0000000     1.789445      0.00000     0.000000
#>      TCGA-AB-2870 TCGA-AB-2871 TCGA-AB-2872 TCGA-AB-2873 TCGA-AB-2874
#> Sig1     8.246054   10.2282136    8.1532092    0.0000000    12.190903
#> Sig2     1.021975    0.0000000    1.0929494    0.8432031     1.325328
#> Sig3     0.000000    0.9991224    0.6175842    0.0000000     0.000000
#>       TCGA-AB-2875 TCGA-AB-2876 TCGA-AB-2877 TCGA-AB-2878 TCGA-AB-2879
#> Sig1  8.470661e+00    8.4706633     16.89437     8.896894     0.000000
#> Sig2 5.913232e-110    0.0000000      0.00000     1.201039     1.502287
#> Sig3  0.000000e+00    0.6175842      0.00000     0.000000     0.000000
#>       TCGA-AB-2880 TCGA-AB-2881 TCGA-AB-2882 TCGA-AB-2883 TCGA-AB-2884
#> Sig1 1.235164e-322    6.3526692 1.445791e+01    0.9927296 3.888454e+00
#> Sig2  8.432031e-01    1.0883744 4.560191e-76    0.0000000 2.203639e-39
#> Sig3  0.000000e+00    0.6175842 0.000000e+00    0.0000000 0.000000e+00
#>      TCGA-AB-2885 TCGA-AB-2886 TCGA-AB-2887 TCGA-AB-2888  TCGA-AB-2889
#> Sig1    8.9073972    5.9509303   10.3653880    2.4425961  4.828035e+00
#> Sig2    1.1932619    2.4567550    1.9786151    1.7993660 5.774234e-181
#> Sig3    0.6175842    0.6175842    0.6175842    0.9991224  0.000000e+00
#>      TCGA-AB-2890 TCGA-AB-2891 TCGA-AB-2892  TCGA-AB-2894 TCGA-AB-2895
#> Sig1     6.186261     14.45791    0.0000000 3.676412e-294 1.362955e+01
#> Sig2     1.209040      0.00000    0.8432031  2.557518e+00 8.840184e-40
#> Sig3     0.000000      0.00000    0.0000000  6.175842e-01 0.000000e+00
#>      TCGA-AB-2896 TCGA-AB-2897 TCGA-AB-2898 TCGA-AB-2899 TCGA-AB-2900
#> Sig1    0.0000000     5.755611 1.527800e+01   10.6489069     8.251140
#> Sig2    0.8432031     0.000000 5.323681e-49    1.1980002     3.517694
#> Sig3    0.0000000     0.000000 0.000000e+00    0.9991224     1.301536
#>      TCGA-AB-2901 TCGA-AB-2903  TCGA-AB-2904  TCGA-AB-2905 TCGA-AB-2906
#> Sig1    4.8280363    0.9927296  1.926265e+01  1.769108e+01     9.720077
#> Sig2    0.0000000    0.0000000 2.187761e-214 5.939764e-131     2.357371
#> Sig3    0.9991224    0.0000000  0.000000e+00  6.175842e-01     0.000000
#>      TCGA-AB-2907 TCGA-AB-2908 TCGA-AB-2910 TCGA-AB-2911  TCGA-AB-2912
#> Sig1   13.5567752   11.8948111     8.855350     0.000000  1.689437e+01
#> Sig2    0.9062519    2.5862835     1.231618     1.502287 3.153891e-200
#> Sig3    0.0000000    0.6175842     0.000000     0.000000  9.991224e-01
#>      TCGA-AB-2913 TCGA-AB-2914 TCGA-AB-2915 TCGA-AB-2916 TCGA-AB-2917
#> Sig1     9.786516   14.3223352   15.9753593    6.5606953     8.573237
#> Sig2     2.319960    0.9607836    1.5099390    2.1254757     2.502946
#> Sig3     0.000000    0.0000000    0.6175842    0.6175842     1.301536
#>      TCGA-AB-2918  TCGA-AB-2919 TCGA-AB-2920 TCGA-AB-2921 TCGA-AB-2922
#> Sig1    0.0000000  9.354442e+00    6.6716303     6.136202   10.2282111
#> Sig2    0.0000000 1.014412e-309    0.0000000     1.839660    0.0000000
#> Sig3    0.6175842  6.175842e-01    0.9991224     0.000000    0.6175842
#>      TCGA-AB-2923 TCGA-AB-2924 TCGA-AB-2925  TCGA-AB-2926 TCGA-AB-2927
#> Sig1    13.891391     8.470661     9.020734  1.022821e+01   19.2094666
#> Sig2     3.293236     0.000000     1.108138 6.587231e-260    1.5431292
#> Sig3     0.000000     0.000000     1.301536  9.991224e-01    0.6175842
#>       TCGA-AB-2928 TCGA-AB-2929  TCGA-AB-2930 TCGA-AB-2931 TCGA-AB-2932
#> Sig1  6.671630e+00 1.279268e+01  5.755611e+00    8.4706612    3.7445125
#> Sig2 2.684157e-171 1.761763e-34 8.236704e-159    0.8432031    0.9521251
#> Sig3  9.991224e-01 6.175842e-01  0.000000e+00    0.0000000    0.0000000
#>       TCGA-AB-2933 TCGA-AB-2934 TCGA-AB-2935 TCGA-AB-2936 TCGA-AB-2937
#> Sig1 1.236288e-193 6.671632e+00    7.5765150     3.502649     7.216986
#> Sig2  8.432031e-01 4.927690e-88    0.0000000     1.677474     1.121426
#> Sig3  0.000000e+00 0.000000e+00    0.9991224     1.559996     0.000000
#>       TCGA-AB-2938 TCGA-AB-2939 TCGA-AB-2940  TCGA-AB-2941 TCGA-AB-2943
#> Sig1  1.609008e+01    8.2183533     0.000000  3.888454e+00    11.472776
#> Sig2 2.845440e-118    3.1269896     1.415341 1.465335e-193     1.823157
#> Sig3  0.000000e+00    0.6175842     0.000000  0.000000e+00     0.000000
#>      TCGA-AB-2945 TCGA-AB-2946 TCGA-AB-2947 TCGA-AB-2948  TCGA-AB-2949
#> Sig1     6.987679     0.000000 1.971331e+00    0.9927296  1.194702e+01
#> Sig2     2.360461     2.062173 5.870589e-15    0.0000000 1.260654e-100
#> Sig3     0.000000     0.000000 0.000000e+00    0.0000000  0.000000e+00
#>       TCGA-AB-2950 TCGA-AB-2952  TCGA-AB-2954  TCGA-AB-2955  TCGA-AB-2956
#> Sig1  1.022821e+01    9.7613145 9.448370e-193  1.445791e+01  2.936390e+00
#> Sig2 2.453169e-112    1.8118137  8.432031e-01 8.910310e-223 1.573494e-265
#> Sig3  0.000000e+00    0.6175842  0.000000e+00  6.175842e-01  6.175842e-01
#>      TCGA-AB-2957 TCGA-AB-2959 TCGA-AB-2963 TCGA-AB-2964 TCGA-AB-2965
#> Sig1     1.971331    18.810115    13.357456   13.9903626     5.342211
#> Sig2     0.000000     1.836134     1.072568    1.8284110     1.153707
#> Sig3     0.000000     1.301536     0.000000    0.6175842     0.000000
#>      TCGA-AB-2966 TCGA-AB-2967 TCGA-AB-2968 TCGA-AB-2970 TCGA-AB-2971
#> Sig1   11.0923000     4.433947    13.556775     5.755611     7.074532
#> Sig2    2.0621725     1.747535     0.906252     0.000000     1.823187
#> Sig3    0.6175842     0.000000     0.000000     0.000000     0.000000
#>      TCGA-AB-2972 TCGA-AB-2973 TCGA-AB-2974 TCGA-AB-2975 TCGA-AB-2976
#> Sig1    12.937095    2.9363899 6.671630e+00    0.0000000     9.531549
#> Sig2     2.476086    0.8432031 6.915328e-56    0.8432031     2.919068
#> Sig3     1.789445    0.0000000 0.000000e+00    0.6175842     1.301536
#>       TCGA-AB-2977 TCGA-AB-2978 TCGA-AB-2979 TCGA-AB-2980 TCGA-AB-2981
#> Sig1  5.755611e+00    8.8304127 5.755611e+00     2.442596     3.888454
#> Sig2 3.861237e-108    3.6427370 4.190173e-09     1.799366     0.000000
#> Sig3  1.301536e+00    0.6175842 0.000000e+00     0.000000     0.000000
#>       TCGA-AB-2982  TCGA-AB-2983 TCGA-AB-2984  TCGA-AB-2985 TCGA-AB-2986
#> Sig1 9.149848e-113  1.194702e+01     4.417923  3.888454e+00     2.707318
#> Sig2  1.502287e+00 1.943517e-103     1.757164  0.000000e+00     3.876860
#> Sig3  0.000000e+00  6.175842e-01     1.301536 9.462957e-136     0.000000
#>       TCGA-AB-2987 TCGA-AB-2988 TCGA-AB-2989  TCGA-AB-2990 TCGA-AB-2991
#> Sig1  5.755611e+00     7.211722    5.3658163  6.671630e+00    5.4772078
#> Sig2 3.903205e-153     2.269104    1.7479646 2.948246e-171    1.0556929
#> Sig3  0.000000e+00     1.301536    0.6175842  6.175842e-01    0.6175842
#>      TCGA-AB-2992 TCGA-AB-2993 TCGA-AB-2994 TCGA-AB-2995 TCGA-AB-2996
#> Sig1    1.0322958     8.470661     4.638691 4.828035e+00   12.3790817
#> Sig2    3.4041932     0.000000     2.166985 1.903190e-49    1.1821923
#> Sig3    0.6175842     0.000000     1.301536 6.175842e-01    0.9991224
#>       TCGA-AB-2997 TCGA-AB-2998 TCGA-AB-2999 TCGA-AB-3000 TCGA-AB-3001
#> Sig1  7.576515e+00 8.470661e+00 9.354442e+00     2.328076     5.508604
#> Sig2 7.747525e-167 5.913427e-34 1.005178e-96     1.273999     3.525891
#> Sig3  1.301536e+00 0.000000e+00 6.175842e-01     0.000000     0.000000
#>      TCGA-AB-3002 TCGA-AB-3005 TCGA-AB-3006 TCGA-AB-3007 TCGA-AB-3008
#> Sig1    20.724448   13.4206714   10.8655343 6.671630e+00     3.888454
#> Sig2     2.117832    1.6500955    3.1186304 3.817377e-87     0.000000
#> Sig3     0.000000    0.6175842    0.6175842 0.000000e+00     0.000000
#>      TCGA-AB-3009  TCGA-AB-3011 TCGA-AB-3012
#> Sig1    22.771898  5.755611e+00     4.532856
#> Sig2     2.743100 5.182146e-149     1.687536
#> Sig3     1.789445  0.000000e+00     0.000000
#> 
#> $Exposure.norm
#>       TCGA-AB-2802  TCGA-AB-2803  TCGA-AB-2804  TCGA-AB-2805 TCGA-AB-2806
#> Sig1  1.000000e+00  9.508473e-01  1.000000e+00  1.000000e+00    0.8127088
#> Sig2 6.005186e-125 2.316673e-150 3.588567e-218 6.397887e-146    0.1872912
#> Sig3  0.000000e+00  4.915268e-02  0.000000e+00  0.000000e+00    0.0000000
#>      TCGA-AB-2807 TCGA-AB-2808 TCGA-AB-2809 TCGA-AB-2810 TCGA-AB-2812
#> Sig1   0.91388266 1.000000e+00    0.7286538   0.91736932    0.7763853
#> Sig2   0.08611734 4.467578e-72    0.2713462   0.00000000    0.2236147
#> Sig3   0.00000000 0.000000e+00    0.0000000   0.08263068    0.0000000
#>      TCGA-AB-2813 TCGA-AB-2814 TCGA-AB-2816  TCGA-AB-2817 TCGA-AB-2818
#> Sig1   0.76758172    0.2770661 9.030966e-01  9.320458e-01     0.894156
#> Sig2   0.18648539    0.5496429 1.935987e-94 5.236989e-117     0.105844
#> Sig3   0.04593289    0.1732910 9.690339e-02  6.795417e-02     0.000000
#>      TCGA-AB-2819 TCGA-AB-2820 TCGA-AB-2821  TCGA-AB-2822 TCGA-AB-2824
#> Sig1   0.69696999   0.92052614    0.8570704  9.662682e-01     0.874238
#> Sig2   0.22388652   0.07947386    0.0000000 4.166763e-187     0.125762
#> Sig3   0.07914348   0.00000000    0.1429296  3.373179e-02     0.000000
#>      TCGA-AB-2825 TCGA-AB-2826 TCGA-AB-2827 TCGA-AB-2828 TCGA-AB-2829
#> Sig1            0            1    0.6633455   0.94305773            1
#> Sig2            1            0    0.2155173   0.00000000            0
#> Sig3            0            0    0.1211373   0.05694227            0
#>      TCGA-AB-2830 TCGA-AB-2831 TCGA-AB-2832 TCGA-AB-2833 TCGA-AB-2834
#> Sig1    0.8781722            1 7.931972e-28 9.152742e-01            0
#> Sig2    0.0725795            0 1.000000e+00 7.398245e-41            1
#> Sig3    0.0492483            0 0.000000e+00 8.472575e-02            0
#>      TCGA-AB-2835 TCGA-AB-2836  TCGA-AB-2838 TCGA-AB-2839  TCGA-AB-2841
#> Sig1            0    0.6164821  9.415349e-01   0.80936510  1.000000e+00
#> Sig2            0    0.0000000 3.054307e-140   0.15031033 3.201578e-106
#> Sig3            1    0.3835179  5.846514e-02   0.04032457  0.000000e+00
#>      TCGA-AB-2842 TCGA-AB-2843 TCGA-AB-2844 TCGA-AB-2845 TCGA-AB-2846
#> Sig1            1    0.6910555   0.90349967 1.000000e+00   0.72032071
#> Sig2            0    0.1225431   0.00000000 7.640792e-73   0.21876794
#> Sig3            0    0.1864014   0.09650033 0.000000e+00   0.06091135
#>      TCGA-AB-2847 TCGA-AB-2849  TCGA-AB-2850 TCGA-AB-2851  TCGA-AB-2853
#> Sig1    0.8486674    0.7699552  1.000000e+00    0.8235432  9.152742e-01
#> Sig2    0.1513326    0.2300448 3.254968e-106    0.1764568 6.218503e-169
#> Sig3    0.0000000    0.0000000  0.000000e+00    0.0000000  8.472575e-02
#>      TCGA-AB-2854 TCGA-AB-2855 TCGA-AB-2857 TCGA-AB-2858 TCGA-AB-2859
#> Sig1    0.8840999            1   0.95084732    0.7726168    0.8271813
#> Sig2    0.1159001            0   0.00000000    0.1724106    0.1728187
#> Sig3    0.0000000            0   0.04915268    0.0549726    0.0000000
#>      TCGA-AB-2860  TCGA-AB-2861 TCGA-AB-2862 TCGA-AB-2863 TCGA-AB-2864
#> Sig1    0.4261161  9.630359e-01    0.8011203   0.78915021   0.84002287
#> Sig2    0.3815997 8.841481e-263    0.1988797   0.16790667   0.08244079
#> Sig3    0.1922842  3.696413e-02    0.0000000   0.04294312   0.07753635
#>      TCGA-AB-2865 TCGA-AB-2866 TCGA-AB-2867 TCGA-AB-2868 TCGA-AB-2869
#> Sig1   0.86060799            1    0.7628321   0.90977074    0.7398999
#> Sig2   0.07455772            0    0.0000000   0.09022926    0.2601001
#> Sig3   0.06483429            0    0.2371679   0.00000000    0.0000000
#>      TCGA-AB-2870 TCGA-AB-2871 TCGA-AB-2872 TCGA-AB-2873 TCGA-AB-2874
#> Sig1    0.8897311   0.91100984   0.82658372            0   0.90194545
#> Sig2    0.1102689   0.00000000   0.11080473            1   0.09805455
#> Sig3    0.0000000   0.08899016   0.06261155            0   0.00000000
#>       TCGA-AB-2875 TCGA-AB-2876 TCGA-AB-2877 TCGA-AB-2878 TCGA-AB-2879
#> Sig1  1.000000e+00   0.93204584            1    0.8810609            0
#> Sig2 6.980839e-111   0.00000000            0    0.1189391            1
#> Sig3  0.000000e+00   0.06795416            0    0.0000000            0
#>       TCGA-AB-2880 TCGA-AB-2881 TCGA-AB-2882 TCGA-AB-2883 TCGA-AB-2884
#> Sig1 1.482197e-322    0.7883066 1.000000e+00            1 1.000000e+00
#> Sig2  1.000000e+00    0.1350570 3.154115e-77            0 5.667135e-40
#> Sig3  0.000000e+00    0.0766364 0.000000e+00            0 0.000000e+00
#>      TCGA-AB-2885 TCGA-AB-2886 TCGA-AB-2887 TCGA-AB-2888 TCGA-AB-2889
#> Sig1   0.83105010   0.65936317   0.79970051    0.4660478  1.00000e+00
#> Sig2   0.11132999   0.27220849   0.15265222    0.3433194 1.19598e-181
#> Sig3   0.05761991   0.06842834   0.04764726    0.1906328  0.00000e+00
#>      TCGA-AB-2890 TCGA-AB-2891 TCGA-AB-2892  TCGA-AB-2894 TCGA-AB-2895
#> Sig1    0.8365124            1            0 1.157888e-294 1.000000e+00
#> Sig2    0.1634876            0            1  8.054916e-01 6.486041e-41
#> Sig3    0.0000000            0            0  1.945084e-01 0.000000e+00
#>      TCGA-AB-2896 TCGA-AB-2897 TCGA-AB-2898 TCGA-AB-2899 TCGA-AB-2900
#> Sig1            0            1  1.00000e+00   0.82896485   0.63128587
#> Sig2            1            0  3.48454e-50   0.09325840   0.26913496
#> Sig3            0            0  0.00000e+00   0.07777675   0.09957916
#>      TCGA-AB-2901 TCGA-AB-2903  TCGA-AB-2904  TCGA-AB-2905 TCGA-AB-2906
#> Sig1    0.8285404            1  1.000000e+00  9.662682e-01    0.8048122
#> Sig2    0.0000000            0 1.135753e-215 3.244236e-132    0.1951878
#> Sig3    0.1714596            0  0.000000e+00  3.373179e-02    0.0000000
#>      TCGA-AB-2907 TCGA-AB-2908 TCGA-AB-2910 TCGA-AB-2911  TCGA-AB-2912
#> Sig1   0.93734009   0.78780477       0.8779            0  9.441628e-01
#> Sig2   0.06265991   0.17129204       0.1221            1 1.762592e-201
#> Sig3   0.00000000   0.04090319       0.0000            0  5.583721e-02
#>      TCGA-AB-2913 TCGA-AB-2914 TCGA-AB-2915 TCGA-AB-2916 TCGA-AB-2917
#> Sig1    0.8083704   0.93713433   0.88247600    0.7051664    0.6926346
#> Sig2    0.1916296   0.06286567   0.08340876    0.2284535    0.2022138
#> Sig3    0.0000000   0.00000000   0.03411524    0.0663801    0.1051516
#>      TCGA-AB-2918  TCGA-AB-2919 TCGA-AB-2920 TCGA-AB-2921 TCGA-AB-2922
#> Sig1            0  9.380683e-01    0.8697491    0.7693465   0.94305773
#> Sig2            0 1.017257e-310    0.0000000    0.2306535   0.00000000
#> Sig3            1  6.193166e-02    0.1302509    0.0000000   0.05694227
#>      TCGA-AB-2923 TCGA-AB-2924 TCGA-AB-2925  TCGA-AB-2926 TCGA-AB-2927
#> Sig1    0.8083615            1   0.78918737  9.110098e-01   0.89889119
#> Sig2    0.1916385            0   0.09694649 5.867138e-261   0.07220946
#> Sig3    0.0000000            0   0.11386614  8.899018e-02   0.02889934
#>       TCGA-AB-2928 TCGA-AB-2929  TCGA-AB-2930 TCGA-AB-2931 TCGA-AB-2932
#> Sig1  8.697491e-01 9.539469e-01  1.000000e+00   0.90946796    0.7972752
#> Sig2 3.499209e-172 1.313742e-35 1.431074e-159   0.09053204    0.2027248
#> Sig3  1.302509e-01 4.605309e-02  0.000000e+00   0.00000000    0.0000000
#>      TCGA-AB-2933 TCGA-AB-2934 TCGA-AB-2935 TCGA-AB-2936 TCGA-AB-2937
#> Sig1 1.46618e-193 1.000000e+00    0.8834929    0.5196717    0.8655108
#> Sig2  1.00000e+00 7.386034e-89    0.0000000    0.2488790    0.1344892
#> Sig3  0.00000e+00 0.000000e+00    0.1165071    0.2314493    0.0000000
#>       TCGA-AB-2938 TCGA-AB-2939 TCGA-AB-2940  TCGA-AB-2941 TCGA-AB-2943
#> Sig1  1.000000e+00   0.68698516            0  1.000000e+00    0.8628786
#> Sig2 1.768444e-119   0.26139001            1 3.768426e-194    0.1371214
#> Sig3  0.000000e+00   0.05162484            0  0.000000e+00    0.0000000
#>      TCGA-AB-2945 TCGA-AB-2946 TCGA-AB-2947 TCGA-AB-2948  TCGA-AB-2949
#> Sig1     0.747494            0 1.000000e+00            1  1.000000e+00
#> Sig2     0.252506            1 2.977983e-15            0 1.055203e-101
#> Sig3     0.000000            0 0.000000e+00            0  0.000000e+00
#>       TCGA-AB-2950 TCGA-AB-2952  TCGA-AB-2954  TCGA-AB-2955  TCGA-AB-2956
#> Sig1  1.000000e+00   0.80071732 1.120533e-192  9.590339e-01  8.262272e-01
#> Sig2 2.398434e-113   0.14862246  1.000000e+00 5.910460e-224 4.427420e-266
#> Sig3  0.000000e+00   0.05066022  0.000000e+00  4.096610e-02  1.737728e-01
#>      TCGA-AB-2957 TCGA-AB-2959 TCGA-AB-2963 TCGA-AB-2964 TCGA-AB-2965
#> Sig1            1   0.85703933   0.92567109   0.85118387     0.822395
#> Sig2            0   0.08365918   0.07432891   0.11124186     0.177605
#> Sig3            0   0.05930150   0.00000000   0.03757427     0.000000
#>      TCGA-AB-2966 TCGA-AB-2967 TCGA-AB-2968 TCGA-AB-2970 TCGA-AB-2971
#> Sig1   0.80542073    0.7172952   0.93734009            1     0.795095
#> Sig2   0.14973599    0.2827048   0.06265991            0     0.204905
#> Sig3   0.04484328    0.0000000   0.00000000            0     0.000000
#>      TCGA-AB-2972 TCGA-AB-2973 TCGA-AB-2974 TCGA-AB-2975 TCGA-AB-2976
#> Sig1    0.7520419    0.7769064 1.000000e+00    0.0000000   0.69309500
#> Sig2    0.1439365    0.2230936 1.036527e-56    0.5772251   0.21226263
#> Sig3    0.1040216    0.0000000 0.000000e+00    0.4227749   0.09464238
#>       TCGA-AB-2977 TCGA-AB-2978 TCGA-AB-2979 TCGA-AB-2980 TCGA-AB-2981
#> Sig1  8.155719e-01    0.6745544 1.000000e+00    0.5758175            1
#> Sig2 5.471385e-109    0.2782684 7.280153e-10    0.4241825            0
#> Sig3  1.844281e-01    0.0471772 0.000000e+00    0.0000000            0
#>       TCGA-AB-2982  TCGA-AB-2983 TCGA-AB-2984  TCGA-AB-2985 TCGA-AB-2986
#> Sig1 6.090612e-113  9.508473e-01    0.5908981  1.000000e+00    0.4111854
#> Sig2  1.000000e+00 1.546819e-104    0.2350211  0.000000e+00    0.5888146
#> Sig3  0.000000e+00  4.915268e-02    0.1740808 2.433604e-136    0.0000000
#>       TCGA-AB-2987 TCGA-AB-2988 TCGA-AB-2989  TCGA-AB-2990 TCGA-AB-2991
#> Sig1  1.000000e+00    0.6688443   0.69403220  9.152742e-01   0.76599110
#> Sig2 6.781566e-154    0.2104459   0.22608745 4.044669e-172   0.14763934
#> Sig3  0.000000e+00    0.1207098   0.07988035  8.472575e-02   0.08636955
#>      TCGA-AB-2992 TCGA-AB-2993 TCGA-AB-2994 TCGA-AB-2995 TCGA-AB-2996
#> Sig1    0.2042503            1    0.5721684 8.865906e-01   0.85018851
#> Sig2    0.6735544            0    0.2672910 3.494901e-50   0.08119231
#> Sig3    0.1221953            0    0.1605406 1.134094e-01   0.06861918
#>       TCGA-AB-2997 TCGA-AB-2998 TCGA-AB-2999 TCGA-AB-3000 TCGA-AB-3001
#> Sig1  8.533984e-01 1.000000e+00 9.380683e-01    0.6463152    0.6097301
#> Sig2 8.726606e-168 6.981069e-35 1.007997e-97    0.3536848    0.3902699
#> Sig3  1.466016e-01 0.000000e+00 6.193166e-02    0.0000000    0.0000000
#>      TCGA-AB-3002 TCGA-AB-3005 TCGA-AB-3006 TCGA-AB-3007 TCGA-AB-3008
#> Sig1   0.90728454   0.85545456   0.74412554 1.000000e+00            1
#> Sig2   0.09271546   0.10517966   0.21357924 5.721806e-88            0
#> Sig3   0.00000000   0.03936578   0.04229522 0.000000e+00            0
#>      TCGA-AB-3009  TCGA-AB-3011 TCGA-AB-3012
#> Sig1   0.83399974  1.000000e+00    0.7287091
#> Sig2   0.10046350 9.003641e-150    0.2712909
#> Sig3   0.06553677  0.000000e+00    0.0000000
#> 
#> $K
#> [1] 3
#> 
#> $Raw
#> $Raw$summary_run
#> # A tibble: 3 × 5
#>     Run     K posterior file                           similarity_to_ref
#>   <int> <int>     <dbl> <chr>                                      <dbl>
#> 1     3     3     9645. /tmp/Rtmp1yZ0Ob/BayesNMF.3.rds             0.672
#> 2     2     2     9753. /tmp/Rtmp1yZ0Ob/BayesNMF.2.rds             0.640
#> 3     1     3     9627. /tmp/Rtmp1yZ0Ob/BayesNMF.1.rds             0.605
#> 
#> $Raw$W
#>                  Sig1          Sig2          Sig3
#> A[C>A]A  3.043035e-81  4.542514e-01 5.643264e-316
#> A[C>A]C  4.654683e-01 4.023656e-122 6.725382e-243
#> A[C>A]G  6.420252e-02 2.945150e-117  0.000000e+00
#> A[C>A]T  1.284050e-01 7.424752e-186 1.742975e-184
#> A[C>G]A  4.654364e-41  3.740894e-01  0.000000e+00
#> A[C>G]C  9.630379e-02  8.048469e-78 3.084143e-313
#> A[C>G]G  1.284050e-01 7.027470e-295  0.000000e+00
#> A[C>G]T  2.141285e-09  2.137653e-01  0.000000e+00
#> A[C>T]A  7.040247e-01  1.907124e-01 7.695633e-262
#> A[C>T]C  7.271123e-01  1.522766e-01 2.878057e-121
#> A[C>T]G  3.707696e+00  3.418294e-46 1.917769e-253
#> A[C>T]T  4.173164e-01 6.093083e-128  0.000000e+00
#> A[T>A]A  0.000000e+00  5.109356e-02  0.000000e+00
#> A[T>A]C  1.926076e-01  6.911758e-34  6.023969e-53
#> A[T>A]G  9.630379e-02  0.000000e+00 2.836994e-294
#> A[T>A]T  6.420252e-02 2.233794e-266  0.000000e+00
#> A[T>C]A  4.494177e-01 1.660118e-129  0.000000e+00
#> A[T>C]C  1.926076e-01  0.000000e+00 5.480453e-172
#> A[T>C]G  4.333670e-01 8.525354e-228 3.870278e-253
#> A[T>C]T  4.075355e-01  1.498863e-01 3.113099e-310
#> A[T>G]A 1.686888e-120  1.603240e-01  0.000000e+00
#> A[T>G]C  3.627914e-92 4.940656e-324  0.000000e+00
#> A[T>G]G  4.815189e-02  0.000000e+00  0.000000e+00
#> A[T>G]T  4.815189e-02 1.204591e-205  0.000000e+00
#> C[C>A]A  2.889114e-01  1.986747e-82 1.930103e-287
#> C[C>A]C  2.407595e-01 1.521622e-249 1.319884e-272
#> C[C>A]G  2.247088e-01 1.953000e-262  0.000000e+00
#> C[C>A]T  2.889114e-01  0.000000e+00  0.000000e+00
#> C[C>G]A  2.086582e-01 1.918377e-262 3.444416e-134
#> C[C>G]C  6.420252e-02 7.764801e-258  0.000000e+00
#> C[C>G]G  2.407595e-01  1.161650e-16  6.342897e-86
#> C[C>G]T  1.123544e-01  0.000000e+00  0.000000e+00
#> C[C>T]A  6.580759e-01 2.299379e-170  0.000000e+00
#> C[C>T]C  4.815189e-01  1.889538e-11  0.000000e+00
#> C[C>T]G  3.290379e+00  6.838678e-73 3.760483e-189
#> C[C>T]T  9.630379e-01 1.353618e-179 4.487963e-252
#> C[T>A]A  1.723269e-52  0.000000e+00  2.168570e-02
#> C[T>A]C 1.740984e-130  2.672067e-01 1.055760e-202
#> C[T>A]G  1.926076e-01 3.515030e-141 3.442128e-214
#> C[T>A]T  8.025316e-02  0.000000e+00  0.000000e+00
#> C[T>C]A  7.037852e-89  1.870447e-01  0.000000e+00
#> C[T>C]C  2.889114e-01 4.042925e-119  0.000000e+00
#> C[T>C]G  2.340602e-01  1.180355e-01  0.000000e+00
#> C[T>C]T  1.543631e-01  1.438305e-01 1.029771e-122
#> C[T>G]A 1.285113e-110  2.672067e-02  0.000000e+00
#> C[T>G]C  1.284050e-01 8.497567e-258 8.273788e-196
#> C[T>G]G  1.605063e-01 5.928578e-189  0.000000e+00
#> C[T>G]T  8.025316e-02 2.725280e-275 2.444118e-112
#> G[C>A]A  3.284934e-29  4.542514e-01 2.316763e-204
#> G[C>A]C  3.258293e-27  2.404860e-01 3.559425e-196
#> G[C>A]G  7.192379e-55  1.336033e-01  0.000000e+00
#> G[C>A]T  1.444557e-01 1.856816e-143  0.000000e+00
#> G[C>G]A  1.765569e-01  3.091853e-21 2.394603e-251
#> G[C>G]C  1.123544e-01 2.770312e-244 5.622693e-261
#> G[C>G]G  1.123544e-01 3.994366e-267  0.000000e+00
#> G[C>G]T  0.000000e+00  1.603240e-01 2.734295e-314
#> G[C>T]A  2.077701e-73  7.214581e-01 7.251785e-213
#> G[C>T]C  7.864809e-01 2.102439e-178 5.074682e-202
#> G[C>T]G  2.416254e+00  1.027690e+00 1.017039e-259
#> G[C>T]T  8.185822e-01 6.835872e-190 7.816576e-275
#> G[T>A]A 4.493148e-131  5.002529e-86  5.675181e-02
#> G[T>A]C  5.002728e-80  2.404860e-01 3.001989e-166
#> G[T>A]G  6.420252e-02  3.324105e-82 2.735032e-277
#> G[T>A]T  9.630379e-02 4.776164e-157 1.184337e-309
#> G[T>C]A  4.199929e-02  9.040469e-02  0.000000e+00
#> G[T>C]C  1.016785e-92  2.404860e-01  3.901567e-98
#> G[T>C]G  1.765569e-01 3.385453e-307  0.000000e+00
#> G[T>C]T  1.926076e-01  4.439522e-91 1.144100e-212
#> G[T>G]A 2.861684e-100 9.922698e-161  1.135036e-01
#> G[T>G]C  4.815189e-02 2.819095e-159 9.608941e-294
#> G[T>G]G  9.630379e-02  0.000000e+00  0.000000e+00
#> G[T>G]T 4.997157e-113  5.344134e-02  0.000000e+00
#> T[C>A]A  1.979637e-07  2.228391e-01  3.038481e-99
#> T[C>A]C  2.634899e-13  6.145754e-01 2.521513e-269
#> T[C>A]G  1.284050e-01  2.361354e-26  0.000000e+00
#> T[C>A]T  4.494177e-01  5.056641e-27 9.368356e-198
#> T[C>G]A  8.025316e-02 8.952395e-202 4.087418e-127
#> T[C>G]C  1.284050e-01 1.846080e-216 3.237392e-152
#> T[C>G]G  1.765569e-01 8.580557e-131 5.162857e-216
#> T[C>G]T  1.006910e-01  7.285825e-02 6.916919e-323
#> T[C>T]A  6.901771e-01 2.289818e-176 2.985626e-211
#> T[C>T]C  4.931792e-01  3.279570e-01 3.656086e-322
#> T[C>T]G  4.048266e-24 4.155401e-105  3.149726e+00
#> T[C>T]T  4.333670e-01 3.180703e-223 6.479641e-227
#> T[T>A]A  0.000000e+00  0.000000e+00  2.837591e-02
#> T[T>A]C  6.420252e-02 4.406374e-263  2.534224e-65
#> T[T>A]G  0.000000e+00 7.613289e-245  1.135036e-01
#> T[T>A]T  8.025316e-02  1.071400e-23 5.351164e-106
#> T[T>C]A  0.000000e+00  5.344134e-02 9.881313e-324
#> T[T>C]C  1.444557e-01 2.845275e-103 4.369035e-296
#> T[T>C]G  2.191296e-75  2.672067e-01  0.000000e+00
#> T[T>C]T  1.867933e-28 2.745515e-166  5.675181e-01
#> T[T>G]A  4.815189e-02 1.272427e-154  0.000000e+00
#> T[T>G]C  8.025316e-02 3.990026e-165  0.000000e+00
#> T[T>G]G 6.802541e-316  5.344134e-02  0.000000e+00
#> T[T>G]T 8.844672e-123 4.345921e-177  1.418795e-01
#> 
#> $Raw$H
#>      TCGA-AB-2802  TCGA-AB-2803  TCGA-AB-2804  TCGA-AB-2805 TCGA-AB-2806
#> Sig1  3.72185e-01  4.753360e-01  2.654440e-01  5.089821e-01    0.4157716
#> Sig2 7.52596e-125 3.899701e-150 3.207529e-218 1.096518e-145    0.3226368
#> Sig3  0.00000e+00  1.472913e-01  0.000000e+00  0.000000e+00    0.0000000
#>      TCGA-AB-2807 TCGA-AB-2808 TCGA-AB-2809 TCGA-AB-2810 TCGA-AB-2812
#> Sig1    0.8329589 3.370220e-01    0.1078739    0.4413291    0.2193407
#> Sig2    0.2643019 5.069988e-72    0.1352681    0.0000000    0.2127251
#> Sig3    0.0000000 0.000000e+00    0.0000000    0.2382866    0.0000000
#>      TCGA-AB-2813 TCGA-AB-2814 TCGA-AB-2816  TCGA-AB-2817 TCGA-AB-2818
#> Sig1    0.4106180   0.06355752 2.289984e-01  3.370220e-01    0.3301210
#> Sig2    0.3359191   0.42456159 1.653017e-94 6.376454e-117    0.1315839
#> Sig3    0.1472913   0.23828663 1.472913e-01  1.472913e-01    0.0000000
#>      TCGA-AB-2819 TCGA-AB-2820 TCGA-AB-2821  TCGA-AB-2822 TCGA-AB-2824
#> Sig1    0.3500730    0.4671930    0.3721850  7.038748e-01   0.13439411
#> Sig2    0.3786596    0.1358192    0.0000000 1.022052e-186   0.06509935
#> Sig3    0.2382866    0.0000000    0.3720527  1.472913e-01   0.00000000
#>      TCGA-AB-2825 TCGA-AB-2826 TCGA-AB-2827 TCGA-AB-2828 TCGA-AB-2829
#> Sig1    0.0000000    0.1547099    0.2835694    0.4069496     0.372185
#> Sig2    0.1129665    0.0000000    0.3102259    0.0000000     0.000000
#> Sig3    0.0000000    0.0000000    0.3104111    0.1472913     0.000000
#>      TCGA-AB-2830 TCGA-AB-2831 TCGA-AB-2832 TCGA-AB-2833  TCGA-AB-2834
#> Sig1    0.4381528   0.07843329 1.525447e-28 2.654440e-01  0.000000e+00
#> Sig2    0.1219373   0.00000000 6.475784e-01 7.224819e-41 4.280424e-245
#> Sig3    0.1472913   0.00000000 0.000000e+00 1.472913e-01  0.000000e+00
#>      TCGA-AB-2835 TCGA-AB-2836  TCGA-AB-2838 TCGA-AB-2839  TCGA-AB-2841
#> Sig1    0.0000000   0.03949771  6.401756e-01    0.4931873  1.547099e-01
#> Sig2    0.0000000   0.00000000 6.992814e-140    0.3084131 1.667856e-106
#> Sig3    0.2382866   0.14729131  2.382866e-01    0.1472913  0.000000e+00
#>      TCGA-AB-2842 TCGA-AB-2843 TCGA-AB-2844 TCGA-AB-2845 TCGA-AB-2846
#> Sig1   0.07843329    0.2215616    0.3721850 1.168301e-01    0.2905794
#> Sig2   0.00000000    0.1322961    0.0000000 3.005866e-73    0.2971660
#> Sig3   0.00000000    0.3582374    0.2382866 0.000000e+00    0.1472913
#>      TCGA-AB-2847 TCGA-AB-2849  TCGA-AB-2850 TCGA-AB-2851  TCGA-AB-2853
#> Sig1    0.3084557    0.5860036  1.547099e-01    0.1842936  2.654440e-01
#> Sig2    0.1852100    0.5895544 1.695669e-106    0.1329655 6.072732e-169
#> Sig3    0.0000000    0.0000000  0.000000e+00    0.0000000  1.472913e-01
#>      TCGA-AB-2854 TCGA-AB-2855 TCGA-AB-2857 TCGA-AB-2858 TCGA-AB-2859
#> Sig1    0.3555607    0.1168301    0.4753360    0.3453465    0.3247417
#> Sig2    0.1569541    0.0000000    0.0000000    0.2594965    0.2284571
#> Sig3    0.0000000    0.0000000    0.1472913    0.1472913    0.0000000
#>      TCGA-AB-2860  TCGA-AB-2861 TCGA-AB-2862 TCGA-AB-2863 TCGA-AB-2864
#> Sig1    0.1147577  6.401756e-01    0.3100330    0.4515472    0.4306708
#> Sig2    0.3460496 1.979057e-262    0.2591655    0.3235102    0.1423223
#> Sig3    0.3104111  1.472913e-01    0.0000000    0.1472913    0.2382866
#>      TCGA-AB-2865 TCGA-AB-2866 TCGA-AB-2867 TCGA-AB-2868 TCGA-AB-2869
#> Sig1    0.5276675   0.03949771    0.2289984    0.4301677    0.2531224
#> Sig2    0.1539303   0.00000000    0.0000000    0.1436580    0.2996226
#> Sig3    0.2382866   0.00000000    0.4267753    0.0000000    0.0000000
#>      TCGA-AB-2870 TCGA-AB-2871 TCGA-AB-2872 TCGA-AB-2873 TCGA-AB-2874
#> Sig1    0.3280856    0.4069497    0.3243916    0.0000000    0.4850392
#> Sig2    0.1369171    0.0000000    0.1464258    0.1129665    0.1775583
#> Sig3    0.0000000    0.2382866    0.1472913    0.0000000    0.0000000
#>       TCGA-AB-2875 TCGA-AB-2876 TCGA-AB-2877 TCGA-AB-2878 TCGA-AB-2879
#> Sig1  3.370220e-01    0.3370221    0.6721758    0.3539805     0.000000
#> Sig2 7.922139e-111    0.0000000    0.0000000    0.1609069     0.201266
#> Sig3  0.000000e+00    0.1472913    0.0000000    0.0000000     0.000000
#>       TCGA-AB-2880 TCGA-AB-2881 TCGA-AB-2882 TCGA-AB-2883 TCGA-AB-2884
#> Sig1 4.940656e-324    0.2527535 5.752365e-01   0.03949771 1.547099e-01
#> Sig2  1.129665e-01    0.1458129 6.109428e-77   0.00000000 2.952283e-40
#> Sig3  0.000000e+00    0.1472913 0.000000e+00   0.00000000 0.000000e+00
#>      TCGA-AB-2885 TCGA-AB-2886 TCGA-AB-2887 TCGA-AB-2888  TCGA-AB-2889
#> Sig1    0.3543984    0.2367695    0.4124075   0.09718352  1.920929e-01
#> Sig2    0.1598650    0.3291390    0.2650812   0.24106660 7.735918e-182
#> Sig3    0.1472913    0.1472913    0.1472913   0.23828663  0.000000e+00
#>      TCGA-AB-2890 TCGA-AB-2891 TCGA-AB-2892  TCGA-AB-2894 TCGA-AB-2895
#> Sig1    0.2461326    0.5752365    0.0000000 1.462733e-295 5.422787e-01
#> Sig2    0.1619788    0.0000000    0.1129665  3.426386e-01 1.184347e-40
#> Sig3    0.0000000    0.0000000    0.0000000  1.472913e-01 0.000000e+00
#>      TCGA-AB-2896 TCGA-AB-2897 TCGA-AB-2898 TCGA-AB-2899 TCGA-AB-2900
#> Sig1    0.0000000    0.2289984 6.078656e-01    0.4236878    0.3282879
#> Sig2    0.1129665    0.0000000 7.132299e-50    0.1604998    0.4712762
#> Sig3    0.0000000    0.0000000 0.000000e+00    0.2382866    0.3104111
#>      TCGA-AB-2901 TCGA-AB-2903  TCGA-AB-2904  TCGA-AB-2905 TCGA-AB-2906
#> Sig1    0.1920930   0.03949771  7.664025e-01  7.038748e-01    0.3867325
#> Sig2    0.0000000   0.00000000 2.931011e-215 7.957685e-132    0.3158242
#> Sig3    0.2382866   0.00000000  0.000000e+00  1.472913e-01    0.0000000
#>      TCGA-AB-2907 TCGA-AB-2908 TCGA-AB-2910 TCGA-AB-2911  TCGA-AB-2912
#> Sig1    0.5393831    0.4732586    0.3523276     0.000000  6.721758e-01
#> Sig2    0.1214134    0.3464924    0.1650037     0.201266 4.225365e-201
#> Sig3    0.0000000    0.1472913    0.0000000     0.000000  2.382866e-01
#>      TCGA-AB-2913 TCGA-AB-2914 TCGA-AB-2915 TCGA-AB-2916 TCGA-AB-2917
#> Sig1    0.3893759    0.5698425    0.6356113    0.2610302    0.3411032
#> Sig2    0.3108121    0.1287191    0.2022912    0.2847565    0.3353274
#> Sig3    0.0000000    0.0000000    0.1472913    0.1472913    0.3104111
#>      TCGA-AB-2918  TCGA-AB-2919 TCGA-AB-2920 TCGA-AB-2921 TCGA-AB-2922
#> Sig1    0.0000000  3.721850e-01    0.2654440    0.2441409    0.4069496
#> Sig2    0.0000000 1.359038e-310    0.0000000    0.2464649    0.0000000
#> Sig3    0.1472913  1.472913e-01    0.2382866    0.0000000    0.1472913
#>      TCGA-AB-2923 TCGA-AB-2924 TCGA-AB-2925  TCGA-AB-2926 TCGA-AB-2927
#> Sig1    0.5526965     0.337022    0.3589078  4.069496e-01    0.7642867
#> Sig2    0.4412049     0.000000    0.1484607 8.825116e-261    0.2067378
#> Sig3    0.0000000     0.000000    0.3104111  2.382866e-01    0.1472913
#>       TCGA-AB-2928 TCGA-AB-2929  TCGA-AB-2930 TCGA-AB-2931 TCGA-AB-2932
#> Sig1  2.654440e-01 5.089821e-01  2.289984e-01    0.3370220    0.1489828
#> Sig2 3.596048e-172 2.360289e-35 1.103497e-159    0.1129665    0.1275591
#> Sig3  2.382866e-01 1.472913e-01  0.000000e+00    0.0000000    0.0000000
#>       TCGA-AB-2933 TCGA-AB-2934 TCGA-AB-2935 TCGA-AB-2936 TCGA-AB-2937
#> Sig1 4.918816e-195 2.654441e-01    0.3014466    0.1393598    0.2871421
#> Sig2  1.129665e-01 6.601778e-89    0.0000000    0.2247363    0.1502409
#> Sig3  0.000000e+00 0.000000e+00    0.2382866    0.3720527    0.0000000
#>       TCGA-AB-2938 TCGA-AB-2939 TCGA-AB-2940  TCGA-AB-2941 TCGA-AB-2943
#> Sig1  6.401756e-01    0.3269835    0.0000000  1.547099e-01    0.4564671
#> Sig2 3.812123e-119    0.4189324    0.1896175 1.963155e-194    0.2442539
#> Sig3  0.000000e+00    0.1472913    0.0000000  0.000000e+00    0.0000000
#>      TCGA-AB-2945 TCGA-AB-2946 TCGA-AB-2947 TCGA-AB-2948  TCGA-AB-2949
#> Sig1    0.2780186    0.0000000 7.843329e-02   0.03949771  4.753360e-01
#> Sig2    0.3162382    0.2762756 7.865008e-16   0.00000000 1.688937e-101
#> Sig3    0.0000000    0.0000000 0.000000e+00   0.00000000  0.000000e+00
#>       TCGA-AB-2950 TCGA-AB-2952  TCGA-AB-2954  TCGA-AB-2955  TCGA-AB-2956
#> Sig1  4.069496e-01    0.3883732 3.759221e-194  5.752365e-01  1.168301e-01
#> Sig2 3.286586e-113    0.2427342  1.129665e-01 1.193742e-223 2.108058e-266
#> Sig3  0.000000e+00    0.1472913  0.000000e+00  1.472913e-01  1.472913e-01
#>      TCGA-AB-2957 TCGA-AB-2959 TCGA-AB-2963 TCGA-AB-2964 TCGA-AB-2965
#> Sig1   0.07843329    0.7483977    0.5314528    0.5566343    0.2125505
#> Sig2   0.00000000    0.2459925    0.1436952    0.2449578    0.1545657
#> Sig3   0.00000000    0.3104111    0.0000000    0.1472913    0.0000000
#>      TCGA-AB-2966 TCGA-AB-2967 TCGA-AB-2968 TCGA-AB-2970 TCGA-AB-2971
#> Sig1    0.4413291    0.1764134    0.5393831    0.2289984    0.2814743
#> Sig2    0.2762756    0.2341226    0.1214134    0.0000000    0.2442580
#> Sig3    0.1472913    0.0000000    0.0000000    0.0000000    0.0000000
#>      TCGA-AB-2972 TCGA-AB-2973 TCGA-AB-2974 TCGA-AB-2975 TCGA-AB-2976
#> Sig1    0.5147279    0.1168301 2.654440e-01    0.0000000    0.3792315
#> Sig2    0.3317289    0.1129665 9.264677e-57    0.1129665    0.3910765
#> Sig3    0.4267753    0.0000000 0.000000e+00    0.1472913    0.3104111
#>       TCGA-AB-2977 TCGA-AB-2978 TCGA-AB-2979 TCGA-AB-2980 TCGA-AB-2981
#> Sig1  2.289984e-01    0.3513354 2.289984e-01   0.09718352    0.1547099
#> Sig2 5.173018e-109    0.4880287 5.613703e-10   0.24106660    0.0000000
#> Sig3  3.104111e-01    0.1472913 0.000000e+00   0.00000000    0.0000000
#>       TCGA-AB-2982  TCGA-AB-2983 TCGA-AB-2984  TCGA-AB-2985 TCGA-AB-2986
#> Sig1 3.640448e-114  4.753360e-01    0.1757758  1.547099e-01    0.1077160
#> Sig2  2.012660e-01 2.603790e-104    0.2354127  0.000000e+00    0.5193948
#> Sig3  0.000000e+00  1.472913e-01    0.3104111 2.256877e-136    0.0000000
#>       TCGA-AB-2987 TCGA-AB-2988 TCGA-AB-2989  TCGA-AB-2990 TCGA-AB-2991
#> Sig1  2.289984e-01    0.2869327    0.2134896  2.654440e-01    0.2179216
#> Sig2 5.229244e-154    0.3039989    0.2341802 3.949855e-172    0.1414344
#> Sig3  0.000000e+00    0.3104111    0.1472913  1.472913e-01    0.1472913
#>      TCGA-AB-2992 TCGA-AB-2993 TCGA-AB-2994 TCGA-AB-2995 TCGA-AB-2996
#> Sig1   0.04107193     0.337022    0.1845595 1.920929e-01    0.4925263
#> Sig2   0.45607023     0.000000    0.2903176 2.549762e-50    0.1583819
#> Sig3   0.14729131     0.000000    0.3104111 1.472913e-01    0.2382866
#>       TCGA-AB-2997 TCGA-AB-2998 TCGA-AB-2999 TCGA-AB-3000 TCGA-AB-3001
#> Sig1  3.014466e-01  3.37022e-01 3.721850e-01   0.09262709    0.2191707
#> Sig2 1.037960e-167  7.92240e-35 1.346667e-97   0.17068156    0.4723745
#> Sig3  3.104111e-01  0.00000e+00 1.472913e-01   0.00000000    0.0000000
#>      TCGA-AB-3002 TCGA-AB-3005 TCGA-AB-3006 TCGA-AB-3007 TCGA-AB-3008
#> Sig1    0.8245632    0.5339680    0.4323068 2.654440e-01    0.1547099
#> Sig2    0.2837325    0.2210684    0.4178125 5.114258e-88    0.0000000
#> Sig3    0.0000000    0.1472913    0.1472913 0.000000e+00    0.0000000
#>      TCGA-AB-3009  TCGA-AB-3011 TCGA-AB-3012
#> Sig1    0.9060250  2.289984e-01    0.1803486
#> Sig2    0.3675015 6.942680e-150    0.2260844
#> Sig3    0.4267753  0.000000e+00    0.0000000
#> 
#> $Raw$best_run
#> [1] 3
#> 
#> 
#> attr(,"class")
#> [1] "Signature"
#> attr(,"nrun")
#> [1] 3
#> attr(,"method")
#> [1] "L1W.L2H"
#> attr(,"call_method")
#> [1] "BayesianNMF"
# }