Tally a Genomic Alteration Object

Tally a variation object like MAF, CopyNumber and return a matrix for NMF de-composition and more. This is a generic function, so it can be further extended to other mutation cases. Please read details about how to set sex for identifying copy number signatures. Please read https://osf.io/s93d5/ for the generation of SBS, DBS and ID (INDEL) components. Of note, many options are designed for method "M" only, and they are highlighted by bold fonts (you can ignore them if you don't use "M" method).

sig_tally(object, ...)

# S3 method for CopyNumber
sig_tally(
  object,
  method = "Wang",
  ignore_chrs = NULL,
  indices = NULL,
  feature_setting = sigminer::CN.features,
  type = c("probability", "count"),
  reference_components = FALSE,
  cores = 1,
  seed = 123456,
  min_comp = 2,
  max_comp = 15,
  min_prior = 0.001,
  model_selection = "BIC",
  threshold = 0.1,
  nrep = 1,
  niter = 1000,
  keep_only_matrix = FALSE,
  ...
)

# S3 method for MAF
sig_tally(
  object,
  mode = c("SBS", "DBS", "ID", "ALL"),
  ref_genome = "BSgenome.Hsapiens.UCSC.hg19",
  genome_build = NULL,
  add_trans_bias = FALSE,
  ignore_chrs = NULL,
  use_syn = TRUE,
  keep_only_matrix = FALSE,
  ...
)

Arguments

object
a CopyNumber object or MAF object.

...
custom setting for operating object. Detail see S3 method for corresponding class (e.g. CopyNumber).

method
method for feature classfication, can be one of "Macintyre" ("M"), "Wang" ("W").

ignore_chrs
Chromsomes to ignore from analysis. e.g. chrX and chrY.

indices
integer vector indicating segments to keep.

feature_setting
a data.frame used for classification. Only used when method is "Wang" ("W"). Default is CN.features. Users can also set custom input with "feature", "min" and "max" columns available. Valid features can be printed by unique(CN.features$feature).

type
one of "probability", "count". Default is "probability", return a matrix with the sum of posterior probabilities for each components. If set to 'count', return a matrix with event count assigned to each components. The result for both types should be close. Only used when method is "Macintyre".

reference_components
default is FALSE, calculate mixture components from CopyNumber object. Only used when method is "Macintyre".

cores
number of computer cores to run this task. You can use future::availableCores() function to check how many cores you can use.

seed
seed number. Only used when method is "Macintyre".

min_comp
minimal number of components to fit, default is 2. Can also be a vector with length 6, which apply to each feature. Only used when method is "Macintyre".

max_comp
maximal number of components to fit, default is 15. Can also be a vector with length 6, which apply to each feature. Only used when method is "Macintyre".

min_prior
the minimum relative size of components, default is 0.001. Details about custom setting please refer to flexmix package. Only used when method is "Macintyre".

model_selection
model selection strategy, default is 'BIC'. Details about custom setting please refer to flexmix package. Only used when method is "Macintyre".

threshold
default is 0.1. Sometimes, the result components include adjacent distributions with similar mu (two and more distribution are very close), we use this threshold to obtain a more meaningful fit with less components. Only used when method is "Macintyre".

nrep
number of run times for each value of component, keep only the solution with maximum likelihood. Only used when method is "Macintyre".

niter
the maximum number of iterations. Only used when method is "Macintyre".

keep_only_matrix
if TRUE, keep only matrix for signature extraction. For a MAF object, this will just return the most useful matrix.

mode
type of mutation matrix to extract, can be one of 'SBS', 'DBS' and 'ID'.

ref_genome
'BSgenome.Hsapiens.UCSC.hg19', 'BSgenome.Hsapiens.UCSC.hg38' and 'BSgenome.Mmusculus.UCSC.mm10' etc.

genome_build
genome build 'hg19', 'hg38' or "mm10", if not set, guess it by ref_genome.

add_trans_bias
if TRUE, consider transcriptional bias categories. 'T:' for Transcribed (the variant is on the transcribed strand); 'U:' for Un-transcribed (the variant is on the untranscribed strand); 'B:' for Bi-directional (the variant is on both strand and is transcribed either way); 'N:' for Non-transcribed (the variant is in a non-coding region and is untranslated); 'Q:' for Questionable. NOTE: the result counts of 'B' and 'N' labels are a little different from SigProfilerMatrixGenerator, the reason is unknown (may be caused by annotation file).

use_syn
Logical. Whether to include synonymous variants in analysis. Defaults to TRUE

Value

a list contains a matrix used for NMF de-composition.

Details

For identifying copy number signatures, we have to derive copy number features firstly. Due to the difference of copy number values in sex chromosomes between male and female, we have to do an extra step if we don't want to ignore them.

I create two options to control this, the default values are shown as the following, you can use the same way to set (per R session).

options(sigminer.sex = "female", sigminer.copynumber.max = NA_integer_)

If your cohort are all females, you can totally ignore this.

If your cohort are all males, set sigminer.sex to 'male' and sigminer.copynumber.max to a proper value (the best is consistent with read_copynumber).

If your cohort contains both males and females, set sigminer.sex as a data.frame with two columns "sample" and "sex". And set sigminer.copynumber.max to a proper value (the best is consistent with read_copynumber).

Methods (by class)

CopyNumber: Returns copy number features, components and component-by-sample matrix

MAF: Returns SBS mutation sample-by-component matrix and APOBEC enrichment

References

Macintyre, Geoff, et al. "Copy number signatures and mutational processes in ovarian carcinoma." Nature genetics 50.9 (2018): 1262.

Wang, Shixiang, et al. "Copy number signature analyses in prostate cancer reveal distinct etiologies and clinical outcomes." medRxiv (2020).

Mayakonda, Anand, et al. "Maftools: efficient and comprehensive analysis of somatic variants in cancer." Genome research 28.11 (2018): 1747-1756.

Roberts SA, Lawrence MS, Klimczak LJ, et al. An APOBEC Cytidine Deaminase Mutagenesis Pattern is Widespread in Human Cancers. Nature genetics. 2013;45(9):970-976. doi:10.1038/ng.2702.

Bergstrom EN, Huang MN, Mahto U, Barnes M, Stratton MR, Rozen SG, Alexandrov LB: SigProfilerMatrixGenerator: a tool for visualizing and exploring patterns of small mutational events. BMC Genomics 2019, 20:685 https://bmcgenomics.biomedcentral.com/articles/10.1186/s12864-019-6041-2

See also

Examples

# Load copy number object
load(system.file("extdata", "toy_copynumber.RData",
  package = "sigminer", mustWork = TRUE
))
# \donttest{
# Use method designed by Wang, Shixiang et al.
cn_tally_W <- sig_tally(cn, method = "W")

#> ℹ [2020-07-07 16:50:34]: Started.

#> ℹ [2020-07-07 16:50:34]: Step: getting copy number features.

#> ℹ [2020-07-07 16:50:34]: Getting breakpoint count per 10 Mb...

#> ℹ [2020-07-07 16:50:34]: Getting breakpoint count per chromosome arm...

#> ℹ [2020-07-07 16:50:34]: Getting copy number...

#> ℹ [2020-07-07 16:50:34]: Getting change-point copy number change...

#> ℹ [2020-07-07 16:50:35]: Getting length of chains of oscillating copy number...

#> ℹ [2020-07-07 16:50:35]: Getting (log10 based) segment size...

#> ℹ [2020-07-07 16:50:35]: Getting the minimal number of chromosome with 50% CNV...

#> ℹ [2020-07-07 16:50:35]: Getting burden of chromosome...

#> ✓ [2020-07-07 16:50:35]: Gotten.

#> ℹ [2020-07-07 16:50:35]: Step: generating copy number components.

#> ✓ [2020-07-07 16:50:35]: `feature_setting` checked.

#> ℹ [2020-07-07 16:50:35]: Step: counting components.

#> ✓ [2020-07-07 16:50:36]: Counted.

#> ℹ [2020-07-07 16:50:36]: Step: generating components by sample matrix.

#> ✓ [2020-07-07 16:50:36]: Matrix generated.

#> ℹ [2020-07-07 16:50:36]: 1.678 secs elapsed.

# Use method designed by Macintyre et al.
cn_tally_M <- sig_tally(cn, method = "M")

#> ℹ [2020-07-07 16:50:36]: Started.

#> ℹ [2020-07-07 16:50:36]: Step: getting copy number features.

#> ℹ [2020-07-07 16:50:36]: Getting segment size...

#> ℹ [2020-07-07 16:50:36]: Getting breakpoint count per 10 Mb...

#> ℹ [2020-07-07 16:50:36]: Getting length of chains of oscillating copy number...

#> ℹ [2020-07-07 16:50:36]: Getting breakpoint count per chromosome arm...

#> ℹ [2020-07-07 16:50:36]: Getting change-point copy number change...

#> ℹ [2020-07-07 16:50:36]: Getting copy number...

#> ✓ [2020-07-07 16:50:37]: Gotten.

#> ℹ [2020-07-07 16:50:37]: Step: fitting copy number components.

#> ℹ [2020-07-07 16:50:37]: Fitting feature: segsize.

#> 载入需要的程辑包：globals

#> 
#> 载入程辑包：‘globals’

#> The following object is masked from ‘package:R.utils’:
#> 
#>     cleanup

#> 载入需要的程辑包：future

#> 
#> 载入程辑包：‘future’

#> The following object is masked from ‘package:rtracklayer’:
#> 
#>     values

#> The following object is masked from ‘package:GenomicRanges’:
#> 
#>     values

#> The following object is masked from ‘package:IRanges’:
#> 
#>     values

#> The following object is masked from ‘package:S4Vectors’:
#> 
#>     values

#> The following object is masked from ‘package:NMF’:
#> 
#>     run

#> 载入需要的程辑包：lattice

#> 2 : *
#> 3 : *
#> 4 : *
#> 5 : *
#> 6 : *
#> 7 : *
#> 8 : *
#> 9 : *
#> 10 : *
#> 11 : *
#> 12 : *
#> 13 : *
#> 14 : *
#> 15 : *

#> ✓ [2020-07-07 16:50:45]: Select 10 according to BIC.

#> ✓ [2020-07-07 16:50:45]: Finally, select 10 after passing threshold 0.1.

#> ℹ [2020-07-07 16:50:45]: Fitting feature: bp10MB.

#> 2 : *
#> 3 : *
#> 4 : *
#> 5 : *
#> 6 : *
#> 7 : *
#> 8 : *
#> 9 : *
#> 10 : *
#> 11 : *
#> 12 : *
#> 13 : *
#> 14 : *
#> 15 : *

#> ✓ [2020-07-07 16:50:49]: Select 2 according to BIC.

#> ✓ [2020-07-07 16:50:49]: Finally, select 2 after passing threshold 0.1.

#> ℹ [2020-07-07 16:50:49]: Fitting feature: osCN.

#> 2 : *
#> 3 : *
#> 4 : *
#> 5 : *
#> 6 : *
#> 7 : *
#> 8 : *
#> 9 : *
#> 10 : *
#> 11 : *
#> 12 : *
#> 13 : *
#> 14 : *
#> 15 : *

#> ✓ [2020-07-07 16:50:51]: Select 2 according to BIC.

#> ✓ [2020-07-07 16:50:51]: Finally, select 2 after passing threshold 0.1.

#> ℹ [2020-07-07 16:50:51]: Fitting feature: bpchrarm.

#> 2 : *
#> 3 : *
#> 4 : *
#> 5 : *
#> 6 : *
#> 7 : *
#> 8 : *
#> 9 : *
#> 10 : *
#> 11 : *
#> 12 : *
#> 13 : *
#> 14 : *
#> 15 : *

#> ✓ [2020-07-07 16:50:54]: Select 2 according to BIC.

#> ✓ [2020-07-07 16:50:54]: Finally, select 2 after passing threshold 0.1.

#> ℹ [2020-07-07 16:50:54]: Fitting feature: changepoint.

#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     25 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     19 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     14 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     14 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     13 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     15 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     12 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     14 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     15 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     20 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     15 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     14 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     14 Log-likelihood: NaN
#> 2 : *
#> 3 : *
#> 4 : *
#> 5 : *
#> 6 : *
#> 7 : *
#> 8 : *
#> 9 : *
#> 10 : *
#> 11 : *
#> 12 : *
#> 13 : *
#> 14 : *
#> 15 : *

#> ℹ [2020-07-07 16:50:55]: Fitting feature: copynumber.

#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     12 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     10 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>      9 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     12 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     11 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     10 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     10 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     12 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     10 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     10 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     11 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     11 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     10 Log-likelihood: NaN
#> 2 : *
#> 3 : *
#> 4 : *
#> 5 : *
#> 6 : *
#> 7 : *
#> 8 : *
#> 9 : *
#> 10 : *
#> 11 : *
#> 12 : *
#> 13 : *
#> 14 : *
#> 15 : *

#> ✓ [2020-07-07 16:50:55]: Components fitted.

#> ℹ [2020-07-07 16:50:55]: Step: calculating the sum of posterior probabilities.

#> ℹ [2020-07-07 16:50:56]: Calculating the sum of posterior probabilities for bp10MB.

#> ℹ [2020-07-07 16:50:56]: Calculating the sum of posterior probabilities for copynumber.

#> ℹ [2020-07-07 16:50:56]: Calculating the sum of posterior probabilities for changepoint.

#> ℹ [2020-07-07 16:50:56]: Calculating the sum of posterior probabilities for bpchrarm.

#> ℹ [2020-07-07 16:50:56]: Calculating the sum of posterior probabilities for osCN.

#> ℹ [2020-07-07 16:50:56]: Calculating the sum of posterior probabilities for segsize.

#> ✓ [2020-07-07 16:50:56]: Matrix generated.

#> ℹ [2020-07-07 16:50:56]: 20.399 secs elapsed.

# }
# \donttest{
# Prepare SBS signature analysis
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read_maf(maf = laml.maf)

#> -Reading
#> -Validating
#> -Silent variants: 475 
#> -Summarizing
#> -Processing clinical data
#> --Missing clinical data
#> -Finished in 0.217s elapsed (0.209s cpu) 

if (require("BSgenome.Hsapiens.UCSC.hg19")) {
  mt_tally <- sig_tally(
    laml,
    ref_genome = "BSgenome.Hsapiens.UCSC.hg19",
    use_syn = TRUE
  )
  mt_tally$nmf_matrix[1:5, 1:5]

  ## Use strand bias categories
  mt_tally <- sig_tally(
    laml,
    ref_genome = "BSgenome.Hsapiens.UCSC.hg19",
    use_syn = TRUE, add_trans_bias = TRUE
  )
  ## Test it by enrichment analysis
  enrich_component_strand_bias(mt_tally$nmf_matrix)
  enrich_component_strand_bias(mt_tally$all_matrices$SBS_24)
} else {
  message("Please install package 'BSgenome.Hsapiens.UCSC.hg19' firstly!")
}

#> ℹ [2020-07-07 16:50:56]: Started.

#> ✓ [2020-07-07 16:50:56]: Reference genome loaded.

#> ✓ [2020-07-07 16:50:56]: Variants from MAF object queried.

#> ✓ [2020-07-07 16:50:56]: Chromosome names checked.

#> ✓ [2020-07-07 16:50:56]: Sex chromosomes properly handled.

#> ✓ [2020-07-07 16:50:56]: Variant start and end position checked.

#> ✓ [2020-07-07 16:50:56]: Variant data for matrix generation preprocessed.

#> ℹ [2020-07-07 16:50:56]: SBS matrix generation - start.

#> ℹ [2020-07-07 16:50:56]: Extracting 5' and 3' adjacent bases.

#> ℹ [2020-07-07 16:50:57]: Extracting +/- 20bp around mutated bases for background C>T estimation.

#> ℹ [2020-07-07 16:50:58]: Estimating APOBEC enrichment scores.

#> ℹ [2020-07-07 16:50:58]: Performing one-way Fisher's test for APOBEC enrichment.

#> ✓ [2020-07-07 16:50:58]: APOBEC related mutations are enriched in 3.297% of samples (APOBEC enrichment score > 2; 6 of 182 samples)

#> ℹ [2020-07-07 16:50:58]: Creating SBS sample-by-component matrices.

#> ✓ [2020-07-07 16:50:58]: SBS-6 matrix created.

#> ✓ [2020-07-07 16:50:58]: SBS-96 matrix created.

#> ✓ [2020-07-07 16:50:58]: SBS-1536 matrix created.

#> ℹ [2020-07-07 16:50:58]: Return SBS-96 as major matrix.

#> ✓ [2020-07-07 16:50:58]: Done.

#> ℹ [2020-07-07 16:50:58]: 1.65 secs elapsed.

#> ℹ [2020-07-07 16:50:58]: Started.

#> ✓ [2020-07-07 16:50:58]: Reference genome loaded.

#> ✓ [2020-07-07 16:50:58]: Variants from MAF object queried.

#> ✓ [2020-07-07 16:50:58]: Chromosome names checked.

#> ✓ [2020-07-07 16:50:58]: Sex chromosomes properly handled.

#> ✓ [2020-07-07 16:50:58]: Variant start and end position checked.

#> ✓ [2020-07-07 16:50:58]: Variant data for matrix generation preprocessed.

#> ℹ [2020-07-07 16:50:58]: SBS matrix generation - start.

#> ℹ [2020-07-07 16:50:58]: Extracting 5' and 3' adjacent bases.

#> ℹ [2020-07-07 16:50:58]: Extracting +/- 20bp around mutated bases for background C>T estimation.

#> ℹ [2020-07-07 16:50:59]: Estimating APOBEC enrichment scores.

#> ℹ [2020-07-07 16:50:59]: Performing one-way Fisher's test for APOBEC enrichment.

#> ✓ [2020-07-07 16:50:59]: APOBEC related mutations are enriched in 3.297% of samples (APOBEC enrichment score > 2; 6 of 182 samples)

#> ℹ [2020-07-07 16:50:59]: Creating SBS sample-by-component matrices.

#> ✓ [2020-07-07 16:50:59]: SBS-6 matrix created.

#> ✓ [2020-07-07 16:50:59]: SBS-96 matrix created.

#> ✓ [2020-07-07 16:50:59]: SBS-1536 matrix created.

#> ✓ [2020-07-07 16:50:59]: SBS-24 (6x4) matrix created.

#> ✓ [2020-07-07 16:50:59]: SBS-384 (96x4) matrix created.

#> ✓ [2020-07-07 16:50:59]: SBS-6144 (1536x4) matrix created.

#> ℹ [2020-07-07 16:50:59]: Return SBS-384 as major matrix.

#> ✓ [2020-07-07 16:50:59]: Done.

#> ℹ [2020-07-07 16:50:59]: 1.381 secs elapsed.

#>             sample component N_T N_U Total_T Total_U Enrichment    p_value
#>    1: TCGA-AB-2838       C>T   3   7      12       8  0.4285714 0.01976661
#>    2: TCGA-AB-3002       C>T   5  11      13      14  0.4545455 0.05423790
#>    3: TCGA-AB-2950       C>T   4   2       4       6  2.0000000 0.07619048
#>    4: TCGA-AB-2878       C>T   3   4       8       4  0.7500000 0.08080808
#>    5: TCGA-AB-2863       T>C   3   0       8       9        Inf 0.08235294
#>   ---                                                                     
#> 1154: TCGA-AB-3012       T>A   0   0       5       2        NaN 1.00000000
#> 1155: TCGA-AB-3012       C>G   0   0       5       2        NaN 1.00000000
#> 1156: TCGA-AB-2876       C>T   3   4       5       6  0.7500000 1.00000000
#> 1157: TCGA-AB-2920       C>T   6   2       7       2  3.0000000 1.00000000
#> 1158: TCGA-AB-2920       C>G   1   0       7       2        Inf 1.00000000
#>             fdr
#>    1: 0.1185997
#>    2: 0.3254274
#>    3: 0.4571429
#>    4: 0.4848485
#>    5: 0.4941176
#>   ---          
#> 1154: 1.0000000
#> 1155: 1.0000000
#> 1156: 1.0000000
#> 1157: 1.0000000
#> 1158: 1.0000000

# }