An Unified Interface to Extract Signatures

This function provides an unified interface to signature extractor implemented in sigminer. If you determine a specific approach, please also read the documentation of corresponding extractor.

"repeated_nmf" - sig_extract

"bayes_nmf" - sig_auto_extract

"bootstrap_nmf" - bp_extract_signatures

"sigprofiler" - sigprofiler

sig_unify_extract(
  nmf_matrix,
  range = 2:5,
  nrun = 10,
  approach = c("repeated_nmf", "bayes_nmf", "bootstrap_nmf", "sigprofiler"),
  cores = 1L,
  ...
)

Arguments

...
other parameters passing to signature extractor based on the approach setting. Check the "Description" part.

Value

Result dependent on the approach setting.

See also

Examples

# \donttest{
load(system.file("extdata", "toy_copynumber_tally_M.RData",
  package = "sigminer", mustWork = TRUE
))
# Extract copy number signatures
# It is same as sig_extract(cn_tally_M$nmf_matrix, 2, nrun = 1)
res <- sig_unify_extract(cn_tally_M$nmf_matrix, 2, nrun = 1)
# }