Extract Signatures with SigProfiler

This function provides an interface to software SigProfiler. More please see https://github.com/AlexandrovLab/SigProfilerExtractor. Typically, a reference genome is not required because the input is a matrix (my understanding). If you are using refitting result by SigProfiler, please make sure you have input the matrix same order as examples at https://github.com/AlexandrovLab/SigProfilerMatrixGenerator/tree/master/SigProfilerMatrixGenerator/references/matrix/BRCA_example. If not, use sigprofiler_reorder() firstly.

sigprofiler_extract(
  nmf_matrix,
  output,
  output_matrix_only = FALSE,
  range = 2:5,
  nrun = 10L,
  refit = FALSE,
  refit_plot = FALSE,
  is_exome = FALSE,
  init_method = c("random", "nndsvd_min", "nndsvd", "nndsvda", "nndsvdar"),
  cores = -1L,
  genome_build = c("hg19", "hg38", "T2T", "mm10", "mm9", "ce11"),
  use_conda = FALSE,
  py_path = NULL,
  sigprofiler_version = "1.1.3"
)

sigprofiler_import(
  output,
  order_by_expo = FALSE,
  type = c("suggest", "refit", "all")
)

sigprofiler_reorder(
  nmf_matrix,
  type = c("SBS96", "SBS6", "SBS12", "SBS192", "SBS1536", "SBS3072", "DBS78", "DBS312",
    "DBS1248", "DBS4992")
)

Arguments

nmf_matrix

a matrix used for NMF decomposition with rows indicate samples and columns indicate components.

output

output directory.

output_matrix_only

if TRUE, only generate matrix file for SigProfiler so user can call SigProfiler with the input by himself.

range

signature number range, i.e. 2:5.

nrun

the number of iteration to be performed to extract each signature number.

refit

if TRUE, then refit the denovo signatures with nnls. Same meaning as optimize option in sig_extract or sig_auto_extract.

refit_plot

if TRUE, SigProfiler will make denovo to COSMIC sigantures decompostion plots. However, this may fail due to some matrix cannot be identified by SigProfiler plot program.

is_exome

if TRUE, the exomes will be extracted.

init_method

the initialization algorithm for W and H matrix of NMF. Options are 'random', 'nndsvd', 'nndsvda', 'nndsvdar', 'alexandrov-lab-custom' and 'nndsvd_min'.

cores

number of cores used for computation.

genome_build

I think this option is useless when input is matrix, keep it in case it is useful.

use_conda

if TRUE, create an independent conda environment to run SigProfiler.

py_path

path to Python executable file, e.g. '/Users/wsx/anaconda3/bin/python'.

sigprofiler_version

version of SigProfilerExtractor. If this package is not installed, the specified package will be installed. If this package is installed, this option is useless.

order_by_expo

if TRUE, order the import signatures by their exposures, e.g. the signature contributed the most exposure in all samples will be named as Sig1.

type

mutational signature type.

Value

For sigprofiler_extract(), returns nothing. See output directory.

For sigprofiler_import(), a list containing Signature object.

A NMF matrix for input of sigprofiler_extract().

Examples

if (FALSE) {
  load(system.file("extdata", "toy_copynumber_tally_W.RData",
    package = "sigminer", mustWork = TRUE
  ))

  reticulate::conda_list()

  sigprofiler_extract(cn_tally_W$nmf_matrix, "~/test/test_sigminer",
    use_conda = TRUE
  )

  sigprofiler_extract(cn_tally_W$nmf_matrix, "~/test/test_sigminer",
    use_conda = FALSE, py_path = "/Users/wsx/anaconda3/bin/python"
  )
}

data("simulated_catalogs")
sigprofiler_reorder(t(simulated_catalogs$set1))
#> Downloading reference file...
#> Reordering...
#> Done
#>           A[C>A]A A[C>A]C A[C>A]G A[C>A]T A[C>G]A A[C>G]C A[C>G]G A[C>G]T
#> Sample_1      911     761      88     744     487     248      80     472
#> Sample_2      195     175      19     174      70      33       3      72
#> Sample_3       95      51      12      55      54      32       9      57
#> Sample_4      131      71      14      77      85      43      16      85
#> Sample_5       33      10       2      14       8       4       2      10
#> Sample_6       30      18       2      19      10       6       2      11
#> Sample_7       28       9       2      13       7       5       2       8
#> Sample_8       48      23       4      23      11       9       4      14
#> Sample_9       70      48       7      57      71      40      13      68
#> Sample_10     227     153      19     160      99      57      20     101
#> Sample_11      47      32       7      29      33      21       6      35
#> Sample_12    1077     520      94     584     531     291     127     538
#> Sample_13      37      21       3      24       9       4       1       9
#> Sample_14      26      16       4      16      20      11       3      20
#> Sample_15     119      57      13      69      70      42      13      73
#> Sample_16     214      76      15     100      58      34      18      63
#> Sample_17     184     140      32     115      57      46       5      63
#> Sample_18      93      81       7      87      26      15       2      26
#> Sample_19     653     255      60     310     173     119      46     193
#> Sample_20      59      20       5      26       8       7       3      10
#> Sample_21     288     250      23     245     161      81      27     160
#> Sample_22     600     431      76     400     474     279      99     477
#> Sample_23     127      43       7      62      17      14       7      22
#> Sample_24     779     435      76     467     511     281     108     517
#> Sample_25     221      61      12      96      15      19      11      26
#> Sample_26      56      52       6      49      17      12       1      20
#> Sample_27     158     109      15     115      57      29       7      58
#> Sample_28      86      29       6      39      22      15       7      24
#> Sample_29      55      45       5      46      18      11       2      18
#> Sample_30      42      36       4      37      12       7       1      12
#>           A[C>T]A A[C>T]C A[C>T]G A[C>T]T A[T>A]A A[T>A]C A[T>A]G A[T>A]T
#> Sample_1      675     282     349     586     381     365     358     534
#> Sample_2      248     103     389     171      85      91      79     147
#> Sample_3      194     103     372     113      41      45      44      64
#> Sample_4      109      54      48      76      52      47      49      56
#> Sample_5       13       5       9       9       4       3       4       4
#> Sample_6       25      12      42      18       7       7       7      11
#> Sample_7       20      10      35      13       5       5       6       7
#> Sample_8       48      24     220      25      10       6       9      16
#> Sample_9      104      52      73      74      31      32      33      41
#> Sample_10     160      77     123     140      79      68      90     115
#> Sample_11     103      55     251      62      26      23      30      38
#> Sample_12     632     322     284     490     305     263     301     324
#> Sample_13      19       8      21      16      11      10      10      16
#> Sample_14      51      25      39      31      17      16      18      20
#> Sample_15     138      75     186      96      50      42      60      64
#> Sample_16      73      37      42      61      39      31      39      39
#> Sample_17     560     270    1928     266      97     106      73     161
#> Sample_18      76      28     141      66      34      34      36      64
#> Sample_19     481     255    1004     320     181     146     201     222
#> Sample_20      33      17      78      21      12      10      13      14
#> Sample_21     205      85     186     187     112     103     109     171
#> Sample_22     869     466    1702     590     289     244     320     382
#> Sample_23      56      29      98      45      19      15      24      28
#> Sample_24     630     339     406     495     289     237     320     340
#> Sample_25     125      62     409      75      20      15      25      34
#> Sample_26     138      66     449      77      21      26      18      50
#> Sample_27     114      49      79      93      65      60      64      91
#> Sample_28      39      22      32      32      18      14      22      21
#> Sample_29      59      25     135      42      22      22      22      37
#> Sample_30      53      23      84      37      18      20      17      32
#>           A[T>C]A A[T>C]C A[T>C]G A[T>C]T A[T>G]A A[T>G]C A[T>G]G A[T>G]T
#> Sample_1     1085     704    1014     896     107      51     245     206
#> Sample_2      250     155     270     189      24      13      39      33
#> Sample_3      190      98     219     124      13      15      23      26
#> Sample_4      211     125     202     148      15      11      32      28
#> Sample_5        5       3       4       7       1       1       3       2
#> Sample_6       12       7      15      13       2       2       5       4
#> Sample_7       16       8      18      13       2       3       4       7
#> Sample_8       27      13      21      22       3       4       6       7
#> Sample_9       83      69      79      81       9      33      39     146
#> Sample_10     184      82     142     180      26      13      52      32
#> Sample_11     102      42      91      75      10       8      15      14
#> Sample_12    1073     659    1034     831      82      65     229     180
#> Sample_13      34      21      33      25       3       2       5       3
#> Sample_14      89      48      84      58       5       4       8       8
#> Sample_15     177      81     146     140      18      23      34      71
#> Sample_16     132      85     132     103       9       9      28      20
#> Sample_17     636     419     699     353      26      46      40     130
#> Sample_18      49      37      36      61      11      16      23      71
#> Sample_19     815     432     747     560      56      49      95      80
#> Sample_20      60      37      59      39       3       4       6      11
#> Sample_21     166      91     133     201      34      13      76      49
#> Sample_22     925     412     783     769     102      73     207     170
#> Sample_23      45      26      48      44       5      10      14      29
#> Sample_24     812     359     665     718      97      64     208     146
#> Sample_25      48      43      64      47       4      27      22      92
#> Sample_26      53      34      80      41       6      11      13      32
#> Sample_27     225     129     205     164      19       9      30      22
#> Sample_28      66      32      58      52       6       5      12       8
#> Sample_29      69      48      64      54       7      10      14      41
#> Sample_30      67      45      74      47       5       5      10      18
#>           C[C>A]A C[C>A]C C[C>A]G C[C>A]T C[C>G]A C[C>G]C C[C>G]G C[C>G]T
#> Sample_1      883     725      84     696     349     295      66     462
#> Sample_2      225     230      44     337      61      38       4      72
#> Sample_3      142     127      55     266      39      30      12      52
#> Sample_4      170      88      32     101      67      48      18      91
#> Sample_5       55      15      10      25       8       5       4      13
#> Sample_6       41      25       9      42       7       7       3      11
#> Sample_7       45      17      11      35       5       5       3       8
#> Sample_8       60      25      11      34       8       8       5      13
#> Sample_9       55      41       9      48      54      39      12      67
#> Sample_10     229     125      32     169      68      66      22      94
#> Sample_11      37      35      14      71      23      19       7      29
#> Sample_12    1464     645     265     795     370     332     141     554
#> Sample_13      53      27       9      35       8       6       2      11
#> Sample_14      28      22       8      30      15      11       3      18
#> Sample_15     128      48      29      89      56      42      18      71
#> Sample_16     336     114      63     159      40      41      25      69
#> Sample_17     247     298      82     420      34      27       4      34
#> Sample_18      79      63       4      70      23      16       2      24
#> Sample_19     925     334     188     494     115     127      71     177
#> Sample_20      94      32      19      45       5       7       6      10
#> Sample_21     255     216      23     255     123      95      23     164
#> Sample_22     453     366     100     551     319     285      94     426
#> Sample_23     200      69      42     127      11      16      13      22
#> Sample_24     847     386     158     574     376     308     119     520
#> Sample_25     370     119      79     221       9      17      22      27
#> Sample_26      68      98      26     187      13       8       1      15
#> Sample_27     184     112      26     128      47      35      10      59
#> Sample_28     127      38      25      64      14      17      10      24
#> Sample_29      50      42       5      41      14      11       2      16
#> Sample_30      49      50      10      71       8       8       1       9
#>           C[C>T]A C[C>T]C C[C>T]G C[C>T]T C[T>A]A C[T>A]C C[T>A]G C[T>A]T
#> Sample_1      556     312     245     783     290     550     427     683
#> Sample_2      175      95     352     209      61     117      87     138
#> Sample_3      136     105     365     121      19      40      43      42
#> Sample_4      135      78      55     124      28      59      52      63
#> Sample_5       26       9      15      14       2       4       5       5
#> Sample_6       18      10      42      18       5      10       9      13
#> Sample_7       19      12      37      13       2       6       7      10
#> Sample_8       41      31     130      33       4       8       9      10
#> Sample_9      148      56      84      98      21      75      54     173
#> Sample_10     140      92      92     158      59      88     100     128
#> Sample_11      90      67     176      75      13      18      27      26
#> Sample_12     738     437     371     711     182     380     352     431
#> Sample_13      19      10      20      22       7      14      11      16
#> Sample_14      65      35      39      55       8      17      15      17
#> Sample_15     169     103     142     124      25      49      62     100
#> Sample_16     100      56      69      89      19      48      46      52
#> Sample_17     481     315    1331     433      28     137      57     138
#> Sample_18      69      26      90      72      28      63      47     124
#> Sample_19     557     380     712     487      74     160     182     178
#> Sample_20      43      26      59      37       4      14      12      17
#> Sample_21     116      77     127     183      96     149     138     207
#> Sample_22     736     566    1120     692     185     271     335     380
#> Sample_23      58      34     108      45       9      26      29      47
#> Sample_24     654     448     358     583     184     272     352     384
#> Sample_25     151      74     341      90       4      43      37     101
#> Sample_26      65      53     355      67      14      34      23      51
#> Sample_27     115      68      62     135      40      75      65      90
#> Sample_28      49      33      37      41       9      16      23      21
#> Sample_29      58      29      85      59      15      39      26      67
#> Sample_30      36      21      78      47      12      30      20      40
#>           C[T>C]A C[T>C]C C[T>C]G C[T>C]T C[T>G]A C[T>G]C C[T>G]G C[T>G]T
#> Sample_1      643     944     828     875      79     197     310     498
#> Sample_2      144     185     243     164      16      38      60      71
#> Sample_3      100      97     198     107       8      32      48      75
#> Sample_4      121     151     159     140      10      31      47      55
#> Sample_5        2       5       3       5       0       2       5       7
#> Sample_6        6       9      14       8       1       4       8       7
#> Sample_7        8      13      19      25       1       6       8      50
#> Sample_8        9      14      15      15       1       6       9       8
#> Sample_9       65     208     160     468       7      98      85    1112
#> Sample_10      74      95      93     112      16      28      62      67
#> Sample_11      42      37      68      53       5      13      22      36
#> Sample_12     621     835     816     744      62     200     327     323
#> Sample_13      19      26      27      23       2       5       8       7
#> Sample_14      49      53      67      55       4      10      14      21
#> Sample_15      80     128     139     259      10      52      60     488
#> Sample_16      78     111     107      92       7      27      45      38
#> Sample_17     399     531     671     594      17     124     110     617
#> Sample_18      29     108      74     234       6      46      43     541
#> Sample_19     413     480     571     491      33      99     157     198
#> Sample_20      35      50      53      56       2      11      14      58
#> Sample_21      73     126      95     121      22      40      85      76
#> Sample_22     395     429     554     540      62     147     264     346
#> Sample_23      22      51      58      96       3      24      31     208
#> Sample_24     345     397     444     483      60     135     267     318
#> Sample_25      34     130     127     291       1      68      65     721
#> Sample_26      30      51     101      88       3      27      30     185
#> Sample_27     121     154     160     146      12      25      42      49
#> Sample_28      30      36      43      39       3      10      19      21
#> Sample_29      43      91      76     152       4      29      27     289
#> Sample_30      43      65      75      80       4      16      18     107
#>           G[C>A]A G[C>A]C G[C>A]G G[C>A]T G[C>G]A G[C>G]C G[C>G]G G[C>G]T
#> Sample_1      548     445      47     493     232     192      27     337
#> Sample_2      133     110      16     132      27      30       2      47
#> Sample_3      137      67      24     114      23      31       4      33
#> Sample_4      169      66      19     112      39      33       6      55
#> Sample_5       74      12       6      43       4       3       1       6
#> Sample_6       44      14       4      31       5       5       1       8
#> Sample_7       61      12       6      38       3       3       1       5
#> Sample_8       80      19       6      47       5       7       1       8
#> Sample_9       61      39      11      50      36      28       5      52
#> Sample_10     241     103      22     177      47      41       8      64
#> Sample_11      37      31       9      34      15      17       3      18
#> Sample_12    1651     476     150    1048     257     217      40     362
#> Sample_13      57      16       5      36       4       3       0       6
#> Sample_14      20      18       5      16       9       8       1      10
#> Sample_15     175      63      23     115      33      29       6      41
#> Sample_16     430      82      34     252      28      23       5      40
#> Sample_17     157     137      34     114      20      47       0      29
#> Sample_18      71      42       5      57      12       9       1      19
#> Sample_19    1201     290     115     704      80      81      16      90
#> Sample_20     125      24      11      70       4       4       1       4
#> Sample_21     169     137      12     171      76      64      10     116
#> Sample_22     397     343      72     378     227     222      37     301
#> Sample_23     276      49      23     164       8       9       2      12
#> Sample_24     966     379     108     688     243     211      41     331
#> Sample_25     534      81      43     305       8      10       2      14
#> Sample_26      44      39       8      51       6      14       0      14
#> Sample_27     170      77      17     117      25      21       3      33
#> Sample_28     172      35      15     102      11      10       2      13
#> Sample_29      42      27       4      31       9       8       1      12
#> Sample_30      31      24       4      28       5       6       0       8
#>           G[C>T]A G[C>T]C G[C>T]G G[C>T]T G[T>A]A G[T>A]C G[T>A]G G[T>A]T
#> Sample_1      439     275     221     403     192     243     263     407
#> Sample_2      290     298     496     248      42      58      49      96
#> Sample_3      309     340     518     258      24      28      27      29
#> Sample_4      108      82      53      90      25      29      33      36
#> Sample_5       13       8      11       7       2       2       2       4
#> Sample_6       32      32      54      25       4       5       5       9
#> Sample_7       27      29      47      22       3       3       4       5
#> Sample_8       41      40     136      29       4       4       5       8
#> Sample_9       72      48      63      54      17      27      39      39
#> Sample_10     109      84      88     104      52      44      60      79
#> Sample_11     118     120     212      99      17      12      18      15
#> Sample_12     590     424     276     474     155     183     215     250
#> Sample_13      16      16      25      15       5       6       6      11
#> Sample_14      55      46      41      46       9       9      10       9
#> Sample_15     119     108     145     106      33      25      39      34
#> Sample_16      68      52      50      53      19      23      24      34
#> Sample_17     681     714    1573     537      21      64      39      67
#> Sample_18      24      18      90      26      19      24      29      50
#> Sample_19     440     415     707     387      98      87     102     114
#> Sample_20      29      27      55      24       5       7       6       9
#> Sample_21     145     115     173     130      64      68      84     130
#> Sample_22     856     783    1260     718     182     146     224     211
#> Sample_23      62      69     123      53      12      13      15      22
#> Sample_24     575     462     361     497     186     147     222     216
#> Sample_25     131     144     364      97       8      17      18      30
#> Sample_26     202     234     479     163       9      18      14      29
#> Sample_27      81      67      69      82      34      36      37      58
#> Sample_28      35      32      31      32      12       9      12      13
#> Sample_29      33      28      84      30      10      15      16      27
#> Sample_30      59      63     106      53       9      14      12      22
#>           G[T>C]A G[T>C]C G[T>C]G G[T>C]T G[T>G]A G[T>G]C G[T>G]G G[T>G]T
#> Sample_1     1015     567     759     820      70      82     326     270
#> Sample_2      251     132     192     185       5       9      34      46
#> Sample_3      158      83     133     112       6       7      30      28
#> Sample_4      196     104     151     154      11      10      50      29
#> Sample_5        2       3       5       5       1       1       8       4
#> Sample_6        8       6       9       7       2       2       9       5
#> Sample_7       10       8      11      10       1       2       7      14
#> Sample_8       18      13      12      20       2       2      12       5
#> Sample_9       36      65      50      69      10      47      59     280
#> Sample_10      80      57      75      91      16       9      70      39
#> Sample_11      57      33      46      50       4       2      16      11
#> Sample_12     958     539     777     793      91      79     415     205
#> Sample_13      34      18      27      27       1       1       7       5
#> Sample_14      78      39      59      60       2       1       8       8
#> Sample_15      84      67      81      96       9      17      44     123
#> Sample_16     128      73     111     110      12       9      62      28
#> Sample_17     739     386     526     539       1      31      35     175
#> Sample_18      19      35      24      41       2      20      24     141
#> Sample_19     677     365     538     567      27       8     149      85
#> Sample_20      61      34      49      50       2       2      12      18
#> Sample_21      80      64      70      89      23      23     104      61
#> Sample_22     485     307     393     459      72      55     285     148
#> Sample_23      23      24      34      33       4       8      30      58
#> Sample_24     386     259     359     400      78      52     324     150
#> Sample_25      32      51      55      61       6      26      54     186
#> Sample_26      51      34      44      37       1      10      13      53
#> Sample_27     203     106     154     162       6       3      32      28
#> Sample_28      43      26      40      41       4       1      22      10
#> Sample_29      62      44      49      59       2      11      16      77
#> Sample_30      71      39      55      54       1       5       9      32
#>           T[C>A]A T[C>A]C T[C>A]G T[C>A]T T[C>G]A T[C>G]C T[C>G]G T[C>G]T
#> Sample_1      621     617      71     813     310     367      35     555
#> Sample_2      424     224      33     332    1479     329      51    1822
#> Sample_3      223     129      34     206     710     193      29     847
#> Sample_4      296     175      36     244    1051     293      42    1266
#> Sample_5      103      50      13     100     326      73      13     409
#> Sample_6       35      24       6      53      23       9       2      42
#> Sample_7       54      30       9      67     114      32       5     136
#> Sample_8       58      43      12      88      17      13       2      32
#> Sample_9       83      65      18      89     235      97      13     318
#> Sample_10     221     161      37     314      95      75      12     180
#> Sample_11      92      57      15      82     240      68      12     307
#> Sample_12    1269     979     239    1748    1504     759     105    2013
#> Sample_13      73      40       9      73     207      52       7     245
#> Sample_14      80      45      10      55     279      64      11     357
#> Sample_15     289     154      42     257     929     253      39    1117
#> Sample_16     278     192      56     421     234     111      16     302
#> Sample_17     216     226      52     244     126      47       6     200
#> Sample_18     103      61      11     118     190      50       6     236
#> Sample_19     903     620     188    1252     817     343      52    1003
#> Sample_20      73      53      17     118       5      12       2       8
#> Sample_21     329     218      24     327     869     286      38    1091
#> Sample_22     498     482     107     646     220     362      52     465
#> Sample_23     155     101      35     266       7      23       4      17
#> Sample_24    1312     812     188    1329    3550    1145     170    4391
#> Sample_25     302     193      71     515      51      36       8      98
#> Sample_26      77      50      10      86     183      47       6     220
#> Sample_27     284     163      31     255     837     219      31     992
#> Sample_28     103      72      24     168      10      24       4      20
#> Sample_29      43      37       7      58      11      12       1      19
#> Sample_30      33      28       5      45       6       6       0       8
#>           T[C>T]A T[C>T]C T[C>T]G T[C>T]T T[T>A]A T[T>A]C T[T>A]G T[T>A]T
#> Sample_1      435     394     129     454     385     483     257     714
#> Sample_2     2045     415     339     997      88     101      50     156
#> Sample_3      368     141     181     136      40      42      28      47
#> Sample_4      494     147      51     168      49      68      36      71
#> Sample_5      525     101      59     251       4       4       2       4
#> Sample_6      199      47      40     112       7       7       5      12
#> Sample_7       60      22      23      22       6       4       3       6
#> Sample_8      130      53     113      79      12       7       5      12
#> Sample_9      179     133      87     137      29      37      22      57
#> Sample_10     640     219     117     395      85      70      57     130
#> Sample_11     391     131     144     207      27      19      18      28
#> Sample_12    1027     643     237     533     298     389     222     452
#> Sample_13      99      22      15      33      11      14       6      18
#> Sample_14     592     138      74     302      16      22      12      21
#> Sample_15     468     188     129     190      51      41      36      58
#> Sample_16     170      89      47      75      41      51      27      55
#> Sample_17    1286     496     978     788     107     160      47     140
#> Sample_18     204      76      92     135      39      34      20      73
#> Sample_19     704     480     523     369     201     205     123     219
#> Sample_20      37      32      44      25      14      17       8      16
#> Sample_21     402     137      84     193     113     106      75     207
#> Sample_22     413     661     753     434     293     242     213     384
#> Sample_23      56      53      65      41      23      17      13      27
#> Sample_24    1721     742     289     690     281     249     216     376
#> Sample_25     583     205     293     345      28      21      10      28
#> Sample_26     124      60     204      69      23      19      10      37
#> Sample_27     412     117      52     155      67      78      42     105
#> Sample_28      38      43      24      26      20      18      14      22
#> Sample_29      46      43      70      51      25      28      13      44
#> Sample_30      31      24      38      30      20      25      11      36
#>           T[T>C]A T[T>C]C T[T>C]G T[T>C]T T[T>G]A T[T>G]C T[T>G]G T[T>G]T
#> Sample_1      928     744     583     834     126     166     252     405
#> Sample_2      203     149     143     174      19      27      34      69
#> Sample_3      130      98     111     106      23      28      24      65
#> Sample_4      180     140     118     148      25      30      33      57
#> Sample_5        2       6       2       3       1       3       2       8
#> Sample_6        6       8       6       7       2       4       5      12
#> Sample_7        8      10       9       9       2       5       3      15
#> Sample_8       15      14      14      14       4       6       7      14
#> Sample_9       47      85      53      73      18      58      48     296
#> Sample_10      84      87      58     102      30      36      53      87
#> Sample_11      52      40      44      49      15      15      17      34
#> Sample_12     903     771     606     776     148     202     239     349
#> Sample_13      29      23      19      25       2       4       4       8
#> Sample_14      70      51      48      58       8       9       8      18
#> Sample_15      87      91      73      96      26      43      37     151
#> Sample_16     117     108      80      99      15      27      25      52
#> Sample_17     610     428     460     460      36      66      43     204
#> Sample_18      19      39      20      41       5      26      24     147
#> Sample_19     611     486     431     510      76     105      85     187
#> Sample_20      53      44      38      43       4       9       5      23
#> Sample_21      80      89      50     109      37      44      81     124
#> Sample_22     486     422     381     492     164     172     242     457
#> Sample_23      20      35      24      27       5      19      12      58
#> Sample_24     417     405     300     442     157     176     230     396
#> Sample_25      24      65      46      37       4      42      19     146
#> Sample_26      31      27      42      29       6      15      14      68
#> Sample_27     181     136     114     155      19      24      26      48
#> Sample_28      41      39      30      38       8      12      11      25
#> Sample_29      55      51      41      54       5      17      15      64
#> Sample_30      59      46      43      51       4      10       9      37