Get Sample Groups from Signature Decomposition Information

One of key results from signature analysis is to cluster samples into different groups. This function takes Signature object as input and return the membership in each cluster.

get_groups(
  Signature,
  method = c("exposure", "consensus", "samples"),
  match_consensus = FALSE
)

Arguments

Signature
a Signature object obtained either from sig_extract or sig_auto_extract.

method
grouping method, more see details, could be one of the following:

'exposure' - the default method, which assigns a sample into a group whose signature exposure is dominant.

'consensus' - returns the cluster membership based on the hierarchical clustering of the consensus matrix, it can only be used for the result obtained by sig_extract() with multiple runs using NMF package.

'samples' - returns the cluster membership based on the contribution of signature to each sample, it can only be used for the result obtained by sig_extract() using NMF package.

match_consensus
only used when the method is 'consensus'. If TRUE, the result will match order as shown in consensus map.

Value

a data.table object

Details

Users may find there are bigger differences between using method 'samples' and 'exposure' but they use a similar idear to find dominant signature, here goes the reason:

Method 'samples' using data directly from NMF decomposition, this means the two matrix W (basis matrix or signature matrix) and H (coefficient matrix or exposure matrix) are the results of NMF. For method 'exposure', it uses the signature exposure loading matrix. In this situation, each signture represents a number of mutations (alterations) about implementation please see source code of sig_extract() function.

See also

Examples

# \donttest{
# Load copy number prepare object
load(system.file("extdata", "toy_copynumber_prepare.RData",
  package = "sigminer", mustWork = TRUE
))
# Extract copy number signatures
library(NMF)
sig <- sig_extract(cn_prepare$nmf_matrix, 2, nrun = 10)

# Methods 'consensus' and 'samples' are from NMF::predict()
get_groups(sig)

#>                           sample group    weight enrich_sig
#>  1: TCGA-05-4417-01A-22D-1854-01     1 0.9512572       Sig1
#>  2: TCGA-A8-A07S-01A-11D-A036-01     2 0.9638279       Sig2
#>  3: TCGA-B6-A0X5-01A-21D-A107-01     2 0.9839765       Sig2
#>  4: TCGA-A5-A0G2-01A-11D-A042-01     2 0.7475197       Sig2
#>  5: TCGA-DF-A2KN-01A-11D-A17U-01     2 0.9603335       Sig2
#>  6: TCGA-06-0644-01A-02D-0310-01     2 0.8005602       Sig2
#>  7: TCGA-19-2621-01B-01D-0911-01     2 0.5913322       Sig2
#>  8: TCGA-26-6174-01A-21D-1842-01     2 0.9033573       Sig2
#>  9: TCGA-CV-7432-01A-11D-2128-01     2 0.5650982       Sig2
#> 10: TCGA-99-7458-01A-11D-2035-01     2 0.5241406       Sig2

get_groups(sig, method = "consensus", match_consensus = TRUE)

#>                           sample group silhouette_width enrich_sig
#>  1: TCGA-05-4417-01A-22D-1854-01     1            1.000       Sig1
#>  2: TCGA-99-7458-01A-11D-2035-01     1            1.000       Sig1
#>  3: TCGA-19-2621-01B-01D-0911-01     1            1.000       Sig1
#>  4: TCGA-CV-7432-01A-11D-2128-01     1            1.000       Sig1
#>  5: TCGA-26-6174-01A-21D-1842-01     2            1.000       Sig2
#>  6: TCGA-06-0644-01A-02D-0310-01     2            1.000       Sig2
#>  7: TCGA-DF-A2KN-01A-11D-A17U-01     2            0.810       Sig2
#>  8: TCGA-A5-A0G2-01A-11D-A042-01     2            0.810       Sig2
#>  9: TCGA-A8-A07S-01A-11D-A036-01     2            0.733       Sig2
#> 10: TCGA-B6-A0X5-01A-21D-A107-01     2            0.833       Sig2

get_groups(sig, method = "samples")

#>                           sample group silhouette_width  prob enrich_sig
#>  1: TCGA-A8-A07S-01A-11D-A036-01     1                1 0.946       Sig2
#>  2: TCGA-B6-A0X5-01A-21D-A107-01     1                1 0.976       Sig2
#>  3: TCGA-A5-A0G2-01A-11D-A042-01     1                1 0.659       Sig2
#>  4: TCGA-DF-A2KN-01A-11D-A17U-01     1                1 0.940       Sig2
#>  5: TCGA-06-0644-01A-02D-0310-01     1                1 0.724       Sig2
#>  6: TCGA-19-2621-01B-01D-0911-01     2                1 0.514       Sig1
#>  7: TCGA-26-6174-01A-21D-1842-01     1                1 0.859       Sig2
#>  8: TCGA-CV-7432-01A-11D-2128-01     2                1 0.541       Sig1
#>  9: TCGA-05-4417-01A-22D-1854-01     2                1 0.968       Sig1
#> 10: TCGA-99-7458-01A-11D-2035-01     2                1 0.582       Sig1

# }