Get Signature Exposure from 'Signature' Object

The expected number of mutations (or copy number segment records) with each signature was determined after a scaling transformation V ~ WH = W'H' where W' = WU' and H' = UH. The scaling matrix U is a KxK diagnal matrix (K is signature number, U' is the inverse of U) with the element corresponding to the L1-norm of column vectors of W (ie. the sum of the elements of the vector). As a result, the k-th row vector of the final matrix H' represents the absolute exposure (activity) of the k-th process across samples (e.g., for SBS, the estimated (or expected) number of mutations generated by the k-th process). Of note, for copy number signatures, only components of feature CN was used for calculating H'.

get_sig_exposure(
  Signature,
  type = c("absolute", "relative"),
  rel_threshold = 0.01
)

Arguments

Signature

a Signature object obtained either from sig_extract or sig_auto_extract, or just a raw exposure matrix with column representing samples (patients) and row representing signatures.

type

'absolute' for signature exposure and 'relative' for signature relative exposure.

rel_threshold

only used when type is 'relative', relative exposure less than (<=) this value will be set to 0 and thus all signature exposures may not sum to 1. This is similar to this argument in sig_fit.

Value

a data.table

References

Kim, Jaegil, et al. "Somatic ERCC2 mutations are associated with a distinct genomic signature in urothelial tumors." Nature genetics 48.6 (2016): 600.

Author

Shixiang Wang w_shixiang@163.com

Examples

# Load mutational signature
load(system.file("extdata", "toy_mutational_signature.RData",
  package = "sigminer", mustWork = TRUE
))
# Get signature exposure
expo1 <- get_sig_exposure(sig2)
expo1
#>            sample      Sig1      Sig2      Sig3
#>   1: TCGA-AB-2802 0.0000000  7.602352 1.2566069
#>   2: TCGA-AB-2803 0.0000000 12.007936 0.7412961
#>   3: TCGA-AB-2804 0.0000000  4.838900 0.0000000
#>   4: TCGA-AB-2805 0.2000784 12.007936 0.0000000
#>   5: TCGA-AB-2806 0.2923331 10.273787 0.0000000
#>  ---                                           
#> 184: TCGA-AB-3007 0.0000000  6.691893 0.0000000
#> 185: TCGA-AB-3008 0.2000784  2.940509 0.0000000
#> 186: TCGA-AB-3009 0.0000000 25.388449 2.4198934
#> 187: TCGA-AB-3011 0.0000000  4.838900 0.0000000
#> 188: TCGA-AB-3012 0.0000000  5.770868 0.0000000
expo2 <- get_sig_exposure(sig2, type = "relative")
expo2
#>            sample       Sig1      Sig2       Sig3
#>   1: TCGA-AB-2802 0.00000000 0.8581541 0.14184589
#>   2: TCGA-AB-2803 0.00000000 0.9418556 0.05814437
#>   3: TCGA-AB-2804 0.00000000 1.0000000 0.00000000
#>   4: TCGA-AB-2805 0.01638910 0.9836109 0.00000000
#>   5: TCGA-AB-2806 0.02766703 0.9723330 0.00000000
#>  ---                                             
#> 182: TCGA-AB-3007 0.00000000 1.0000000 0.00000000
#> 183: TCGA-AB-3008 0.06370731 0.9362927 0.00000000
#> 184: TCGA-AB-3009 0.00000000 0.9129796 0.08702041
#> 185: TCGA-AB-3011 0.00000000 1.0000000 0.00000000
#> 186: TCGA-AB-3012 0.00000000 1.0000000 0.00000000