Calculate Similarity between Identified Signatures and Reference Signatures

The reference signatures can be either a Signature object specified by Ref argument or known COSMIC signatures specified by sig_db argument. Two COSMIC databases are used for comparisons - "legacy" which includes 30 signaures, and "SBS" - which includes updated/refined 65 signatures. This function is modified from compareSignatures() in maftools package.

get_sig_similarity(
  Signature,
  Ref = NULL,
  sig_db = "legacy",
  db_type = c("", "human-exome", "human-genome"),
  method = "cosine",
  normalize = c("row", "feature"),
  feature_setting = sigminer::CN.features,
  pattern_to_rm = NULL,
  verbose = TRUE
)

Arguments

Signature
a Signature object or a component-by-signature matrix (sum of each column is 1). More please see examples.

Ref
default is NULL, can be a same object as Signature.

sig_db
can be 'legacy' (for COSMIC v2 'SBS'), 'SBS', 'DBS' and 'ID'. Default 'legacy'.

db_type
only used when sig_db is enabled. "" for keeping default, "human-exome" for transforming to exome frequency of component, and "human-genome" for transforming to whole genome frequency of component.

method
default is 'cosine' for cosine similarity.

normalize
one of "row" and "feature". "row" is typically used for common mutational signatures. "feature" is designed by me to use when input are copy number signatures.

feature_setting
a data.frame used for classification. Only used when method is "Wang" ("W"). Default is CN.features. Users can also set custom input with "feature", "min" and "max" columns available. Valid features can be printed by unique(CN.features$feature).

pattern_to_rm
patterns for removing some features/components in similarity calculation. A vector of component name is also accepted. The remove operation will be done after normalization. Default is NULL.

verbose
if TRUE, print extra info.

Value

a list containing smilarities, aetiologies if available, and best match.

Examples

# Load mutational signature
load(system.file("extdata", "toy_mutational_signature.RData",
  package = "sigminer", mustWork = TRUE
))

s1 <- get_sig_similarity(sig2, Ref = sig2)

#> -Comparing against Custom signatures

#> ------------------------------------

#> --Found Sig1 most similar to Sig1 [similarity: 1]

#> --Found Sig2 most similar to Sig2 [similarity: 1]

#> --Found Sig3 most similar to Sig3 [similarity: 1]

#> ------------------------------------

#> Return result invisiblely.

s1

#> $similarity
#>      Sig1 Sig2 Sig3
#> Sig1    1    0    0
#> Sig2    0    1    0
#> Sig3    0    0    1
#> 
#> $aetiology_db
#> [1] NA
#> 
#> $best_match
#> $best_match$Sig1
#> $best_match$Sig1$aetiology
#> [1] NA
#> 
#> $best_match$Sig1$best_match
#> [1] "Best match: Sig1 [similarity: 1]"
#> 
#> 
#> $best_match$Sig2
#> $best_match$Sig2$aetiology
#> [1] NA
#> 
#> $best_match$Sig2$best_match
#> [1] "Best match: Sig2 [similarity: 1]"
#> 
#> 
#> $best_match$Sig3
#> $best_match$Sig3$aetiology
#> [1] NA
#> 
#> $best_match$Sig3$best_match
#> [1] "Best match: Sig3 [similarity: 1]"
#> 
#> 
#> 
#> attr(,"class")
#> [1] "similarity" "list"      

s2 <- get_sig_similarity(sig2)

#> -Comparing against COSMIC signatures

#> ------------------------------------

#> --Found Sig1 most similar to COSMIC_28

#>    Aetiology: Unknown [similarity: 0.393]

#> --Found Sig2 most similar to COSMIC_1

#>    Aetiology: spontaneous deamination of 5-methylcytosine [similarity: 0.913]

#> --Found Sig3 most similar to COSMIC_10

#>    Aetiology: defects in polymerase POLE [similarity: 0.531]

#> ------------------------------------

#> Return result invisiblely.

s2

#> $similarity
#>      COSMIC_1 COSMIC_2 COSMIC_3 COSMIC_4 COSMIC_5 COSMIC_6 COSMIC_7 COSMIC_8
#> Sig1    0.018    0.001    0.221    0.097    0.146    0.029    0.004    0.113
#> Sig2    0.913    0.164    0.334    0.287    0.525    0.833    0.261    0.409
#> Sig3    0.327    0.097    0.094    0.060    0.168    0.243    0.252    0.093
#>      COSMIC_9 COSMIC_10 COSMIC_11 COSMIC_12 COSMIC_13 COSMIC_14 COSMIC_15
#> Sig1    0.333     0.070     0.002     0.035     0.008     0.015     0.021
#> Sig2    0.295     0.160     0.358     0.250     0.076     0.609     0.607
#> Sig3    0.099     0.531     0.045     0.113     0.025     0.103     0.109
#>      COSMIC_16 COSMIC_17 COSMIC_18 COSMIC_19 COSMIC_20 COSMIC_21 COSMIC_22
#> Sig1     0.131     0.342     0.041     0.019     0.015     0.012     0.020
#> Sig2     0.319     0.120     0.239     0.566     0.461     0.264     0.083
#> Sig3     0.107     0.078     0.081     0.095     0.250     0.117     0.040
#>      COSMIC_23 COSMIC_24 COSMIC_25 COSMIC_26 COSMIC_27 COSMIC_28 COSMIC_29
#> Sig1     0.001     0.105     0.107     0.027     0.004     0.393     0.042
#> Sig2     0.356     0.236     0.393     0.335     0.081     0.065     0.382
#> Sig3     0.045     0.079     0.159     0.146     0.031     0.042     0.121
#>      COSMIC_30
#> Sig1     0.050
#> Sig2     0.463
#> Sig3     0.135
#> 
#> $aetiology_db
#> $aetiology_db[[1]]
#>  [1] "spontaneous deamination of 5-methylcytosine"
#>  [2] "APOBEC Cytidine Deaminase (C>T)"            
#>  [3] "defects in DNA-DSB repair by HR"            
#>  [4] "exposure to tobacco (smoking) mutagens"     
#>  [5] "Unknown"                                    
#>  [6] "defective DNA mismatch repair"              
#>  [7] "UV exposure"                                
#>  [8] "Unknown"                                    
#>  [9] "defects in polymerase-eta"                  
#> [10] "defects in polymerase POLE"                 
#> [11] "exposure to alkylating agents"              
#> [12] "Unknown"                                    
#> [13] "APOBEC Cytidine Deaminase (C>G)"            
#> [14] "Unknown"                                    
#> [15] "defective DNA mismatch repair"              
#> [16] "Unknown"                                    
#> [17] "Unknown"                                    
#> [18] "Unknown"                                    
#> [19] "Unknown"                                    
#> [20] "defective DNA mismatch repair"              
#> [21] "unknown"                                    
#> [22] "exposure to aristolochic acid"              
#> [23] "Unknown"                                    
#> [24] "exposures to aflatoxin"                     
#> [25] "Unknown"                                    
#> [26] "defective DNA mismatch repair"              
#> [27] "Unknown"                                    
#> [28] "Unknown"                                    
#> [29] "exposure to tobacco (chewing) mutagens"     
#> [30] "Unknown"                                    
#> 
#> 
#> $best_match
#> $best_match$Sig1
#> $best_match$Sig1$aetiology
#> [1] "Unknown"
#> 
#> $best_match$Sig1$best_match
#> [1] "Best match: COSMIC_28 [similarity: 0.393]"
#> 
#> 
#> $best_match$Sig2
#> $best_match$Sig2$aetiology
#> [1] "spontaneous deamination of 5-methylcytosine"
#> 
#> $best_match$Sig2$best_match
#> [1] "Best match: COSMIC_1 [similarity: 0.913]"
#> 
#> 
#> $best_match$Sig3
#> $best_match$Sig3$aetiology
#> [1] "defects in polymerase POLE"
#> 
#> $best_match$Sig3$best_match
#> [1] "Best match: COSMIC_10 [similarity: 0.531]"
#> 
#> 
#> 
#> attr(,"class")
#> [1] "similarity" "list"      

s3 <- get_sig_similarity(sig2, sig_db = "SBS")

#> -Comparing against COSMIC signatures

#> ------------------------------------

#> --Found Sig1 most similar to SBS28

#>    Aetiology: Unknown [similarity: 0.383]

#> --Found Sig2 most similar to SBS1

#>    Aetiology: spontaneous or enzymatic deamination of 5-methylcytosine [similarity: 0.868]

#> --Found Sig3 most similar to SBS10b

#>    Aetiology: Polymerase epsilon exonuclease domain mutations [similarity: 0.865]

#> ------------------------------------

#> Return result invisiblely.

s3

#> $similarity
#>       SBS1  SBS2  SBS3  SBS4  SBS5  SBS6 SBS7a SBS7b SBS7c SBS7d  SBS8  SBS9
#> Sig1 0.001 0.000 0.198 0.062 0.117 0.003 0.001 0.004 0.070 0.011 0.094 0.350
#> Sig2 0.868 0.141 0.340 0.219 0.492 0.836 0.220 0.281 0.095 0.111 0.219 0.209
#> Sig3 0.218 0.067 0.090 0.042 0.208 0.232 0.159 0.158 0.011 0.016 0.043 0.079
#>      SBS10a SBS10b SBS11 SBS12 SBS13 SBS14 SBS15 SBS16 SBS17a SBS17b SBS18
#> Sig1  0.011  0.070 0.001 0.011 0.007 0.004 0.010 0.046  0.011  0.373 0.043
#> Sig2  0.096  0.080 0.316 0.154 0.048 0.187 0.674 0.176  0.061  0.008 0.273
#> Sig3  0.019  0.865 0.014 0.075 0.021 0.033 0.118 0.033  0.040  0.005 0.096
#>      SBS19 SBS20 SBS21 SBS22 SBS23 SBS24 SBS25 SBS26 SBS27 SBS28 SBS29 SBS30
#> Sig1 0.013 0.015 0.004 0.021 0.001 0.133 0.107 0.030 0.004 0.383 0.054 0.013
#> Sig2 0.317 0.255 0.103 0.054 0.355 0.260 0.393 0.178 0.081 0.023 0.282 0.389
#> Sig3 0.055 0.061 0.043 0.034 0.046 0.069 0.159 0.099 0.031 0.011 0.100 0.090
#>      SBS31 SBS32 SBS33 SBS34 SBS35 SBS36 SBS37 SBS38 SBS39 SBS40 SBS41 SBS42
#> Sig1 0.028 0.016 0.006 0.007 0.061 0.014 0.112 0.014 0.249 0.201 0.232 0.031
#> Sig2 0.304 0.429 0.108 0.049 0.228 0.171 0.199 0.115 0.260 0.347 0.192 0.338
#> Sig3 0.053 0.067 0.036 0.008 0.023 0.065 0.073 0.015 0.075 0.101 0.056 0.049
#>      SBS43 SBS44 SBS45 SBS46 SBS47 SBS48 SBS49 SBS50 SBS51 SBS52 SBS53 SBS54
#> Sig1 0.051 0.020 0.007 0.012 0.024 0.001 0.005 0.014 0.027 0.008 0.065 0.014
#> Sig2 0.059 0.417 0.127 0.219 0.100 0.027 0.068 0.237 0.184 0.051 0.135 0.170
#> Sig3 0.030 0.132 0.024 0.128 0.027 0.017 0.007 0.035 0.068 0.099 0.042 0.106
#>      SBS55 SBS56 SBS57 SBS58 SBS59 SBS60 SBS84 SBS85
#> Sig1 0.054 0.029 0.186 0.051 0.026 0.025 0.032 0.057
#> Sig2 0.076 0.098 0.237 0.265 0.056 0.031 0.297 0.114
#> Sig3 0.026 0.049 0.042 0.034 0.003 0.006 0.080 0.027
#> 
#> $aetiology_db
#> $aetiology_db[[1]]
#>  [1] "spontaneous or enzymatic deamination of 5-methylcytosine"                           
#>  [2] "APOBEC Cytidine Deaminase (C>T)"                                                    
#>  [3] "Defects in DNA-DSB repair by HR"                                                    
#>  [4] "exposure to tobacco (smoking) mutagens"                                             
#>  [5] "Unknown"                                                                            
#>  [6] "defective DNA mismatch repair"                                                      
#>  [7] "UV exposure"                                                                        
#>  [8] "UV exposure"                                                                        
#>  [9] "UV exposure"                                                                        
#> [10] "UV exposure"                                                                        
#> [11] "Unknown"                                                                            
#> [12] "defects in polymerase-eta"                                                          
#> [13] "Polymerase epsilon exonuclease domain mutations"                                    
#> [14] "Polymerase epsilon exonuclease domain mutations"                                    
#> [15] "exposure to alkylating agents"                                                      
#> [16] "Unknown"                                                                            
#> [17] "APOBEC Cytidine Deaminase (C>G)"                                                    
#> [18] "Concurrent polymerase epsilon mutation and defective DNA mismatch repair"           
#> [19] "Defective DNA mismatch repair"                                                      
#> [20] "Unknown"                                                                            
#> [21] "Unknown"                                                                            
#> [22] "Unknown"                                                                            
#> [23] "Possibly damage by reactive oxygen species"                                         
#> [24] "Unknown"                                                                            
#> [25] "Concurrent POLD1 mutations and defective DNA mismatch repair"                       
#> [26] "DNA mismatch repair deficiency"                                                     
#> [27] "exposure to aristolochic acid"                                                      
#> [28] "Unknown"                                                                            
#> [29] "exposures to aflatoxin"                                                             
#> [30] "Possible exposure to chemotherapy"                                                  
#> [31] "Defective DNA mismatch repair"                                                      
#> [32] "Unknown"                                                                            
#> [33] "exposure to tobacco (chewing) mutagens"                                             
#> [34] "deficiency in base excision repair due to inactivating mutations in NTHL1"          
#> [35] "Prior chemotherapy treatment with platinum drugs"                                   
#> [36] "Prior treatment with azathioprine to induce immunosuppression"                      
#> [37] "Unknown"                                                                            
#> [38] "Unknown"                                                                            
#> [39] "Prior chemotherapy treatment with platinum drugs."                                  
#> [40] "Defective base excision repair due to biallelic germline or somatic MUTYH mutations"
#> [41] "Unknown"                                                                            
#> [42] "Potential indirect damage from UV-light"                                            
#> [43] "Unknown"                                                                            
#> [44] "Unknown"                                                                            
#> [45] "Unknown"                                                                            
#> [46] "Occupational exposure to haloalkanes"                                               
#> [47] "Defective DNA mismatch repair"                                                      
#> [48] "Activity of activation-induced cytidine deaminase (AID)"                            
#> [49] "Indirect effects of activation-induced cytidine deaminase (AID)"                    
#> [50] "Possible sequencing artefact."                                                      
#> [51] "Possible sequencing artefact"                                                       
#> [52] "Possible sequencing artefact"                                                       
#> [53] "Possible sequencing artefact"                                                       
#> [54] "Possible sequencing artefact"                                                       
#> [55] "Possible sequencing artefact"                                                       
#> [56] "Possible sequencing artefact"                                                       
#> [57] "Possible sequencing artefact"                                                       
#> [58] "Possible sequencing artefact"                                                       
#> [59] "Possible sequencing artefact"                                                       
#> [60] "Possible sequencing artefact"                                                       
#> [61] "Possible sequencing artefact"                                                       
#> [62] "Possible sequencing artefact"                                                       
#> [63] "Possible sequencing artefact"                                                       
#> [64] "Possible sequencing artefact"                                                       
#> [65] "Possible sequencing artefact"                                                       
#> [66] "Possible sequencing artefact"                                                       
#> [67] "Possible sequencing artefact"                                                       
#> 
#> 
#> $best_match
#> $best_match$Sig1
#> $best_match$Sig1$aetiology
#> [1] "Unknown"
#> 
#> $best_match$Sig1$best_match
#> [1] "Best match: SBS28 [similarity: 0.383]"
#> 
#> 
#> $best_match$Sig2
#> $best_match$Sig2$aetiology
#> [1] "spontaneous or enzymatic deamination of 5-methylcytosine"
#> 
#> $best_match$Sig2$best_match
#> [1] "Best match: SBS1 [similarity: 0.868]"
#> 
#> 
#> $best_match$Sig3
#> $best_match$Sig3$aetiology
#> [1] "Polymerase epsilon exonuclease domain mutations"
#> 
#> $best_match$Sig3$best_match
#> [1] "Best match: SBS10b [similarity: 0.865]"
#> 
#> 
#> 
#> attr(,"class")
#> [1] "similarity" "list"      

## Remove some components
## in similarity calculation
s4 <- get_sig_similarity(sig2,
  Ref = sig2,
  pattern_to_rm = c("T[T>G]C", "T[T>G]G", "T[T>G]T")
)

#> -Comparing against Custom signatures

#> ------------------------------------

#> --Found Sig1 most similar to Sig1 [similarity: 1]

#> --Found Sig2 most similar to Sig2 [similarity: 1]

#> --Found Sig3 most similar to Sig3 [similarity: 1]

#> ------------------------------------

#> Return result invisiblely.

s4

#> $similarity
#>      Sig1 Sig2 Sig3
#> Sig1    1    0    0
#> Sig2    0    1    0
#> Sig3    0    0    1
#> 
#> $aetiology_db
#> [1] NA
#> 
#> $best_match
#> $best_match$Sig1
#> $best_match$Sig1$aetiology
#> [1] NA
#> 
#> $best_match$Sig1$best_match
#> [1] "Best match: Sig1 [similarity: 1]"
#> 
#> 
#> $best_match$Sig2
#> $best_match$Sig2$aetiology
#> [1] NA
#> 
#> $best_match$Sig2$best_match
#> [1] "Best match: Sig2 [similarity: 1]"
#> 
#> 
#> $best_match$Sig3
#> $best_match$Sig3$aetiology
#> [1] NA
#> 
#> $best_match$Sig3$best_match
#> [1] "Best match: Sig3 [similarity: 1]"
#> 
#> 
#> 
#> attr(,"class")
#> [1] "similarity" "list"      

## Same to DBS and ID signatures