Calculate Similarity between Identified Signatures and Reference Signatures

The reference signatures can be either a Signature object specified by Ref argument or known COSMIC signatures specified by sig_db argument. Two COSMIC databases are used for comparisons - "legacy" which includes 30 signaures, and "SBS" - which includes updated/refined 65 signatures. This function is modified from compareSignatures() in maftools package. NOTE: all reference signatures are generated from gold standard tool: SigProfiler.

get_sig_similarity(
  Signature,
  Ref = NULL,
  sig_db = c("SBS", "legacy", "DBS", "ID", "TSB", "SBS_Nik_lab", "RS_Nik_lab",
    "RS_BRCA560", "RS_USARC", "CNS_USARC", "CNS_TCGA", "CNS_TCGA176", "CNS_PCAWG176",
    "SBS_hg19", "SBS_hg38", "SBS_mm9", "SBS_mm10", "DBS_hg19", "DBS_hg38", "DBS_mm9",
    "DBS_mm10", "SBS_Nik_lab_Organ", "RS_Nik_lab_Organ", "latest_SBS_GRCh37",
    "latest_DBS_GRCh37", "latest_ID_GRCh37", "latest_SBS_GRCh38", "latest_DBS_GRCh38",
    "latest_SBS_mm9", "latest_DBS_mm9", "latest_SBS_mm10", "latest_DBS_mm10",
    "latest_SBS_rn6", "latest_DBS_rn6", "latest_CN_GRCh37", 
    
    "latest_RNA-SBS_GRCh37", "latest_SV_GRCh38"),
  db_type = c("", "human-exome", "human-genome"),
  method = "cosine",
  normalize = c("row", "feature"),
  feature_setting = sigminer::CN.features,
  set_order = TRUE,
  pattern_to_rm = NULL,
  verbose = TRUE
)

Arguments

Signature

a Signature object or a component-by-signature matrix/data.frame (sum of each column is 1) or a normalized component-by-sample matrix/data.frame (sum of each column is 1). More please see examples.

Ref

default is NULL, can be a same object as Signature.

sig_db

default 'legacy', it can be 'legacy' (for COSMIC v2 'SBS'), 'SBS', 'DBS', 'ID' and 'TSB' (for COSMIV v3.1 signatures) for small scale mutations. For more specific details, it can also be 'SBS_hg19', 'SBS_hg38', 'SBS_mm9', 'SBS_mm10', 'DBS_hg19', 'DBS_hg38', 'DBS_mm9', 'DBS_mm10' to use COSMIC v3 reference signatures from Alexandrov, Ludmil B., et al. (2020) (reference #1). In addition, it can be one of "SBS_Nik_lab_Organ", "RS_Nik_lab_Organ", "SBS_Nik_lab", "RS_Nik_lab" to refer reference signatures from Degasperi, Andrea, et al. (2020) (reference #2); "RS_BRCA560", "RS_USARC" to reference signatures from BRCA560 and USARC cohorts; "CNS_USARC" (40 categories), "CNS_TCGA" (48 categories) to reference copy number signatures from USARC cohort and TCGA; "CNS_TCGA176" (176 categories) and "CNS_PCAWG176" (176 categories) to reference copy number signatures from PCAWG and TCGA separately. UPDATE, the latest version of reference version can be automatically downloaded and loaded from https://cancer.sanger.ac.uk/signatures/downloads/ when a option with latest_ prefix is specified (e.g. "latest_SBS_GRCh37"). Note: the signature profile for different genome builds are basically same. And specific database (e.g. 'SBS_mm10') contains less signatures than all COSMIC signatures (because some signatures are not detected from Alexandrov, Ludmil B., et al. (2020)). For all available options, check the parameter setting.

db_type

only used when sig_db is enabled. "" for keeping default, "human-exome" for transforming to exome frequency of component, and "human-genome" for transforming to whole genome frequency of component. Currently only works for 'SBS'.

method

default is 'cosine' for cosine similarity.

normalize

one of "row" and "feature". "row" is typically used for common mutational signatures. "feature" is designed by me to use when input are copy number signatures.

feature_setting

a data.frame used for classification. Only used when method is "Wang" ("W"). Default is CN.features. Users can also set custom input with "feature", "min" and "max" columns available. Valid features can be printed by unique(CN.features$feature).

set_order

if TRUE, order the return similarity matrix.

pattern_to_rm

patterns for removing some features/components in similarity calculation. A vector of component name is also accepted. The remove operation will be done after normalization. Default is NULL.

verbose

if TRUE, print extra info.

Value

a list containing smilarities, aetiologies if available, best match and RSS.

References

Alexandrov, Ludmil B., et al. "The repertoire of mutational signatures in human cancer." Nature 578.7793 (2020): 94-101.

Degasperi, Andrea, et al. "A practical framework and online tool for mutational signature analyses show intertissue variation and driver dependencies." Nature cancer 1.2 (2020): 249-263.

Steele, Christopher D., et al. "Undifferentiated sarcomas develop through distinct evolutionary pathways." Cancer Cell 35.3 (2019): 441-456.

Nik-Zainal, Serena, et al. "Landscape of somatic mutations in 560 breast cancer whole-genome sequences." Nature 534.7605 (2016): 47-54.

Steele, Christopher D., et al. "Signatures of copy number alterations in human cancer." Nature 606.7916 (2022): 984-991.

Author

Shixiang Wang w_shixiang@163.com

Examples

# Load mutational signature
load(system.file("extdata", "toy_mutational_signature.RData",
  package = "sigminer", mustWork = TRUE
))

s1 <- get_sig_similarity(sig2, Ref = sig2)
#> -Comparing against Custom signatures/profiles
#> ---------------------------------------------
#> --Found Sig1 most similar to Sig1 [similarity: 1]
#> --Found Sig2 most similar to Sig2 [similarity: 1]
#> --Found Sig3 most similar to Sig3 [similarity: 1]
#> ------------------------------------
#> Return result invisiblely.
s1
#> $similarity
#>      Sig1 Sig2 Sig3
#> Sig1    1    0    0
#> Sig2    0    1    0
#> Sig3    0    0    1
#> 
#> $aetiology_db
#> [1] NA
#> 
#> $best_match
#> $best_match$Sig1
#> $best_match$Sig1$aetiology
#> [1] NA
#> 
#> $best_match$Sig1$best_match
#> [1] "Best match: Sig1 [similarity: 1]"
#> 
#> 
#> $best_match$Sig2
#> $best_match$Sig2$aetiology
#> [1] NA
#> 
#> $best_match$Sig2$best_match
#> [1] "Best match: Sig2 [similarity: 1]"
#> 
#> 
#> $best_match$Sig3
#> $best_match$Sig3$aetiology
#> [1] NA
#> 
#> $best_match$Sig3$best_match
#> [1] "Best match: Sig3 [similarity: 1]"
#> 
#> 
#> 
#> $rss
#>          Sig1     Sig2     Sig3
#> Sig1 0.000000 0.190663 0.430982
#> Sig2 0.190663 0.000000 0.347182
#> Sig3 0.430982 0.347182 0.000000
#> 
#> attr(,"class")
#> [1] "similarity" "list"      

s2 <- get_sig_similarity(sig2)
#> -Comparing against COSMIC signatures
#> ------------------------------------
#> --Found Sig1 most similar to SBS28
#>    Aetiology: Unknown [similarity: 0.383]
#> --Found Sig2 most similar to SBS1
#>    Aetiology: Spontaneous deamination of 5-methylcytosine (clock-like signature) [similarity: 0.868]
#> --Found Sig3 most similar to SBS10b
#>    Aetiology: Polymerase epsilon exonuclease domain mutations [similarity: 0.865]
#> ------------------------------------
#> Return result invisiblely.
s2
#> $similarity
#>      SBS28  SBS1 SBS10b  SBS2  SBS3  SBS4  SBS5  SBS6 SBS7a SBS7b SBS7c SBS7d
#> Sig1 0.383 0.001  0.070 0.000 0.198 0.062 0.117 0.003 0.001 0.004 0.070 0.011
#> Sig2 0.023 0.868  0.080 0.141 0.340 0.219 0.492 0.836 0.220 0.281 0.095 0.111
#> Sig3 0.011 0.218  0.865 0.067 0.090 0.042 0.208 0.232 0.159 0.158 0.011 0.016
#>       SBS8  SBS9 SBS10a SBS11 SBS12 SBS13 SBS14 SBS15 SBS16 SBS17a SBS17b SBS18
#> Sig1 0.094 0.350  0.011 0.001 0.011 0.007 0.004 0.010 0.046  0.011  0.373 0.043
#> Sig2 0.219 0.209  0.096 0.316 0.154 0.048 0.187 0.674 0.176  0.061  0.008 0.273
#> Sig3 0.043 0.079  0.019 0.014 0.075 0.021 0.033 0.118 0.033  0.040  0.005 0.096
#>      SBS19 SBS20 SBS21 SBS22 SBS23 SBS24 SBS25 SBS26 SBS27 SBS29 SBS30 SBS31
#> Sig1 0.013 0.015 0.004 0.021 0.001 0.133 0.107 0.030 0.004 0.054 0.013 0.028
#> Sig2 0.317 0.255 0.103 0.054 0.355 0.260 0.393 0.178 0.081 0.282 0.389 0.304
#> Sig3 0.055 0.061 0.043 0.034 0.046 0.069 0.159 0.099 0.031 0.100 0.090 0.053
#>      SBS32 SBS33 SBS34 SBS35 SBS36 SBS37 SBS38 SBS39 SBS40 SBS41 SBS42 SBS43
#> Sig1 0.016 0.006 0.007 0.061 0.014 0.112 0.014 0.249 0.201 0.232 0.031 0.051
#> Sig2 0.429 0.108 0.049 0.228 0.171 0.199 0.115 0.260 0.347 0.192 0.338 0.059
#> Sig3 0.067 0.036 0.008 0.023 0.065 0.073 0.015 0.075 0.101 0.056 0.049 0.030
#>      SBS44 SBS45 SBS46 SBS47 SBS48 SBS49 SBS50 SBS51 SBS52 SBS53 SBS54 SBS55
#> Sig1 0.020 0.007 0.012 0.024 0.001 0.005 0.014 0.027 0.008 0.065 0.014 0.054
#> Sig2 0.417 0.127 0.219 0.100 0.027 0.068 0.237 0.184 0.051 0.135 0.170 0.076
#> Sig3 0.132 0.024 0.128 0.027 0.017 0.007 0.035 0.068 0.099 0.042 0.106 0.026
#>      SBS56 SBS57 SBS58 SBS59 SBS60 SBS84 SBS85 SBS86 SBS87 SBS88 SBS89 SBS90
#> Sig1 0.029 0.186 0.051 0.026 0.025 0.032 0.057 0.111 0.119 0.166 0.141 0.040
#> Sig2 0.098 0.237 0.265 0.056 0.031 0.297 0.114 0.194 0.735 0.158 0.325 0.016
#> Sig3 0.049 0.042 0.034 0.003 0.006 0.080 0.027 0.069 0.228 0.051 0.066 0.004
#> 
#> $aetiology_db
#> $aetiology_db[[1]]
#>  [1] "Spontaneous deamination of 5-methylcytosine (clock-like signature)"      
#>  [2] "Activity of APOBEC family of cytidine deaminases"                        
#>  [3] "Defective homologous recombination DNA damage repair"                    
#>  [4] "Tobacco smoking"                                                         
#>  [5] "Unknown (clock-like signature)"                                          
#>  [6] "Defective DNA mismatch repair"                                           
#>  [7] "Ultraviolet light exposure"                                              
#>  [8] "Ultraviolet light exposure"                                              
#>  [9] "Ultraviolet light exposure"                                              
#> [10] "Ultraviolet light exposure"                                              
#> [11] "Unknown"                                                                 
#> [12] "Polimerase eta somatic hypermutation activity"                           
#> [13] "Polymerase epsilon exonuclease domain mutations"                         
#> [14] "Polymerase epsilon exonuclease domain mutations"                         
#> [15] "Temozolomide treatment"                                                  
#> [16] "Unknown"                                                                 
#> [17] "Activity of APOBEC family of cytidine deaminases"                        
#> [18] "Concurrent polymerase epsilon mutation and defective DNA mismatch repair"
#> [19] "Defective DNA mismatch repair"                                           
#> [20] "Unknown"                                                                 
#> [21] "Unknown"                                                                 
#> [22] "Unknown"                                                                 
#> [23] "Damage by reactive oxygen species"                                       
#> [24] "Unknown"                                                                 
#> [25] "Concurrent POLD1 mutations and defective DNA mismatch repair"            
#> [26] "Defective DNA mismatch repair"                                           
#> [27] "Aristolochic acid exposure"                                              
#> [28] "Unknown"                                                                 
#> [29] "Aflatoxin exposure"                                                      
#> [30] "Chemotherapy treatment"                                                  
#> [31] "Defective DNA mismatch repair"                                           
#> [32] "Possible sequencing artefact"                                            
#> [33] "Unknown"                                                                 
#> [34] "Tobacco chewing"                                                         
#> [35] "Defective DNA base excision repair due to NTHL1 mutations"               
#> [36] "Platinum chemotherapy treatment"                                         
#> [37] "Azathioprine treatment"                                                  
#> [38] "Unknown"                                                                 
#> [39] "Unknown"                                                                 
#> [40] "Platinum chemotherapy treatment"                                         
#> [41] "Defective DNA base excision repair due to MUTYH mutations"               
#> [42] "Unknown"                                                                 
#> [43] "Indirect effect of ultraviolet light"                                    
#> [44] "Unknown"                                                                 
#> [45] "Unknown"                                                                 
#> [46] "Unknown"                                                                 
#> [47] "Haloalkane exposure"                                                     
#> [48] "Possible sequencing artefact"                                            
#> [49] "Defective DNA mismatch repair"                                           
#> [50] "Possible sequencing artefact"                                            
#> [51] "Possible sequencing artefact"                                            
#> [52] "Possible sequencing artefact"                                            
#> [53] "Possible sequencing artefact"                                            
#> [54] "Possible sequencing artefact"                                            
#> [55] "Possible sequencing artefact"                                            
#> [56] "Possible sequencing artefact"                                            
#> [57] "Possible sequencing artefact"                                            
#> [58] "Possible sequencing artefact"                                            
#> [59] "Possible sequencing artefact"                                            
#> [60] "Possible sequencing artefact"                                            
#> [61] "Possible sequencing artefact"                                            
#> [62] "Possible sequencing artefact"                                            
#> [63] "Possible sequencing artefact"                                            
#> [64] "Possible sequencing artefact"                                            
#> [65] "Possible sequencing artefact"                                            
#> [66] "Activity of activation-induced cytidine deaminase (AID)"                 
#> [67] "Indirect effects of activation-induced cytidine deaminase (AID)"         
#> [68] "Unknown chemotherapy treatment"                                          
#> [69] "Thiopurine chemotherapy treatment"                                       
#> [70] "Colibactin exposure (E.coli bacteria carrying pks pathogenicity island)" 
#> [71] "Unknown"                                                                 
#> [72] "Duocarmycin exposure"                                                    
#> 
#> 
#> $best_match
#> $best_match$Sig1
#> $best_match$Sig1$aetiology
#> [1] "Unknown"
#> 
#> $best_match$Sig1$best_match
#> [1] "Best match: SBS28 [similarity: 0.383]"
#> 
#> 
#> $best_match$Sig2
#> $best_match$Sig2$aetiology
#> [1] "Spontaneous deamination of 5-methylcytosine (clock-like signature)"
#> 
#> $best_match$Sig2$best_match
#> [1] "Best match: SBS1 [similarity: 0.868]"
#> 
#> 
#> $best_match$Sig3
#> $best_match$Sig3$aetiology
#> [1] "Polymerase epsilon exonuclease domain mutations"
#> 
#> $best_match$Sig3$best_match
#> [1] "Best match: SBS10b [similarity: 0.865]"
#> 
#> 
#> 
#> $rss
#>          SBS1     SBS2     SBS3     SBS4     SBS5     SBS6    SBS7a    SBS7b
#> Sig1 0.369193 0.525837 0.133773 0.163154 0.142928 0.233049 0.329982 0.257810
#> Sig2 0.092395 0.401493 0.048795 0.069176 0.040832 0.029963 0.201746 0.129839
#> Sig3 0.412095 0.637387 0.296138 0.319891 0.281398 0.312334 0.410916 0.355830
#>         SBS7c    SBS7d     SBS8     SBS9   SBS10a   SBS10b    SBS11    SBS12
#> Sig1 0.220593 0.259290 0.151807 0.121283 0.592288 0.339654 0.249398 0.184606
#> Sig2 0.139360 0.160424 0.062998 0.063645 0.483733 0.262953 0.116867 0.086949
#> Sig3 0.389688 0.412737 0.312035 0.305754 0.740139 0.073998 0.400920 0.325061
#>         SBS13    SBS14    SBS15    SBS16   SBS17a   SBS17b    SBS18    SBS19
#> Sig1 0.384435 0.254322 0.251461 0.214292 0.361085 0.313559 0.181550 0.249426
#> Sig2 0.291947 0.141861 0.063769 0.116602 0.267755 0.388287 0.076398 0.119212
#> Sig3 0.531917 0.399662 0.366821 0.370351 0.501123 0.627313 0.321777 0.388976
#>         SBS20    SBS21    SBS22    SBS23    SBS24    SBS25    SBS26    SBS27
#> Sig1 0.222261 0.237466 0.209514 0.250671 0.159083 0.145363 0.183891 0.284051
#> Sig2 0.106840 0.139489 0.123211 0.111862 0.070913 0.047459 0.086505 0.186965
#> Sig3 0.362647 0.379998 0.360274 0.390581 0.320568 0.289040 0.321052 0.428821
#>         SBS28    SBS29    SBS30    SBS31    SBS32    SBS33    SBS34    SBS35
#> Sig1 0.251169 0.176705 0.217281 0.190168 0.198065 0.291670 0.326219 0.167234
#> Sig2 0.306440 0.073145 0.084439 0.077665 0.067240 0.189799 0.234887 0.071536
#> Sig3 0.546324 0.318320 0.348499 0.337970 0.339405 0.434520 0.481561 0.327852
#>         SBS36    SBS37    SBS38    SBS39    SBS40    SBS41    SBS42    SBS43
#> Sig1 0.222163 0.155024 0.321314 0.131135 0.133988 0.141395 0.180887 0.219344
#> Sig2 0.118089 0.069663 0.219050 0.056606 0.048896 0.073695 0.067748 0.138797
#> Sig3 0.360991 0.312379 0.475377 0.302408 0.295547 0.319400 0.330777 0.377444
#>         SBS44    SBS45    SBS46    SBS47    SBS48    SBS49    SBS50    SBS51
#> Sig1 0.182577 0.269907 0.176080 0.205584 0.674224 0.460338 0.197968 0.187056
#> Sig2 0.059625 0.166457 0.073658 0.114100 0.581815 0.360637 0.089241 0.088074
#> Sig3 0.310896 0.418861 0.306514 0.359080 0.817393 0.614301 0.347555 0.331071
#>         SBS52    SBS53    SBS54    SBS55    SBS56    SBS57    SBS58    SBS59
#> Sig1 0.378140 0.250307 0.239819 0.256845 0.345687 0.165649 0.185579 0.405001
#> Sig2 0.285849 0.161257 0.133842 0.174820 0.251045 0.089001 0.081505 0.317869
#> Sig3 0.484724 0.407754 0.362176 0.417739 0.487844 0.345214 0.342362 0.569894
#>         SBS60    SBS84    SBS85    SBS86    SBS87    SBS88    SBS89    SBS90
#> Sig1 0.571066 0.204442 0.175558 0.171859 0.180572 0.179788 0.141616 0.305934
#> Sig2 0.490093 0.089916 0.089645 0.086394 0.032164 0.109855 0.051612 0.231820
#> Sig3 0.735723 0.343883 0.335085 0.330248 0.296362 0.355069 0.302726 0.473481
#> 
#> attr(,"class")
#> [1] "similarity" "list"      
s3 <- get_sig_similarity(sig2, sig_db = "SBS")
#> -Comparing against COSMIC signatures
#> ------------------------------------
#> --Found Sig1 most similar to SBS28
#>    Aetiology: Unknown [similarity: 0.383]
#> --Found Sig2 most similar to SBS1
#>    Aetiology: Spontaneous deamination of 5-methylcytosine (clock-like signature) [similarity: 0.868]
#> --Found Sig3 most similar to SBS10b
#>    Aetiology: Polymerase epsilon exonuclease domain mutations [similarity: 0.865]
#> ------------------------------------
#> Return result invisiblely.
s3
#> $similarity
#>      SBS28  SBS1 SBS10b  SBS2  SBS3  SBS4  SBS5  SBS6 SBS7a SBS7b SBS7c SBS7d
#> Sig1 0.383 0.001  0.070 0.000 0.198 0.062 0.117 0.003 0.001 0.004 0.070 0.011
#> Sig2 0.023 0.868  0.080 0.141 0.340 0.219 0.492 0.836 0.220 0.281 0.095 0.111
#> Sig3 0.011 0.218  0.865 0.067 0.090 0.042 0.208 0.232 0.159 0.158 0.011 0.016
#>       SBS8  SBS9 SBS10a SBS11 SBS12 SBS13 SBS14 SBS15 SBS16 SBS17a SBS17b SBS18
#> Sig1 0.094 0.350  0.011 0.001 0.011 0.007 0.004 0.010 0.046  0.011  0.373 0.043
#> Sig2 0.219 0.209  0.096 0.316 0.154 0.048 0.187 0.674 0.176  0.061  0.008 0.273
#> Sig3 0.043 0.079  0.019 0.014 0.075 0.021 0.033 0.118 0.033  0.040  0.005 0.096
#>      SBS19 SBS20 SBS21 SBS22 SBS23 SBS24 SBS25 SBS26 SBS27 SBS29 SBS30 SBS31
#> Sig1 0.013 0.015 0.004 0.021 0.001 0.133 0.107 0.030 0.004 0.054 0.013 0.028
#> Sig2 0.317 0.255 0.103 0.054 0.355 0.260 0.393 0.178 0.081 0.282 0.389 0.304
#> Sig3 0.055 0.061 0.043 0.034 0.046 0.069 0.159 0.099 0.031 0.100 0.090 0.053
#>      SBS32 SBS33 SBS34 SBS35 SBS36 SBS37 SBS38 SBS39 SBS40 SBS41 SBS42 SBS43
#> Sig1 0.016 0.006 0.007 0.061 0.014 0.112 0.014 0.249 0.201 0.232 0.031 0.051
#> Sig2 0.429 0.108 0.049 0.228 0.171 0.199 0.115 0.260 0.347 0.192 0.338 0.059
#> Sig3 0.067 0.036 0.008 0.023 0.065 0.073 0.015 0.075 0.101 0.056 0.049 0.030
#>      SBS44 SBS45 SBS46 SBS47 SBS48 SBS49 SBS50 SBS51 SBS52 SBS53 SBS54 SBS55
#> Sig1 0.020 0.007 0.012 0.024 0.001 0.005 0.014 0.027 0.008 0.065 0.014 0.054
#> Sig2 0.417 0.127 0.219 0.100 0.027 0.068 0.237 0.184 0.051 0.135 0.170 0.076
#> Sig3 0.132 0.024 0.128 0.027 0.017 0.007 0.035 0.068 0.099 0.042 0.106 0.026
#>      SBS56 SBS57 SBS58 SBS59 SBS60 SBS84 SBS85 SBS86 SBS87 SBS88 SBS89 SBS90
#> Sig1 0.029 0.186 0.051 0.026 0.025 0.032 0.057 0.111 0.119 0.166 0.141 0.040
#> Sig2 0.098 0.237 0.265 0.056 0.031 0.297 0.114 0.194 0.735 0.158 0.325 0.016
#> Sig3 0.049 0.042 0.034 0.003 0.006 0.080 0.027 0.069 0.228 0.051 0.066 0.004
#> 
#> $aetiology_db
#> $aetiology_db[[1]]
#>  [1] "Spontaneous deamination of 5-methylcytosine (clock-like signature)"      
#>  [2] "Activity of APOBEC family of cytidine deaminases"                        
#>  [3] "Defective homologous recombination DNA damage repair"                    
#>  [4] "Tobacco smoking"                                                         
#>  [5] "Unknown (clock-like signature)"                                          
#>  [6] "Defective DNA mismatch repair"                                           
#>  [7] "Ultraviolet light exposure"                                              
#>  [8] "Ultraviolet light exposure"                                              
#>  [9] "Ultraviolet light exposure"                                              
#> [10] "Ultraviolet light exposure"                                              
#> [11] "Unknown"                                                                 
#> [12] "Polimerase eta somatic hypermutation activity"                           
#> [13] "Polymerase epsilon exonuclease domain mutations"                         
#> [14] "Polymerase epsilon exonuclease domain mutations"                         
#> [15] "Temozolomide treatment"                                                  
#> [16] "Unknown"                                                                 
#> [17] "Activity of APOBEC family of cytidine deaminases"                        
#> [18] "Concurrent polymerase epsilon mutation and defective DNA mismatch repair"
#> [19] "Defective DNA mismatch repair"                                           
#> [20] "Unknown"                                                                 
#> [21] "Unknown"                                                                 
#> [22] "Unknown"                                                                 
#> [23] "Damage by reactive oxygen species"                                       
#> [24] "Unknown"                                                                 
#> [25] "Concurrent POLD1 mutations and defective DNA mismatch repair"            
#> [26] "Defective DNA mismatch repair"                                           
#> [27] "Aristolochic acid exposure"                                              
#> [28] "Unknown"                                                                 
#> [29] "Aflatoxin exposure"                                                      
#> [30] "Chemotherapy treatment"                                                  
#> [31] "Defective DNA mismatch repair"                                           
#> [32] "Possible sequencing artefact"                                            
#> [33] "Unknown"                                                                 
#> [34] "Tobacco chewing"                                                         
#> [35] "Defective DNA base excision repair due to NTHL1 mutations"               
#> [36] "Platinum chemotherapy treatment"                                         
#> [37] "Azathioprine treatment"                                                  
#> [38] "Unknown"                                                                 
#> [39] "Unknown"                                                                 
#> [40] "Platinum chemotherapy treatment"                                         
#> [41] "Defective DNA base excision repair due to MUTYH mutations"               
#> [42] "Unknown"                                                                 
#> [43] "Indirect effect of ultraviolet light"                                    
#> [44] "Unknown"                                                                 
#> [45] "Unknown"                                                                 
#> [46] "Unknown"                                                                 
#> [47] "Haloalkane exposure"                                                     
#> [48] "Possible sequencing artefact"                                            
#> [49] "Defective DNA mismatch repair"                                           
#> [50] "Possible sequencing artefact"                                            
#> [51] "Possible sequencing artefact"                                            
#> [52] "Possible sequencing artefact"                                            
#> [53] "Possible sequencing artefact"                                            
#> [54] "Possible sequencing artefact"                                            
#> [55] "Possible sequencing artefact"                                            
#> [56] "Possible sequencing artefact"                                            
#> [57] "Possible sequencing artefact"                                            
#> [58] "Possible sequencing artefact"                                            
#> [59] "Possible sequencing artefact"                                            
#> [60] "Possible sequencing artefact"                                            
#> [61] "Possible sequencing artefact"                                            
#> [62] "Possible sequencing artefact"                                            
#> [63] "Possible sequencing artefact"                                            
#> [64] "Possible sequencing artefact"                                            
#> [65] "Possible sequencing artefact"                                            
#> [66] "Activity of activation-induced cytidine deaminase (AID)"                 
#> [67] "Indirect effects of activation-induced cytidine deaminase (AID)"         
#> [68] "Unknown chemotherapy treatment"                                          
#> [69] "Thiopurine chemotherapy treatment"                                       
#> [70] "Colibactin exposure (E.coli bacteria carrying pks pathogenicity island)" 
#> [71] "Unknown"                                                                 
#> [72] "Duocarmycin exposure"                                                    
#> 
#> 
#> $best_match
#> $best_match$Sig1
#> $best_match$Sig1$aetiology
#> [1] "Unknown"
#> 
#> $best_match$Sig1$best_match
#> [1] "Best match: SBS28 [similarity: 0.383]"
#> 
#> 
#> $best_match$Sig2
#> $best_match$Sig2$aetiology
#> [1] "Spontaneous deamination of 5-methylcytosine (clock-like signature)"
#> 
#> $best_match$Sig2$best_match
#> [1] "Best match: SBS1 [similarity: 0.868]"
#> 
#> 
#> $best_match$Sig3
#> $best_match$Sig3$aetiology
#> [1] "Polymerase epsilon exonuclease domain mutations"
#> 
#> $best_match$Sig3$best_match
#> [1] "Best match: SBS10b [similarity: 0.865]"
#> 
#> 
#> 
#> $rss
#>          SBS1     SBS2     SBS3     SBS4     SBS5     SBS6    SBS7a    SBS7b
#> Sig1 0.369193 0.525837 0.133773 0.163154 0.142928 0.233049 0.329982 0.257810
#> Sig2 0.092395 0.401493 0.048795 0.069176 0.040832 0.029963 0.201746 0.129839
#> Sig3 0.412095 0.637387 0.296138 0.319891 0.281398 0.312334 0.410916 0.355830
#>         SBS7c    SBS7d     SBS8     SBS9   SBS10a   SBS10b    SBS11    SBS12
#> Sig1 0.220593 0.259290 0.151807 0.121283 0.592288 0.339654 0.249398 0.184606
#> Sig2 0.139360 0.160424 0.062998 0.063645 0.483733 0.262953 0.116867 0.086949
#> Sig3 0.389688 0.412737 0.312035 0.305754 0.740139 0.073998 0.400920 0.325061
#>         SBS13    SBS14    SBS15    SBS16   SBS17a   SBS17b    SBS18    SBS19
#> Sig1 0.384435 0.254322 0.251461 0.214292 0.361085 0.313559 0.181550 0.249426
#> Sig2 0.291947 0.141861 0.063769 0.116602 0.267755 0.388287 0.076398 0.119212
#> Sig3 0.531917 0.399662 0.366821 0.370351 0.501123 0.627313 0.321777 0.388976
#>         SBS20    SBS21    SBS22    SBS23    SBS24    SBS25    SBS26    SBS27
#> Sig1 0.222261 0.237466 0.209514 0.250671 0.159083 0.145363 0.183891 0.284051
#> Sig2 0.106840 0.139489 0.123211 0.111862 0.070913 0.047459 0.086505 0.186965
#> Sig3 0.362647 0.379998 0.360274 0.390581 0.320568 0.289040 0.321052 0.428821
#>         SBS28    SBS29    SBS30    SBS31    SBS32    SBS33    SBS34    SBS35
#> Sig1 0.251169 0.176705 0.217281 0.190168 0.198065 0.291670 0.326219 0.167234
#> Sig2 0.306440 0.073145 0.084439 0.077665 0.067240 0.189799 0.234887 0.071536
#> Sig3 0.546324 0.318320 0.348499 0.337970 0.339405 0.434520 0.481561 0.327852
#>         SBS36    SBS37    SBS38    SBS39    SBS40    SBS41    SBS42    SBS43
#> Sig1 0.222163 0.155024 0.321314 0.131135 0.133988 0.141395 0.180887 0.219344
#> Sig2 0.118089 0.069663 0.219050 0.056606 0.048896 0.073695 0.067748 0.138797
#> Sig3 0.360991 0.312379 0.475377 0.302408 0.295547 0.319400 0.330777 0.377444
#>         SBS44    SBS45    SBS46    SBS47    SBS48    SBS49    SBS50    SBS51
#> Sig1 0.182577 0.269907 0.176080 0.205584 0.674224 0.460338 0.197968 0.187056
#> Sig2 0.059625 0.166457 0.073658 0.114100 0.581815 0.360637 0.089241 0.088074
#> Sig3 0.310896 0.418861 0.306514 0.359080 0.817393 0.614301 0.347555 0.331071
#>         SBS52    SBS53    SBS54    SBS55    SBS56    SBS57    SBS58    SBS59
#> Sig1 0.378140 0.250307 0.239819 0.256845 0.345687 0.165649 0.185579 0.405001
#> Sig2 0.285849 0.161257 0.133842 0.174820 0.251045 0.089001 0.081505 0.317869
#> Sig3 0.484724 0.407754 0.362176 0.417739 0.487844 0.345214 0.342362 0.569894
#>         SBS60    SBS84    SBS85    SBS86    SBS87    SBS88    SBS89    SBS90
#> Sig1 0.571066 0.204442 0.175558 0.171859 0.180572 0.179788 0.141616 0.305934
#> Sig2 0.490093 0.089916 0.089645 0.086394 0.032164 0.109855 0.051612 0.231820
#> Sig3 0.735723 0.343883 0.335085 0.330248 0.296362 0.355069 0.302726 0.473481
#> 
#> attr(,"class")
#> [1] "similarity" "list"      

# Set order for result similarity matrix
s4 <- get_sig_similarity(sig2, sig_db = "SBS", set_order = TRUE)
#> -Comparing against COSMIC signatures
#> ------------------------------------
#> --Found Sig1 most similar to SBS28
#>    Aetiology: Unknown [similarity: 0.383]
#> --Found Sig2 most similar to SBS1
#>    Aetiology: Spontaneous deamination of 5-methylcytosine (clock-like signature) [similarity: 0.868]
#> --Found Sig3 most similar to SBS10b
#>    Aetiology: Polymerase epsilon exonuclease domain mutations [similarity: 0.865]
#> ------------------------------------
#> Return result invisiblely.
s4
#> $similarity
#>      SBS28  SBS1 SBS10b  SBS2  SBS3  SBS4  SBS5  SBS6 SBS7a SBS7b SBS7c SBS7d
#> Sig1 0.383 0.001  0.070 0.000 0.198 0.062 0.117 0.003 0.001 0.004 0.070 0.011
#> Sig2 0.023 0.868  0.080 0.141 0.340 0.219 0.492 0.836 0.220 0.281 0.095 0.111
#> Sig3 0.011 0.218  0.865 0.067 0.090 0.042 0.208 0.232 0.159 0.158 0.011 0.016
#>       SBS8  SBS9 SBS10a SBS11 SBS12 SBS13 SBS14 SBS15 SBS16 SBS17a SBS17b SBS18
#> Sig1 0.094 0.350  0.011 0.001 0.011 0.007 0.004 0.010 0.046  0.011  0.373 0.043
#> Sig2 0.219 0.209  0.096 0.316 0.154 0.048 0.187 0.674 0.176  0.061  0.008 0.273
#> Sig3 0.043 0.079  0.019 0.014 0.075 0.021 0.033 0.118 0.033  0.040  0.005 0.096
#>      SBS19 SBS20 SBS21 SBS22 SBS23 SBS24 SBS25 SBS26 SBS27 SBS29 SBS30 SBS31
#> Sig1 0.013 0.015 0.004 0.021 0.001 0.133 0.107 0.030 0.004 0.054 0.013 0.028
#> Sig2 0.317 0.255 0.103 0.054 0.355 0.260 0.393 0.178 0.081 0.282 0.389 0.304
#> Sig3 0.055 0.061 0.043 0.034 0.046 0.069 0.159 0.099 0.031 0.100 0.090 0.053
#>      SBS32 SBS33 SBS34 SBS35 SBS36 SBS37 SBS38 SBS39 SBS40 SBS41 SBS42 SBS43
#> Sig1 0.016 0.006 0.007 0.061 0.014 0.112 0.014 0.249 0.201 0.232 0.031 0.051
#> Sig2 0.429 0.108 0.049 0.228 0.171 0.199 0.115 0.260 0.347 0.192 0.338 0.059
#> Sig3 0.067 0.036 0.008 0.023 0.065 0.073 0.015 0.075 0.101 0.056 0.049 0.030
#>      SBS44 SBS45 SBS46 SBS47 SBS48 SBS49 SBS50 SBS51 SBS52 SBS53 SBS54 SBS55
#> Sig1 0.020 0.007 0.012 0.024 0.001 0.005 0.014 0.027 0.008 0.065 0.014 0.054
#> Sig2 0.417 0.127 0.219 0.100 0.027 0.068 0.237 0.184 0.051 0.135 0.170 0.076
#> Sig3 0.132 0.024 0.128 0.027 0.017 0.007 0.035 0.068 0.099 0.042 0.106 0.026
#>      SBS56 SBS57 SBS58 SBS59 SBS60 SBS84 SBS85 SBS86 SBS87 SBS88 SBS89 SBS90
#> Sig1 0.029 0.186 0.051 0.026 0.025 0.032 0.057 0.111 0.119 0.166 0.141 0.040
#> Sig2 0.098 0.237 0.265 0.056 0.031 0.297 0.114 0.194 0.735 0.158 0.325 0.016
#> Sig3 0.049 0.042 0.034 0.003 0.006 0.080 0.027 0.069 0.228 0.051 0.066 0.004
#> 
#> $aetiology_db
#> $aetiology_db[[1]]
#>  [1] "Spontaneous deamination of 5-methylcytosine (clock-like signature)"      
#>  [2] "Activity of APOBEC family of cytidine deaminases"                        
#>  [3] "Defective homologous recombination DNA damage repair"                    
#>  [4] "Tobacco smoking"                                                         
#>  [5] "Unknown (clock-like signature)"                                          
#>  [6] "Defective DNA mismatch repair"                                           
#>  [7] "Ultraviolet light exposure"                                              
#>  [8] "Ultraviolet light exposure"                                              
#>  [9] "Ultraviolet light exposure"                                              
#> [10] "Ultraviolet light exposure"                                              
#> [11] "Unknown"                                                                 
#> [12] "Polimerase eta somatic hypermutation activity"                           
#> [13] "Polymerase epsilon exonuclease domain mutations"                         
#> [14] "Polymerase epsilon exonuclease domain mutations"                         
#> [15] "Temozolomide treatment"                                                  
#> [16] "Unknown"                                                                 
#> [17] "Activity of APOBEC family of cytidine deaminases"                        
#> [18] "Concurrent polymerase epsilon mutation and defective DNA mismatch repair"
#> [19] "Defective DNA mismatch repair"                                           
#> [20] "Unknown"                                                                 
#> [21] "Unknown"                                                                 
#> [22] "Unknown"                                                                 
#> [23] "Damage by reactive oxygen species"                                       
#> [24] "Unknown"                                                                 
#> [25] "Concurrent POLD1 mutations and defective DNA mismatch repair"            
#> [26] "Defective DNA mismatch repair"                                           
#> [27] "Aristolochic acid exposure"                                              
#> [28] "Unknown"                                                                 
#> [29] "Aflatoxin exposure"                                                      
#> [30] "Chemotherapy treatment"                                                  
#> [31] "Defective DNA mismatch repair"                                           
#> [32] "Possible sequencing artefact"                                            
#> [33] "Unknown"                                                                 
#> [34] "Tobacco chewing"                                                         
#> [35] "Defective DNA base excision repair due to NTHL1 mutations"               
#> [36] "Platinum chemotherapy treatment"                                         
#> [37] "Azathioprine treatment"                                                  
#> [38] "Unknown"                                                                 
#> [39] "Unknown"                                                                 
#> [40] "Platinum chemotherapy treatment"                                         
#> [41] "Defective DNA base excision repair due to MUTYH mutations"               
#> [42] "Unknown"                                                                 
#> [43] "Indirect effect of ultraviolet light"                                    
#> [44] "Unknown"                                                                 
#> [45] "Unknown"                                                                 
#> [46] "Unknown"                                                                 
#> [47] "Haloalkane exposure"                                                     
#> [48] "Possible sequencing artefact"                                            
#> [49] "Defective DNA mismatch repair"                                           
#> [50] "Possible sequencing artefact"                                            
#> [51] "Possible sequencing artefact"                                            
#> [52] "Possible sequencing artefact"                                            
#> [53] "Possible sequencing artefact"                                            
#> [54] "Possible sequencing artefact"                                            
#> [55] "Possible sequencing artefact"                                            
#> [56] "Possible sequencing artefact"                                            
#> [57] "Possible sequencing artefact"                                            
#> [58] "Possible sequencing artefact"                                            
#> [59] "Possible sequencing artefact"                                            
#> [60] "Possible sequencing artefact"                                            
#> [61] "Possible sequencing artefact"                                            
#> [62] "Possible sequencing artefact"                                            
#> [63] "Possible sequencing artefact"                                            
#> [64] "Possible sequencing artefact"                                            
#> [65] "Possible sequencing artefact"                                            
#> [66] "Activity of activation-induced cytidine deaminase (AID)"                 
#> [67] "Indirect effects of activation-induced cytidine deaminase (AID)"         
#> [68] "Unknown chemotherapy treatment"                                          
#> [69] "Thiopurine chemotherapy treatment"                                       
#> [70] "Colibactin exposure (E.coli bacteria carrying pks pathogenicity island)" 
#> [71] "Unknown"                                                                 
#> [72] "Duocarmycin exposure"                                                    
#> 
#> 
#> $best_match
#> $best_match$Sig1
#> $best_match$Sig1$aetiology
#> [1] "Unknown"
#> 
#> $best_match$Sig1$best_match
#> [1] "Best match: SBS28 [similarity: 0.383]"
#> 
#> 
#> $best_match$Sig2
#> $best_match$Sig2$aetiology
#> [1] "Spontaneous deamination of 5-methylcytosine (clock-like signature)"
#> 
#> $best_match$Sig2$best_match
#> [1] "Best match: SBS1 [similarity: 0.868]"
#> 
#> 
#> $best_match$Sig3
#> $best_match$Sig3$aetiology
#> [1] "Polymerase epsilon exonuclease domain mutations"
#> 
#> $best_match$Sig3$best_match
#> [1] "Best match: SBS10b [similarity: 0.865]"
#> 
#> 
#> 
#> $rss
#>          SBS1     SBS2     SBS3     SBS4     SBS5     SBS6    SBS7a    SBS7b
#> Sig1 0.369193 0.525837 0.133773 0.163154 0.142928 0.233049 0.329982 0.257810
#> Sig2 0.092395 0.401493 0.048795 0.069176 0.040832 0.029963 0.201746 0.129839
#> Sig3 0.412095 0.637387 0.296138 0.319891 0.281398 0.312334 0.410916 0.355830
#>         SBS7c    SBS7d     SBS8     SBS9   SBS10a   SBS10b    SBS11    SBS12
#> Sig1 0.220593 0.259290 0.151807 0.121283 0.592288 0.339654 0.249398 0.184606
#> Sig2 0.139360 0.160424 0.062998 0.063645 0.483733 0.262953 0.116867 0.086949
#> Sig3 0.389688 0.412737 0.312035 0.305754 0.740139 0.073998 0.400920 0.325061
#>         SBS13    SBS14    SBS15    SBS16   SBS17a   SBS17b    SBS18    SBS19
#> Sig1 0.384435 0.254322 0.251461 0.214292 0.361085 0.313559 0.181550 0.249426
#> Sig2 0.291947 0.141861 0.063769 0.116602 0.267755 0.388287 0.076398 0.119212
#> Sig3 0.531917 0.399662 0.366821 0.370351 0.501123 0.627313 0.321777 0.388976
#>         SBS20    SBS21    SBS22    SBS23    SBS24    SBS25    SBS26    SBS27
#> Sig1 0.222261 0.237466 0.209514 0.250671 0.159083 0.145363 0.183891 0.284051
#> Sig2 0.106840 0.139489 0.123211 0.111862 0.070913 0.047459 0.086505 0.186965
#> Sig3 0.362647 0.379998 0.360274 0.390581 0.320568 0.289040 0.321052 0.428821
#>         SBS28    SBS29    SBS30    SBS31    SBS32    SBS33    SBS34    SBS35
#> Sig1 0.251169 0.176705 0.217281 0.190168 0.198065 0.291670 0.326219 0.167234
#> Sig2 0.306440 0.073145 0.084439 0.077665 0.067240 0.189799 0.234887 0.071536
#> Sig3 0.546324 0.318320 0.348499 0.337970 0.339405 0.434520 0.481561 0.327852
#>         SBS36    SBS37    SBS38    SBS39    SBS40    SBS41    SBS42    SBS43
#> Sig1 0.222163 0.155024 0.321314 0.131135 0.133988 0.141395 0.180887 0.219344
#> Sig2 0.118089 0.069663 0.219050 0.056606 0.048896 0.073695 0.067748 0.138797
#> Sig3 0.360991 0.312379 0.475377 0.302408 0.295547 0.319400 0.330777 0.377444
#>         SBS44    SBS45    SBS46    SBS47    SBS48    SBS49    SBS50    SBS51
#> Sig1 0.182577 0.269907 0.176080 0.205584 0.674224 0.460338 0.197968 0.187056
#> Sig2 0.059625 0.166457 0.073658 0.114100 0.581815 0.360637 0.089241 0.088074
#> Sig3 0.310896 0.418861 0.306514 0.359080 0.817393 0.614301 0.347555 0.331071
#>         SBS52    SBS53    SBS54    SBS55    SBS56    SBS57    SBS58    SBS59
#> Sig1 0.378140 0.250307 0.239819 0.256845 0.345687 0.165649 0.185579 0.405001
#> Sig2 0.285849 0.161257 0.133842 0.174820 0.251045 0.089001 0.081505 0.317869
#> Sig3 0.484724 0.407754 0.362176 0.417739 0.487844 0.345214 0.342362 0.569894
#>         SBS60    SBS84    SBS85    SBS86    SBS87    SBS88    SBS89    SBS90
#> Sig1 0.571066 0.204442 0.175558 0.171859 0.180572 0.179788 0.141616 0.305934
#> Sig2 0.490093 0.089916 0.089645 0.086394 0.032164 0.109855 0.051612 0.231820
#> Sig3 0.735723 0.343883 0.335085 0.330248 0.296362 0.355069 0.302726 0.473481
#> 
#> attr(,"class")
#> [1] "similarity" "list"      

## Remove some components
## in similarity calculation
s5 <- get_sig_similarity(sig2,
  Ref = sig2,
  pattern_to_rm = c("T[T>G]C", "T[T>G]G", "T[T>G]T")
)
#> -Comparing against Custom signatures/profiles
#> ---------------------------------------------
#> --Found Sig1 most similar to Sig1 [similarity: 1]
#> --Found Sig2 most similar to Sig2 [similarity: 1]
#> --Found Sig3 most similar to Sig3 [similarity: 1]
#> ------------------------------------
#> Return result invisiblely.
s5
#> $similarity
#>      Sig1 Sig2 Sig3
#> Sig1    1    0    0
#> Sig2    0    1    0
#> Sig3    0    0    1
#> 
#> $aetiology_db
#> [1] NA
#> 
#> $best_match
#> $best_match$Sig1
#> $best_match$Sig1$aetiology
#> [1] NA
#> 
#> $best_match$Sig1$best_match
#> [1] "Best match: Sig1 [similarity: 1]"
#> 
#> 
#> $best_match$Sig2
#> $best_match$Sig2$aetiology
#> [1] NA
#> 
#> $best_match$Sig2$best_match
#> [1] "Best match: Sig2 [similarity: 1]"
#> 
#> 
#> $best_match$Sig3
#> $best_match$Sig3$aetiology
#> [1] NA
#> 
#> $best_match$Sig3$best_match
#> [1] "Best match: Sig3 [similarity: 1]"
#> 
#> 
#> 
#> $rss
#>          Sig1     Sig2     Sig3
#> Sig1 0.000000 0.176481 0.416715
#> Sig2 0.176481 0.000000 0.347079
#> Sig3 0.416715 0.347079 0.000000
#> 
#> attr(,"class")
#> [1] "similarity" "list"      

## Same to DBS and ID signatures
x1 <- get_sig_db("DBS_hg19")
x2 <- get_sig_db("DBS_hg38")
s6 <- get_sig_similarity(x1$db, x2$db)
#> -Comparing against Custom signatures/profiles
#> ---------------------------------------------
#> --Found DBS1 most similar to DBS1 [similarity: 1]
#> --Found DBS2 most similar to DBS2 [similarity: 1]
#> --Found DBS3 most similar to DBS3 [similarity: 1]
#> --Found DBS4 most similar to DBS4 [similarity: 1]
#> --Found DBS5 most similar to DBS5 [similarity: 1]
#> --Found DBS6 most similar to DBS6 [similarity: 1]
#> --Found DBS7 most similar to DBS7 [similarity: 1]
#> --Found DBS8 most similar to DBS8 [similarity: 1]
#> --Found DBS9 most similar to DBS9 [similarity: 1]
#> --Found DBS10 most similar to DBS10 [similarity: 1]
#> --Found DBS11 most similar to DBS11 [similarity: 1]
#> ------------------------------------
#> Return result invisiblely.
s6
#> $similarity
#>        DBS1  DBS2  DBS3  DBS4  DBS5  DBS6  DBS7  DBS8  DBS9 DBS10 DBS11
#> DBS1  1.000 0.004 0.006 0.002 0.002 0.004 0.011 0.018 0.012 0.059 0.905
#> DBS2  0.004 1.000 0.040 0.012 0.011 0.028 0.007 0.112 0.029 0.043 0.018
#> DBS3  0.006 0.039 1.000 0.077 0.074 0.095 0.070 0.083 0.121 0.635 0.042
#> DBS4  0.002 0.012 0.077 1.000 0.060 0.106 0.070 0.115 0.098 0.045 0.013
#> DBS5  0.002 0.011 0.074 0.060 1.000 0.077 0.025 0.086 0.032 0.031 0.026
#> DBS6  0.004 0.028 0.095 0.106 0.077 1.000 0.041 0.090 0.093 0.040 0.049
#> DBS7  0.011 0.007 0.070 0.070 0.025 0.041 1.000 0.171 0.101 0.069 0.045
#> DBS8  0.018 0.111 0.083 0.115 0.086 0.090 0.171 1.000 0.241 0.071 0.049
#> DBS9  0.012 0.029 0.122 0.098 0.032 0.093 0.101 0.242 1.000 0.081 0.192
#> DBS10 0.058 0.042 0.629 0.044 0.031 0.039 0.068 0.070 0.080 1.000 0.090
#> DBS11 0.903 0.018 0.042 0.013 0.026 0.050 0.045 0.049 0.193 0.092 1.000
#> 
#> $aetiology_db
#> [1] NA
#> 
#> $best_match
#> $best_match$DBS1
#> $best_match$DBS1$aetiology
#> [1] NA
#> 
#> $best_match$DBS1$best_match
#> [1] "Best match: DBS1 [similarity: 1]"
#> 
#> 
#> $best_match$DBS2
#> $best_match$DBS2$aetiology
#> [1] NA
#> 
#> $best_match$DBS2$best_match
#> [1] "Best match: DBS2 [similarity: 1]"
#> 
#> 
#> $best_match$DBS3
#> $best_match$DBS3$aetiology
#> [1] NA
#> 
#> $best_match$DBS3$best_match
#> [1] "Best match: DBS3 [similarity: 1]"
#> 
#> 
#> $best_match$DBS4
#> $best_match$DBS4$aetiology
#> [1] NA
#> 
#> $best_match$DBS4$best_match
#> [1] "Best match: DBS4 [similarity: 1]"
#> 
#> 
#> $best_match$DBS5
#> $best_match$DBS5$aetiology
#> [1] NA
#> 
#> $best_match$DBS5$best_match
#> [1] "Best match: DBS5 [similarity: 1]"
#> 
#> 
#> $best_match$DBS6
#> $best_match$DBS6$aetiology
#> [1] NA
#> 
#> $best_match$DBS6$best_match
#> [1] "Best match: DBS6 [similarity: 1]"
#> 
#> 
#> $best_match$DBS7
#> $best_match$DBS7$aetiology
#> [1] NA
#> 
#> $best_match$DBS7$best_match
#> [1] "Best match: DBS7 [similarity: 1]"
#> 
#> 
#> $best_match$DBS8
#> $best_match$DBS8$aetiology
#> [1] NA
#> 
#> $best_match$DBS8$best_match
#> [1] "Best match: DBS8 [similarity: 1]"
#> 
#> 
#> $best_match$DBS9
#> $best_match$DBS9$aetiology
#> [1] NA
#> 
#> $best_match$DBS9$best_match
#> [1] "Best match: DBS9 [similarity: 1]"
#> 
#> 
#> $best_match$DBS10
#> $best_match$DBS10$aetiology
#> [1] NA
#> 
#> $best_match$DBS10$best_match
#> [1] "Best match: DBS10 [similarity: 1]"
#> 
#> 
#> $best_match$DBS11
#> $best_match$DBS11$aetiology
#> [1] NA
#> 
#> $best_match$DBS11$best_match
#> [1] "Best match: DBS11 [similarity: 1]"
#> 
#> 
#> 
#> $rss
#>           DBS1     DBS2     DBS3     DBS4     DBS5     DBS6     DBS7     DBS8
#> DBS1  0.000001 1.274845 0.811053 0.998088 1.008771 0.872793 0.819496 0.880769
#> DBS2  1.275535 0.000000 0.600070 0.793207 0.803739 0.662650 0.624209 0.630502
#> DBS3  0.812674 0.601145 0.000002 0.315999 0.326705 0.194041 0.151426 0.212984
#> DBS4  0.999962 0.794450 0.316209 0.000008 0.502262 0.359550 0.328095 0.371142
#> DBS5  1.010232 0.804571 0.326510 0.501858 0.000000 0.379854 0.351277 0.391872
#> DBS6  0.874332 0.663593 0.193905 0.359186 0.379901 0.000001 0.215563 0.266499
#> DBS7  0.821218 0.625222 0.151482 0.327914 0.351505 0.215728 0.000001 0.202510
#> DBS8  0.882588 0.632074 0.213066 0.371010 0.392142 0.266719 0.202557 0.000000
#> DBS9  0.787424 0.583931 0.112796 0.291311 0.316967 0.174142 0.125736 0.164031
#> DBS10 0.849672 0.666045 0.090767 0.392899 0.408140 0.274887 0.220104 0.282443
#> DBS11 0.301043 0.668610 0.204577 0.394314 0.399372 0.261241 0.214536 0.278222
#>           DBS9    DBS10    DBS11
#> DBS1  0.786055 0.845623 0.298327
#> DBS2  0.583217 0.662877 0.668733
#> DBS3  0.113112 0.088468 0.205699
#> DBS4  0.291864 0.391130 0.395667
#> DBS5  0.317090 0.405990 0.400337
#> DBS6  0.174335 0.272804 0.262280
#> DBS7  0.126095 0.218165 0.215727
#> DBS8  0.164482 0.280516 0.279475
#> DBS9  0.000001 0.184824 0.158164
#> DBS10 0.187116 0.000010 0.261218
#> DBS11 0.157264 0.257965 0.000004
#> 
#> attr(,"class")
#> [1] "similarity" "list"