Get Tidy Parameter from Flexmix Model

When users derive copy number features, it is useful to know the parameters of the fit components, including mean, sd and coefficient of variation. This function is used by sig_derive function and exported to users for extra usage.

get_tidy_parameter(x)

Arguments

x
a flexmix object or a list of flexmix objects.

Value

a tibble.

Examples

# Load copy number prepare object
load(system.file("extdata", "toy_copynumber_prepare.RData",
  package = "sigminer", mustWork = TRUE
))
# Get all parameters
get_tidy_parameter(cn_prepare$components)

#> # A tibble: 19 x 7
#>    feature     components   dist      mean      sd           cv n_obs
#>    <chr>       <chr>        <chr>    <dbl>   <dbl>        <dbl> <dbl>
#>  1 bp10MB      bp10MB1      pois  3.66e-12 1.91e-6 522870.       2839
#>  2 bp10MB      bp10MB2      pois  9.24e- 1 9.61e-1      1.04      161
#>  3 bpchrarm    bpchrarm1    pois  1.22e- 6 1.11e-3    904.        328
#>  4 bpchrarm    bpchrarm2    pois  1.21e+ 0 1.10e+0      0.909     102
#>  5 bpchrarm    bpchrarm3    pois  5.54e+ 0 2.35e+0      0.425      10
#>  6 changepoint changepoint1 norm  9.01e- 1 7.10e-1      0.787     236
#>  7 changepoint changepoint2 norm  5.83e+ 0 3.40e+0      0.583      11
#>  8 copynumber  copynumber1  norm  1.90e+ 0 5.63e-1      0.296     418
#>  9 copynumber  copynumber2  norm  3.55e+ 0 2.89e+0      0.812      49
#> 10 osCN        osCN1        pois  7.68e- 2 2.77e-1      3.61       62
#> 11 osCN        osCN2        pois  4.55e+ 0 2.13e+0      0.469       5
#> 12 segsize     segsize1     norm  3.19e+ 5 2.09e+5      0.654      60
#> 13 segsize     segsize2     norm  5.53e+ 6 3.95e+6      0.714      79
#> 14 segsize     segsize3     norm  2.39e+ 7 1.06e+7      0.442      95
#> 15 segsize     segsize4     norm  7.00e+ 7 2.36e+7      0.337     121
#> 16 segsize     segsize5     norm  7.07e+ 7 2.37e+7      0.335      25
#> 17 segsize     segsize6     norm  1.34e+ 8 1.16e+6      0.00869    23
#> 18 segsize     segsize7     norm  1.55e+ 8 2.69e+7      0.174      56
#> 19 segsize     segsize8     norm  2.43e+ 8 1.40e+6      0.00576     8

# Get parameters for segsize feature
get_tidy_parameter(cn_prepare$components$segsize)

#> # A tibble: 8 x 5
#>   dist        mean        sd      cv n_obs
#>   <chr>      <dbl>     <dbl>   <dbl> <dbl>
#> 1 norm     318820.   208667. 0.654      60
#> 2 norm    5534293.  3953527. 0.714      79
#> 3 norm   23899505. 10561057. 0.442      95
#> 4 norm   69998482. 23578520. 0.337     121
#> 5 norm   70744732. 23706674. 0.335      25
#> 6 norm  133701749.  1161481. 0.00869    23
#> 7 norm  154983944. 26923555. 0.174      56
#> 8 norm  243283418.  1401122. 0.00576     8