Read absolute copy number profile

Read absolute copy number profile for preparing CNV signature analysis.

read_copynumber(input, pattern = NULL, ignore_case = FALSE,
  seg_cols = c("Chromosome", "Start.bp", "End.bp", "modal_cn"),
  samp_col = "sample", use_all = FALSE, min_segnum = 0,
  genome_build = c("hg19", "hg38"), genome_measure = c("called", "wg"),
  clinical_data = NULL, complement = TRUE, verbose = FALSE, ...)

Arguments

input
a data.frame or a file or a directory contains copy number profile.

pattern
an optional regular expression used to select part of files if input is a directory, more detail please see list.files function.

ignore_case
logical. Should pattern-matching be case-insensitive?

seg_cols
four characters used to specify chromosome, start position, end position and copy number value in input, respectively. Default use names from ABSOLUTE calling result.

samp_col
a character used to specify the sample column name. If input is a directory and cannot find samp_col, sample names will use file names (set this parameter to NULL is recommended in this case).

use_all
default is FALSE. If True, use all columns from raw input.

min_segnum
minimal number of copy number segments within a sample.

genome_build
genome build version, should be 'hg19' or 'hg38'.

genome_measure
default is 'called', can be 'wg' or 'called'. Set 'called' will use autosomo called segments size to compute total size for CNA burden calculation, this option is useful for WES and target sequencing. Set 'wg' will use autosome size from genome build, this option is useful for WGS, SNP etc..

clinical_data
a data.frame representing clinical data associated with each sample in copy number profile.

complement
if TRUE, complement chromosome does not show in input data with normal copy 2 and force use_all to FALSE (no matter what user input).

verbose
print extra messages.

...
other parameters pass to data.table::fread()

Value

a CopyNumber object

See also

Examples

# NOT RUN {
# Load toy dataset of absolute copynumber profile
load(system.file("extdata", "toy_segTab.RData",
  package = "sigminer", mustWork = TRUE
))
cn <- read_copynumber(segTabs,
  seg_cols = c("chromosome", "start", "end", "segVal"),
  genome_build = "hg19", complement = FALSE, verbose = TRUE
)
# }