Read Absolute Copy Number Profile

Read absolute copy number profile for preparing CNV signature analysis. See detail part of sig_tally() to see how to handle sex to get correct summary.

read_copynumber(
  input,
  pattern = NULL,
  ignore_case = FALSE,
  seg_cols = c("Chromosome", "Start.bp", "End.bp", "modal_cn"),
  samp_col = "sample",
  add_loh = FALSE,
  loh_min_len = 10000,
  loh_min_frac = 0.05,
  join_adj_seg = TRUE,
  skip_annotation = FALSE,
  use_all = add_loh,
  min_segnum = 0L,
  max_copynumber = 20L,
  genome_build = c("hg19", "hg38", "T2T", "mm10", "mm9", "ce11"),
  genome_measure = c("called", "wg"),
  complement = FALSE,
  ...
)

Arguments

input

a data.frame or a file or a directory contains copy number profile.

pattern

an optional regular expression used to select part of files if input is a directory, more detail please see list.files() function.

ignore_case

logical. Should pattern-matching be case-insensitive?

seg_cols

four strings used to specify chromosome, start position, end position and copy number value in input, respectively. Default use names from ABSOLUTE calling result.

samp_col

a character used to specify the sample column name. If input is a directory and cannot find samp_col, sample names will use file names (set this parameter to NULL is recommended in this case).

add_loh

if TRUE, add LOH labels to segments. NOTE a column 'minor_cn' must exist to indicate minor allele copy number value. Sex chromosome will not be labeled.

loh_min_len

The length cut-off for labeling a segment as 'LOH'. Default is 10Kb.

loh_min_frac

When join_adj_seg set to TRUE, only the length fraction of LOH region is larger than this value will be labeled as 'LOH'. Default is 30%.

join_adj_seg

if TRUE (default), join adjacent segments with same copy number value. This is helpful for precisely count the number of breakpoint. When set use_all=TRUE, the mean function will be applied to extra numeric columns and unique string columns will be pasted by comma for joined records.

skip_annotation

if TRUE, skip annotation step, it may affect some analysis and visualization functionality, but speed up reading data.

use_all

default is FALSE. If True, use all columns from raw input.

min_segnum

minimal number of copy number segments within a sample.

max_copynumber

bigger copy number within a sample will be reset to this value.

genome_build

genome build version, should be 'hg19', 'hg38', 'mm9' or 'mm10'.

genome_measure

default is 'called', can be 'wg' or 'called'. Set 'called' will use called segments size to compute total size for CNA burden calculation, this option is useful for WES and target sequencing. Set 'wg' will use autosome size from genome build, this option is useful for WGS, SNP etc..

complement

if TRUE, complement chromosome (except 'Y') does not show in input data with normal copy 2.

...

other parameters pass to data.table::fread()

Value

a CopyNumber object.

See also

Author

Shixiang Wang w_shixiang@163.com

Examples

# Load toy dataset of absolute copynumber profile
load(system.file("extdata", "toy_segTab.RData",
  package = "sigminer", mustWork = TRUE
))

# \donttest{
cn <- read_copynumber(segTabs,
  seg_cols = c("chromosome", "start", "end", "segVal"),
  genome_build = "hg19", complement = FALSE
)
#> ℹ [2025-11-30 14:25:05.844451]: Started.
#> ℹ [2025-11-30 14:25:05.846315]: Genome build  : hg19.
#> ℹ [2025-11-30 14:25:05.84789]: Genome measure: called.
#> ✔ [2025-11-30 14:25:05.852313]: Chromosome size database for build obtained.
#> ℹ [2025-11-30 14:25:05.853929]: Reading input.
#> ✔ [2025-11-30 14:25:05.855543]: A data frame as input detected.
#> ✔ [2025-11-30 14:25:05.85721]: Column names checked.
#> ✔ [2025-11-30 14:25:05.858824]: Column order set.
#> ✔ [2025-11-30 14:25:05.862154]: Chromosomes unified.
#> ✔ [2025-11-30 14:25:05.866184]: Data imported.
#> ℹ [2025-11-30 14:25:05.867784]: Segments info:
#> ℹ [2025-11-30 14:25:05.869392]:     Keep - 467
#> ℹ [2025-11-30 14:25:05.870953]:     Drop - 0
#> ✔ [2025-11-30 14:25:05.872851]: Segments sorted.
#> ℹ [2025-11-30 14:25:05.874424]: Joining adjacent segments with same copy number value. Be patient...
#> ✔ [2025-11-30 14:25:05.957191]: 400 segments left after joining.
#> ✔ [2025-11-30 14:25:05.959095]: Segmental table cleaned.
#> ℹ [2025-11-30 14:25:05.960698]: Annotating.
#> ✔ [2025-11-30 14:25:05.974678]: Annotation done.
#> ℹ [2025-11-30 14:25:05.976339]: Summarizing per sample.
#> ✔ [2025-11-30 14:25:05.987448]: Summarized.
#> ℹ [2025-11-30 14:25:05.989064]: Generating CopyNumber object.
#> ✔ [2025-11-30 14:25:05.990848]: Generated.
#> ℹ [2025-11-30 14:25:05.992558]: Validating object.
#> ✔ [2025-11-30 14:25:05.994158]: Done.
#> ℹ [2025-11-30 14:25:05.995739]: 0.151 secs elapsed.
cn
#> An object of class CopyNumber 
#> =============================
#>                           sample n_of_seg n_of_cnv n_of_amp n_of_del n_of_vchr
#>                           <char>    <int>    <int>    <int>    <int>     <int>
#>  1: TCGA-DF-A2KN-01A-11D-A17U-01       33        6        5        1         4
#>  2: TCGA-19-2621-01B-01D-0911-01       33        8        5        3         5
#>  3: TCGA-B6-A0X5-01A-21D-A107-01       28        8        4        4         2
#>  4: TCGA-A8-A07S-01A-11D-A036-01       38       11        2        9         4
#>  5: TCGA-26-6174-01A-21D-1842-01       43       13        8        5         8
#>  6: TCGA-CV-7432-01A-11D-2128-01       40       16        7        9         9
#>  7: TCGA-06-0644-01A-02D-0310-01       46       19        5       14         8
#>  8: TCGA-A5-A0G2-01A-11D-A042-01       39       21        5       16        10
#>  9: TCGA-99-7458-01A-11D-2035-01       48       26       10       16        13
#> 10: TCGA-05-4417-01A-22D-1854-01       52       37       33        4        17
#>     cna_burden
#>          <num>
#>  1:      0.000
#>  2:      0.099
#>  3:      0.087
#>  4:      0.112
#>  5:      0.119
#>  6:      0.198
#>  7:      0.165
#>  8:      0.393
#>  9:      0.318
#> 10:      0.654
cn_subset <- subset(cn, sample == "TCGA-DF-A2KN-01A-11D-A17U-01")

# Add LOH
set.seed(1234)
segTabs$minor_cn <- sample(c(0, 1), size = nrow(segTabs), replace = TRUE)
cn <- read_copynumber(segTabs,
  seg_cols = c("chromosome", "start", "end", "segVal"),
  genome_measure = "wg", complement = TRUE, add_loh = TRUE
)
#> ℹ [2025-11-30 14:25:06.010834]: Started.
#> ℹ [2025-11-30 14:25:06.012478]: Genome build  : hg19.
#> ℹ [2025-11-30 14:25:06.014039]: Genome measure: wg.
#> ℹ [2025-11-30 14:25:06.01565]: When add_loh is TRUE, use_all is forced to TRUE.
#> Please drop columns you don't want to keep before reading.
#> ✔ [2025-11-30 14:25:06.019599]: Chromosome size database for build obtained.
#> ℹ [2025-11-30 14:25:06.021248]: Reading input.
#> ✔ [2025-11-30 14:25:06.028954]: A data frame as input detected.
#> ✔ [2025-11-30 14:25:06.031399]: Column names checked.
#> ✔ [2025-11-30 14:25:06.033215]: Column order set.
#> ✔ [2025-11-30 14:25:06.036538]: Chromosomes unified.
#> ✔ [2025-11-30 14:25:06.050504]: Value 2 (normal copy) filled to uncalled chromosomes.
#> ✔ [2025-11-30 14:25:06.054363]: Data imported.
#> ℹ [2025-11-30 14:25:06.055968]: Segments info:
#> ℹ [2025-11-30 14:25:06.057596]:     Keep - 477
#> ℹ [2025-11-30 14:25:06.059175]:     Drop - 0
#> ✔ [2025-11-30 14:25:06.061082]: Segments sorted.
#> ℹ [2025-11-30 14:25:06.06264]: Adding LOH labels...
#> ℹ [2025-11-30 14:25:06.064766]: Joining adjacent segments with same copy number value. Be patient...
#> ✔ [2025-11-30 14:25:06.188114]: 410 segments left after joining.
#> ✔ [2025-11-30 14:25:06.190287]: Segmental table cleaned.
#> ℹ [2025-11-30 14:25:06.191985]: Annotating.
#> ✔ [2025-11-30 14:25:06.206178]: Annotation done.
#> ℹ [2025-11-30 14:25:06.207818]: Summarizing per sample.
#> ✔ [2025-11-30 14:25:06.221452]: Summarized.
#> ℹ [2025-11-30 14:25:06.223144]: Generating CopyNumber object.
#> ✔ [2025-11-30 14:25:06.224955]: Generated.
#> ℹ [2025-11-30 14:25:06.226552]: Validating object.
#> ✔ [2025-11-30 14:25:06.228141]: Done.
#> ℹ [2025-11-30 14:25:06.229746]: 0.219 secs elapsed.
# Use tally method "S" (Steele et al.)
tally_s <- sig_tally(cn, method = "S")
#> ℹ [2025-11-30 14:25:06.231884]: Started.
#> ℹ [2025-11-30 14:25:06.235482]: When you use method 'S', please make sure you have set 'join_adj_seg' to FALSE and 'add_loh' to TRUE in 'read_copynumber() in the previous step!
#> ✔ [2025-11-30 14:25:06.456921]: Matrix generated.
#> ℹ [2025-11-30 14:25:06.459281]: 0.227 secs elapsed.

tab_file <- system.file("extdata", "metastatic_tumor.segtab.txt",
  package = "sigminer", mustWork = TRUE
)
cn2 <- read_copynumber(tab_file)
#> ℹ [2025-11-30 14:25:06.462441]: Started.
#> ℹ [2025-11-30 14:25:06.464114]: Genome build  : hg19.
#> ℹ [2025-11-30 14:25:06.465684]: Genome measure: called.
#> ✔ [2025-11-30 14:25:06.470135]: Chromosome size database for build obtained.
#> ℹ [2025-11-30 14:25:06.471721]: Reading input.
#> ✔ [2025-11-30 14:25:06.473284]: A file as input detected.
#> ✔ [2025-11-30 14:25:06.475599]: Column names checked.
#> ✔ [2025-11-30 14:25:06.477364]: Column order set.
#> ✔ [2025-11-30 14:25:06.480472]: Chromosomes unified.
#> ✔ [2025-11-30 14:25:06.484475]: Data imported.
#> ℹ [2025-11-30 14:25:06.486061]: Segments info:
#> ℹ [2025-11-30 14:25:06.487596]:     Keep - 201
#> ℹ [2025-11-30 14:25:06.489137]:     Drop - 0
#> ✔ [2025-11-30 14:25:06.490946]: Segments sorted.
#> ℹ [2025-11-30 14:25:06.492526]: Joining adjacent segments with same copy number value. Be patient...
#> ✔ [2025-11-30 14:25:06.509899]: 180 segments left after joining.
#> ✔ [2025-11-30 14:25:06.511667]: Segmental table cleaned.
#> ℹ [2025-11-30 14:25:06.513233]: Annotating.
#> ✔ [2025-11-30 14:25:06.526558]: Annotation done.
#> ℹ [2025-11-30 14:25:06.528202]: Summarizing per sample.
#> ✔ [2025-11-30 14:25:06.538819]: Summarized.
#> ℹ [2025-11-30 14:25:06.540427]: Generating CopyNumber object.
#> ✔ [2025-11-30 14:25:06.542213]: Generated.
#> ℹ [2025-11-30 14:25:06.543761]: Validating object.
#> ✔ [2025-11-30 14:25:06.545309]: Done.
#> ℹ [2025-11-30 14:25:06.546874]: 0.084 secs elapsed.
cn2
#> An object of class CopyNumber 
#> =============================
#>              sample n_of_seg n_of_cnv n_of_amp n_of_del n_of_vchr cna_burden
#>              <char>    <int>    <int>    <int>    <int>     <int>      <num>
#> 1: metastatic_tumor      180      110       92       18        21      0.361
# }