Read Absolute Copy Number Profile

Read absolute copy number profile for preparing CNV signature analysis.

read_copynumber(
  input,
  pattern = NULL,
  ignore_case = FALSE,
  seg_cols = c("Chromosome", "Start.bp", "End.bp", "modal_cn"),
  samp_col = "sample",
  join_adj_seg = TRUE,
  use_all = FALSE,
  min_segnum = 0L,
  max_copynumber = 20L,
  genome_build = c("hg19", "hg38"),
  genome_measure = c("called", "wg"),
  complement = TRUE,
  verbose = FALSE,
  ...
)

Arguments

input
a data.frame or a file or a directory contains copy number profile.

pattern
an optional regular expression used to select part of files if input is a directory, more detail please see list.files function.

ignore_case
logical. Should pattern-matching be case-insensitive?

seg_cols
four characters used to specify chromosome, start position, end position and copy number value in input, respectively. Default use names from ABSOLUTE calling result.

samp_col
a character used to specify the sample column name. If input is a directory and cannot find samp_col, sample names will use file names (set this parameter to NULL is recommended in this case).

join_adj_seg
if TRUE (default), join adjacent segments with same copy number value. This is helpful for precisely count the number of breakpoint.

use_all
default is FALSE. If True, use all columns from raw input.

min_segnum
minimal number of copy number segments within a sample.

max_copynumber
bigger copy number within a sample will be reset to this value.

genome_build
genome build version, should be 'hg19' or 'hg38'.

genome_measure
default is 'called', can be 'wg' or 'called'. Set 'called' will use called segments size to compute total size for CNA burden calculation, this option is useful for WES and target sequencing. Set 'wg' will use autosome size from genome build, this option is useful for WGS, SNP etc..

complement
if TRUE, complement chromosome (except 'Y') does not show in input data with normal copy 2 and force use_all to FALSE (no matter what user input).

verbose
print extra messages.

...
other parameters pass to data.table::fread()

Value

a CopyNumber object

See also

Examples

# Load toy dataset of absolute copynumber profile
load(system.file("extdata", "toy_segTab.RData",
  package = "sigminer", mustWork = TRUE
))
cn <- read_copynumber(segTabs,
  seg_cols = c("chromosome", "start", "end", "segVal"),
  genome_build = "hg19", complement = FALSE, verbose = TRUE
)

#> Treating input as a data frame...

#> Checking chromosome names...

#> Segments info:

#>     Keep - 467

#>   Filter - 0

#> Joining adjacent segments with same copy number value...

#> Anotating...

#> 
#> Summarizing per sample...

#> Done!

cn

#> An object of class CopyNumber 
#> =============================
#>                           sample n_of_seg n_of_cnv n_of_amp n_of_del n_of_vchr
#>  1: TCGA-DF-A2KN-01A-11D-A17U-01       33        6        5        1         4
#>  2: TCGA-19-2621-01B-01D-0911-01       33        8        5        3         5
#>  3: TCGA-B6-A0X5-01A-21D-A107-01       28        8        4        4         2
#>  4: TCGA-A8-A07S-01A-11D-A036-01       38       11        2        9         4
#>  5: TCGA-26-6174-01A-21D-1842-01       43       13        8        5         8
#>  6: TCGA-CV-7432-01A-11D-2128-01       40       16        7        9         9
#>  7: TCGA-06-0644-01A-02D-0310-01       46       19        5       14         8
#>  8: TCGA-A5-A0G2-01A-11D-A042-01       39       21        5       16        10
#>  9: TCGA-99-7458-01A-11D-2035-01       48       26       10       16        13
#> 10: TCGA-05-4417-01A-22D-1854-01       52       37       33        4        17
#>     cna_burden
#>  1:      0.000
#>  2:      0.099
#>  3:      0.087
#>  4:      0.112
#>  5:      0.119
#>  6:      0.198
#>  7:      0.165
#>  8:      0.393
#>  9:      0.318
#> 10:      0.654

tab_file <- system.file("extdata", "metastatic_tumor.segtab.txt",
  package = "sigminer", mustWork = TRUE
)
cn2 <- read_copynumber(tab_file)
cn2

#> An object of class CopyNumber 
#> =============================
#>              sample n_of_seg n_of_cnv n_of_amp n_of_del n_of_vchr cna_burden
#> 1: metastatic_tumor      180      110       92       18        21      0.361