Score Copy Number Profile

Returns quantification of copy number profile and events including tandem duplication and Chromothripisis etc. Only copy number data from autosome is used here. Some of the quantification methods are rough, you use at your risk. You should do some extra work to check the result scores.

scoring(object, TD_size_cutoff = c(1000, 1e+05, 2e+06))

Arguments

object
a object of CopyNumber.

TD_size_cutoff
a length-3 numeric vector used to specify the start, midpoint, end segment size for determining tandem duplication size range, midpoint is used to split TD into short TD and long TD. Default is 1Kb to 100Kb for short TD, 100Kb to 2Mb for long TD.

Value

a data.table with following scores:

cnaBurden: CNA burden representing the altered genomic fraction as previously reported.

cnaLoad: CNA load representing the quantity of copy number alteration.

MACN: mean altered copy number (MACN) reflecting the property of altered copy number segments, calculated as $$MACN = \frac{\sum_{i} CN_i}{N_{cnv}}$$ where \(CN_i\) is the copy number of altered segment \(i\), \(N_{cnv}\) is the number of CNV.

weightedMACN: same as MACN but weighted with segment length. $$MACN_{weighted} = \frac{\sum_{i} (CN_i \times L_{i})}{ \sum_{i} L_{i} }$$ where \(L_{i}\) is the length of altered copy number segment \(i\).

Ploidy: ploidy, the formula is same as weightedMACN but using all copy number segments instead of altered copy number segments.

TDP: tandem duplication phenotype score from https://www.pnas.org/content/113/17/E2373, the threshold k in reference is omitted. $$TDP = - \frac{\sum_{chr} |TD_{obs}-TD_{exp}|}{TD_{total}}$$ where \(TD_{total}\) is the number of TD, \(TD_{obs}\) and \(TD_exp\) are observed number of TD and expected number of TD for each chromosome.

sTDP: TDP score for short TD.

lTDP: TDP score for long TD.

TD: tandem duplication score, TD represents segment with copy number greater than 2. $$TD = \frac{TD_{total}}{\sum_{chr} |TD_{obs}-TD_{exp}|+1}$$

sTD: TD score for short TD.

lTD: TD score for long TD.

Chromothripisis: according to reference http://dx.doi.org/10.1016/j.cell.2013.02.023, Chromothripsis frequently leads to massive loss of segments on the affected chromosome with segmental losses being interspersed with regions displaying normal (disomic) copy-number (e.g., copy-number states oscillating between copy-number = 1 and copy-number = 2), form tens to hundreds of locally clustered DNA rearrangements. Most of methods use both SV and CNV to infer chromothripsis, here we roughly quantify it with $$\sum_{chr}{N_{OsCN}^2}$$ where \(N_{OsCN}\) is the number of oscillating copy number pattern "2-1-2" for each chromosome.

Examples

# Load copy number object
load(system.file("extdata", "toy_copynumber.RData",
  package = "sigminer", mustWork = TRUE
))

scoring(cn)

#>                           sample    cnaBurden cnaLoad MACN weightedMACN Ploidy
#>  1: TCGA-05-4417-01A-22D-1854-01 0.6535735305    5.36 3.85         3.53   3.00
#>  2: TCGA-06-0644-01A-02D-0310-01 0.1651843925    4.32 1.47         2.10   2.02
#>  3: TCGA-19-2621-01B-01D-0911-01 0.0991005307    3.91 3.07         1.10   1.91
#>  4: TCGA-26-6174-01A-21D-1842-01 0.1186037707    3.81 2.08         2.60   2.07
#>  5: TCGA-99-7458-01A-11D-2035-01 0.3180319443    4.75 2.27         1.63   1.88
#>  6: TCGA-A5-A0G2-01A-11D-A042-01 0.3931655541    4.86 1.29         1.62   1.85
#>  7: TCGA-A8-A07S-01A-11D-A036-01 0.1123575193    3.58 1.45         2.17   2.02
#>  8: TCGA-B6-A0X5-01A-21D-A107-01 0.0872390204    3.32 1.89         1.94   1.99
#>  9: TCGA-CV-7432-01A-11D-2128-01 0.1976249853    4.25 2.39         2.02   2.00
#> 10: TCGA-DF-A2KN-01A-11D-A17U-01 0.0003297153    2.81 2.83         2.98   2.00
#>           TDP      sTDP      lTDP        TD       sTD       lTD Chromothripisis
#>  1: -1.818182       NaN -1.818182 0.4313725 0.0000000 0.4313725               0
#>  2: -1.909091       NaN -1.909091 0.3437500 0.0000000 0.3437500              13
#>  3: -1.818182       NaN -1.818182 0.4954955 0.0000000 0.4954955               0
#>  4: -1.545455 -1.909091 -1.636364 0.5840708 0.3437500 0.5445545               9
#>  5: -1.545455       NaN -1.545455 0.5840708 0.0000000 0.5840708              10
#>  6: -1.909091 -1.909091       NaN 0.3437500 0.3437500 0.0000000               6
#>  7:       NaN       NaN       NaN 0.0000000 0.0000000 0.0000000              34
#>  8: -1.909091       NaN -1.909091 0.4459459 0.0000000 0.4459459               0
#>  9: -1.909091       NaN -1.909091 0.3437500 0.0000000 0.3437500               7
#> 10: -1.727273 -1.818182 -1.818182 0.5188679 0.4313725 0.4647887               1