Score Copy Number Profile

Returns quantification of copy number profile and events including tandem duplication and Chromothripisis etc. Only copy number data from autosome is used here. Some of the quantification methods are rough, you use at your risk. You should do some extra work to check the result scores.

scoring(object, TD_size_cutoff = c(1000, 100000, 2000000), TD_cn_cutoff = Inf)

Arguments

object

a object of CopyNumber.

TD_size_cutoff

a length-3 numeric vector used to specify the start, midpoint, end segment size for determining tandem duplication size range, midpoint is used to split TD into short TD and long TD. Default is 1Kb to 100Kb for short TD, 100Kb to 2Mb for long TD.

TD_cn_cutoff

a number defining the maximum copy number of TD, default is Inf, i.e. no cutoff.

Value

a data.table with following scores:

cnaBurden: CNA burden representing the altered genomic fraction as previously reported.

cnaLoad: CNA load representing the quantity of copy number alteration.

MACN: mean altered copy number (MACN) reflecting the property of altered copy number segments, calculated as $$MACN = \frac{\sum_{i} CN_i}{N_{cnv}}$$ where \(CN_i\) is the copy number of altered segment \(i\), \(N_{cnv}\) is the number of CNV.

weightedMACN: same as MACN but weighted with segment length. $$MACN_{weighted} = \frac{\sum_{i} (CN_i \times L_{i})}{ \sum_{i} L_{i} }$$ where \(L_{i}\) is the length of altered copy number segment \(i\).

Ploidy: ploidy, the formula is same as weightedMACN but using all copy number segments instead of altered copy number segments.

TDP_pnas: tandem duplication phenotype score from https://www.pnas.org/doi/10.1073/pnas.1520010113, the threshold k in reference is omitted. $$TDP = - \frac{\sum_{chr} |TD_{obs}-TD_{exp}|}{TD_{total}}$$ where \(TD_{total}\) is the number of TD, \(TD_{obs}\) and \(TD_exp\) are observed number of TD and expected number of TD for each chromosome.

TDP: tandem duplication score used defined by our group work, TD represents segment with copy number greater than 2. $$TD = \frac{TD_{total}}{\sum_{chr} |TD_{obs}-TD_{exp}|+1}$$

sTDP: TDP score for short TD.

lTDP: TDP score for long TD.

TDP_size : TDP region size (Mb).

sTDP_size: sTDP region size (Mb).

lTDP_size: lTDP region size(Mb).

Chromoth_state: chromothripsis state score, according to reference doi:10.1016/j.cell.2013.02.023 , chromothripsis frequently leads to massive loss of segments on the affected chromosome with segmental losses being interspersed with regions displaying normal (disomic) copy-number (e.g., copy-number states oscillating between copy-number = 1 and copy-number = 2), form tens to hundreds of locally clustered DNA rearrangements. Most of methods use both SV and CNV to infer chromothripsis, here we roughly quantify it with $$\sum_{chr}{N_{OsCN}^2}$$ where \(N_{OsCN}\) is the number of oscillating copy number pattern "2-1-2" for each chromosome.

Examples

# Load copy number object
load(system.file("extdata", "toy_copynumber.RData",
  package = "sigminer", mustWork = TRUE
))

d <- scoring(cn)
d
#>                           sample cnaBurden cnaLoad  MACN weightedMACN Ploidy
#>                           <char>     <num>   <num> <num>        <num>  <num>
#>  1: TCGA-DF-A2KN-01A-11D-A17U-01     0.000    2.81  2.83         2.98   2.00
#>  2: TCGA-B6-A0X5-01A-21D-A107-01     0.087    3.17  2.00         1.94   1.99
#>  3: TCGA-19-2621-01B-01D-0911-01     0.099    3.17  4.62         1.10   1.91
#>  4: TCGA-A8-A07S-01A-11D-A036-01     0.112    3.58  1.45         2.17   2.02
#>  5: TCGA-26-6174-01A-21D-1842-01     0.119    3.81  2.08         2.60   2.07
#>  6: TCGA-06-0644-01A-02D-0310-01     0.165    4.32  1.47         2.10   2.02
#>  7: TCGA-CV-7432-01A-11D-2128-01     0.198    4.09  2.69         2.02   2.00
#>  8: TCGA-99-7458-01A-11D-2035-01     0.318    4.75  2.27         1.63   1.88
#>  9: TCGA-A5-A0G2-01A-11D-A042-01     0.393    4.46  1.29         1.62   1.85
#> 10: TCGA-05-4417-01A-22D-1854-01     0.654    5.25  3.89         3.53   3.00
#>     TD_count       TDP      sTDP      lTDP TDP_size sTDP_size lTDP_size
#>        <int>     <num>     <num>     <num>    <num>     <num>     <num>
#>  1:        5 0.5188679 0.4313725 0.4647887 0.790283  0.103883  0.686400
#>  2:        3 0.4459459 0.0000000 0.4459459 1.812822  0.000000  1.812822
#>  3:        5 0.4954955 0.0000000 0.4954955 4.189389  0.000000  4.189389
#>  4:        0 0.0000000 0.0000000 0.0000000 0.000000  0.000000  0.000000
#>  5:        6 0.5840708 0.3437500 0.5445545 2.285822  0.095838  2.189984
#>  6:        1 0.3437500 0.0000000 0.3437500 0.101453  0.000000  0.101453
#>  7:        1 0.3437500 0.0000000 0.3437500 0.846245  0.000000  0.846245
#>  8:        6 0.5840708 0.0000000 0.5840708 5.223909  0.000000  5.223909
#>  9:        1 0.3437500 0.3437500 0.0000000 0.091632  0.091632  0.000000
#> 10:        2 0.4313725 0.0000000 0.4313725 0.452862  0.000000  0.452862
#>      TDP_pnas Chromoth_state
#>         <num>          <num>
#>  1: -1.727273              1
#>  2: -1.909091              0
#>  3: -1.818182              3
#>  4:       NaN             34
#>  5: -1.545455              9
#>  6: -1.909091             13
#>  7: -1.909091              7
#>  8: -1.545455             10
#>  9: -1.909091              6
#> 10: -1.818182              0

d2 <- scoring(cn, TD_cn_cutoff = 4L)
d2
#>                           sample cnaBurden cnaLoad  MACN weightedMACN Ploidy
#>                           <char>     <num>   <num> <num>        <num>  <num>
#>  1: TCGA-DF-A2KN-01A-11D-A17U-01     0.000    2.81  2.83         2.98   2.00
#>  2: TCGA-B6-A0X5-01A-21D-A107-01     0.087    3.17  2.00         1.94   1.99
#>  3: TCGA-19-2621-01B-01D-0911-01     0.099    3.17  4.62         1.10   1.91
#>  4: TCGA-A8-A07S-01A-11D-A036-01     0.112    3.58  1.45         2.17   2.02
#>  5: TCGA-26-6174-01A-21D-1842-01     0.119    3.81  2.08         2.60   2.07
#>  6: TCGA-06-0644-01A-02D-0310-01     0.165    4.32  1.47         2.10   2.02
#>  7: TCGA-CV-7432-01A-11D-2128-01     0.198    4.09  2.69         2.02   2.00
#>  8: TCGA-99-7458-01A-11D-2035-01     0.318    4.75  2.27         1.63   1.88
#>  9: TCGA-A5-A0G2-01A-11D-A042-01     0.393    4.46  1.29         1.62   1.85
#> 10: TCGA-05-4417-01A-22D-1854-01     0.654    5.25  3.89         3.53   3.00
#>     TD_count       TDP      sTDP      lTDP TDP_size sTDP_size lTDP_size
#>        <int>     <num>     <num>     <num>    <num>     <num>     <num>
#>  1:        5 0.5188679 0.4313725 0.4647887 0.790283  0.103883  0.686400
#>  2:        3 0.4459459 0.0000000 0.4459459 1.812822  0.000000  1.812822
#>  3:        3 0.4647887 0.0000000 0.4647887 2.272032  0.000000  2.272032
#>  4:        0 0.0000000 0.0000000 0.0000000 0.000000  0.000000  0.000000
#>  5:        6 0.5840708 0.3437500 0.5445545 2.285822  0.095838  2.189984
#>  6:        1 0.3437500 0.0000000 0.3437500 0.101453  0.000000  0.101453
#>  7:        1 0.3437500 0.0000000 0.3437500 0.846245  0.000000  0.846245
#>  8:        4 0.5301205 0.0000000 0.5301205 2.555474  0.000000  2.555474
#>  9:        1 0.3437500 0.3437500 0.0000000 0.091632  0.091632  0.000000
#> 10:        0 0.0000000 0.0000000 0.0000000 0.000000  0.000000  0.000000
#>      TDP_pnas Chromoth_state
#>         <num>          <num>
#>  1: -1.727273              1
#>  2: -1.909091              0
#>  3: -1.818182              3
#>  4:       NaN             34
#>  5: -1.545455              9
#>  6: -1.909091             13
#>  7: -1.909091              7
#>  8: -1.636364             10
#>  9: -1.909091              6
#> 10:       NaN              0