Extract Signatures through the Automatic Relevance Determination Technique

A bayesian variant of NMF algorithm to enable optimal inferences for the number of signatures through the automatic relevance determination technique. This functions delevers highly interpretable and sparse representations for both signature profiles and attributions at a balance between data fitting and model complexity (this method may introduce more signatures than expected, especially for copy number signatures (thus I don't recommend you to use this feature to extract copy number signatures)). See detail part and references for more.

sig_auto_extract(
  nmf_matrix = NULL,
  result_prefix = "BayesNMF",
  destdir = tempdir(),
  method = c("L1W.L2H", "L1KL", "L2KL"),
  strategy = c("stable", "optimal", "ms"),
  ref_sigs = NULL,
  K0 = 25,
  nrun = 10,
  niter = 200000,
  tol = 0.0000001,
  cores = 1,
  optimize = FALSE,
  skip = FALSE,
  recover = FALSE
)

Arguments

nmf_matrix

a matrix used for NMF decomposition with rows indicate samples and columns indicate components.

result_prefix

prefix for result data files.

destdir

path to save data runs, default is tempdir().

method

default is "L1W.L2H", which uses an exponential prior for W and a half-normal prior for H (This method is used by PCAWG project, see reference #3). You can also use "L1KL" to set expoential priors for both W and H, and "L2KL" to set half-normal priors for both W and H. The latter two methods are originally implemented by SignatureAnalyzer software.

strategy

the selection strategy for returned data. Set 'stable' for getting optimal result from the most frequent K. Set 'optimal' for getting optimal result from all Ks. Set 'ms' for getting result with maximum mean cosine similarity with provided reference signatures. See ref_sigs option for details. If you want select other solution, please check get_bayesian_result.

ref_sigs

A Signature object or matrix or string for specifying reference signatures, only used when strategy = 'ms'. See Signature and sig_db options in get_sig_similarity for details.

K0

number of initial signatures.

nrun

number of independent simulations.

niter

the maximum number of iterations.

tol

tolerance for convergence.

cores

number of cpu cores to run NMF.

optimize

if TRUE, then refit the denovo signatures with QP method, see sig_fit.

skip

if TRUE, it will skip running a previous stored result. This can be used to extend run times, e.g. you try running 10 times firstly and then you want to extend it to 20 times.

recover

if TRUE, try to recover result from previous runs based on input result_prefix, destdir and nrun. This is pretty useful for reproducing result. Please use skip if you want to recover an unfinished job.

Value

a list with Signature class.

Details

There are three methods available in this function: "L1W.L2H", "L1KL" and "L2KL". They use different priors for the bayesian variant of NMF algorithm (see method parameter) written by reference #1 and implemented in SignatureAnalyzer software (reference #2).

I copied source code for the three methods from Broad Institute and supplementary files of reference #3, and wrote this higher function. It is more friendly for users to extract, visualize and analyze signatures by combining with other powerful functions in sigminer package. Besides, I implemented parallel computation to speed up the calculation process and a similar input and output structure like sig_extract().

References

Tan, Vincent YF, and Cédric Févotte. "Automatic relevance determination in nonnegative matrix factorization with the/spl beta/-divergence." IEEE Transactions on Pattern Analysis and Machine Intelligence 35.7 (2012): 1592-1605.

Kim, Jaegil, et al. "Somatic ERCC2 mutations are associated with a distinct genomic signature in urothelial tumors." Nature genetics 48.6 (2016): 600.

Alexandrov, Ludmil, et al. "The repertoire of mutational signatures in human cancer." BioRxiv (2018): 322859.

See also

Author

Shixiang Wang

Examples

# \donttest{
load(system.file("extdata", "toy_copynumber_tally_W.RData",
  package = "sigminer", mustWork = TRUE
))
res <- sig_auto_extract(cn_tally_W$nmf_matrix, result_prefix = "Test_copynumber", nrun = 1)
#> Select Run 1, which K = 2 as best solution.
# At default, all run files are stored in tempdir()
dir(tempdir(), pattern = "Test_copynumber")
#> [1] "Test_copynumber.1.rds"

laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read_maf(maf = laml.maf)
#> -Reading
#> -Validating
#> -Silent variants: 475 
#> -Summarizing
#> -Processing clinical data
#> --Missing clinical data
#> -Finished in 0.159s elapsed (0.201s cpu) 
mt_tally <- sig_tally(
  laml,
  ref_genome = "BSgenome.Hsapiens.UCSC.hg19",
  use_syn = TRUE
)
#> ℹ [2024-03-08 02:37:17.043402]: Started.
#> ℹ [2024-03-08 02:37:17.047416]: We would assume you marked all variants' position in + strand.
#> ✔ [2024-03-08 02:37:17.049053]: Reference genome loaded.
#> ✔ [2024-03-08 02:37:17.053121]: Variants from MAF object queried.
#> ✔ [2024-03-08 02:37:17.055398]: Chromosome names checked.
#> ✔ [2024-03-08 02:37:17.058443]: Sex chromosomes properly handled.
#> ✔ [2024-03-08 02:37:17.060012]: Only variants located in standard chromosomes (1:22, X, Y, M/MT) are kept.
#> ✔ [2024-03-08 02:37:17.06488]: Variant start and end position checked.
#> ✔ [2024-03-08 02:37:17.068601]: Variant data for matrix generation preprocessed.
#> ℹ [2024-03-08 02:37:17.069931]: SBS matrix generation - start.
#> ℹ [2024-03-08 02:37:17.071712]: Extracting 5' and 3' adjacent bases.
#> ℹ [2024-03-08 02:37:17.346889]: Extracting +/- 20bp around mutated bases for background C>T estimation.
#> ℹ [2024-03-08 02:37:17.678005]: Estimating APOBEC enrichment scores.
#> ℹ [2024-03-08 02:37:17.679862]: Performing one-way Fisher's test for APOBEC enrichment.
#> ✔ [2024-03-08 02:37:17.752696]: APOBEC related mutations are enriched in 3.297% of samples (APOBEC enrichment score > 2; 6 of 182 samples)
#> ℹ [2024-03-08 02:37:17.756081]: Creating SBS sample-by-component matrices.
#> ✔ [2024-03-08 02:37:17.760361]: SBS-6 matrix created.
#> ✔ [2024-03-08 02:37:17.765128]: SBS-96 matrix created.
#> ✔ [2024-03-08 02:37:17.778998]: SBS-1536 matrix created.
#> ℹ [2024-03-08 02:37:17.780383]: Return SBS-96 as major matrix.
#> ✔ [2024-03-08 02:37:17.782782]: Done.
#> ℹ [2024-03-08 02:37:17.784163]: 0.741 secs elapsed.

x <- sig_auto_extract(mt_tally$nmf_matrix,
  strategy = "ms", nrun = 3, ref_sigs = "legacy"
)
#> The follow samples dropped due to null catalogue:
#> 	TCGA-AB-2823, TCGA-AB-2840, TCGA-AB-2848, TCGA-AB-2909, TCGA-AB-2942
#> Solutions ordered by mean cosine similarity to references:
#> Run#2:K#2 Run#3:K#2 Run#1:K#3 
#> 0.7180000 0.6755000 0.6113333 
#> Details:
#> $`Run#1:K#3`
#>  COSMIC_1 COSMIC_11 COSMIC_16 
#>     0.959     0.457     0.418 
#> 
#> $`Run#2:K#2`
#>  COSMIC_1 COSMIC_10 
#>     0.911     0.525 
#> 
#> $`Run#3:K#2`
#>  COSMIC_1 COSMIC_19 
#>     0.960     0.391 
#> 
#> Select Run 2, which K = 2 as best solution.
x
#> $Signature
#>                  Sig1          Sig2
#> A[C>A]A  1.580849e+01  3.354537e-43
#> A[C>A]C  2.696742e+01 1.711302e-146
#> A[C>A]G  3.719645e+00  0.000000e+00
#> A[C>A]T 4.369948e-101  5.629483e+00
#> A[C>G]A  1.301876e+01  5.218037e-14
#> A[C>G]C  5.579467e+00  1.790773e-42
#> A[C>G]G  7.439289e+00 6.623960e-295
#> A[C>G]T  7.439289e+00  1.665300e-92
#> A[C>T]A  4.742547e+01  2.507544e-99
#> A[C>T]C  4.742547e+01 3.365275e-110
#> A[C>T]G  2.148095e+02  2.046864e-70
#> A[C>T]T  2.417769e+01 2.494609e-141
#> A[T>A]A 4.435226e-169  2.646557e-74
#> A[T>A]C  1.115893e+01 2.473039e-100
#> A[T>A]G 2.512097e-127  4.222112e+00
#> A[T>A]T  3.719645e+00 6.916129e-319
#> A[T>C]A  2.603751e+01  1.886028e-76
#> A[T>C]C  1.115893e+01  9.907872e-10
#> A[T>C]G  2.510759e+01  6.336350e-06
#> A[T>C]T  2.882725e+01  2.323630e-22
#> A[T>G]A  5.579467e+00  1.370245e-53
#> A[T>G]C  0.000000e+00  1.407371e+00
#> A[T>G]G  2.789733e+00 9.268124e-134
#> A[T>G]T  2.789733e+00 1.507417e-242
#> C[C>A]A  1.673840e+01  0.000000e+00
#> C[C>A]C 3.781832e-122  1.055528e+01
#> C[C>A]G  1.301876e+01 5.710966e-238
#> C[C>A]T  1.673840e+01  2.687738e-70
#> C[C>G]A  1.208884e+01  7.292559e-15
#> C[C>G]C  3.719645e+00  0.000000e+00
#> C[C>G]G  1.394862e+01  3.939853e-05
#> C[C>G]T  6.509378e+00  1.243146e-20
#> C[C>T]A  3.812636e+01 7.525394e-244
#> C[C>T]C  2.789733e+01 3.656930e-123
#> C[C>T]G  1.850063e+02  4.256959e+00
#> C[C>T]T  5.579467e+01  1.355036e-88
#> C[T>A]A 5.343519e-104  1.152078e-88
#> C[T>A]C  4.471475e+00  3.653185e+00
#> C[T>A]G  1.115893e+01 3.531131e-108
#> C[T>A]T  4.649556e+00  0.000000e+00
#> C[T>C]A  2.283856e-36  4.925797e+00
#> C[T>C]C  1.673840e+01 2.607810e-110
#> C[T>C]G 1.157591e-150  1.337002e+01
#> C[T>C]T  1.394867e+01  2.381950e-36
#> C[T>G]A  9.299111e-01  0.000000e+00
#> C[T>G]C 4.128422e-101  5.629483e+00
#> C[T>G]G  9.299111e+00  0.000000e+00
#> C[T>G]T  4.649556e+00 1.981648e-149
#> G[C>A]A  1.191639e+01  2.945243e+00
#> G[C>A]C  8.369200e+00 6.440726e-203
#> G[C>A]G  4.649556e+00  0.000000e+00
#> G[C>A]T 4.759607e-164  6.333168e+00
#> G[C>G]A  1.022902e+01 1.431919e-204
#> G[C>G]C  6.509378e+00 7.469539e-147
#> G[C>G]G  6.509378e+00  0.000000e+00
#> G[C>G]T  1.906852e-13  4.222112e+00
#> G[C>T]A 1.629196e-100  1.899950e+01
#> G[C>T]C  4.556565e+01  4.293728e-92
#> G[C>T]G  1.716289e+02  3.120988e+00
#> G[C>T]T  4.742547e+01  6.360692e-34
#> G[T>A]A 4.011672e-306  1.407371e+00
#> G[T>A]C 5.721564e-171  6.333168e+00
#> G[T>A]G  3.719645e+00 1.387479e-103
#> G[T>A]T  3.265340e-25  4.222112e+00
#> G[T>C]A  5.690453e-39  4.222112e+00
#> G[T>C]C 5.647862e-148  6.333168e+00
#> G[T>C]G  1.022902e+01 3.978697e-195
#> G[T>C]T 2.777612e-129  8.444224e+00
#> G[T>G]A  4.800309e-95  2.814741e+00
#> G[T>G]C  2.789733e+00 7.173285e-226
#> G[T>G]G  5.579467e+00  7.593900e-26
#> G[T>G]T  1.859822e+00  0.000000e+00
#> T[C>A]A  4.511174e+00  5.734199e+00
#> T[C>A]C  2.138796e+01 6.320816e-116
#> T[C>A]G  7.439289e+00 5.186619e-135
#> T[C>A]T  2.603751e+01 2.240206e-180
#> T[C>G]A  4.649556e+00 7.740270e-287
#> T[C>G]C  1.032515e-31  5.629483e+00
#> T[C>G]G  1.022902e+01 1.902746e-259
#> T[C>G]T 3.475291e-185  6.333168e+00
#> T[C>T]A  3.998613e+01  3.426904e-05
#> T[C>T]C  3.998618e+01 5.934640e-151
#> T[C>T]G  2.473771e-32  7.810907e+01
#> T[C>T]T  2.510760e+01  5.659784e-27
#> T[T>A]A  0.000000e+00 7.710426e-294
#> T[T>A]C  3.719645e+00 1.628685e-111
#> T[T>A]G  0.000000e+00  2.814741e+00
#> T[T>A]T  4.649556e+00 8.122779e-107
#> T[T>C]A  0.000000e+00  1.407371e+00
#> T[T>C]C  8.369200e+00 7.358783e-268
#> T[T>C]G  9.299111e+00  0.000000e+00
#> T[T>C]T  1.859822e+01 3.315674e-136
#> T[T>G]A  2.789733e+00  0.000000e+00
#> T[T>G]C  4.649556e+00 4.438833e-146
#> T[T>G]G  1.859822e+00  0.000000e+00
#> T[T>G]T  4.649556e+00  0.000000e+00
#> 
#> $Signature.norm
#>                  Sig1          Sig2
#> A[C>A]A  1.012514e-02  1.503766e-45
#> A[C>A]C  1.727229e-02 7.671396e-149
#> A[C>A]G  2.382386e-03  0.000000e+00
#> A[C>A]T 2.798897e-104  2.523575e-02
#> A[C>G]A  8.338349e-03  2.339132e-16
#> A[C>G]C  3.573578e-03  8.027646e-45
#> A[C>G]G  4.764771e-03 2.969377e-297
#> A[C>G]T  4.764771e-03  7.465178e-95
#> A[C>T]A  3.037542e-02 1.124078e-101
#> A[C>T]C  3.037542e-02 1.508580e-112
#> A[C>T]G  1.375828e-01  9.175645e-73
#> A[C>T]T  1.548551e-02 1.118279e-143
#> A[T>A]A 2.840706e-172  1.186394e-76
#> A[T>A]C  7.147157e-03 1.108610e-102
#> A[T>A]G 1.608966e-130  1.892681e-02
#> A[T>A]T  2.382386e-03 3.102732e-321
#> A[T>C]A  1.667670e-02  8.454653e-79
#> A[T>C]C  7.147157e-03  4.441484e-12
#> A[T>C]G  1.608110e-02  2.840448e-08
#> A[T>C]T  1.846349e-02  1.041633e-24
#> A[T>G]A  3.573578e-03  6.142513e-56
#> A[T>G]C  0.000000e+00  6.308937e-03
#> A[T>G]G  1.786789e-03 4.154699e-136
#> A[T>G]T  1.786789e-03 6.757422e-245
#> C[C>A]A  1.072073e-02  0.000000e+00
#> C[C>A]C 2.422216e-125  4.731703e-02
#> C[C>A]G  8.338349e-03 2.560102e-240
#> C[C>A]T  1.072073e-02  1.204854e-72
#> C[C>G]A  7.742753e-03  3.269096e-17
#> C[C>G]C  2.382386e-03  0.000000e+00
#> C[C>G]G  8.933912e-03  1.766151e-07
#> C[C>G]T  4.169175e-03  5.572753e-23
#> C[C>T]A  2.441945e-02 3.373471e-246
#> C[C>T]C  1.786789e-02 1.639322e-125
#> C[C>T]G  1.184942e-01  1.908302e-02
#> C[C>T]T  3.573578e-02  6.074330e-91
#> C[T>A]A 3.422457e-107  5.164515e-91
#> C[T>A]C  2.863923e-03  1.637643e-02
#> C[T>A]G  7.147157e-03 1.582929e-110
#> C[T>A]T  2.977982e-03  0.000000e+00
#> C[T>C]A  1.462781e-39  2.208128e-02
#> C[T>C]C  1.072073e-02 1.169024e-112
#> C[T>C]G 7.414224e-154  5.993490e-02
#> C[T>C]T  8.933946e-03  1.067777e-38
#> C[T>G]A  5.955964e-04  0.000000e+00
#> C[T>G]C 2.644202e-104  2.523575e-02
#> C[T>G]G  5.955964e-03  0.000000e+00
#> C[T>G]T  2.977982e-03 8.883300e-152
#> G[C>A]A  7.632297e-03  1.320289e-02
#> G[C>A]C  5.360367e-03 2.887238e-205
#> G[C>A]G  2.977982e-03  0.000000e+00
#> G[C>A]T 3.048468e-167  2.839022e-02
#> G[C>G]A  6.551560e-03 6.418983e-207
#> G[C>G]C  4.169175e-03 3.348432e-149
#> G[C>G]G  4.169175e-03  0.000000e+00
#> G[C>G]T  1.221315e-16  1.892681e-02
#> G[C>T]A 1.043479e-103  8.517065e-02
#> G[C>T]C  2.918422e-02  1.924785e-94
#> G[C>T]G  1.099261e-01  1.399071e-02
#> G[C>T]T  3.037542e-02  2.851360e-36
#> G[T>A]A 2.569425e-309  6.308937e-03
#> G[T>A]C 3.664590e-174  2.839022e-02
#> G[T>A]G  2.382386e-03 6.219766e-106
#> G[T>A]T  2.091409e-28  1.892681e-02
#> G[T>C]A  3.644663e-42  1.892681e-02
#> G[T>C]C 3.617385e-151  2.839022e-02
#> G[T>C]G  6.551560e-03 1.783563e-197
#> G[T>C]T 1.779025e-132  3.785362e-02
#> G[T>G]A  3.074537e-98  1.261787e-02
#> G[T>G]C  1.786789e-03 3.215628e-228
#> G[T>G]G  3.573578e-03  3.404180e-28
#> G[T>G]T  1.191193e-03  0.000000e+00
#> T[C>A]A  2.889350e-03  2.570517e-02
#> T[C>A]C  1.369872e-02 2.833485e-118
#> T[C>A]G  4.764771e-03 2.325049e-137
#> T[C>A]T  1.667670e-02 1.004236e-182
#> T[C>G]A  2.977982e-03 3.469795e-289
#> T[C>G]C  6.613130e-35  2.523575e-02
#> T[C>G]G  6.551560e-03 8.529596e-262
#> T[C>G]T 2.225880e-188  2.839022e-02
#> T[C>T]A  2.561062e-02  1.536207e-07
#> T[C>T]C  2.561064e-02 2.660370e-153
#> T[C>T]G  1.584419e-35  3.501460e-01
#> T[C>T]T  1.608110e-02  2.537158e-29
#> T[T>A]A  0.000000e+00 3.456417e-296
#> T[T>A]C  2.382386e-03 7.301042e-114
#> T[T>A]G  0.000000e+00  1.261787e-02
#> T[T>A]T  2.977982e-03 3.641266e-109
#> T[T>C]A  0.000000e+00  6.308937e-03
#> T[T>C]C  5.360367e-03 3.298783e-270
#> T[T>C]G  5.955964e-03  0.000000e+00
#> T[T>C]T  1.191193e-02 1.486345e-138
#> T[T>G]A  1.786789e-03  0.000000e+00
#> T[T>G]C  2.977982e-03 1.989832e-148
#> T[T>G]G  1.191193e-03  0.000000e+00
#> T[T>G]T  2.977982e-03  0.000000e+00
#> 
#> $Exposure
#>      TCGA-AB-2802 TCGA-AB-2803 TCGA-AB-2804 TCGA-AB-2805  TCGA-AB-2806
#> Sig1     7.565056    11.144377     6.704882     12.90647  1.375751e+01
#> Sig2     1.602612     1.590532     0.000000      0.00000 2.149314e-143
#>      TCGA-AB-2807 TCGA-AB-2808 TCGA-AB-2809 TCGA-AB-2810 TCGA-AB-2812
#> Sig1    21.788890     8.523119     3.900144    11.056323     5.050140
#> Sig2     1.625003     0.000000     0.000000     1.654508     2.041469
#>      TCGA-AB-2813 TCGA-AB-2814 TCGA-AB-2816 TCGA-AB-2817 TCGA-AB-2818
#> Sig1    10.191330     5.316748    5.7395393     7.595464     9.417752
#> Sig2     3.249495     2.439561    0.9001467     1.581263     0.000000
#>      TCGA-AB-2819 TCGA-AB-2820 TCGA-AB-2821 TCGA-AB-2822 TCGA-AB-2824
#> Sig1    10.394879    11.144377   12.0472295   17.8922669    2.8670370
#> Sig2     3.141335     1.590532    0.8665078    0.8712128    0.9264135
#>       TCGA-AB-2825 TCGA-AB-2826 TCGA-AB-2827 TCGA-AB-2828 TCGA-AB-2829
#> Sig1  9.935202e-01     3.900144     8.271511   10.2912059     9.417752
#> Sig2 5.179432e-266     0.000000     3.319841    0.8768405     0.000000
#>      TCGA-AB-2830 TCGA-AB-2831 TCGA-AB-2832 TCGA-AB-2833 TCGA-AB-2834
#> Sig1   11.9170384     1.974409     6.926579     5.682234            0
#> Sig2    0.9799417     0.000000     2.027561     1.632172            0
#>      TCGA-AB-2835 TCGA-AB-2836 TCGA-AB-2838 TCGA-AB-2839 TCGA-AB-2841
#> Sig1    0.0000000    0.8356419    16.227792     16.26361     3.900144
#> Sig2    0.8665074    0.9866850     1.592541      0.00000     0.000000
#>      TCGA-AB-2842 TCGA-AB-2843 TCGA-AB-2844 TCGA-AB-2845 TCGA-AB-2846
#> Sig1     1.974409     6.668020     9.370132     2.943137     7.749221
#> Sig2     0.000000     2.188532     1.598945     0.000000     2.628345
#>      TCGA-AB-2847 TCGA-AB-2849 TCGA-AB-2850 TCGA-AB-2851 TCGA-AB-2853
#> Sig1     9.417752    19.586227     3.900144    4.8458470     4.652840
#> Sig2     0.000000     2.106501     0.000000    0.8665074     2.280446
#>      TCGA-AB-2854 TCGA-AB-2855 TCGA-AB-2857 TCGA-AB-2858 TCGA-AB-2859
#> Sig1     10.30313     2.943137   12.0372345   11.1795332     10.30313
#> Sig2      0.00000     0.000000    0.8753459    0.8665074      0.00000
#>      TCGA-AB-2860 TCGA-AB-2861 TCGA-AB-2862 TCGA-AB-2863 TCGA-AB-2864
#> Sig1     3.928365    15.411590    9.3686729    13.688220    10.830846
#> Sig2     3.628944     1.583421    0.9084147     1.616937     2.386487
#>      TCGA-AB-2865 TCGA-AB-2866 TCGA-AB-2867 TCGA-AB-2868 TCGA-AB-2869
#> Sig1    13.736396    0.9935202     6.647778     9.352557    8.4584162
#> Sig2     2.179991    0.0000000     2.733573     2.207129    0.9210788
#>      TCGA-AB-2870 TCGA-AB-2871 TCGA-AB-2872  TCGA-AB-2873 TCGA-AB-2874
#> Sig1     9.417752    10.551917    9.3938505  9.935202e-01     13.75751
#> Sig2     0.000000     2.003496    0.8869785 6.081454e-265      0.00000
#>      TCGA-AB-2875 TCGA-AB-2876 TCGA-AB-2877 TCGA-AB-2878 TCGA-AB-2879
#> Sig1     8.523119     7.859771    15.380805     8.457922     1.974409
#> Sig2     0.000000     2.013726     1.607035     2.206701     0.000000
#>      TCGA-AB-2880 TCGA-AB-2881 TCGA-AB-2882 TCGA-AB-2883 TCGA-AB-2884
#> Sig1    0.9935202     7.446140     14.60056    0.9935202    2.9068126
#> Sig2    0.0000000     1.008908      0.00000    0.0000000    0.8952299
#>      TCGA-AB-2885 TCGA-AB-2886 TCGA-AB-2887 TCGA-AB-2888 TCGA-AB-2889
#> Sig1     8.427898     9.234471    10.237697     2.837042     4.845847
#> Sig2     2.225354     1.020549     2.739729     2.757865     0.000000
#>      TCGA-AB-2890 TCGA-AB-2891 TCGA-AB-2892 TCGA-AB-2894 TCGA-AB-2895
#> Sig1     7.618933    12.873120    0.9935202    3.8444234     13.75751
#> Sig2     0.000000     1.589695    0.0000000    0.9110001      0.00000
#>       TCGA-AB-2896 TCGA-AB-2897 TCGA-AB-2898 TCGA-AB-2899 TCGA-AB-2900
#> Sig1  9.935202e-01    4.8187282     15.43586     8.290437   15.3422201
#> Sig2 8.087159e-265    0.8884882      0.00000     3.753497    0.9515714
#>      TCGA-AB-2901 TCGA-AB-2903 TCGA-AB-2904 TCGA-AB-2905 TCGA-AB-2906
#> Sig1     4.985998    0.9935202     19.50301    16.200047    12.016572
#> Sig2     2.080922    0.0000000      0.00000     2.209301     0.893547
#>      TCGA-AB-2907 TCGA-AB-2908 TCGA-AB-2910 TCGA-AB-2911 TCGA-AB-2912
#> Sig1     14.60056    13.497791    9.3656766     0.000000    15.327947
#> Sig2      0.00000     2.335997    0.9109571     1.564709     2.767746
#>      TCGA-AB-2913 TCGA-AB-2914  TCGA-AB-2915 TCGA-AB-2916 TCGA-AB-2917
#> Sig1     9.353787 1.543586e+01  1.870353e+01     7.578480    10.109971
#> Sig2     2.738508 2.178232e-60 2.783721e-256     2.191001     3.292121
#>      TCGA-AB-2918 TCGA-AB-2919 TCGA-AB-2920 TCGA-AB-2921 TCGA-AB-2922
#> Sig1    0.9935202     8.496909     6.635605     7.578292   10.3031296
#> Sig2    0.0000000     1.583391     1.612854     0.900518    0.8665075
#>       TCGA-AB-2923 TCGA-AB-2924 TCGA-AB-2925 TCGA-AB-2926 TCGA-AB-2927
#> Sig1  1.870353e+01     8.523119    10.256387     9.385565    20.880738
#> Sig2 1.589733e-251     0.000000     2.195811     2.186317     1.058007
#>      TCGA-AB-2928 TCGA-AB-2929 TCGA-AB-2930 TCGA-AB-2931 TCGA-AB-2932
#> Sig1     6.606419     13.75751     4.818730     7.527022     4.845847
#> Sig2     1.632969      0.00000     0.888487     1.629160     0.000000
#>      TCGA-AB-2933 TCGA-AB-2934 TCGA-AB-2935 TCGA-AB-2936 TCGA-AB-2937
#> Sig1    0.0000000     3.854757     3.627150     4.817387     5.669693
#> Sig2    0.8665074     2.726260     3.765531     3.204524     2.232866
#>      TCGA-AB-2938 TCGA-AB-2939 TCGA-AB-2940 TCGA-AB-2941 TCGA-AB-2943
#> Sig1     16.26361    11.821088    1.1363803    3.8659869     13.75751
#> Sig2      0.00000     1.729667    0.7540511    0.8938488      0.00000
#>      TCGA-AB-2945  TCGA-AB-2946 TCGA-AB-2947 TCGA-AB-2948  TCGA-AB-2949
#> Sig1     7.571351 2.701675e-121     1.974409    0.9935202  1.204723e+01
#> Sig2     1.598201  2.165931e+00     0.000000    0.0000000 3.621365e-284
#>      TCGA-AB-2950  TCGA-AB-2952 TCGA-AB-2954 TCGA-AB-2955 TCGA-AB-2956
#> Sig1     8.447855  1.290647e+01    0.0000000 1.543586e+01    2.9431367
#> Sig2     1.618128 5.350731e-321    0.8665074 5.184884e-92    0.8665074
#>      TCGA-AB-2957 TCGA-AB-2959 TCGA-AB-2963 TCGA-AB-2964 TCGA-AB-2965
#> Sig1     1.974409    14.998437    11.982753    13.736724    5.7628939
#> Sig2     0.000000     5.379565     2.207924     2.714486    0.8810699
#>      TCGA-AB-2966 TCGA-AB-2967  TCGA-AB-2968 TCGA-AB-2970 TCGA-AB-2971
#> Sig1    12.688388     6.704882  1.460056e+01     5.780637    8.5231179
#> Sig2     1.725426     0.000000 3.193239e-299     0.000000    0.8665079
#>      TCGA-AB-2972 TCGA-AB-2973 TCGA-AB-2974 TCGA-AB-2975 TCGA-AB-2976
#> Sig1    15.349214     3.900144     6.704882    0.8722250     8.383488
#> Sig2     3.687111     0.000000     0.000000    0.9592003     4.895018
#>      TCGA-AB-2977 TCGA-AB-2978 TCGA-AB-2979 TCGA-AB-2980 TCGA-AB-2981
#> Sig1     4.781991    14.139875     5.780637     4.845847     3.900144
#> Sig2     2.736504     1.263689     0.000000     0.000000     0.000000
#>      TCGA-AB-2982 TCGA-AB-2983 TCGA-AB-2984 TCGA-AB-2985 TCGA-AB-2986
#> Sig1     1.974409     9.345499     7.061468     3.900144     5.541645
#> Sig2     0.000000     2.743221     1.941213     0.000000     2.308923
#>       TCGA-AB-2987 TCGA-AB-2988 TCGA-AB-2989 TCGA-AB-2990 TCGA-AB-2991
#> Sig1  5.780637e+00   12.0270176    7.5850242    6.6699851     5.699194
#> Sig2 5.202874e-309    0.8843575    0.8949067    0.8954135     1.620623
#>      TCGA-AB-2992 TCGA-AB-2993 TCGA-AB-2994 TCGA-AB-2995 TCGA-AB-2996
#> Sig1 3.663365e-65     5.726490     7.475932     5.780637    12.842051
#> Sig2 4.461327e+00     2.198722     2.253942     0.000000     2.208296
#>      TCGA-AB-2997 TCGA-AB-2998 TCGA-AB-2999 TCGA-AB-3000 TCGA-AB-3001
#> Sig1     5.637528     8.523119     10.30313    2.9431367     8.427898
#> Sig2     3.262006     0.000000      0.00000    0.8665074     2.225354
#>      TCGA-AB-3002 TCGA-AB-3005 TCGA-AB-3006 TCGA-AB-3007 TCGA-AB-3008
#> Sig1    20.214553    13.697656    13.662395     4.767976 3.900144e+00
#> Sig2     2.754235     2.205679     2.228936     1.617581 4.224728e-62
#>      TCGA-AB-3009 TCGA-AB-3011 TCGA-AB-3012
#> Sig1    25.110309     5.780637     6.704882
#> Sig2     3.961001     0.000000     0.000000
#> 
#> $Exposure.norm
#>      TCGA-AB-2802 TCGA-AB-2803 TCGA-AB-2804 TCGA-AB-2805  TCGA-AB-2806
#> Sig1    0.8251887    0.8751046            1            1  1.000000e+00
#> Sig2    0.1748113    0.1248954            0            0 1.562284e-144
#>      TCGA-AB-2807 TCGA-AB-2808 TCGA-AB-2809 TCGA-AB-2810 TCGA-AB-2812
#> Sig1   0.93059662            1            1    0.8698348    0.7121289
#> Sig2   0.06940338            0            0    0.1301652    0.2878711
#>      TCGA-AB-2813 TCGA-AB-2814 TCGA-AB-2816 TCGA-AB-2817 TCGA-AB-2818
#> Sig1    0.7582369     0.685474    0.8644293    0.8276877            1
#> Sig2    0.2417631     0.314526    0.1355707    0.1723123            0
#>      TCGA-AB-2819 TCGA-AB-2820 TCGA-AB-2821 TCGA-AB-2822 TCGA-AB-2824
#> Sig1     0.767931    0.8751046   0.93290031    0.9535687    0.7557861
#> Sig2     0.232069    0.1248954   0.06709969    0.0464313    0.2442139
#>       TCGA-AB-2825 TCGA-AB-2826 TCGA-AB-2827 TCGA-AB-2828 TCGA-AB-2829
#> Sig1  1.000000e+00            1    0.7135933   0.92148667            1
#> Sig2 5.213213e-266            0    0.2864067   0.07851333            0
#>      TCGA-AB-2830 TCGA-AB-2831 TCGA-AB-2832 TCGA-AB-2833 TCGA-AB-2834
#> Sig1   0.92401774            1    0.7735616    0.7768552          NaN
#> Sig2   0.07598226            0    0.2264384    0.2231448          NaN
#>      TCGA-AB-2835 TCGA-AB-2836 TCGA-AB-2838 TCGA-AB-2839 TCGA-AB-2841
#> Sig1            0    0.4585576   0.91063347            1            1
#> Sig2            1    0.5414424   0.08936653            0            0
#>      TCGA-AB-2842 TCGA-AB-2843 TCGA-AB-2844 TCGA-AB-2845 TCGA-AB-2846
#> Sig1            1    0.7528912    0.8542316            1    0.7467282
#> Sig2            0    0.2471088    0.1457684            0    0.2532718
#>      TCGA-AB-2847 TCGA-AB-2849 TCGA-AB-2850 TCGA-AB-2851 TCGA-AB-2853
#> Sig1            1   0.90289369            1    0.8483099    0.6710873
#> Sig2            0   0.09710631            0    0.1516901    0.3289127
#>      TCGA-AB-2854 TCGA-AB-2855 TCGA-AB-2857 TCGA-AB-2858 TCGA-AB-2859
#> Sig1            1            1   0.93220984   0.92806704            1
#> Sig2            0            0   0.06779016   0.07193296            0
#>      TCGA-AB-2860 TCGA-AB-2861 TCGA-AB-2862 TCGA-AB-2863 TCGA-AB-2864
#> Sig1    0.5198101   0.90683024   0.91160777    0.8943534    0.8194427
#> Sig2    0.4801899   0.09316976   0.08839223    0.1056466    0.1805573
#>      TCGA-AB-2865 TCGA-AB-2866 TCGA-AB-2867 TCGA-AB-2868 TCGA-AB-2869
#> Sig1    0.8630348            1    0.7086163    0.8090667   0.90179868
#> Sig2    0.1369652            0    0.2913837    0.1909333   0.09820132
#>      TCGA-AB-2870 TCGA-AB-2871 TCGA-AB-2872  TCGA-AB-2873 TCGA-AB-2874
#> Sig1            1    0.8404277     0.913725  1.000000e+00            1
#> Sig2            0    0.1595723     0.086275 6.121117e-265            0
#>      TCGA-AB-2875 TCGA-AB-2876 TCGA-AB-2877 TCGA-AB-2878 TCGA-AB-2879
#> Sig1            1    0.7960473   0.90540088    0.7930822            1
#> Sig2            0    0.2039527   0.09459912    0.2069178            0
#>      TCGA-AB-2880 TCGA-AB-2881 TCGA-AB-2882 TCGA-AB-2883 TCGA-AB-2884
#> Sig1            1    0.8806739            1            1    0.7645397
#> Sig2            0    0.1193261            0            0    0.2354603
#>      TCGA-AB-2885 TCGA-AB-2886 TCGA-AB-2887 TCGA-AB-2888 TCGA-AB-2889
#> Sig1    0.7911103   0.90048298     0.788885    0.5070758            1
#> Sig2    0.2088897   0.09951702     0.211115    0.4929242            0
#>      TCGA-AB-2890 TCGA-AB-2891 TCGA-AB-2892 TCGA-AB-2894 TCGA-AB-2895
#> Sig1            1     0.890084            1    0.8084292            1
#> Sig2            0     0.109916            0    0.1915708            0
#>       TCGA-AB-2896 TCGA-AB-2897 TCGA-AB-2898 TCGA-AB-2899 TCGA-AB-2900
#> Sig1  1.000000e+00     0.844322            1    0.6883496   0.94159914
#> Sig2 8.139904e-265     0.155678            0    0.3116504   0.05840086
#>      TCGA-AB-2901 TCGA-AB-2903 TCGA-AB-2904 TCGA-AB-2905 TCGA-AB-2906
#> Sig1    0.7055405            1            1    0.8799902   0.93078708
#> Sig2    0.2944595            0            0    0.1200098   0.06921292
#>      TCGA-AB-2907 TCGA-AB-2908 TCGA-AB-2910 TCGA-AB-2911 TCGA-AB-2912
#> Sig1            1    0.8524676   0.91135647            0    0.8470495
#> Sig2            0    0.1475324   0.08864353            1    0.1529505
#>      TCGA-AB-2913 TCGA-AB-2914 TCGA-AB-2915 TCGA-AB-2916 TCGA-AB-2917
#> Sig1    0.7735328 1.000000e+00  1.00000e+00      0.77573    0.7543577
#> Sig2    0.2264672 1.411151e-61 1.48834e-257      0.22427    0.2456423
#>      TCGA-AB-2918 TCGA-AB-2919 TCGA-AB-2920 TCGA-AB-2921 TCGA-AB-2922
#> Sig1            1    0.8429223     0.804466    0.8937919   0.92242295
#> Sig2            0    0.1570777     0.195534    0.1062081   0.07757705
#>      TCGA-AB-2923 TCGA-AB-2924 TCGA-AB-2925 TCGA-AB-2926 TCGA-AB-2927
#> Sig1  1.00000e+00            1    0.8236608    0.8110664   0.95177448
#> Sig2 8.49964e-253            0    0.1763392    0.1889336   0.04822552
#>      TCGA-AB-2928 TCGA-AB-2929 TCGA-AB-2930 TCGA-AB-2931 TCGA-AB-2932
#> Sig1    0.8018095            1    0.8443222    0.8220699            1
#> Sig2    0.1981905            0    0.1556778    0.1779301            0
#>      TCGA-AB-2933 TCGA-AB-2934 TCGA-AB-2935 TCGA-AB-2936 TCGA-AB-2937
#> Sig1            0    0.5857388    0.4906407    0.6005286    0.7174502
#> Sig2            1    0.4142612    0.5093593    0.3994714    0.2825498
#>      TCGA-AB-2938 TCGA-AB-2939 TCGA-AB-2940 TCGA-AB-2941 TCGA-AB-2943
#> Sig1            1    0.8723564    0.6011222    0.8122102            1
#> Sig2            0    0.1276436    0.3988778    0.1877898            0
#>      TCGA-AB-2945 TCGA-AB-2946 TCGA-AB-2947 TCGA-AB-2948  TCGA-AB-2949
#> Sig1    0.8257057 1.24735e-121            1            1  1.000000e+00
#> Sig2    0.1742943  1.00000e+00            0            0 3.005974e-285
#>      TCGA-AB-2950  TCGA-AB-2952 TCGA-AB-2954 TCGA-AB-2955 TCGA-AB-2956
#> Sig1    0.8392479  1.000000e+00            0 1.000000e+00     0.772549
#> Sig2    0.1607521 4.150151e-322            1 3.358986e-93     0.227451
#>      TCGA-AB-2957 TCGA-AB-2959 TCGA-AB-2963 TCGA-AB-2964 TCGA-AB-2965
#> Sig1            1    0.7360112    0.8444102    0.8349978    0.8673879
#> Sig2            0    0.2639888    0.1555898    0.1650022    0.1326121
#>      TCGA-AB-2966 TCGA-AB-2967  TCGA-AB-2968 TCGA-AB-2970 TCGA-AB-2971
#> Sig1    0.8802936            1  1.000000e+00            1   0.90771646
#> Sig2    0.1197064            0 2.187065e-300            0   0.09228354
#>      TCGA-AB-2972 TCGA-AB-2973 TCGA-AB-2974 TCGA-AB-2975 TCGA-AB-2976
#> Sig1    0.8063118            1            1    0.4762548    0.6313578
#> Sig2    0.1936882            0            0    0.5237452    0.3686422
#>      TCGA-AB-2977 TCGA-AB-2978 TCGA-AB-2979 TCGA-AB-2980 TCGA-AB-2981
#> Sig1    0.6360303   0.91796125            1            1            1
#> Sig2    0.3639697   0.08203875            0            0            0
#>      TCGA-AB-2982 TCGA-AB-2983 TCGA-AB-2984 TCGA-AB-2985 TCGA-AB-2986
#> Sig1            1     0.773076    0.7843739            1     0.705891
#> Sig2            0     0.226924    0.2156261            0     0.294109
#>       TCGA-AB-2987 TCGA-AB-2988 TCGA-AB-2989 TCGA-AB-2990 TCGA-AB-2991
#> Sig1  1.000000e+00   0.93150555    0.8944677    0.8816436    0.7785979
#> Sig2 9.000521e-310   0.06849445    0.1055323    0.1183564    0.2214021
#>      TCGA-AB-2992 TCGA-AB-2993 TCGA-AB-2994 TCGA-AB-2995 TCGA-AB-2996
#> Sig1 8.211379e-66    0.7225662    0.7683483            1    0.8532727
#> Sig2 1.000000e+00    0.2774338    0.2316517            0    0.1467273
#>      TCGA-AB-2997 TCGA-AB-2998 TCGA-AB-2999 TCGA-AB-3000 TCGA-AB-3001
#> Sig1    0.6334632            1            1     0.772549    0.7911103
#> Sig2    0.3665368            0            0     0.227451    0.2088897
#>      TCGA-AB-3002 TCGA-AB-3005 TCGA-AB-3006 TCGA-AB-3007 TCGA-AB-3008
#> Sig1    0.8800879    0.8613071    0.8597389    0.7466813 1.000000e+00
#> Sig2    0.1199121    0.1386929    0.1402611    0.2533187 1.083224e-62
#>      TCGA-AB-3009 TCGA-AB-3011 TCGA-AB-3012
#> Sig1    0.8637488            1            1
#> Sig2    0.1362512            0            0
#> 
#> $K
#> [1] 2
#> 
#> $Raw
#> $Raw$summary_run
#> # A tibble: 3 × 5
#>     Run     K posterior file                           similarity_to_ref
#>   <int> <int>     <dbl> <chr>                                      <dbl>
#> 1     2     2     9714. /tmp/RtmpXZeTRm/BayesNMF.2.rds             0.718
#> 2     3     2     9711. /tmp/RtmpXZeTRm/BayesNMF.3.rds             0.676
#> 3     1     3     9676. /tmp/RtmpXZeTRm/BayesNMF.1.rds             0.611
#> 
#> $Raw$W
#>                  Sig1          Sig2
#> A[C>A]A  2.649894e-01  1.215678e-44
#> A[C>A]C  4.520407e-01 6.201727e-148
#> A[C>A]G  6.235044e-02  0.000000e+00
#> A[C>A]T 7.325113e-103  2.040114e-01
#> A[C>G]A  2.182266e-01  1.891007e-15
#> A[C>G]C  9.352567e-02  6.489727e-44
#> A[C>G]G  1.247009e-01 2.400511e-296
#> A[C>G]T  1.247009e-01  6.035016e-94
#> A[C>T]A  7.949682e-01 9.087293e-101
#> A[C>T]C  7.949682e-01 1.219569e-111
#> A[C>T]G  3.600738e+00  7.417795e-72
#> A[C>T]T  4.052779e-01 9.040417e-143
#> A[T>A]A 7.434535e-171  9.591072e-76
#> A[T>A]C  1.870513e-01 8.962249e-102
#> A[T>A]G 4.210896e-129  1.530086e-01
#> A[T>A]T  6.235044e-02 2.506395e-320
#> A[T>C]A  4.364531e-01  6.834930e-78
#> A[T>C]C  1.870513e-01  3.590594e-11
#> A[T>C]G  4.208654e-01  2.296281e-07
#> A[T>C]T  4.832159e-01  8.420791e-24
#> A[T>G]A  9.352567e-02  4.965744e-55
#> A[T>G]C  0.000000e+00  5.100285e-02
#> A[T>G]G  4.676283e-02 3.358751e-135
#> A[T>G]T  4.676283e-02 5.462850e-244
#> C[C>A]A  2.805770e-01  0.000000e+00
#> C[C>A]C 6.339286e-124  3.825214e-01
#> C[C>A]G  2.182266e-01 2.069643e-239
#> C[C>A]T  2.805770e-01  9.740311e-72
#> C[C>G]A  2.026389e-01  2.642810e-16
#> C[C>G]C  6.235044e-02  0.000000e+00
#> C[C>G]G  2.338133e-01  1.427795e-06
#> C[C>G]T  1.091133e-01  4.505137e-22
#> C[C>T]A  6.390921e-01 2.727189e-245
#> C[C>T]C  4.676283e-01 1.325265e-124
#> C[C>T]G  3.101163e+00  1.542714e-01
#> C[C>T]T  9.352567e-01  4.910624e-90
#> C[T>A]A 8.957060e-106  4.175109e-90
#> C[T>A]C  7.495298e-02  1.323907e-01
#> C[T>A]G  1.870513e-01 1.279675e-109
#> C[T>A]T  7.793806e-02  0.000000e+00
#> C[T>C]A  3.828307e-38  1.785100e-01
#> C[T>C]C  2.805770e-01 9.450653e-112
#> C[T>C]G 1.940409e-152  4.845271e-01
#> C[T>C]T  2.338142e-01  8.632143e-38
#> C[T>G]A  1.558761e-02  0.000000e+00
#> C[T>G]C 6.920257e-103  2.040114e-01
#> C[T>G]G  1.558761e-01  0.000000e+00
#> C[T>G]T  7.793806e-02 7.181457e-151
#> G[C>A]A  1.997481e-01  1.067351e-01
#> G[C>A]C  1.402885e-01 2.334107e-204
#> G[C>A]G  7.793806e-02  0.000000e+00
#> G[C>A]T 7.978279e-166  2.295128e-01
#> G[C>G]A  1.714637e-01 5.189249e-206
#> G[C>G]C  1.091133e-01 2.706947e-148
#> G[C>G]G  1.091133e-01  0.000000e+00
#> G[C>G]T  3.196356e-15  1.530086e-01
#> G[C>T]A 2.730935e-102  6.885385e-01
#> G[C>T]C  7.637929e-01  1.556039e-93
#> G[C>T]G  2.876924e+00  1.131040e-01
#> G[C>T]T  7.949682e-01  2.305103e-35
#> G[T>A]A 6.724555e-308  5.100285e-02
#> G[T>A]C 9.590757e-173  2.295128e-01
#> G[T>A]G  6.235044e-02 5.028197e-105
#> G[T>A]T  5.473518e-27  1.530086e-01
#> G[T>C]A  9.538606e-41  1.530086e-01
#> G[T>C]C 9.467214e-150  2.295128e-01
#> G[T>C]G  1.714637e-01 1.441872e-196
#> G[T>C]T 4.655964e-131  3.060171e-01
#> G[T>G]A  8.046505e-97  1.020057e-01
#> G[T>G]C  4.676283e-02 2.599585e-227
#> G[T>G]G  9.352567e-02  2.752015e-27
#> G[T>G]T  3.117522e-02  0.000000e+00
#> T[C>A]A  7.561843e-02  2.078063e-01
#> T[C>A]C  3.585151e-01 2.290652e-117
#> T[C>A]G  1.247009e-01 1.879621e-136
#> T[C>A]T  4.364531e-01 8.118464e-182
#> T[C>G]A  7.793806e-02 2.805059e-288
#> T[C>G]C  1.730751e-33  2.040114e-01
#> T[C>G]G  1.714637e-01 6.895515e-261
#> T[C>G]T 5.825448e-187  2.295128e-01
#> T[C>T]A  6.702665e-01  1.241904e-06
#> T[C>T]C  6.702673e-01 2.150703e-152
#> T[C>T]G  4.146651e-34  2.830658e+00
#> T[C>T]T  4.208655e-01  2.051095e-28
#> T[T>A]A  0.000000e+00 2.794244e-295
#> T[T>A]C  6.235044e-02 5.902324e-113
#> T[T>A]G  0.000000e+00  1.020057e-01
#> T[T>A]T  7.793806e-02 2.943680e-108
#> T[T>C]A  0.000000e+00  5.100285e-02
#> T[T>C]C  1.402885e-01 2.666809e-269
#> T[T>C]G  1.558761e-01  0.000000e+00
#> T[T>C]T  3.117522e-01 1.201594e-137
#> T[T>G]A  4.676283e-02  0.000000e+00
#> T[T>G]C  7.793806e-02 1.608625e-147
#> T[T>G]G  3.117522e-02  0.000000e+00
#> T[T>G]T  7.793806e-02  0.000000e+00
#> 
#> $Raw$H
#>      TCGA-AB-2802 TCGA-AB-2803 TCGA-AB-2804 TCGA-AB-2805  TCGA-AB-2806
#> Sig1    0.2890578    0.4258222    0.2561909    0.4931512  5.256689e-01
#> Sig2    0.1982395    0.1967452    0.0000000    0.0000000 2.658653e-144
#>      TCGA-AB-2807 TCGA-AB-2808 TCGA-AB-2809 TCGA-AB-2810 TCGA-AB-2812
#> Sig1    0.8325448    0.3256649    0.1490229    0.4224577    0.1929638
#> Sig2    0.2010092    0.0000000    0.0000000    0.2046589    0.2525252
#>      TCGA-AB-2813 TCGA-AB-2814 TCGA-AB-2816 TCGA-AB-2817 TCGA-AB-2818
#> Sig1    0.3894066    0.2031508    0.2193055    0.2902196    0.3598485
#> Sig2    0.4019552    0.3017682    0.1113461    0.1955987    0.0000000
#>      TCGA-AB-2819 TCGA-AB-2820 TCGA-AB-2821 TCGA-AB-2822 TCGA-AB-2824
#> Sig1    0.3971842    0.4258222    0.4603198    0.6836564    0.1095483
#> Sig2    0.3885760    0.1967452    0.1071850    0.1077670    0.1145953
#>       TCGA-AB-2825 TCGA-AB-2826 TCGA-AB-2827 TCGA-AB-2828 TCGA-AB-2829
#> Sig1  3.796201e-02    0.1490229    0.3160511    0.3932229    0.3598485
#> Sig2 6.406840e-267    0.0000000    0.4106568    0.1084632    0.0000000
#>      TCGA-AB-2830 TCGA-AB-2831 TCGA-AB-2832 TCGA-AB-2833 TCGA-AB-2834
#> Sig1    0.4553453   0.07544138    0.2646618    0.2171159            0
#> Sig2    0.1212166   0.00000000    0.2508047    0.2018960            0
#>      TCGA-AB-2835 TCGA-AB-2836 TCGA-AB-2838 TCGA-AB-2839 TCGA-AB-2841
#> Sig1     0.000000   0.03192954    0.6200574     0.621426    0.1490229
#> Sig2     0.107185   0.12205070    0.1969938     0.000000    0.0000000
#>      TCGA-AB-2842 TCGA-AB-2843 TCGA-AB-2844 TCGA-AB-2845 TCGA-AB-2846
#> Sig1   0.07544138    0.2547824    0.3580290    0.1124561    0.2960946
#> Sig2   0.00000000    0.2707165    0.1977859    0.0000000    0.3251203
#>      TCGA-AB-2847 TCGA-AB-2849 TCGA-AB-2850 TCGA-AB-2851 TCGA-AB-2853
#> Sig1    0.3598485    0.7483819    0.1490229    0.1851579    0.1777832
#> Sig2    0.0000000    0.2605694    0.0000000    0.1071850    0.2820860
#>      TCGA-AB-2854 TCGA-AB-2855 TCGA-AB-2857 TCGA-AB-2858 TCGA-AB-2859
#> Sig1    0.3936785    0.1124561    0.4599379    0.4271655    0.3936785
#> Sig2    0.0000000    0.0000000    0.1082783    0.1071850    0.0000000
#>      TCGA-AB-2860 TCGA-AB-2861 TCGA-AB-2862 TCGA-AB-2863 TCGA-AB-2864
#> Sig1    0.1501013    0.5888707    0.3579732    0.5230214    0.4138423
#> Sig2    0.4488921    0.1958656    0.1123688    0.2000115    0.2952030
#>      TCGA-AB-2865 TCGA-AB-2866 TCGA-AB-2867 TCGA-AB-2868 TCGA-AB-2869
#> Sig1    0.5248622   0.03796201    0.2540089    0.3573575    0.3231927
#> Sig2    0.2696600   0.00000000    0.3381368    0.2730168    0.1139354
#>      TCGA-AB-2870 TCGA-AB-2871 TCGA-AB-2872  TCGA-AB-2873 TCGA-AB-2874
#> Sig1    0.3598485    0.4031845    0.3589353  3.796201e-02    0.5256689
#> Sig2    0.0000000    0.2478279    0.1097172 7.522621e-266    0.0000000
#>      TCGA-AB-2875 TCGA-AB-2876 TCGA-AB-2877 TCGA-AB-2878 TCGA-AB-2879
#> Sig1    0.3256649    0.3003187    0.5876944    0.3231738   0.07544138
#> Sig2    0.0000000    0.2490934    0.1987865    0.2729639   0.00000000
#>      TCGA-AB-2880 TCGA-AB-2881 TCGA-AB-2882 TCGA-AB-2883 TCGA-AB-2884
#> Sig1   0.03796201    0.2845140    0.5578817   0.03796201    0.1110681
#> Sig2   0.00000000    0.1247996    0.0000000   0.00000000    0.1107379
#>      TCGA-AB-2885 TCGA-AB-2886 TCGA-AB-2887 TCGA-AB-2888 TCGA-AB-2889
#> Sig1    0.3220266    0.3528454    0.3911783    0.1084022    0.1851579
#> Sig2    0.2752713    0.1262396    0.3388982    0.3411417    0.0000000
#>      TCGA-AB-2890 TCGA-AB-2891 TCGA-AB-2892 TCGA-AB-2894 TCGA-AB-2895
#> Sig1    0.2911164    0.4918767   0.03796201    0.1468939    0.5256689
#> Sig2    0.0000000    0.1966417   0.00000000    0.1126887    0.0000000
#>       TCGA-AB-2896 TCGA-AB-2897 TCGA-AB-2898 TCGA-AB-2899 TCGA-AB-2900
#> Sig1  3.796201e-02    0.1841217     0.589798    0.3167743    0.5862201
#> Sig2 1.000363e-265    0.1099040     0.000000    0.4642991    0.1177072
#>      TCGA-AB-2901 TCGA-AB-2903 TCGA-AB-2904 TCGA-AB-2905 TCGA-AB-2906
#> Sig1    0.1905130   0.03796201    0.7452021    0.6189973    0.4591484
#> Sig2    0.2574053   0.00000000    0.0000000    0.2732856    0.1105297
#>      TCGA-AB-2907 TCGA-AB-2908 TCGA-AB-2910 TCGA-AB-2911 TCGA-AB-2912
#> Sig1    0.5578817    0.5157452    0.3578588     0.000000    0.5856747
#> Sig2    0.0000000    0.2889575    0.1126833     0.193551    0.3423639
#>      TCGA-AB-2913 TCGA-AB-2914  TCGA-AB-2915 TCGA-AB-2916 TCGA-AB-2917
#> Sig1    0.3574045 5.897980e-01  7.146543e-01    0.2895707    0.3862979
#> Sig2    0.3387473 2.694424e-61 3.443400e-257    0.2710219    0.4072280
#>      TCGA-AB-2918 TCGA-AB-2919 TCGA-AB-2920 TCGA-AB-2921 TCGA-AB-2922
#> Sig1   0.03796201    0.3246635    0.2535438    0.2895635    0.3936785
#> Sig2   0.00000000    0.1958618    0.1995064    0.1113920    0.1071850
#>       TCGA-AB-2923 TCGA-AB-2924 TCGA-AB-2925 TCGA-AB-2926 TCGA-AB-2927
#> Sig1  7.146543e-01    0.3256649    0.3918924    0.3586187    0.7978446
#> Sig2 1.966463e-252    0.0000000    0.2716169    0.2704425    0.1308731
#>      TCGA-AB-2928 TCGA-AB-2929 TCGA-AB-2930 TCGA-AB-2931 TCGA-AB-2932
#> Sig1    0.2524286    0.5256689    0.1841217    0.2876045    0.1851579
#> Sig2    0.2019945    0.0000000    0.1099038    0.2015234    0.0000000
#>      TCGA-AB-2933 TCGA-AB-2934 TCGA-AB-2935 TCGA-AB-2936 TCGA-AB-2937
#> Sig1     0.000000    0.1472887    0.1385919    0.1840704    0.2166367
#> Sig2     0.107185    0.3372322    0.4657877    0.3963923    0.2762005
#>      TCGA-AB-2938 TCGA-AB-2939 TCGA-AB-2940 TCGA-AB-2941 TCGA-AB-2943
#> Sig1     0.621426    0.4516790   0.04342063    0.1477178    0.5256689
#> Sig2     0.000000    0.2139559   0.09327441    0.1105671    0.0000000
#>      TCGA-AB-2945  TCGA-AB-2946 TCGA-AB-2947 TCGA-AB-2948  TCGA-AB-2949
#> Sig1    0.2892983 1.032299e-122   0.07544138   0.03796201  4.603198e-01
#> Sig2    0.1976939  2.679208e-01   0.00000000   0.00000000 4.479547e-285
#>      TCGA-AB-2950  TCGA-AB-2952 TCGA-AB-2954 TCGA-AB-2955 TCGA-AB-2956
#> Sig1    0.3227892  4.931512e-01     0.000000 5.897980e-01    0.1124561
#> Sig2    0.2001588 6.620480e-322     0.107185 6.413585e-93    0.1071850
#>      TCGA-AB-2957 TCGA-AB-2959 TCGA-AB-2963 TCGA-AB-2964 TCGA-AB-2965
#> Sig1   0.07544138    0.5730843    0.4578562    0.5248747    0.2201979
#> Sig2   0.00000000    0.6654400    0.2731152    0.3357758    0.1089864
#>      TCGA-AB-2966 TCGA-AB-2967  TCGA-AB-2968 TCGA-AB-2970 TCGA-AB-2971
#> Sig1    0.4848182    0.2561909  5.578817e-01    0.2208758    0.3256649
#> Sig2    0.2134313    0.0000000 3.949964e-300    0.0000000    0.1071851
#>      TCGA-AB-2972 TCGA-AB-2973 TCGA-AB-2974 TCGA-AB-2975 TCGA-AB-2976
#> Sig1    0.5864873    0.1490229    0.2561909   0.03332737    0.3203297
#> Sig2    0.4560873    0.0000000    0.0000000   0.11865091    0.6055026
#>      TCGA-AB-2977 TCGA-AB-2978 TCGA-AB-2979 TCGA-AB-2980 TCGA-AB-2981
#> Sig1    0.1827180    0.5402790    0.2208758    0.1851579    0.1490229
#> Sig2    0.3384994    0.1563155    0.0000000    0.0000000    0.0000000
#>      TCGA-AB-2982 TCGA-AB-2983 TCGA-AB-2984 TCGA-AB-2985 TCGA-AB-2986
#> Sig1   0.07544138    0.3570878    0.2698159    0.1490229    0.2117440
#> Sig2   0.00000000    0.3393302    0.2401237    0.0000000    0.2856086
#>       TCGA-AB-2987 TCGA-AB-2988 TCGA-AB-2989 TCGA-AB-2990 TCGA-AB-2991
#> Sig1  2.208758e-01    0.4595475    0.2898207    0.2548575    0.2177639
#> Sig2 6.435837e-310    0.1093930    0.1106979    0.1107606    0.2004674
#>      TCGA-AB-2992 TCGA-AB-2993 TCGA-AB-2994 TCGA-AB-2995 TCGA-AB-2996
#> Sig1 1.399757e-66    0.2188069    0.2856524    0.2208758    0.4906896
#> Sig2 5.518561e-01    0.2719770    0.2788075    0.0000000    0.2731613
#>      TCGA-AB-2997 TCGA-AB-2998 TCGA-AB-2999 TCGA-AB-3000 TCGA-AB-3001
#> Sig1    0.2154077    0.3256649    0.3936785    0.1124561    0.3220266
#> Sig2    0.4035028    0.0000000    0.0000000    0.1071850    0.2752713
#>      TCGA-AB-3002 TCGA-AB-3005 TCGA-AB-3006 TCGA-AB-3007 TCGA-AB-3008
#> Sig1    0.7723900    0.5233819    0.5220347    0.1821824 1.490229e-01
#> Sig2    0.3406926    0.2728375    0.2757143    0.2000911 5.225893e-63
#>      TCGA-AB-3009 TCGA-AB-3011 TCGA-AB-3012
#> Sig1    0.9594549    0.2208758    0.2561909
#> Sig2    0.4899668    0.0000000    0.0000000
#> 
#> $Raw$best_run
#> [1] 2
#> 
#> 
#> attr(,"class")
#> [1] "Signature"
#> attr(,"nrun")
#> [1] 3
#> attr(,"method")
#> [1] "L1W.L2H"
#> attr(,"call_method")
#> [1] "BayesianNMF"
# }