Estimate Signature Number

Use NMF package to evaluate the optimal number of signatures. This is used along with sig_extract. Users should library(NMF) firstly.

sig_estimate(
  nmf_matrix,
  range = 2:5,
  nrun = 10,
  keep_nmfObj = FALSE,
  what = "all",
  cores = 1,
  seed = 123456,
  use_random = TRUE,
  save_plots = FALSE,
  plot_basename = file.path(tempdir(), "nmf"),
  method = "brunet",
  pConstant = NULL,
  verbose = FALSE
)

Arguments

nmf_matrix
a matrix used for NMF decomposition with rows indicate samples and columns indicate components.

range
a numeric vector containing the ranks of factorization to try. Note that duplicates are removed and values are sorted in increasing order. The results are notably returned in this order.

nrun
a numeric giving the number of run to perform for each value in range, nrun set to 30~50 is enough to achieve robust result.

keep_nmfObj
default is FALSE, if TRUE, keep NMF objects from runs, and the result may be huge.

what
a character vector whose elements partially match one of the following item, which correspond to the measures computed by summary on each multi-run NMF result: ‘all’, ‘cophenetic’, ‘rss’, ‘residuals’, ‘dispersion’, ‘evar’, ‘silhouette’ (and more specific .coef, .basis, .consensus), ‘sparseness’ (and more specific .coef, .basis). It specifies which measure must be plotted (what='all' plots all the measures).

cores
number of cpu cores to run NMF.

seed
specification of the starting point or seeding method, which will compute a starting point, usually using data from the target matrix in order to provide a good guess.

use_random
Should generate random data from input to test measurements. Default is TRUE.

save_plots
if TRUE, save plots to local machine. Of note, if pdf file consensus map has extra blank page, please call pdf(..., onefile=FALSE), it may fix this problem, see https://stackoverflow.com/questions/12481267/in-r-how-to-prevent-blank-page-in-pdf-when-using-gridbase-to-embed-subplot-insi.

plot_basename
when save plots, set custom basename for file path.

method
specification of the NMF algorithm. Use 'brunet' as default. Available methods for nmf decompositions are 'brunet', 'lee', 'ls-nmf', 'nsNMF', 'offset'.

pConstant
A small positive value (like 1e-9) to add to the matrix. Use it ONLY if the functions throws an non-conformable arrays error.

verbose
if TRUE, print extra message.

Value

a list contains information of NMF run and rank survey.

Details

The most common approach is to choose the smallest rank for which cophenetic correlation coefficient starts decreasing (Used by this function). Another approach is to choose the rank for which the plot of the residual sum of squares (RSS) between the input matrix and its estimate shows an inflection point. More custom features please directly use NMF::nmfEstimateRank.

References

Gaujoux, Renaud, and Cathal Seoighe. "A flexible R package for nonnegative matrix factorization." BMC bioinformatics 11.1 (2010): 367.

See also

Examples

# Load copy number prepare object
load(system.file("extdata", "toy_copynumber_prepare.RData",
  package = "sigminer", mustWork = TRUE
))
if (FALSE) {
library(NMF)
cn_estimate <- sig_estimate(cn_prepare$nmf_matrix,
  cores = 1, nrun = 5,
  verbose = TRUE
)
}