Exposure Instability Analysis of Signature Exposures with Bootstrapping

Read sig_fit_bootstrap for more option setting.

sig_fit_bootstrap_batch(
  catalogue_matrix,
  methods = c("QP"),
  n = 100L,
  min_count = 1L,
  p_val_thresholds = c(0.05),
  use_parallel = FALSE,
  seed = 123456L,
  job_id = NULL,
  result_dir = tempdir(),
  ...
)

Arguments

catalogue_matrix

a numeric matrix V with row representing components and columns representing samples, typically you can get nmf_matrix from sig_tally() and transpose it by t().

methods

a subset of c("NNLS", "QP", "SA").

n

the number of bootstrap replicates.

min_count

minimal exposure in a sample, default is 1. Any patient has total exposure less than this value will be filtered out.

p_val_thresholds

a vector of relative exposure threshold for calculating p values.

use_parallel

if TRUE, use parallel computation based on furrr package. It can also be an integer for specifying cores.

seed

random seed to reproduce the result.

job_id

a job ID, default is NULL, can be a string. When not NULL, all bootstrapped results will be saved to local machine location defined by result_dir. This is very useful for running more than 10 times for more than 100 samples.

result_dir

see above, default is temp directory defined by R.

...

other common parameters passing to sig_fit_bootstrap, including sig, sig_index, sig_db, db_type, mode, auto_reduce etc.

Value

a list of data.table.

See also

Examples

# For mutational signatures ----------------
# SBS is used for illustration, similar
# operations can be applied to DBS, INDEL, CN, RS, etc.

# Load simulated data
data("simulated_catalogs")
data = simulated_catalogs$set1
data[1:5, 1:5]
#>         Sample_1 Sample_2 Sample_3 Sample_4 Sample_5
#> A[C>A]A      911      195       95      131       33
#> A[C>A]C      761      175       51       71       10
#> A[C>A]G       88       19       12       14        2
#> A[C>A]T      744      174       55       77       14
#> C[C>A]A      883      225      142      170       55

# Fitting with COSMIC reference signatures
# \donttest{
# Generally set n = 100
rv = sig_fit_bootstrap_batch(data,
  sig_index = c(1, 5, 9, 2, 13),
   sig_db = "SBS", n = 10)
#> ℹ [2024-05-12 07:01:10.710772]: Batch Bootstrap Signature Exposure Analysis Started.
#> ℹ [2024-05-12 07:01:10.712387]: Samples to be filtered out: 
#> ℹ [2024-05-12 07:01:10.713824]: Finding optimal exposures (&errors) for different methods.
#> ℹ [2024-05-12 07:01:10.715212]: Calling method `QP`.
#> ℹ [2024-05-12 07:01:10.718397]: Started.
#> ✔ [2024-05-12 07:01:10.719799]: Signature index detected.
#> ℹ [2024-05-12 07:01:10.721163]: Checking signature database in package.
#> ℹ [2024-05-12 07:01:10.724283]: Checking signature index.
#> ℹ [2024-05-12 07:01:10.725726]: Valid index for db 'SBS':
#> 1 2 3 4 5 6 7a 7b 7c 7d 8 9 10a 10b 11 12 13 14 15 16 17a 17b 18 19 20 21 22 23 24 25 26 27 28 29 30 31 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48 49 50 51 52 53 54 55 56 57 58 59 60 84 85 86 87 88 89 90
#> ✔ [2024-05-12 07:01:10.727181]: Database and index checked.
#> ✔ [2024-05-12 07:01:10.728675]: Signature normalized.
#> ℹ [2024-05-12 07:01:10.730066]: Checking row number for catalog matrix and signature matrix.
#> ✔ [2024-05-12 07:01:10.731443]: Checked.
#> ℹ [2024-05-12 07:01:10.732859]: Checking rownames for catalog matrix and signature matrix.
#> ℹ [2024-05-12 07:01:10.734327]: Matrix V and W don't have same orders. Try reordering...
#> ✔ [2024-05-12 07:01:10.73585]: Checked.
#> ✔ [2024-05-12 07:01:10.737198]: Method 'QP' detected.
#> ✔ [2024-05-12 07:01:10.73853]: Corresponding function generated.
#> ℹ [2024-05-12 07:01:10.739902]: Calling function.
#> ℹ [2024-05-12 07:01:10.741511]: Fitting sample: Sample_1
#> ℹ [2024-05-12 07:01:10.742987]: Fitting sample: Sample_2
#> ℹ [2024-05-12 07:01:10.744415]: Fitting sample: Sample_3
#> ℹ [2024-05-12 07:01:10.745835]: Fitting sample: Sample_4
#> ℹ [2024-05-12 07:01:10.747268]: Fitting sample: Sample_5
#> ℹ [2024-05-12 07:01:10.758644]: Fitting sample: Sample_6
#> ℹ [2024-05-12 07:01:10.760172]: Fitting sample: Sample_7
#> ℹ [2024-05-12 07:01:10.761592]: Fitting sample: Sample_8
#> ℹ [2024-05-12 07:01:10.76301]: Fitting sample: Sample_9
#> ℹ [2024-05-12 07:01:10.764406]: Fitting sample: Sample_10
#> ℹ [2024-05-12 07:01:10.765795]: Fitting sample: Sample_11
#> ℹ [2024-05-12 07:01:10.767211]: Fitting sample: Sample_12
#> ℹ [2024-05-12 07:01:10.768608]: Fitting sample: Sample_13
#> ℹ [2024-05-12 07:01:10.770122]: Fitting sample: Sample_14
#> ℹ [2024-05-12 07:01:10.771557]: Fitting sample: Sample_15
#> ℹ [2024-05-12 07:01:10.772959]: Fitting sample: Sample_16
#> ℹ [2024-05-12 07:01:10.774357]: Fitting sample: Sample_17
#> ℹ [2024-05-12 07:01:10.77578]: Fitting sample: Sample_18
#> ℹ [2024-05-12 07:01:10.777172]: Fitting sample: Sample_19
#> ℹ [2024-05-12 07:01:10.778559]: Fitting sample: Sample_20
#> ℹ [2024-05-12 07:01:10.77999]: Fitting sample: Sample_21
#> ℹ [2024-05-12 07:01:10.781425]: Fitting sample: Sample_22
#> ℹ [2024-05-12 07:01:10.782841]: Fitting sample: Sample_23
#> ℹ [2024-05-12 07:01:10.784241]: Fitting sample: Sample_24
#> ℹ [2024-05-12 07:01:10.785656]: Fitting sample: Sample_25
#> ℹ [2024-05-12 07:01:10.787075]: Fitting sample: Sample_26
#> ℹ [2024-05-12 07:01:10.788455]: Fitting sample: Sample_27
#> ℹ [2024-05-12 07:01:10.78984]: Fitting sample: Sample_28
#> ℹ [2024-05-12 07:01:10.791275]: Fitting sample: Sample_29
#> ℹ [2024-05-12 07:01:10.792659]: Fitting sample: Sample_30
#> ✔ [2024-05-12 07:01:10.794041]: Done.
#> ℹ [2024-05-12 07:01:10.795382]: Generating output signature exposures.
#> ✔ [2024-05-12 07:01:10.805305]: Done.
#> ℹ [2024-05-12 07:01:10.806687]: Calculating errors (Frobenius Norm).
#> ✔ [2024-05-12 07:01:10.809014]: Done.
#> ℹ [2024-05-12 07:01:10.810423]: 0.092 secs elapsed.
#> ℹ [2024-05-12 07:01:10.813304]: Getting bootstrap exposures (&errors/similarity) for different methods.
#> ℹ [2024-05-12 07:01:10.81466]: This step is time consuming, please be patient.
#> ℹ [2024-05-12 07:01:10.816225]: Processing sample `Sample_1`.
#> ℹ [2024-05-12 07:01:12.004578]: Processing sample `Sample_2`.
#> ℹ [2024-05-12 07:01:13.155352]: Processing sample `Sample_3`.
#> ℹ [2024-05-12 07:01:14.296069]: Processing sample `Sample_4`.
#> ℹ [2024-05-12 07:01:15.438757]: Processing sample `Sample_5`.
#> ℹ [2024-05-12 07:01:16.575728]: Processing sample `Sample_6`.
#> ℹ [2024-05-12 07:01:17.707898]: Processing sample `Sample_7`.
#> ℹ [2024-05-12 07:01:18.840453]: Processing sample `Sample_8`.
#> ℹ [2024-05-12 07:01:19.97191]: Processing sample `Sample_9`.
#> ℹ [2024-05-12 07:01:21.112866]: Processing sample `Sample_10`.
#> ℹ [2024-05-12 07:01:22.254404]: Processing sample `Sample_11`.
#> ℹ [2024-05-12 07:01:23.393846]: Processing sample `Sample_12`.
#> ℹ [2024-05-12 07:01:24.574183]: Processing sample `Sample_13`.
#> ℹ [2024-05-12 07:01:25.710462]: Processing sample `Sample_14`.
#> ℹ [2024-05-12 07:01:26.844918]: Processing sample `Sample_15`.
#> ℹ [2024-05-12 07:01:27.986786]: Processing sample `Sample_16`.
#> ℹ [2024-05-12 07:01:29.133231]: Processing sample `Sample_17`.
#> ℹ [2024-05-12 07:01:30.315032]: Processing sample `Sample_18`.
#> ℹ [2024-05-12 07:01:31.454454]: Processing sample `Sample_19`.
#> ℹ [2024-05-12 07:01:32.6195]: Processing sample `Sample_20`.
#> ℹ [2024-05-12 07:01:33.759698]: Processing sample `Sample_21`.
#> ℹ [2024-05-12 07:01:34.90626]: Processing sample `Sample_22`.
#> ℹ [2024-05-12 07:01:36.076817]: Processing sample `Sample_23`.
#> ℹ [2024-05-12 07:01:37.216271]: Processing sample `Sample_24`.
#> ℹ [2024-05-12 07:01:38.395275]: Processing sample `Sample_25`.
#> ℹ [2024-05-12 07:01:39.542415]: Processing sample `Sample_26`.
#> ℹ [2024-05-12 07:01:40.691327]: Processing sample `Sample_27`.
#> ℹ [2024-05-12 07:01:41.836666]: Processing sample `Sample_28`.
#> ℹ [2024-05-12 07:01:42.969244]: Processing sample `Sample_29`.
#> ℹ [2024-05-12 07:01:44.106462]: Processing sample `Sample_30`.
#> ✔ [2024-05-12 07:01:45.242204]: Gotten.
#> ℹ [2024-05-12 07:01:45.245088]: Reporting p values...
#> ℹ [2024-05-12 07:01:45.246629]: Started.
#> ✔ [2024-05-12 07:01:45.24799]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.250015]: Done.
#> ℹ [2024-05-12 07:01:45.25142]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.252795]: Started.
#> ✔ [2024-05-12 07:01:45.254096]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.256017]: Done.
#> ℹ [2024-05-12 07:01:45.257391]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.258749]: Started.
#> ✔ [2024-05-12 07:01:45.260096]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.261976]: Done.
#> ℹ [2024-05-12 07:01:45.263387]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.264746]: Started.
#> ✔ [2024-05-12 07:01:45.266074]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.267997]: Done.
#> ℹ [2024-05-12 07:01:45.269371]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.270729]: Started.
#> ✔ [2024-05-12 07:01:45.272062]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.273945]: Done.
#> ℹ [2024-05-12 07:01:45.27534]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.276697]: Started.
#> ✔ [2024-05-12 07:01:45.278016]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.279959]: Done.
#> ℹ [2024-05-12 07:01:45.281327]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.282671]: Started.
#> ✔ [2024-05-12 07:01:45.284005]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.285901]: Done.
#> ℹ [2024-05-12 07:01:45.287288]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.288639]: Started.
#> ✔ [2024-05-12 07:01:45.289953]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.291863]: Done.
#> ℹ [2024-05-12 07:01:45.293248]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.294601]: Started.
#> ✔ [2024-05-12 07:01:45.296023]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.297921]: Done.
#> ℹ [2024-05-12 07:01:45.299356]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.300709]: Started.
#> ✔ [2024-05-12 07:01:45.302041]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.303971]: Done.
#> ℹ [2024-05-12 07:01:45.305389]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.306745]: Started.
#> ✔ [2024-05-12 07:01:45.308089]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.309985]: Done.
#> ℹ [2024-05-12 07:01:45.311373]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.312733]: Started.
#> ✔ [2024-05-12 07:01:45.314045]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.315958]: Done.
#> ℹ [2024-05-12 07:01:45.317336]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.318686]: Started.
#> ✔ [2024-05-12 07:01:45.320031]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.321915]: Done.
#> ℹ [2024-05-12 07:01:45.323311]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.324676]: Started.
#> ✔ [2024-05-12 07:01:45.325995]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.327898]: Done.
#> ℹ [2024-05-12 07:01:45.329277]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.33063]: Started.
#> ✔ [2024-05-12 07:01:45.331975]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.333865]: Done.
#> ℹ [2024-05-12 07:01:45.335262]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.336627]: Started.
#> ✔ [2024-05-12 07:01:45.337939]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.339876]: Done.
#> ℹ [2024-05-12 07:01:45.341244]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.342595]: Started.
#> ✔ [2024-05-12 07:01:45.343934]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.345821]: Done.
#> ℹ [2024-05-12 07:01:45.347208]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.348562]: Started.
#> ✔ [2024-05-12 07:01:45.349894]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.351836]: Done.
#> ℹ [2024-05-12 07:01:45.353211]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.354554]: Started.
#> ✔ [2024-05-12 07:01:45.35589]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.357795]: Done.
#> ℹ [2024-05-12 07:01:45.359195]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.360565]: Started.
#> ✔ [2024-05-12 07:01:45.361882]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.363798]: Done.
#> ℹ [2024-05-12 07:01:45.365168]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.366522]: Started.
#> ✔ [2024-05-12 07:01:45.367884]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.369774]: Done.
#> ℹ [2024-05-12 07:01:45.371199]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.372565]: Started.
#> ✔ [2024-05-12 07:01:45.373875]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.37579]: Done.
#> ℹ [2024-05-12 07:01:45.377167]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.378516]: Started.
#> ✔ [2024-05-12 07:01:45.379864]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.381742]: Done.
#> ℹ [2024-05-12 07:01:45.383135]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.384486]: Started.
#> ✔ [2024-05-12 07:01:45.3858]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.387726]: Done.
#> ℹ [2024-05-12 07:01:45.389101]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.390458]: Started.
#> ✔ [2024-05-12 07:01:45.391792]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.393763]: Done.
#> ℹ [2024-05-12 07:01:45.395154]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.396513]: Started.
#> ✔ [2024-05-12 07:01:45.397822]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.39977]: Done.
#> ℹ [2024-05-12 07:01:45.401146]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.402503]: Started.
#> ✔ [2024-05-12 07:01:45.40386]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.405751]: Done.
#> ℹ [2024-05-12 07:01:45.407145]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.408504]: Started.
#> ✔ [2024-05-12 07:01:45.409819]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.411727]: Done.
#> ℹ [2024-05-12 07:01:45.413097]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.414451]: Started.
#> ✔ [2024-05-12 07:01:45.415786]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.417719]: Done.
#> ℹ [2024-05-12 07:01:45.419107]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:45.420475]: Started.
#> ✔ [2024-05-12 07:01:45.421792]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:45.423722]: Done.
#> ℹ [2024-05-12 07:01:45.425094]: 0.005 secs elapsed.
#> ✔ [2024-05-12 07:01:45.426448]: Done.
#> ℹ [2024-05-12 07:01:45.427789]: Cleaning results...
#> ✔ [2024-05-12 07:01:45.657689]: Outputing.
#> ℹ [2024-05-12 07:01:45.659873]: Total 34.949 secs elapsed.
rv
#> $expo
#>       method    sample    sig   exposure    type
#>       <char>    <char> <fctr>      <num>  <char>
#>    1:     QP  Sample_1   SBS1    0.00000 optimal
#>    2:     QP  Sample_2   SBS1 1115.51740 optimal
#>    3:     QP  Sample_3   SBS1 1181.67357 optimal
#>    4:     QP  Sample_4   SBS1    0.00000 optimal
#>    5:     QP  Sample_5   SBS1   27.23392 optimal
#>   ---                                           
#> 1646:     QP Sample_30   SBS1  230.74630  Rep_10
#> 1647:     QP Sample_30   SBS5 2062.72035  Rep_10
#> 1648:     QP Sample_30   SBS9  625.53335  Rep_10
#> 1649:     QP Sample_30   SBS2    0.00000  Rep_10
#> 1650:     QP Sample_30  SBS13    0.00000  Rep_10
#> 
#> $error
#>      method    sample   errors    type
#>      <char>    <char>    <num>  <char>
#>   1:     QP  Sample_1 2679.128 optimal
#>   2:     QP  Sample_2  702.209 optimal
#>   3:     QP  Sample_3  542.401 optimal
#>   4:     QP  Sample_4  431.054 optimal
#>   5:     QP  Sample_5  132.440 optimal
#>  ---                                  
#> 326:     QP Sample_30  184.349   Rep_6
#> 327:     QP Sample_30  185.102   Rep_7
#> 328:     QP Sample_30  184.294   Rep_8
#> 329:     QP Sample_30  184.066   Rep_9
#> 330:     QP Sample_30  184.032  Rep_10
#> 
#> $cosine
#>      method    sample   cosine    type
#>      <char>    <char>    <num>  <char>
#>   1:     QP  Sample_1 0.870282 optimal
#>   2:     QP  Sample_2 0.982053 optimal
#>   3:     QP  Sample_3 0.950683 optimal
#>   4:     QP  Sample_4 0.977040 optimal
#>   5:     QP  Sample_5 0.986924 optimal
#>  ---                                  
#> 326:     QP Sample_30 0.885355   Rep_6
#> 327:     QP Sample_30 0.864371   Rep_7
#> 328:     QP Sample_30 0.865994   Rep_8
#> 329:     QP Sample_30 0.868274   Rep_9
#> 330:     QP Sample_30 0.878272  Rep_10
#> 
#> $p_val
#>         sample method threshold    sig      p_value
#>         <char> <char>    <char> <char>        <num>
#>   1:  Sample_1     QP      0.05   SBS1 1.000000e+00
#>   2:  Sample_1     QP      0.05   SBS5 2.070854e-23
#>   3:  Sample_1     QP      0.05   SBS9 2.441497e-16
#>   4:  Sample_1     QP      0.05   SBS2 1.000000e+00
#>   5:  Sample_1     QP      0.05  SBS13 1.000000e+00
#>  ---                                               
#> 146: Sample_30     QP      0.05   SBS1 2.906989e-07
#> 147: Sample_30     QP      0.05   SBS5 2.195250e-19
#> 148: Sample_30     QP      0.05   SBS9 1.485679e-12
#> 149: Sample_30     QP      0.05   SBS2 1.000000e+00
#> 150: Sample_30     QP      0.05  SBS13 1.000000e+00
#> 
# }

# For general purpose --------------------
W <- matrix(c(1, 2, 3, 4, 5, 6), ncol = 2)
colnames(W) <- c("sig1", "sig2")
W <- apply(W, 2, function(x) x / sum(x))

H <- matrix(c(2, 5, 3, 6, 1, 9, 1, 2), ncol = 4)
colnames(H) <- paste0("samp", 1:4)

V <- W %*% H
V
#>         samp1 samp2    samp3 samp4
#> [1,] 1.666667   2.1 2.566667   0.7
#> [2,] 2.333333   3.0 3.333333   1.0
#> [3,] 3.000000   3.9 4.100000   1.3

if (requireNamespace("quadprog")) {
  z10 <- sig_fit_bootstrap_batch(V, sig = W, n = 10)
  z10
}
#> ℹ [2024-05-12 07:01:45.673453]: Batch Bootstrap Signature Exposure Analysis Started.
#> ℹ [2024-05-12 07:01:45.674956]: Samples to be filtered out: 
#> ℹ [2024-05-12 07:01:45.676385]: Finding optimal exposures (&errors) for different methods.
#> ℹ [2024-05-12 07:01:45.677743]: Calling method `QP`.
#> ℹ [2024-05-12 07:01:45.681046]: Started.
#> ℹ [2024-05-12 07:01:45.682409]: Signature index not detected.
#> ✔ [2024-05-12 07:01:45.683819]: Signature matrix/data.frame detected.
#> ✔ [2024-05-12 07:01:45.685188]: Database and index checked.
#> ✔ [2024-05-12 07:01:45.686627]: Signature normalized.
#> ℹ [2024-05-12 07:01:45.688039]: Checking row number for catalog matrix and signature matrix.
#> ✔ [2024-05-12 07:01:45.689421]: Checked.
#> ✔ [2024-05-12 07:01:45.690767]: Method 'QP' detected.
#> ✔ [2024-05-12 07:01:45.692172]: Corresponding function generated.
#> ℹ [2024-05-12 07:01:45.693514]: Calling function.
#> ℹ [2024-05-12 07:01:45.695014]: Fitting sample: samp1
#> ℹ [2024-05-12 07:01:45.696465]: Fitting sample: samp2
#> ℹ [2024-05-12 07:01:45.6979]: Fitting sample: samp3
#> ℹ [2024-05-12 07:01:45.699355]: Fitting sample: samp4
#> ✔ [2024-05-12 07:01:45.700772]: Done.
#> ℹ [2024-05-12 07:01:45.70212]: Generating output signature exposures.
#> ✔ [2024-05-12 07:01:45.710862]: Done.
#> ℹ [2024-05-12 07:01:45.712321]: Calculating errors (Frobenius Norm).
#> ✔ [2024-05-12 07:01:45.713778]: Done.
#> ℹ [2024-05-12 07:01:45.715249]: 0.034 secs elapsed.
#> ℹ [2024-05-12 07:01:45.71804]: Getting bootstrap exposures (&errors/similarity) for different methods.
#> ℹ [2024-05-12 07:01:45.719713]: This step is time consuming, please be patient.
#> ℹ [2024-05-12 07:01:45.721224]: Processing sample `samp1`.
#> ℹ [2024-05-12 07:01:46.776308]: Processing sample `samp2`.
#> ℹ [2024-05-12 07:01:47.819568]: Processing sample `samp3`.
#> ℹ [2024-05-12 07:01:48.855972]: Processing sample `samp4`.
#> ✔ [2024-05-12 07:01:49.897231]: Gotten.
#> ℹ [2024-05-12 07:01:49.899719]: Reporting p values...
#> ℹ [2024-05-12 07:01:49.901404]: Started.
#> ✔ [2024-05-12 07:01:49.902883]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:49.904747]: Done.
#> ℹ [2024-05-12 07:01:49.906188]: 0.005 secs elapsed.
#> ℹ [2024-05-12 07:01:49.907641]: Started.
#> ✔ [2024-05-12 07:01:49.908995]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:49.910697]: Done.
#> ℹ [2024-05-12 07:01:49.912124]: 0.004 secs elapsed.
#> ℹ [2024-05-12 07:01:49.913544]: Started.
#> ✔ [2024-05-12 07:01:49.91491]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:49.916591]: Done.
#> ℹ [2024-05-12 07:01:49.917986]: 0.004 secs elapsed.
#> ℹ [2024-05-12 07:01:49.919404]: Started.
#> ✔ [2024-05-12 07:01:49.920735]: Batch mode enabled.
#> ✔ [2024-05-12 07:01:49.922384]: Done.
#> ℹ [2024-05-12 07:01:49.923796]: 0.004 secs elapsed.
#> ✔ [2024-05-12 07:01:49.925171]: Done.
#> ℹ [2024-05-12 07:01:49.926509]: Cleaning results...
#> ✔ [2024-05-12 07:01:49.961895]: Outputing.
#> ℹ [2024-05-12 07:01:49.963842]: Total 4.29 secs elapsed.
#> $expo
#>     method sample    sig  exposure    type
#>     <char> <char> <fctr>     <num>  <char>
#>  1:     QP  samp1   sig1  2.000000 optimal
#>  2:     QP  samp2   sig1  3.000000 optimal
#>  3:     QP  samp3   sig1  1.000000 optimal
#>  4:     QP  samp4   sig1  1.000000 optimal
#>  5:     QP  samp1   sig2  5.000000 optimal
#>  6:     QP  samp2   sig2  6.000000 optimal
#>  7:     QP  samp3   sig2  9.000000 optimal
#>  8:     QP  samp4   sig2  2.000000 optimal
#>  9:     QP  samp1   sig1  0.333333   Rep_1
#> 10:     QP  samp1   sig2  6.666667   Rep_1
#> 11:     QP  samp1   sig1  0.000000   Rep_2
#> 12:     QP  samp1   sig2  7.000000   Rep_2
#> 13:     QP  samp1   sig1  7.000000   Rep_3
#> 14:     QP  samp1   sig2  0.000000   Rep_3
#> 15:     QP  samp1   sig1  0.000000   Rep_4
#> 16:     QP  samp1   sig2  7.000000   Rep_4
#> 17:     QP  samp1   sig1  0.000000   Rep_5
#> 18:     QP  samp1   sig2  7.000000   Rep_5
#> 19:     QP  samp1   sig1  0.000000   Rep_6
#> 20:     QP  samp1   sig2  7.000000   Rep_6
#> 21:     QP  samp1   sig1  5.333333   Rep_7
#> 22:     QP  samp1   sig2  1.666667   Rep_7
#> 23:     QP  samp1   sig1  0.333333   Rep_8
#> 24:     QP  samp1   sig2  6.666667   Rep_8
#> 25:     QP  samp1   sig1  7.000000   Rep_9
#> 26:     QP  samp1   sig2  0.000000   Rep_9
#> 27:     QP  samp1   sig1  7.000000  Rep_10
#> 28:     QP  samp1   sig2  0.000000  Rep_10
#> 29:     QP  samp2   sig1  0.000000   Rep_1
#> 30:     QP  samp2   sig2  9.000000   Rep_1
#> 31:     QP  samp2   sig1  9.000000   Rep_2
#> 32:     QP  samp2   sig2  0.000000   Rep_2
#> 33:     QP  samp2   sig1  9.000000   Rep_3
#> 34:     QP  samp2   sig2  0.000000   Rep_3
#> 35:     QP  samp2   sig1  9.000000   Rep_4
#> 36:     QP  samp2   sig2  0.000000   Rep_4
#> 37:     QP  samp2   sig1  0.000000   Rep_5
#> 38:     QP  samp2   sig2  9.000000   Rep_5
#> 39:     QP  samp2   sig1  0.000000   Rep_6
#> 40:     QP  samp2   sig2  9.000000   Rep_6
#> 41:     QP  samp2   sig1  0.000000   Rep_7
#> 42:     QP  samp2   sig2  9.000000   Rep_7
#> 43:     QP  samp2   sig1  0.000000   Rep_8
#> 44:     QP  samp2   sig2  9.000000   Rep_8
#> 45:     QP  samp2   sig1  0.000000   Rep_9
#> 46:     QP  samp2   sig2  9.000000   Rep_9
#> 47:     QP  samp2   sig1  9.000000  Rep_10
#> 48:     QP  samp2   sig2  0.000000  Rep_10
#> 49:     QP  samp3   sig1  0.000000   Rep_1
#> 50:     QP  samp3   sig2 10.000000   Rep_1
#> 51:     QP  samp3   sig1  0.000000   Rep_2
#> 52:     QP  samp3   sig2 10.000000   Rep_2
#> 53:     QP  samp3   sig1  0.000000   Rep_3
#> 54:     QP  samp3   sig2 10.000000   Rep_3
#> 55:     QP  samp3   sig1  3.333333   Rep_4
#> 56:     QP  samp3   sig2  6.666667   Rep_4
#> 57:     QP  samp3   sig1  8.333333   Rep_5
#> 58:     QP  samp3   sig2  1.666667   Rep_5
#> 59:     QP  samp3   sig1  0.000000   Rep_6
#> 60:     QP  samp3   sig2 10.000000   Rep_6
#> 61:     QP  samp3   sig1  8.333333   Rep_7
#> 62:     QP  samp3   sig2  1.666667   Rep_7
#> 63:     QP  samp3   sig1  0.000000   Rep_8
#> 64:     QP  samp3   sig2 10.000000   Rep_8
#> 65:     QP  samp3   sig1  3.333333   Rep_9
#> 66:     QP  samp3   sig2  6.666667   Rep_9
#> 67:     QP  samp3   sig1 10.000000  Rep_10
#> 68:     QP  samp3   sig2  0.000000  Rep_10
#> 69:     QP  samp4   sig1  3.000000   Rep_1
#> 70:     QP  samp4   sig2  0.000000   Rep_1
#> 71:     QP  samp4   sig1  3.000000   Rep_2
#> 72:     QP  samp4   sig2  0.000000   Rep_2
#> 73:     QP  samp4   sig1  3.000000   Rep_3
#> 74:     QP  samp4   sig2  0.000000   Rep_3
#> 75:     QP  samp4   sig1  3.000000   Rep_4
#> 76:     QP  samp4   sig2  0.000000   Rep_4
#> 77:     QP  samp4   sig1  3.000000   Rep_5
#> 78:     QP  samp4   sig2  0.000000   Rep_5
#> 79:     QP  samp4   sig1  0.000000   Rep_6
#> 80:     QP  samp4   sig2  3.000000   Rep_6
#> 81:     QP  samp4   sig1  0.000000   Rep_7
#> 82:     QP  samp4   sig2  3.000000   Rep_7
#> 83:     QP  samp4   sig1  3.000000   Rep_8
#> 84:     QP  samp4   sig2  0.000000   Rep_8
#> 85:     QP  samp4   sig1  3.000000   Rep_9
#> 86:     QP  samp4   sig2  0.000000   Rep_9
#> 87:     QP  samp4   sig1  0.000000  Rep_10
#> 88:     QP  samp4   sig2  3.000000  Rep_10
#>     method sample    sig  exposure    type
#> 
#> $error
#>     method sample errors    type
#>     <char> <char>  <num>  <char>
#>  1:     QP  samp1  0.000 optimal
#>  2:     QP  samp2  0.000 optimal
#>  3:     QP  samp3  0.000 optimal
#>  4:     QP  samp4  0.000 optimal
#>  5:     QP  samp1  0.236   Rep_1
#>  6:     QP  samp1  0.283   Rep_2
#>  7:     QP  samp1  0.707   Rep_3
#>  8:     QP  samp1  0.283   Rep_4
#>  9:     QP  samp1  0.283   Rep_5
#> 10:     QP  samp1  0.283   Rep_6
#> 11:     QP  samp1  0.471   Rep_7
#> 12:     QP  samp1  0.236   Rep_8
#> 13:     QP  samp1  0.707   Rep_9
#> 14:     QP  samp1  0.707  Rep_10
#> 15:     QP  samp2  0.424   Rep_1
#> 16:     QP  samp2  0.849   Rep_2
#> 17:     QP  samp2  0.849   Rep_3
#> 18:     QP  samp2  0.849   Rep_4
#> 19:     QP  samp2  0.424   Rep_5
#> 20:     QP  samp2  0.424   Rep_6
#> 21:     QP  samp2  0.424   Rep_7
#> 22:     QP  samp2  0.424   Rep_8
#> 23:     QP  samp2  0.424   Rep_9
#> 24:     QP  samp2  0.849  Rep_10
#> 25:     QP  samp3  0.141   Rep_1
#> 26:     QP  samp3  0.141   Rep_2
#> 27:     QP  samp3  0.141   Rep_3
#> 28:     QP  samp3  0.330   Rep_4
#> 29:     QP  samp3  1.037   Rep_5
#> 30:     QP  samp3  0.141   Rep_6
#> 31:     QP  samp3  1.037   Rep_7
#> 32:     QP  samp3  0.141   Rep_8
#> 33:     QP  samp3  0.330   Rep_9
#> 34:     QP  samp3  1.273  Rep_10
#> 35:     QP  samp4  0.283   Rep_1
#> 36:     QP  samp4  0.283   Rep_2
#> 37:     QP  samp4  0.283   Rep_3
#> 38:     QP  samp4  0.283   Rep_4
#> 39:     QP  samp4  0.283   Rep_5
#> 40:     QP  samp4  0.141   Rep_6
#> 41:     QP  samp4  0.141   Rep_7
#> 42:     QP  samp4  0.283   Rep_8
#> 43:     QP  samp4  0.283   Rep_9
#> 44:     QP  samp4  0.141  Rep_10
#>     method sample errors    type
#> 
#> $cosine
#>     method sample   cosine    type
#>     <char> <char>    <num>  <char>
#>  1:     QP  samp1 1.000000 optimal
#>  2:     QP  samp2 1.000000 optimal
#>  3:     QP  samp3 1.000000 optimal
#>  4:     QP  samp4 1.000000 optimal
#>  5:     QP  samp1 0.895314   Rep_1
#>  6:     QP  samp1 0.795784   Rep_2
#>  7:     QP  samp1 0.925820   Rep_3
#>  8:     QP  samp1 0.915052   Rep_4
#>  9:     QP  samp1 0.634858   Rep_5
#> 10:     QP  samp1 0.967382   Rep_6
#> 11:     QP  samp1 0.982299   Rep_7
#> 12:     QP  samp1 0.995086   Rep_8
#> 13:     QP  samp1 0.908688   Rep_9
#> 14:     QP  samp1 0.991460  Rep_10
#> 15:     QP  samp2 0.909964   Rep_1
#> 16:     QP  samp2 0.960001   Rep_2
#> 17:     QP  samp2 0.977008   Rep_3
#> 18:     QP  samp2 0.993859   Rep_4
#> 19:     QP  samp2 0.973450   Rep_5
#> 20:     QP  samp2 0.986928   Rep_6
#> 21:     QP  samp2 0.828303   Rep_7
#> 22:     QP  samp2 0.892710   Rep_8
#> 23:     QP  samp2 0.973450   Rep_9
#> 24:     QP  samp2 0.918262  Rep_10
#> 25:     QP  samp3 0.682355   Rep_1
#> 26:     QP  samp3 0.887369   Rep_2
#> 27:     QP  samp3 0.996748   Rep_3
#> 28:     QP  samp3 0.964274   Rep_4
#> 29:     QP  samp3 0.949101   Rep_5
#> 30:     QP  samp3 0.949671   Rep_6
#> 31:     QP  samp3 0.997805   Rep_7
#> 32:     QP  samp3 0.949671   Rep_8
#> 33:     QP  samp3 0.964274   Rep_9
#> 34:     QP  samp3 0.944911  Rep_10
#> 35:     QP  samp4 0.956183   Rep_1
#> 36:     QP  samp4 0.956183   Rep_2
#> 37:     QP  samp4 0.956183   Rep_3
#> 38:     QP  samp4 0.801784   Rep_4
#> 39:     QP  samp4 0.836660   Rep_5
#> 40:     QP  samp4 0.986928   Rep_6
#> 41:     QP  samp4 0.713506   Rep_7
#> 42:     QP  samp4 0.836660   Rep_8
#> 43:     QP  samp4 0.836660   Rep_9
#> 44:     QP  samp4 0.713506  Rep_10
#>     method sample   cosine    type
#> 
#> $p_val
#>    sample method threshold    sig      p_value
#>    <char> <char>    <char> <char>        <num>
#> 1:  samp1     QP      0.05   sig1 0.0277068123
#> 2:  samp1     QP      0.05   sig2 0.0024736045
#> 3:  samp2     QP      0.05   sig1 0.0303471317
#> 4:  samp2     QP      0.05   sig2 0.0041413035
#> 5:  samp3     QP      0.05   sig1 0.0279108829
#> 6:  samp3     QP      0.05   sig2 0.0005029653
#> 7:  samp4     QP      0.05   sig1 0.0010629845
#> 8:  samp4     QP      0.05   sig2 0.0680670588
#>