Exposure Instability Analysis of Signature Exposures with Bootstrap

Exposure Instability Analysis of Signature Exposures with Bootstrap

sig_fit_bootstrap_batch(
  catalogue_matrix,
  methods = c("LS", "QP"),
  n = 100L,
  p_val_thresholds = c(0.05),
  use_parallel = FALSE,
  seed = 123456L,
  ...
)

Arguments

catalogue_matrix
a numeric matrix V with row representing components and columns representing samples, typically you can get nmf_matrix from sig_tally() and transpose it by t().

methods
a subset of c("LS", "QP", "SA").

n
the number of bootstrap replicates.

p_val_thresholds
a vector of relative exposure threshold for calculating p values.

use_parallel
if TRUE, use parallel computation based on furrr package, not implemented yet.

seed
random seed to reproduce the result.

...
other common parameters passing to sig_fit_bootstrap, including sig, sig_index, sig_db, db_type, mode, etc.

Value

a list of data.table.

See also

Examples

W <- matrix(c(1, 2, 3, 4, 5, 6), ncol = 2)
colnames(W) <- c("sig1", "sig2")
W <- apply(W, 2, function(x) x / sum(x))

H <- matrix(c(2, 5, 3, 6, 1, 9, 1, 2), ncol = 4)
colnames(H) <- paste0("samp", 1:4)

V <- W %*% H
V

#>         samp1 samp2    samp3 samp4
#> [1,] 1.666667   2.1 2.566667   0.7
#> [2,] 2.333333   3.0 3.333333   1.0
#> [3,] 3.000000   3.9 4.100000   1.3

if (requireNamespace("lsei") & requireNamespace("quadprog")) {
  z <- sig_fit_bootstrap_batch(V, sig = W, n = 10)
  z
}

#> ℹ [2020-05-10 20:35:14]: Batch Bootstrap Signature Exposure Analysis Started.

#> ℹ [2020-05-10 20:35:14]: Finding optimal exposures (&errors) for different methods.

#> ℹ [2020-05-10 20:35:14]: Calling method `LS`.

#> ℹ [2020-05-10 20:35:14]: Started.

#> ✓ [2020-05-10 20:35:14]: Signature matrix detected.

#> ✓ [2020-05-10 20:35:14]: Database and index checked.

#> ✓ [2020-05-10 20:35:14]: Signature normalized.

#> ℹ [2020-05-10 20:35:14]: Checking row number for catalog matrix and signature matrix.

#> ✓ [2020-05-10 20:35:14]: Checked.

#> ✓ [2020-05-10 20:35:14]: Method 'LS' detected.

#> ✓ [2020-05-10 20:35:14]: Corresponding function generated.

#> ℹ [2020-05-10 20:35:14]: Calling function.

#> ✓ [2020-05-10 20:35:14]: Done.

#> ℹ [2020-05-10 20:35:14]: Generating output signature exposures.

#> ✓ [2020-05-10 20:35:14]: Done.

#> ℹ [2020-05-10 20:35:14]: Calculating errors (Frobenius Norm).

#> ✓ [2020-05-10 20:35:14]: Done.

#> ℹ [2020-05-10 20:35:14]: 0.039 secs elapsed.

#> ℹ [2020-05-10 20:35:14]: Calling method `QP`.

#> ℹ [2020-05-10 20:35:14]: Started.

#> ✓ [2020-05-10 20:35:14]: Signature matrix detected.

#> ✓ [2020-05-10 20:35:14]: Database and index checked.

#> ✓ [2020-05-10 20:35:14]: Signature normalized.

#> ℹ [2020-05-10 20:35:14]: Checking row number for catalog matrix and signature matrix.

#> ✓ [2020-05-10 20:35:14]: Checked.

#> ✓ [2020-05-10 20:35:14]: Method 'QP' detected.

#> ✓ [2020-05-10 20:35:14]: Corresponding function generated.

#> ℹ [2020-05-10 20:35:14]: Calling function.

#> ✓ [2020-05-10 20:35:14]: Done.

#> ℹ [2020-05-10 20:35:14]: Generating output signature exposures.

#> ✓ [2020-05-10 20:35:14]: Done.

#> ℹ [2020-05-10 20:35:14]: Calculating errors (Frobenius Norm).

#> ✓ [2020-05-10 20:35:14]: Done.

#> ℹ [2020-05-10 20:35:14]: 0.037 secs elapsed.

#> ℹ [2020-05-10 20:35:14]: Getting bootstrap exposures (&errors) for different methods.

#> ℹ [2020-05-10 20:35:14]: This step is time consuming, please be patient.

#> ℹ [2020-05-10 20:35:14]: Processing sample `samp1`.

#> ℹ [2020-05-10 20:35:14]: Started.

#> ℹ [2020-05-10 20:35:14]: Checking catalog.

#> ✓ [2020-05-10 20:35:14]: Done.

#> ℹ [2020-05-10 20:35:14]: About to start bootstrap.

#> → Bootstrapping 10 times.

#> → Total 10 times, starting no.1.

#> → Total 10 times, starting no.2.

#> → Total 10 times, starting no.3.

#> → Total 10 times, starting no.4.

#> → Total 10 times, starting no.5.

#> → Total 10 times, starting no.6.

#> → Total 10 times, starting no.7.

#> → Total 10 times, starting no.8.

#> → Total 10 times, starting no.9.

#> → Total 10 times, starting no.10.

#>                                   

#> ✓ [2020-05-10 20:35:14]: Bootstrap done.

#> ✓ [2020-05-10 20:35:14]: Signature exposures collected.

#> ✓ [2020-05-10 20:35:14]: Errors collected.

#> ✓ [2020-05-10 20:35:14]: Done.

#> ℹ [2020-05-10 20:35:14]: 0.47 secs elapsed.

#> ℹ [2020-05-10 20:35:14]: Started.

#> ℹ [2020-05-10 20:35:14]: Checking catalog.

#> ✓ [2020-05-10 20:35:14]: Done.

#> ℹ [2020-05-10 20:35:14]: About to start bootstrap.

#> → Bootstrapping 10 times.

#> → Total 10 times, starting no.1.

#> → Total 10 times, starting no.2.

#> → Total 10 times, starting no.3.

#> → Total 10 times, starting no.4.

#> → Total 10 times, starting no.5.

#> → Total 10 times, starting no.6.

#> → Total 10 times, starting no.7.

#> → Total 10 times, starting no.8.

#> → Total 10 times, starting no.9.

#> → Total 10 times, starting no.10.

#>                                   

#> ✓ [2020-05-10 20:35:15]: Bootstrap done.

#> ✓ [2020-05-10 20:35:15]: Signature exposures collected.

#> ✓ [2020-05-10 20:35:15]: Errors collected.

#> ✓ [2020-05-10 20:35:15]: Done.

#> ℹ [2020-05-10 20:35:15]: 0.494 secs elapsed.

#> ℹ [2020-05-10 20:35:15]: Processing sample `samp2`.

#> ℹ [2020-05-10 20:35:15]: Started.

#> ℹ [2020-05-10 20:35:15]: Checking catalog.

#> ✓ [2020-05-10 20:35:15]: Done.

#> ℹ [2020-05-10 20:35:15]: About to start bootstrap.

#> → Bootstrapping 10 times.

#> → Total 10 times, starting no.1.

#> → Total 10 times, starting no.2.

#> → Total 10 times, starting no.3.

#> → Total 10 times, starting no.4.

#> → Total 10 times, starting no.5.

#> → Total 10 times, starting no.6.

#> → Total 10 times, starting no.7.

#> → Total 10 times, starting no.8.

#> → Total 10 times, starting no.9.

#> → Total 10 times, starting no.10.

#>                                   

#> ✓ [2020-05-10 20:35:15]: Bootstrap done.

#> ✓ [2020-05-10 20:35:15]: Signature exposures collected.

#> ✓ [2020-05-10 20:35:15]: Errors collected.

#> ✓ [2020-05-10 20:35:15]: Done.

#> ℹ [2020-05-10 20:35:15]: 0.458 secs elapsed.

#> ℹ [2020-05-10 20:35:15]: Started.

#> ℹ [2020-05-10 20:35:15]: Checking catalog.

#> ✓ [2020-05-10 20:35:15]: Done.

#> ℹ [2020-05-10 20:35:15]: About to start bootstrap.

#> → Bootstrapping 10 times.

#> → Total 10 times, starting no.1.

#> → Total 10 times, starting no.2.

#> → Total 10 times, starting no.3.

#> → Total 10 times, starting no.4.

#> → Total 10 times, starting no.5.

#> → Total 10 times, starting no.6.

#> → Total 10 times, starting no.7.

#> → Total 10 times, starting no.8.

#> → Total 10 times, starting no.9.

#> → Total 10 times, starting no.10.

#>                                   

#> ✓ [2020-05-10 20:35:16]: Bootstrap done.

#> ✓ [2020-05-10 20:35:16]: Signature exposures collected.

#> ✓ [2020-05-10 20:35:16]: Errors collected.

#> ✓ [2020-05-10 20:35:16]: Done.

#> ℹ [2020-05-10 20:35:16]: 0.46 secs elapsed.

#> ℹ [2020-05-10 20:35:16]: Processing sample `samp3`.

#> ℹ [2020-05-10 20:35:16]: Started.

#> ℹ [2020-05-10 20:35:16]: Checking catalog.

#> ✓ [2020-05-10 20:35:16]: Done.

#> ℹ [2020-05-10 20:35:16]: About to start bootstrap.

#> → Bootstrapping 10 times.

#> → Total 10 times, starting no.1.

#> → Total 10 times, starting no.2.

#> → Total 10 times, starting no.3.

#> → Total 10 times, starting no.4.

#> → Total 10 times, starting no.5.

#> → Total 10 times, starting no.6.

#> → Total 10 times, starting no.7.

#> → Total 10 times, starting no.8.

#> → Total 10 times, starting no.9.

#> → Total 10 times, starting no.10.

#>                                   

#> ✓ [2020-05-10 20:35:16]: Bootstrap done.

#> ✓ [2020-05-10 20:35:16]: Signature exposures collected.

#> ✓ [2020-05-10 20:35:16]: Errors collected.

#> ✓ [2020-05-10 20:35:16]: Done.

#> ℹ [2020-05-10 20:35:16]: 0.461 secs elapsed.

#> ℹ [2020-05-10 20:35:16]: Started.

#> ℹ [2020-05-10 20:35:16]: Checking catalog.

#> ✓ [2020-05-10 20:35:16]: Done.

#> ℹ [2020-05-10 20:35:16]: About to start bootstrap.

#> → Bootstrapping 10 times.

#> → Total 10 times, starting no.1.

#> → Total 10 times, starting no.2.

#> → Total 10 times, starting no.3.

#> → Total 10 times, starting no.4.

#> → Total 10 times, starting no.5.

#> → Total 10 times, starting no.6.

#> → Total 10 times, starting no.7.

#> → Total 10 times, starting no.8.

#> → Total 10 times, starting no.9.

#> → Total 10 times, starting no.10.

#>                                   

#> ✓ [2020-05-10 20:35:17]: Bootstrap done.

#> ✓ [2020-05-10 20:35:17]: Signature exposures collected.

#> ✓ [2020-05-10 20:35:17]: Errors collected.

#> ✓ [2020-05-10 20:35:17]: Done.

#> ℹ [2020-05-10 20:35:17]: 0.457 secs elapsed.

#> ℹ [2020-05-10 20:35:17]: Processing sample `samp4`.

#> ℹ [2020-05-10 20:35:17]: Started.

#> ℹ [2020-05-10 20:35:17]: Checking catalog.

#> ✓ [2020-05-10 20:35:17]: Done.

#> ℹ [2020-05-10 20:35:17]: About to start bootstrap.

#> → Bootstrapping 10 times.

#> → Total 10 times, starting no.1.

#> → Total 10 times, starting no.2.

#> → Total 10 times, starting no.3.

#> → Total 10 times, starting no.4.

#> → Total 10 times, starting no.5.

#> → Total 10 times, starting no.6.

#> → Total 10 times, starting no.7.

#> → Total 10 times, starting no.8.

#> → Total 10 times, starting no.9.

#> → Total 10 times, starting no.10.

#>                                   

#> ✓ [2020-05-10 20:35:17]: Bootstrap done.

#> ✓ [2020-05-10 20:35:17]: Signature exposures collected.

#> ✓ [2020-05-10 20:35:17]: Errors collected.

#> ✓ [2020-05-10 20:35:17]: Done.

#> ℹ [2020-05-10 20:35:17]: 0.454 secs elapsed.

#> ℹ [2020-05-10 20:35:17]: Started.

#> ℹ [2020-05-10 20:35:17]: Checking catalog.

#> ✓ [2020-05-10 20:35:17]: Done.

#> ℹ [2020-05-10 20:35:17]: About to start bootstrap.

#> → Bootstrapping 10 times.

#> → Total 10 times, starting no.1.

#> → Total 10 times, starting no.2.

#> → Total 10 times, starting no.3.

#> → Total 10 times, starting no.4.

#> → Total 10 times, starting no.5.

#> → Total 10 times, starting no.6.

#> → Total 10 times, starting no.7.

#> → Total 10 times, starting no.8.

#> → Total 10 times, starting no.9.

#> → Total 10 times, starting no.10.

#>                                   

#> ✓ [2020-05-10 20:35:17]: Bootstrap done.

#> ✓ [2020-05-10 20:35:17]: Signature exposures collected.

#> ✓ [2020-05-10 20:35:17]: Errors collected.

#> ✓ [2020-05-10 20:35:17]: Done.

#> ℹ [2020-05-10 20:35:17]: 0.468 secs elapsed.

#> ✓ [2020-05-10 20:35:17]: Gotten.

#> ℹ [2020-05-10 20:35:17]: Reporting p values...

#> ℹ [2020-05-10 20:35:17]: Started.

#> ✓ [2020-05-10 20:35:17]: Batch mode enabled.

#> ✓ [2020-05-10 20:35:17]: Done.

#> ℹ [2020-05-10 20:35:17]: 0.006 secs elapsed.

#> ℹ [2020-05-10 20:35:17]: Started.

#> ✓ [2020-05-10 20:35:17]: Batch mode enabled.

#> ✓ [2020-05-10 20:35:17]: Done.

#> ℹ [2020-05-10 20:35:17]: 0.005 secs elapsed.

#> ℹ [2020-05-10 20:35:17]: Started.

#> ✓ [2020-05-10 20:35:17]: Batch mode enabled.

#> ✓ [2020-05-10 20:35:17]: Done.

#> ℹ [2020-05-10 20:35:17]: 0.005 secs elapsed.

#> ℹ [2020-05-10 20:35:17]: Started.

#> ✓ [2020-05-10 20:35:17]: Batch mode enabled.

#> ✓ [2020-05-10 20:35:17]: Done.

#> ℹ [2020-05-10 20:35:17]: 0.005 secs elapsed.

#> ✓ [2020-05-10 20:35:17]: Done.

#> ℹ [2020-05-10 20:35:17]: Cleaning results...

#> ✓ [2020-05-10 20:35:18]: Outputing.

#> ℹ [2020-05-10 20:35:18]: Total 3.986 secs elapsed.

#> $expo
#>      method sample  sig exposure    type
#>   1:     LS  samp1 sig1        2 optimal
#>   2:     LS  samp2 sig1        3 optimal
#>   3:     LS  samp3 sig1        1 optimal
#>   4:     LS  samp4 sig1        1 optimal
#>   5:     LS  samp1 sig2        5 optimal
#>  ---                                    
#> 172:     QP  samp4 sig2        3   Rep_8
#> 173:     QP  samp4 sig1        0   Rep_9
#> 174:     QP  samp4 sig2        3   Rep_9
#> 175:     QP  samp4 sig1        3  Rep_10
#> 176:     QP  samp4 sig2        0  Rep_10
#> 
#> $error
#>     method sample errors    type
#>  1:     LS  samp1  0.000 optimal
#>  2:     LS  samp2  0.000 optimal
#>  3:     LS  samp3  0.000 optimal
#>  4:     LS  samp4  0.000 optimal
#>  5:     QP  samp1  0.000 optimal
#>  6:     QP  samp2  0.000 optimal
#>  7:     QP  samp3  0.000 optimal
#>  8:     QP  samp4  0.000 optimal
#>  9:     LS  samp1  0.283   Rep_1
#> 10:     LS  samp1  0.471   Rep_2
#> 11:     LS  samp1  0.283   Rep_3
#> 12:     LS  samp1  0.707   Rep_4
#> 13:     LS  samp1  0.471   Rep_5
#> 14:     LS  samp1  0.471   Rep_6
#> 15:     LS  samp1  0.471   Rep_7
#> 16:     LS  samp1  0.707   Rep_8
#> 17:     LS  samp1  0.283   Rep_9
#> 18:     LS  samp1  0.471  Rep_10
#> 19:     QP  samp1  0.283   Rep_1
#> 20:     QP  samp1  0.236   Rep_2
#> 21:     QP  samp1  0.283   Rep_3
#> 22:     QP  samp1  0.236   Rep_4
#> 23:     QP  samp1  0.707   Rep_5
#> 24:     QP  samp1  0.471   Rep_6
#> 25:     QP  samp1  0.283   Rep_7
#> 26:     QP  samp1  0.471   Rep_8
#> 27:     QP  samp1  0.236   Rep_9
#> 28:     QP  samp1  0.707  Rep_10
#> 29:     LS  samp2  0.424   Rep_1
#> 30:     LS  samp2  0.849   Rep_2
#> 31:     LS  samp2  0.424   Rep_3
#> 32:     LS  samp2  0.424   Rep_4
#> 33:     LS  samp2  0.849   Rep_5
#> 34:     LS  samp2  0.424   Rep_6
#> 35:     LS  samp2  0.849   Rep_7
#> 36:     LS  samp2  0.141   Rep_8
#> 37:     LS  samp2  0.849   Rep_9
#> 38:     LS  samp2  0.849  Rep_10
#> 39:     QP  samp2  0.141   Rep_1
#> 40:     QP  samp2  0.849   Rep_2
#> 41:     QP  samp2  0.849   Rep_3
#> 42:     QP  samp2  0.141   Rep_4
#> 43:     QP  samp2  0.849   Rep_5
#> 44:     QP  samp2  0.424   Rep_6
#> 45:     QP  samp2  0.424   Rep_7
#> 46:     QP  samp2  0.141   Rep_8
#> 47:     QP  samp2  0.849   Rep_9
#> 48:     QP  samp2  0.849  Rep_10
#> 49:     LS  samp3  1.273   Rep_1
#> 50:     LS  samp3  1.273   Rep_2
#> 51:     LS  samp3  0.141   Rep_3
#> 52:     LS  samp3  1.037   Rep_4
#> 53:     LS  samp3  0.141   Rep_5
#> 54:     LS  samp3  1.037   Rep_6
#> 55:     LS  samp3  0.141   Rep_7
#> 56:     LS  samp3  0.141   Rep_8
#> 57:     LS  samp3  1.273   Rep_9
#> 58:     LS  samp3  0.330  Rep_10
#> 59:     QP  samp3  0.330   Rep_1
#> 60:     QP  samp3  0.141   Rep_2
#> 61:     QP  samp3  0.141   Rep_3
#> 62:     QP  samp3  0.141   Rep_4
#> 63:     QP  samp3  1.273   Rep_5
#> 64:     QP  samp3  0.141   Rep_6
#> 65:     QP  samp3  0.141   Rep_7
#> 66:     QP  samp3  1.273   Rep_8
#> 67:     QP  samp3  0.330   Rep_9
#> 68:     QP  samp3  0.141  Rep_10
#> 69:     LS  samp4  0.141   Rep_1
#> 70:     LS  samp4  0.141   Rep_2
#> 71:     LS  samp4  0.141   Rep_3
#> 72:     LS  samp4  0.283   Rep_4
#> 73:     LS  samp4  0.141   Rep_5
#> 74:     LS  samp4  0.141   Rep_6
#> 75:     LS  samp4  0.283   Rep_7
#> 76:     LS  samp4  0.141   Rep_8
#> 77:     LS  samp4  0.283   Rep_9
#> 78:     LS  samp4  0.283  Rep_10
#> 79:     QP  samp4  0.141   Rep_1
#> 80:     QP  samp4  0.283   Rep_2
#> 81:     QP  samp4  0.283   Rep_3
#> 82:     QP  samp4  0.141   Rep_4
#> 83:     QP  samp4  0.283   Rep_5
#> 84:     QP  samp4  0.283   Rep_6
#> 85:     QP  samp4  0.283   Rep_7
#> 86:     QP  samp4  0.141   Rep_8
#> 87:     QP  samp4  0.141   Rep_9
#> 88:     QP  samp4  0.283  Rep_10
#>     method sample errors    type
#> 
#> $p_val
#>     sample method threshold  sig p_value
#>  1:  samp1     LS      0.05 sig1     0.3
#>  2:  samp1     LS      0.05 sig2     0.2
#>  3:  samp1     QP      0.05 sig1     0.6
#>  4:  samp1     QP      0.05 sig2     0.2
#>  5:  samp2     LS      0.05 sig1     0.4
#>  6:  samp2     LS      0.05 sig2     0.5
#>  7:  samp2     QP      0.05 sig1     0.2
#>  8:  samp2     QP      0.05 sig2     0.5
#>  9:  samp3     LS      0.05 sig1     0.4
#> 10:  samp3     LS      0.05 sig2     0.3
#> 11:  samp3     QP      0.05 sig1     0.6
#> 12:  samp3     QP      0.05 sig2     0.2
#> 13:  samp4     LS      0.05 sig1     0.6
#> 14:  samp4     LS      0.05 sig2     0.4
#> 15:  samp4     QP      0.05 sig1     0.4
#> 16:  samp4     QP      0.05 sig2     0.6
#>