Exposure Instability Analysis of Signature Exposures with Bootstrapping

Exposure Instability Analysis of Signature Exposures with Bootstrapping

sig_fit_bootstrap_batch(
  catalogue_matrix,
  methods = c("QP"),
  n = 100L,
  min_count = 1L,
  p_val_thresholds = c(0.05),
  use_parallel = FALSE,
  seed = 123456L,
  job_id = NULL,
  result_dir = tempdir(),
  ...
)

Arguments

catalogue_matrix
a numeric matrix V with row representing components and columns representing samples, typically you can get nmf_matrix from sig_tally() and transpose it by t().

methods
a subset of c("NNLS", "QP", "SA").

n
the number of bootstrap replicates.

min_count
minimal exposure in a sample, default is 1. Any patient has total exposure less than this value will be filtered out.

p_val_thresholds
a vector of relative exposure threshold for calculating p values.

use_parallel
if TRUE, use parallel computation based on furrr package.

seed
random seed to reproduce the result.

job_id
a job ID, default is NULL, can be a string. When not NULL, all bootstrapped results will be saved to local machine location defined by result_dir. This is very useful for running more than 10 times for more than 100 samples.

result_dir
see above, default is temp directory defined by R.

...
other common parameters passing to sig_fit_bootstrap, including sig, sig_index, sig_db, db_type, mode, etc.

Value

a list of data.table.

See also

Examples

W <- matrix(c(1, 2, 3, 4, 5, 6), ncol = 2)
colnames(W) <- c("sig1", "sig2")
W <- apply(W, 2, function(x) x / sum(x))

H <- matrix(c(2, 5, 3, 6, 1, 9, 1, 2), ncol = 4)
colnames(H) <- paste0("samp", 1:4)

V <- W %*% H
V

#>         samp1 samp2    samp3 samp4
#> [1,] 1.666667   2.1 2.566667   0.7
#> [2,] 2.333333   3.0 3.333333   1.0
#> [3,] 3.000000   3.9 4.100000   1.3

if (requireNamespace("quadprog")) {
  z10 <- sig_fit_bootstrap_batch(V, sig = W, n = 10)
  z10
}

#> ℹ [2020-06-23 16:22:36]: Batch Bootstrap Signature Exposure Analysis Started.

#> ℹ [2020-06-23 16:22:36]: Samples to be filtered out: 

#> ℹ [2020-06-23 16:22:36]: Finding optimal exposures (&errors) for different methods.

#> ℹ [2020-06-23 16:22:36]: Calling method `QP`.

#> ℹ [2020-06-23 16:22:36]: Started.

#> ✓ [2020-06-23 16:22:36]: Signature matrix detected.

#> ✓ [2020-06-23 16:22:36]: Database and index checked.

#> ✓ [2020-06-23 16:22:36]: Signature normalized.

#> ℹ [2020-06-23 16:22:36]: Checking row number for catalog matrix and signature matrix.

#> ✓ [2020-06-23 16:22:36]: Checked.

#> ✓ [2020-06-23 16:22:36]: Method 'QP' detected.

#> ✓ [2020-06-23 16:22:36]: Corresponding function generated.

#> ℹ [2020-06-23 16:22:36]: Calling function.

#> ✓ [2020-06-23 16:22:36]: Done.

#> ℹ [2020-06-23 16:22:36]: Generating output signature exposures.

#> ✓ [2020-06-23 16:22:36]: Done.

#> ℹ [2020-06-23 16:22:36]: Calculating errors (Frobenius Norm).

#> ✓ [2020-06-23 16:22:36]: Done.

#> ℹ [2020-06-23 16:22:36]: 0.043 secs elapsed.

#> ℹ [2020-06-23 16:22:36]: Getting bootstrap exposures (&errors) for different methods.

#> ℹ [2020-06-23 16:22:36]: This step is time consuming, please be patient.

#> ℹ [2020-06-23 16:22:36]: Processing sample `samp1`.

#> ℹ [2020-06-23 16:22:36]: Started.

#> ℹ [2020-06-23 16:22:36]: Checking catalog.

#> ✓ [2020-06-23 16:22:36]: Done.

#> ℹ [2020-06-23 16:22:36]: About to start bootstrap.

#> → Bootstrapping 10 times.

#> → Total 10 times, starting no.1.

#> → Total 10 times, starting no.2.

#> → Total 10 times, starting no.3.

#> → Total 10 times, starting no.4.

#> → Total 10 times, starting no.5.

#> → Total 10 times, starting no.6.

#> → Total 10 times, starting no.7.

#> → Total 10 times, starting no.8.

#> → Total 10 times, starting no.9.

#> → Total 10 times, starting no.10.

#>                                   

#> ✓ [2020-06-23 16:22:36]: Bootstrap done.

#> ✓ [2020-06-23 16:22:36]: Signature exposures collected.

#> ✓ [2020-06-23 16:22:36]: Errors collected.

#> ✓ [2020-06-23 16:22:36]: Done.

#> ℹ [2020-06-23 16:22:36]: 0.503 secs elapsed.

#> ℹ [2020-06-23 16:22:37]: Processing sample `samp2`.

#> ℹ [2020-06-23 16:22:37]: Started.

#> ℹ [2020-06-23 16:22:37]: Checking catalog.

#> ✓ [2020-06-23 16:22:37]: Done.

#> ℹ [2020-06-23 16:22:37]: About to start bootstrap.

#> → Bootstrapping 10 times.

#> → Total 10 times, starting no.1.

#> → Total 10 times, starting no.2.

#> → Total 10 times, starting no.3.

#> → Total 10 times, starting no.4.

#> → Total 10 times, starting no.5.

#> → Total 10 times, starting no.6.

#> → Total 10 times, starting no.7.

#> → Total 10 times, starting no.8.

#> → Total 10 times, starting no.9.

#> → Total 10 times, starting no.10.

#>                                   

#> ✓ [2020-06-23 16:22:38]: Bootstrap done.

#> ✓ [2020-06-23 16:22:38]: Signature exposures collected.

#> ✓ [2020-06-23 16:22:38]: Errors collected.

#> ✓ [2020-06-23 16:22:38]: Done.

#> ℹ [2020-06-23 16:22:38]: 0.499 secs elapsed.

#> ℹ [2020-06-23 16:22:38]: Processing sample `samp3`.

#> ℹ [2020-06-23 16:22:38]: Started.

#> ℹ [2020-06-23 16:22:38]: Checking catalog.

#> ✓ [2020-06-23 16:22:38]: Done.

#> ℹ [2020-06-23 16:22:38]: About to start bootstrap.

#> → Bootstrapping 10 times.

#> → Total 10 times, starting no.1.

#> → Total 10 times, starting no.2.

#> → Total 10 times, starting no.3.

#> → Total 10 times, starting no.4.

#> → Total 10 times, starting no.5.

#> → Total 10 times, starting no.6.

#> → Total 10 times, starting no.7.

#> → Total 10 times, starting no.8.

#> → Total 10 times, starting no.9.

#> → Total 10 times, starting no.10.

#>                                   

#> ✓ [2020-06-23 16:22:39]: Bootstrap done.

#> ✓ [2020-06-23 16:22:39]: Signature exposures collected.

#> ✓ [2020-06-23 16:22:39]: Errors collected.

#> ✓ [2020-06-23 16:22:39]: Done.

#> ℹ [2020-06-23 16:22:39]: 0.498 secs elapsed.

#> ℹ [2020-06-23 16:22:39]: Processing sample `samp4`.

#> ℹ [2020-06-23 16:22:39]: Started.

#> ℹ [2020-06-23 16:22:39]: Checking catalog.

#> ✓ [2020-06-23 16:22:39]: Done.

#> ℹ [2020-06-23 16:22:39]: About to start bootstrap.

#> → Bootstrapping 10 times.

#> → Total 10 times, starting no.1.

#> → Total 10 times, starting no.2.

#> → Total 10 times, starting no.3.

#> → Total 10 times, starting no.4.

#> → Total 10 times, starting no.5.

#> → Total 10 times, starting no.6.

#> → Total 10 times, starting no.7.

#> → Total 10 times, starting no.8.

#> → Total 10 times, starting no.9.

#> → Total 10 times, starting no.10.

#>                                   

#> ✓ [2020-06-23 16:22:40]: Bootstrap done.

#> ✓ [2020-06-23 16:22:40]: Signature exposures collected.

#> ✓ [2020-06-23 16:22:40]: Errors collected.

#> ✓ [2020-06-23 16:22:40]: Done.

#> ℹ [2020-06-23 16:22:40]: 0.5 secs elapsed.

#> ✓ [2020-06-23 16:22:40]: Gotten.

#> ℹ [2020-06-23 16:22:40]: Reporting p values...

#> ℹ [2020-06-23 16:22:40]: Started.

#> ✓ [2020-06-23 16:22:40]: Batch mode enabled.

#> ✓ [2020-06-23 16:22:40]: Done.

#> ℹ [2020-06-23 16:22:40]: 0.006 secs elapsed.

#> ℹ [2020-06-23 16:22:40]: Started.

#> ✓ [2020-06-23 16:22:40]: Batch mode enabled.

#> ✓ [2020-06-23 16:22:40]: Done.

#> ℹ [2020-06-23 16:22:40]: 0.006 secs elapsed.

#> ℹ [2020-06-23 16:22:40]: Started.

#> ✓ [2020-06-23 16:22:40]: Batch mode enabled.

#> ✓ [2020-06-23 16:22:40]: Done.

#> ℹ [2020-06-23 16:22:40]: 0.006 secs elapsed.

#> ℹ [2020-06-23 16:22:40]: Started.

#> ✓ [2020-06-23 16:22:40]: Batch mode enabled.

#> ✓ [2020-06-23 16:22:40]: Done.

#> ℹ [2020-06-23 16:22:40]: 0.006 secs elapsed.

#> ✓ [2020-06-23 16:22:40]: Done.

#> ℹ [2020-06-23 16:22:40]: Cleaning results...

#> ✓ [2020-06-23 16:22:40]: Outputing.

#> ℹ [2020-06-23 16:22:40]: Total 4.618 secs elapsed.

#> $expo
#>     method sample  sig exposure    type
#>  1:     QP  samp1 sig1    2.000 optimal
#>  2:     QP  samp2 sig1    3.000 optimal
#>  3:     QP  samp3 sig1    1.000 optimal
#>  4:     QP  samp4 sig1    1.000 optimal
#>  5:     QP  samp1 sig2    5.000 optimal
#>  6:     QP  samp2 sig2    6.000 optimal
#>  7:     QP  samp3 sig2    9.000 optimal
#>  8:     QP  samp4 sig2    2.000 optimal
#>  9:     QP  samp1 sig1    7.000   Rep_1
#> 10:     QP  samp1 sig2    0.000   Rep_1
#> 11:     QP  samp1 sig1    5.333   Rep_2
#> 12:     QP  samp1 sig2    1.667   Rep_2
#> 13:     QP  samp1 sig1    0.333   Rep_3
#> 14:     QP  samp1 sig2    6.667   Rep_3
#> 15:     QP  samp1 sig1    7.000   Rep_4
#> 16:     QP  samp1 sig2    0.000   Rep_4
#> 17:     QP  samp1 sig1    5.333   Rep_5
#> 18:     QP  samp1 sig2    1.667   Rep_5
#> 19:     QP  samp1 sig1    7.000   Rep_6
#> 20:     QP  samp1 sig2    0.000   Rep_6
#> 21:     QP  samp1 sig1    5.333   Rep_7
#> 22:     QP  samp1 sig2    1.667   Rep_7
#> 23:     QP  samp1 sig1    0.000   Rep_8
#> 24:     QP  samp1 sig2    7.000   Rep_8
#> 25:     QP  samp1 sig1    0.333   Rep_9
#> 26:     QP  samp1 sig2    6.667   Rep_9
#> 27:     QP  samp1 sig1    7.000  Rep_10
#> 28:     QP  samp1 sig2    0.000  Rep_10
#> 29:     QP  samp2 sig1    9.000   Rep_1
#> 30:     QP  samp2 sig2    0.000   Rep_1
#> 31:     QP  samp2 sig1    9.000   Rep_2
#> 32:     QP  samp2 sig2    0.000   Rep_2
#> 33:     QP  samp2 sig1    0.000   Rep_3
#> 34:     QP  samp2 sig2    9.000   Rep_3
#> 35:     QP  samp2 sig1    9.000   Rep_4
#> 36:     QP  samp2 sig2    0.000   Rep_4
#> 37:     QP  samp2 sig1    9.000   Rep_5
#> 38:     QP  samp2 sig2    0.000   Rep_5
#> 39:     QP  samp2 sig1    9.000   Rep_6
#> 40:     QP  samp2 sig2    0.000   Rep_6
#> 41:     QP  samp2 sig1    0.000   Rep_7
#> 42:     QP  samp2 sig2    9.000   Rep_7
#> 43:     QP  samp2 sig1    0.000   Rep_8
#> 44:     QP  samp2 sig2    9.000   Rep_8
#> 45:     QP  samp2 sig1    0.000   Rep_9
#> 46:     QP  samp2 sig2    9.000   Rep_9
#> 47:     QP  samp2 sig1    0.000  Rep_10
#> 48:     QP  samp2 sig2    9.000  Rep_10
#> 49:     QP  samp3 sig1    0.000   Rep_1
#> 50:     QP  samp3 sig2   10.000   Rep_1
#> 51:     QP  samp3 sig1    0.000   Rep_2
#> 52:     QP  samp3 sig2   10.000   Rep_2
#> 53:     QP  samp3 sig1    0.000   Rep_3
#> 54:     QP  samp3 sig2   10.000   Rep_3
#> 55:     QP  samp3 sig1    8.333   Rep_4
#> 56:     QP  samp3 sig2    1.667   Rep_4
#> 57:     QP  samp3 sig1    0.000   Rep_5
#> 58:     QP  samp3 sig2   10.000   Rep_5
#> 59:     QP  samp3 sig1    8.333   Rep_6
#> 60:     QP  samp3 sig2    1.667   Rep_6
#> 61:     QP  samp3 sig1    3.333   Rep_7
#> 62:     QP  samp3 sig2    6.667   Rep_7
#> 63:     QP  samp3 sig1    0.000   Rep_8
#> 64:     QP  samp3 sig2   10.000   Rep_8
#> 65:     QP  samp3 sig1   10.000   Rep_9
#> 66:     QP  samp3 sig2    0.000   Rep_9
#> 67:     QP  samp3 sig1    8.333  Rep_10
#> 68:     QP  samp3 sig2    1.667  Rep_10
#> 69:     QP  samp4 sig1    0.000   Rep_1
#> 70:     QP  samp4 sig2    3.000   Rep_1
#> 71:     QP  samp4 sig1    0.000   Rep_2
#> 72:     QP  samp4 sig2    3.000   Rep_2
#> 73:     QP  samp4 sig1    0.000   Rep_3
#> 74:     QP  samp4 sig2    3.000   Rep_3
#> 75:     QP  samp4 sig1    3.000   Rep_4
#> 76:     QP  samp4 sig2    0.000   Rep_4
#> 77:     QP  samp4 sig1    0.000   Rep_5
#> 78:     QP  samp4 sig2    3.000   Rep_5
#> 79:     QP  samp4 sig1    0.000   Rep_6
#> 80:     QP  samp4 sig2    3.000   Rep_6
#> 81:     QP  samp4 sig1    3.000   Rep_7
#> 82:     QP  samp4 sig2    0.000   Rep_7
#> 83:     QP  samp4 sig1    3.000   Rep_8
#> 84:     QP  samp4 sig2    0.000   Rep_8
#> 85:     QP  samp4 sig1    3.000   Rep_9
#> 86:     QP  samp4 sig2    0.000   Rep_9
#> 87:     QP  samp4 sig1    0.000  Rep_10
#> 88:     QP  samp4 sig2    3.000  Rep_10
#>     method sample  sig exposure    type
#> 
#> $error
#>     method sample errors    type
#>  1:     QP  samp1  0.000 optimal
#>  2:     QP  samp2  0.000 optimal
#>  3:     QP  samp3  0.000 optimal
#>  4:     QP  samp4  0.000 optimal
#>  5:     QP  samp1  0.707   Rep_1
#>  6:     QP  samp1  0.471   Rep_2
#>  7:     QP  samp1  0.236   Rep_3
#>  8:     QP  samp1  0.707   Rep_4
#>  9:     QP  samp1  0.471   Rep_5
#> 10:     QP  samp1  0.707   Rep_6
#> 11:     QP  samp1  0.471   Rep_7
#> 12:     QP  samp1  0.283   Rep_8
#> 13:     QP  samp1  0.236   Rep_9
#> 14:     QP  samp1  0.707  Rep_10
#> 15:     QP  samp2  0.849   Rep_1
#> 16:     QP  samp2  0.849   Rep_2
#> 17:     QP  samp2  0.424   Rep_3
#> 18:     QP  samp2  0.849   Rep_4
#> 19:     QP  samp2  0.849   Rep_5
#> 20:     QP  samp2  0.849   Rep_6
#> 21:     QP  samp2  0.424   Rep_7
#> 22:     QP  samp2  0.424   Rep_8
#> 23:     QP  samp2  0.424   Rep_9
#> 24:     QP  samp2  0.424  Rep_10
#> 25:     QP  samp3  0.141   Rep_1
#> 26:     QP  samp3  0.141   Rep_2
#> 27:     QP  samp3  0.141   Rep_3
#> 28:     QP  samp3  1.037   Rep_4
#> 29:     QP  samp3  0.141   Rep_5
#> 30:     QP  samp3  1.037   Rep_6
#> 31:     QP  samp3  0.330   Rep_7
#> 32:     QP  samp3  0.141   Rep_8
#> 33:     QP  samp3  1.273   Rep_9
#> 34:     QP  samp3  1.037  Rep_10
#> 35:     QP  samp4  0.141   Rep_1
#> 36:     QP  samp4  0.141   Rep_2
#> 37:     QP  samp4  0.141   Rep_3
#> 38:     QP  samp4  0.283   Rep_4
#> 39:     QP  samp4  0.141   Rep_5
#> 40:     QP  samp4  0.141   Rep_6
#> 41:     QP  samp4  0.283   Rep_7
#> 42:     QP  samp4  0.283   Rep_8
#> 43:     QP  samp4  0.283   Rep_9
#> 44:     QP  samp4  0.141  Rep_10
#>     method sample errors    type
#> 
#> $p_val
#>    sample method threshold  sig p_value
#> 1:  samp1     QP      0.05 sig1     0.3
#> 2:  samp1     QP      0.05 sig2     0.4
#> 3:  samp2     QP      0.05 sig1     0.5
#> 4:  samp2     QP      0.05 sig2     0.5
#> 5:  samp3     QP      0.05 sig1     0.5
#> 6:  samp3     QP      0.05 sig2     0.1
#> 7:  samp4     QP      0.05 sig1     0.6
#> 8:  samp4     QP      0.05 sig2     0.4
#>