Tally a Variation Object

Tally a variation object like MAF and return a matrix for NMF de-composition and more. This is a generic function, so it can be further extended to other mutation cases. Please read details about how to set sex for identifying copy number signatures.

sig_tally(object, ...)

# S3 method for CopyNumber
sig_tally(
  object,
  method = "Macintyre",
  ignore_chrs = NULL,
  feature_setting = sigminer::CN.features,
  type = c("probability", "count"),
  reference_components = FALSE,
  cores = 1,
  seed = 123456,
  min_comp = 2,
  max_comp = 15,
  min_prior = 0.001,
  model_selection = "BIC",
  threshold = 0.1,
  nrep = 1,
  niter = 1000,
  keep_only_matrix = FALSE,
  ...
)

# S3 method for MAF
sig_tally(
  object,
  ref_genome = NULL,
  prefix = NULL,
  add = TRUE,
  ignore_chrs = NULL,
  use_syn = TRUE,
  keep_only_matrix = FALSE,
  ...
)

Arguments

object
a CopyNumber object or MAF object.

...
custom setting for operating object. Detail see S3 method for corresponding class (e.g. CopyNumber).

method
method for feature classfication, can be one of "Macintyre" ("M") and "Wang" ("W").

ignore_chrs
Chromsomes to ignore from analysis. e.g. chrX and chrY.

feature_setting
a data.frame used for classification. Only used when method is "Wang" ("W"). Default is CN.features. Users can also set custom input with "feature", "min" and "max" columns available. Valid features can be printed by unique(CN.features$feature).

type
one of "probability", "count". Default is "probability", return a matrix with the sum of posterior probabilities for each components. If set to 'count', return a matrix with event count assigned to each components. The result for both types should be close. Only used when method is "Macintyre".

reference_components
default is FALSE, calculate mixture components from CopyNumber object. Only used when method is "Macintyre".

cores
number of compute cores to run this task. You can use future::availableCores() function to check how many cores you can use.

seed
seed number. Only used when method is "Macintyre".

min_comp
minimal number of components to fit, default is 2. Can also be a vector with length 6, which apply to each feature. Only used when method is "Macintyre".

max_comp
maximal number of components to fit, default is 15. Can also be a vector with length 6, which apply to each feature. Only used when method is "Macintyre".

min_prior
the minimum relative size of components, default is 0.001. Details about custom setting please refer to flexmix package. Only used when method is "Macintyre".

model_selection
model selection strategy, default is 'BIC'. Details about custom setting please refer to flexmix package. Only used when method is "Macintyre".

threshold
default is 0.1. Sometimes, the result components include adjacent distributions with similar mu (two and more distribution are very close), we use this threshold to obtain a more meaningful fit with less components. Only used when method is "Macintyre".

nrep
number of run times for each value of component, keep only the solution with maximum likelihood. Only used when method is "Macintyre".

niter
the maximum number of iterations. Only used when method is "Macintyre".

keep_only_matrix
if TRUE, keep only matrix for signature extraction.

ref_genome
BSgenome object or name of the installed BSgenome package. Example: BSgenome.Hsapiens.UCSC.hg19 Default NULL, tries to auto-detect from installed genomes.

prefix
Prefix to add or remove from contig names in MAF file.

add
If prefix is used, default is to add prefix to contig names in MAF file. If false prefix will be removed from contig names.

use_syn
Logical. Whether to include synonymous variants in analysis. Defaults to TRUE

Value

a list contains a matrix used for NMF de-composition.

Details

For identifying copy number signatures, we have to derive copy number features firstly. Due to the difference of copy number values in sex chromosomes between male and female, we have to do an extra step if we don't want to ignore them.

I create two options to control this, the default values are shown as the following, you can use the same way to set (per R session).

options(sigminer.sex = "female", sigminer.copynumber.max = NA_integer_)

If your cohort are all females, you can totally ignore this.

If your cohort are all males, set sigminer.sex to 'male' and sigminer.copynumber.max to a proper value (the best is consistent with read_copynumber).

If your cohort contains bother males and females, set sigminer.sex as a data.frame with two columns "sample" and "sex". And set sigminer.copynumber.max to a proper value (the best is consistent with read_copynumber).

Methods (by class)

CopyNumber: Returns copy number features, components and component-by-sample matrix

MAF: Returns SBS mutation component-by-sample matrix and APOBEC enrichment

References

Macintyre, Geoff, et al. "Copy number signatures and mutational processes in ovarian carcinoma." Nature genetics 50.9 (2018): 1262.

Mayakonda, Anand, et al. "Maftools: efficient and comprehensive analysis of somatic variants in cancer." Genome research 28.11 (2018): 1747-1756.

See also

Examples

# Load copy number object
load(system.file("extdata", "toy_copynumber.RData",
  package = "sigminer", mustWork = TRUE
))
# \donttest{
# Use method designed by Macintyre et al.
cn_tally_M <- sig_tally(cn)

#> => Step: getting copy number features

#> Getting segment size...

#> Getting breakpoint count per 10 Mb...

#> Getting length of chains of oscillating copy number...

#> Getting breakpoint count per chromosome arm...

#> Getting change-point copy number change...

#> Getting copy number...

#> => Step: fitting copy number components

#> Fitting feature: segsize

#> Loading required package: globals

#> Loading required package: future

#> 
#> Attaching package: 'future'

#> The following object is masked from 'package:NMF':
#> 
#>     run

#> 2 : *
#> 3 : *
#> 4 : *
#> 5 : *
#> 6 : *
#> 7 : *
#> 8 : *
#> 9 : *
#> 10 : *
#> 11 : *
#> 12 : *
#> 13 : *
#> 14 : *
#> 15 : *

#> Select 10 according to BIC

#> Finally, select 10 after passing threshold 0.1

#> ===================

#> Fitting feature: bp10MB

#> 2 : *
#> 3 : *
#> 4 : *
#> 5 : *
#> 6 : *
#> 7 : *
#> 8 : *
#> 9 : *
#> 10 : *
#> 11 : *
#> 12 : *
#> 13 : *
#> 14 : *
#> 15 : *

#> Select 2 according to BIC

#> Finally, select 2 after passing threshold 0.1

#> ===================

#> Fitting feature: osCN

#> 2 : *
#> 3 : *
#> 4 : *
#> 5 : *
#> 6 : *
#> 7 : *
#> 8 : *
#> 9 : *
#> 10 : *
#> 11 : *
#> 12 : *
#> 13 : *
#> 14 : *
#> 15 : *

#> Select 2 according to BIC

#> Finally, select 2 after passing threshold 0.1

#> ===================

#> Fitting feature: bpchrarm

#> 2 : *
#> 3 : *
#> 4 : *
#> 5 : *
#> 6 : *
#> 7 : *
#> 8 : *
#> 9 : *
#> 10 : *
#> 11 : *
#> 12 : *
#> 13 : *
#> 14 : *
#> 15 : *

#> Select 2 according to BIC

#> Finally, select 2 after passing threshold 0.1

#> ===================

#> Fitting feature: changepoint

#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     25 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     19 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     14 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     14 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     13 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     15 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     12 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     14 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     15 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     20 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     15 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     14 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     14 Log-likelihood: NaN
#> 2 : *
#> 3 : *
#> 4 : *
#> 5 : *
#> 6 : *
#> 7 : *
#> 8 : *
#> 9 : *
#> 10 : *
#> 11 : *
#> 12 : *
#> 13 : *
#> 14 : *
#> 15 : *

#> Fitting feature: copynumber

#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     12 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     10 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>      9 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     12 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     11 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     10 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     10 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     12 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     10 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     10 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     11 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     11 Log-likelihood: NaN
#> Error in FLXfit(model = model, concomitant = concomitant, control = control,  : 
#>     10 Log-likelihood: NaN
#> 2 : *
#> 3 : *
#> 4 : *
#> 5 : *
#> 6 : *
#> 7 : *
#> 8 : *
#> 9 : *
#> 10 : *
#> 11 : *
#> 12 : *
#> 13 : *
#> 14 : *
#> 15 : *

#> => Step: calculating the sum of posterior probabilities

#> Calculating the sum of posterior probabilities for bp10MB

#> Calculating the sum of posterior probabilities for copynumber

#> Calculating the sum of posterior probabilities for changepoint

#> Calculating the sum of posterior probabilities for bpchrarm

#> Calculating the sum of posterior probabilities for osCN

#> Calculating the sum of posterior probabilities for segsize

#> => Done.

# }
# Use method designed by Wang, Shixiang et al.
cn_tally_W <- sig_tally(cn, method = "W")

#> => Step: getting copy number features

#> Getting breakpoint count per 10 Mb...

#> Getting breakpoint count per chromosome arm...

#> Getting copy number...

#> Getting change-point copy number change...

#> Getting length of chains of oscillating copy number...

#> Getting (log10 based) segment size...

#> Getting the minimal number of chromosome with 50% CNV...

#> Getting burden of chromosome...

#> => Step: generating copy number components

#> => Step: generating components by sample matrix

#> => Done.

# \donttest{
# Prepare SBS signature analysis
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read_maf(maf = laml.maf)

#> -Reading
#> -Validating
#> -Silent variants: 475 
#> -Summarizing
#> -Processing clinical data
#> --Missing clinical data
#> -Finished in 0.330s elapsed (0.250s cpu) 

if (require("BSgenome.Hsapiens.UCSC.hg19")) {
  mt_tally <- sig_tally(
    laml,
    ref_genome = "BSgenome.Hsapiens.UCSC.hg19",
    prefix = "chr", add = TRUE, useSyn = TRUE
  )
} else {
  message("Please install package 'BSgenome.Hsapiens.UCSC.hg19' firstly!")
}

#> Loading required package: BSgenome.Hsapiens.UCSC.hg19

#> Loading required package: BSgenome

#> Loading required package: S4Vectors

#> Loading required package: stats4

#> 
#> Attaching package: 'S4Vectors'

#> The following object is masked from 'package:future':
#> 
#>     values

#> The following object is masked from 'package:NMF':
#> 
#>     nrun

#> The following object is masked from 'package:pkgmaker':
#> 
#>     new2

#> The following object is masked from 'package:base':
#> 
#>     expand.grid

#> Loading required package: IRanges

#> 
#> Attaching package: 'IRanges'

#> The following object is masked from 'package:purrr':
#> 
#>     reduce

#> The following object is masked from 'package:grDevices':
#> 
#>     windows

#> Loading required package: GenomeInfoDb

#> Loading required package: GenomicRanges

#> Loading required package: Biostrings

#> Loading required package: XVector

#> 
#> Attaching package: 'XVector'

#> The following object is masked from 'package:purrr':
#> 
#>     compact

#> 
#> Attaching package: 'Biostrings'

#> The following object is masked from 'package:base':
#> 
#>     strsplit

#> Loading required package: rtracklayer

#> Warning: Chromosome names in MAF must match chromosome names in reference genome.
#> Ignorinig 101 single nucleotide variants from missing chromosomes chr23

#> -Extracting 5' and 3' adjacent bases
#> -Extracting +/- 20bp around mutated bases for background C>T estimation
#> -Estimating APOBEC enrichment scores
#> --Performing one-way Fisher's test for APOBEC enrichment
#> ---APOBEC related mutations are enriched in  3.315 % of samples (APOBEC enrichment score > 2 ;  6  of  181  samples)
#> -Creating mutation matrix
#> --matrix of dimension 188x96

# }